% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{abbr}
\alias{abbr}
\title{Abbreviations}
\usage{
abbr(x, pattern = "[A-Z]", n = 1L, include = "[a-z]+")
}
\arguments{
\item{x}{a vector of strings to abbreviate}

\item{pattern, n}{the pattern and number of occurrences to use}

\item{include}{an optional pattern to match additional strings surrounded
by word boundaries; any non character will omit}
}
\description{
Extract abbreviations from strings.
}
\examples{
x <- c('United States of America', 'the Red, the White, and the Blue')
abbr(x)
abbr(x, include = NA)
abbr(x, include = 'and')
abbr(x, '[A-Z][A-z]', include = NA)

}
