% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_extra.R
\name{arrows2}
\alias{arrows2}
\title{Add filled arrows to a plot}
\usage{
arrows2(x0, y0, x1 = x0, y1 = y0, size = 1, width = (sqrt(5) - 1)/8/cin,
  curve = 1, code = 2, col = par("fg"), lty = par("lty"),
  lwd = par("lwd"), fill = col, border = fill, sadj = c(0, 0, 0, 0),
  ...)
}
\arguments{
\item{x0, y0}{coordinates of points \strong{from} which to draw}

\item{x1, y1}{coordinates of points \strong{to} which to draw}

\item{size, width}{size parameters for arrows}

\item{curve}{a numeric value greater than 0 giving the curvature of the
arrows; default is \code{1} for staight lines, but values less than or
greater than 1 may be given for}

\item{code}{integer determining \emph{kind} of arrows to be drawn; arrows
are drawn at \code{{x0[i], y0[i]}}, \code{{x1[i], y1[i]}}, or both for
\code{code = 1:3}, respectively, or no heads drawn if \code{code = 0}}

\item{col, lty, lwd}{color, line type, and line width of the segment}

\item{fill, border}{fill and border color of arrow}

\item{sadj}{optional vector of the form \code{{x0,y0,x1,y1}} for adjusting
the \code{\link{segments}} of the arrows}

\item{...}{additional graphical parameters passed to \code{\link{segments}}
and further to \code{\link{par}}}
}
\description{
Draw filled arrows between pairs of points.
}
\examples{
plot.new()
plot.window(0:1, 0:1, asp = 1)
arrows2(0,0,1,1, code = 3, border = 2, fill = 0)
arrows2(.2, 0, .2, 1, code = 1, fill = 4, col = 4)
arrows2(.4, 0, .4, 1, code = 2, fill = 2, size = .5, width = .5)
arrows2(.6, 0, .6, 1, code = 3, curve = 1.5)
arrows2(.8, 0, .8, 1, code = 3, curve = .5, width = .5, lwd = 10, col = 3)

}
\author{
Original: Andreas Ruckstuhl, 19 May 1994; Cosmetic: Martin Machler, June
1998; Modifications: Robert Redd
}
\seealso{
\code{arrows}; \code{\link{carrows}}; \url{https://github.com/cran/sfsmisc}
}

