% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{atrisk_data_}
\alias{atrisk_data_}
\title{Create a list of data frames for at-risk tables}
\usage{
atrisk_data_(object, times, digits)
}
\arguments{
\item{object}{a \code{\link{survfit}} object}

\item{times}{time points}

\item{digits}{digits}
}
\description{
Create a list of data frames for at-risk tables
}
\seealso{
\code{\link{kmplot}}; \code{\link{kmplot_by}}
}
