% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ave_dt}
\alias{ave_dt}
\title{\code{ave} for date objects}
\usage{
ave_dt(x, ..., FUN = identity)
}
\arguments{
\item{x}{an integer or date vector}

\item{...}{grouping variables, typically factors, all of the same length
as \code{x}}

\item{FUN}{function to apply for each factor level combination}
}
\description{
Return date objects when \code{ave} tries to return integers.
}
\examples{
set.seed(1)
dt <- Sys.Date() + sample(1:20, 20)
id <- rep(1:4, each = 5)

ave(dt, id, FUN = min)
ave_dt(dt, id, FUN = min)

ave(seq_along(dt), id, FUN = function(ii) dt[ii])
ave_dt(seq_along(dt), id, FUN = function(ii) dt[ii])

}
\seealso{
\code{\link{ave}}
}
