% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{binconr}
\alias{binconr}
\title{\code{bincon} formatter}
\usage{
binconr(r, n, conf = 0.95, digits = 0, est = TRUE, frac = FALSE,
  show_conf = TRUE, pct.sign = TRUE, method = "exact")
}
\arguments{
\item{r}{number of responses (successes)}

\item{n}{number of observations (trials)}

\item{conf}{level of confidence}

\item{digits}{number of digits}

\item{est}{logical; if \code{TRUE}, includes the point estimate}

\item{frac}{logical; if \code{TRUE}, includes the fraction \code{r/n}}

\item{show_conf}{logical; if \code{TRUE} includes the confidence level}

\item{method}{method to use; see \code{\link{bincon}}}
}
\description{
Binomial confidence interval (\code{\link{bincon}}) formatter.
}
\examples{
binconr(5, 10, .90, est = FALSE)
binconr(45, 53, digits = 1, conf = .975)
binconr(45, 53, show_conf = FALSE, frac = TRUE)

}
\seealso{
\code{\link{bincon}}; \code{\link[Hmisc]{binconf}}
}

