% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_all}
\alias{bind_all}
\title{Bind objects with unequal number of rows or columns}
\usage{
bind_all(..., which)
}
\arguments{
\item{...}{vectors}

\item{which}{joining method; \code{'rbind'} or \code{'cbind'}}
}
\description{
Bind objects with unequal number of rows or columns.
}
\examples{
bind_all(1:5, 1:3, which = 'cbind')
bind_all(1:5, 1:3, which = 'rbind')
}

