% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogdown.R
\name{blogdown}
\alias{blogdown}
\title{Retrieve raw R code from a blogdown web page}
\usage{
blogdown(url)
}
\arguments{
\item{url}{Link to a blogdown web page}
}
\description{
Retrieve raw R code from a blogdown web page
}
\examples{
library(dplyr)
blogdown("https://www.jtimm.net/2019/04/14/lexical-change-procrustes/")

# Same as above but provided to cat for easy viewing
blogdown("https://www.jtimm.net/2019/04/14/lexical-change-procrustes/")  \%>\%
  cat

}
