% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{bni}
\alias{bni}
\title{Two-arm binomial non-inferiority trials}
\usage{
bni(margin, p1, p2 = p1, alpha = 0.05, power = NULL, n = NULL, p = 0.5)
}
\arguments{
\item{margin}{the largest acceptable difference in success rates between
the standard (\code{p1}) and experimental (\code{p2}) arms that would be
consistent with non-inferiority}

\item{p1, p2}{true success rates for the standard and experimental arms,
respectively}

\item{alpha}{type I error (one-sided), i.e., the \code{1 - 2 * alpha}%
confidence interval around the difference between the rates}

\item{power}{the desired power level to rule out the null of inferiority}

\item{n}{the total sample size of both arms}

\item{p}{the proportion of \code{n} on the experimental arm, \code{p2}}
}
\value{
The a numeric vector with the total sample size (\code{n}), sample size
for the standard (\code{n1}) and experimental (\code{n2}) arms, and power
(\code{power}).
}
\description{
Calculate the required sample size or power for a two-arm non-inferiority
design with binomial outcome. Exactly one of \code{power} or \code{n}
should be \code{NULL} to solve for this value given the other parameters.
}
\examples{
bni(0.1, 0.65, 0.85, power = 0.8)
bni(0.1, 0.65, 0.85, n = 50)
bni(0.1, 0.65, 0.85, n = 50, p = 2 / 3) ## 1:2 randomization

}
\references{
Kopecky K and Green S (2012). Noninferiority trials. In: Handbook of
Statistics in Clinical Oncology. Crowley J and Hoering A, eds. CRC Press,
Boca Raton, FL USA.
}
\seealso{
\url{https://stattools.crab.org/R/Binomial_Non_Inferiority.html}

\url{http://www.powerandsamplesize.com/Calculators/Compare-2-Proportions/2-Sample-Non-Inferiority-or-Superiority}
}
