% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{bp.test}
\alias{bp.test}
\title{Box plot tests}
\usage{
bp.test(
  formula,
  data,
  which = NULL,
  at = NULL,
  line = NULL,
  test = wilcox.test,
  ...
)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ group} where
\code{response} is a numeric variable and \code{group} is a factor-like
variable with three or more unique values (groups)}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in \code{formula}; by
default, the variables are taken from \code{environment(formula)}}

\item{which}{an index vector of p-values to display}

\item{at}{numeric vector giving the locations where box plots were drawn}

\item{line}{lines at which to plot \code{test} results}

\item{test}{the test to use for pairwise comparisons}

\item{...}{additional graphical parameters passed to \code{link{par}}}
}
\value{
A matrix of user coordinates where p-values are drawn.
}
\description{
Add pairwise tests to box plots with three or more groups. Using
\code{\link{cuzick.test}} under-the-hood, performs \code{test} for all
pairs of groups and draws results on an existing plot.
}
\examples{
boxplot(mpg ~ gear, mtcars)
bp.test(mpg ~ gear, mtcars)


boxplot(mpg ~ gear, mtcars, at = c(1, 3, 4), ylim = c(10, 55))
bp.test(mpg ~ gear, mtcars, at = c(1, 3, 4), which = c(1, 3, 2), line = -5)


op <- par(mar = par('mar') + c(0, 0, 3, 0))
tplot(mpg ~ interaction(vs, am), mtcars, show.n = FALSE)
bp.test(mpg ~ interaction(vs, am), mtcars)

tplot(mpg ~ interaction(vs, am), mtcars, show.n = FALSE)
bp.test(mpg ~ interaction(vs, am), mtcars, which = c(1, 3, 5))

bp.test(mpg ~ interaction(vs, am), mtcars, which = 6, line = 4,
        col = 'red', fg = 'red', lty = 2, font = 2, test = t.test) -> at
points(at[1], at[2], pch = 1, cex = 5, col = 'red', xpd = NA)
par(op)


## also works for barplots
bp <- barplot(with(mtcars, tapply(mpg, gear, mean)), ylim = c(0, 30))
bp.test(mpg ~ gear, mtcars, at = bp, line = -1, test = t.test)

}
\seealso{
\code{\link{cuzick.test}}; \code{\link{boxplot}}; \code{\link{tplot}};
\code{rawr:::coords}
}
