% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{bpCI}
\alias{bpCI}
\title{Barplot confidence intervals}
\usage{
bpCI(bp, horiz = FALSE, ci = TRUE, ci.u, ci.l, ci.width = 0.5,
  sig = FALSE, pvals, pch, ...)
}
\arguments{
\item{bp}{the return value of \code{\link{barplot}}, i.e., a vector or
matrix (when \code{beside = TRUE}) of all bar (or group) midpoints}

\item{horiz}{logical; if \code{TRUE}, \code{bpCI} assumes horizontal bars}

\item{ci}{logical; draw error bars (must give \code{ci.u}, \code{ci.l})}

\item{ci.u,ci.l}{a numeric vector or matrix having the same dimensions as
\code{bp} giving the upper and lower intervals, respectively}

\item{ci.width}{width of the ends of the error bars, will depend on
\code{range(bp)}}

\item{sig}{logical; if \code{TRUE}, draws group comparisons (must give
\code{pvals} to plot sig stars)}

\item{pvals}{p-values of group comparisons to be displayed as sig stars}

\item{pch}{plotting character to be used for significance; default is
\code{*} and uses same significance codes as \code{\link{printCoefmat}}}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\description{
Add confidence intervals (error bars) and group comparisons to barplots.
}
\examples{
## generate data and p-values
hh <- t(VADeaths)[1:2, 5:1]
ci.l <- hh * 0.85
ci.u <- hh * 1.15
pvals <- pt(apply(hh, 2, diff), 1) / 5:1

bp <- barplot(hh, beside = TRUE, ylim = c(0,100))
bpCI(bp, ci.u = ci.u, ci.l = ci.l, sig = TRUE, pvals = pvals, pch = "+")
mtext("Signif. codes:  0 '+++' 0.001 '++' 0.01 '+' 0.05 '.' 0.1 ' ' 1",
      side = 1, at = par('usr')[2], line = 2, adj = 1, cex = .8, font = 3)


bp <- barplot(hh <- cbind(x = c(465, 91) / 465 * 100,
                          y = c(200, 840) / 840 * 100,
                          z = c(37, 17) / 37 * 100),
              beside = TRUE, width = c(465, 840, 37),
              col = c(1, 2), ylim = c(0,130))

ci.l <- hh * 0.85
ci.u <- hh * 1.15
pvals <- pt(-abs(apply(hh, 2, diff)), 1)

bpCI(bp, ci.u = ci.u, ci.l = ci.l, sig = TRUE, pvals = pvals, ci.width = 100,
     col = 'red', lty = 'dashed', lwd = 2)
mtext("Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1",
      side = 1, at = par('usr')[2], line = 2, adj = 1, cex = .8, font = 3)
}

