% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{ca.test}
\alias{ca.test}
\alias{ca.test.default}
\alias{ca.test.formula}
\title{Cochran-Armitage test for trend}
\usage{
ca.test(x, ...)

\method{ca.test}{default}(x, g, ..., score = NULL, simulate.p.value = FALSE, B = 2000L)

\method{ca.test}{formula}(formula, data, ...)
}
\arguments{
\item{x}{a factor-like vector giving the (unordered) variable (equivalently
  the row variable of a contingency table)

  alternatively, \code{x} can be a \code{2 x c} table or matrix with
  exactly two rows and at least three ordered columns; \code{x} may also
  be a list of the row variable split by the ordered column variable in
  which case the list is assumed to be ordered, i.e.,
  \code{x[[1]] < x[[2]] < ... < x[[c]]}; see examples}

\item{...}{further arguments to be passed to or from methods}

\item{g}{a factor-like vector giving the \emph{ordered} group for each
corresponding element of \code{x}, ignored with a warning if \code{x} is
a list or table; if \code{g} is not a factor, it will be coerced, and
groups will be ordered as sort(unique(g)); see \code{\link{factor}}}

\item{score}{group score for each column, default is \code{1:ncol}}

\item{simulate.p.value}{logical; if \code{TRUE}, p-value is computed using
by Monte Carlo simulation}

\item{B}{an integer specifying the number of replicates used in the Monte
Carlo test}

\item{formula}{a formula of the form \code{row ~ column} where \code{row}
gives the row variable having two unique values and \code{column} gives
the \emph{ordered} column variable}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in \code{formula};
by default the variables are taken from \code{environment(formula)}}
}
\value{
A list with class "\code{htest}" containing the following elements:

\item{\code{statistic}}{the chi-squared test statistic}
\item{\code{parameter}}{the degrees of freedom of the approximate chi-
squared distribution of the test statistic}
\item{\code{p.value}}{the p-value of the test (two-sided)}
\item{\code{method}}{a character string describing the test, and,
optionally, the number of Monte Carlo replications, if applicable}
\item{\code{data.name}}{a character string giving the names of the data}
\item{\code{conf.int}}{optionally (if \code{simulate.p.value = TRUE}),
the 99\% confidence interval of the Monte Carlo p-value}
\item{\code{summary}}{optionally (if \code{simulate.p.value = TRUE}),
a summary of the simulated test statistics}
}
\description{
Performs a Cochran-Armitage chi-squared test for trend in proportions for
a \code{2 x c} contingency table with a nominal row (r == 2) and ordinal
column (c > 2) variable.
}
\examples{
## example from stats::prop.trend.test
smokers  <- c(83, 90, 129, 70)
patients <- c(86, 93, 136, 82)
prop.test(smokers, patients)
prop.trend.test(smokers, patients)

# DescTools::CochranArmitageTest(rbind(smokers, patients - smokers))
ca.test(rbind(smokers, patients - smokers))
ca.test(rbind(smokers, patients - smokers), score = c(0, 0, 1, 2))

## the following are equivalent to call ca.test
dat <- data.frame(x = mtcars$vs, y = mtcars$gear)
ca.test(dat$x, dat$y)
ca.test(x ~ y, dat)
ca.test(split(dat$x, dat$y))
ca.test(table(dat$x, dat$y))


\dontrun{
## simulate p-value with 10k replicates
set.seed(1)
ca.test(rbind(smokers, patients - smokers), simulate.p.value = TRUE, B = 1000)
}

}
\seealso{
\code{\link{prop.trend.test}}; \code{\link{jt.test}} for doubly-ordered tables;
\code{\link{cuzick.test}}; \code{DescTools::CochranArmitageTest}
}
