% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{carrows}
\alias{carrows}
\title{Curved arrows}
\usage{
carrows(p1, p2, arc, degree = FALSE, pad = 0.01 * 1:2, flip = FALSE,
  dir = NULL, col = par("col"), lwd = par("lwd"), lty = par("lty"),
  size = 1, width = size/2, curve = 1, fill = col, border = NA)
}
\arguments{
\item{p1, p2}{vectors of length two giving the \code{{x,y}} coordinates for
two points to draw a connecting arc}

\item{arc}{a vector of length two with the starting and ending positions
to draw the arc, in radians or degrees}

\item{degree}{logical; if \code{TRUE}, \code{arc} should be in degrees;
however, if either element of \code{arc} is greater than \code{2 * pi},
they are assumed to be in degrees and will be converted to radians; set to
\code{TRUE} to avoid conversion of small angles to radians}

\item{pad}{a vector of length two giving 1) padding between the tips of
the arrow/segment and the points and 2) additional padding between the
segment endpoints and tip of arrow--useful for thick lines which may
protrude from under the arrowhead}

\item{flip}{logical; if \code{TRUE}, the arrow will be rotated around the
circle 180 degrees}

\item{dir}{optional vector of directions for arrows; by default, arrows
will point to the nearest endpoint; \code{dir} should be a vector of
\code{1}s and \code{-1}s and will be recycled as necessary; other values
will be ignored and result in no arrows for those positions}

\item{col, lwd, lty}{color, line width, and line type passed to
\code{\link{lines}}}

\item{size, width, curve, fill, border}{additional parameters passed to
\code{\link{arrows2}}}
}
\value{
A list containing the arc endpoints and the center and radius of the
corresponding circle.
}
\description{
Draw an arrow along the arc of a circle.
}
\examples{
plot.new()
plot.window(c(-2,2), c(-2,2))
p <- matrix(c(rep(-1:1, 2), rep(-1:1, each = 2)), ncol = 2)
points(p)

carrows(p1 <- p[2, ], p2 <- p[1, ], pad = .3)
carrows(p1 <- p[4, ], p2 <- p[5, ], dir = c(0, 1), col = 3)
carrows(p1 <- p[6, ], p2 <- p[3, ], lwd = 10, pad = c(.05, .1))
carrows(p1 <- p[1, ], p2 <- p[6, ], flip = TRUE)
carrows(p1 <- p[1, ], p2 <- p[5, ], dir = c(1, 0))

}
\seealso{
\code{\link{arrows2}}; \code{\link{xspline}}
}
