% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{case}
\alias{case}
\title{Letter case}
\usage{
case(x, which = c("first", "upcase", "downcase", "camelcase", "upper",
  "lower", "lowup", "uplow"))
}
\arguments{
\item{x}{a text string}

\item{which}{case to use; can be (unambiguously) abbreviated; see details}
}
\description{
Convert a text string to several common types of letter case.
}
\details{
This function supports the following letter cases:

\itemize{
\item{first}{Only the first letter is uppercase}
\item{upcase}{Each Word Is Uppercase}
\item{downcase}{tHE oPPOSITE oF uPCASE}
\item{camelcase}{TheStringWillBeInCamelCase}
\item{upper}{EQUIVALENT TO \code{toupper}}
\item{lower}{equivalent to \code{tolower}}
\item{lowup}{aLtErNaTiNg cAsEs}
\item{uplow}{ThE OpPoSiTe oF LoWuP}
}
}
\examples{
wh <- eval(formals(case)$which)
x <- 'the quick brown fox'
sapply(wh, case, x = x)

## all cases are vectorized
sapply(wh, case, x = strsplit(x, ' ')[[1]])

}

