% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{catlist}
\alias{catlist}
\title{Concatenate a named list for output}
\usage{
catlist(l)
}
\arguments{
\item{l}{list to concatenate}
}
\description{
Print a \emph{named} \code{list} as a character string with values.
}
\examples{
(l <- list(a = 1, b = 2))

# $a
# [1] 1
#
# $b
# [1] 2

catlist(l) # [1] "a = 1, b = 2"
catlist(par()[1:5])

}
