% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{clc}
\alias{clc}
\title{clc}
\usage{
clc(all = FALSE)
}
\arguments{
\item{all}{logical; if \code{TRUE}, also removes hidden files}
}
\description{
Clear the workspace by removing all objects in \code{\link{ls}}.
}

