% Generated by roxygen2 (4.0.0): do not edit by hand
\name{click.text}
\alias{click.text}
\title{Add text interactively in base \code{R} graphics}
\usage{
click.text(express, col = 'black', cex = NULL, srt = 0, trans = NULL,
           family = 'sans', dev = FALSE, ...)
}
\arguments{
\item{express}{text or \code{\link{expression}}}

\item{col}{colour of text}

\item{cex}{numerical \strong{c}haracter \strong{ex}pansion factor;
multiplied by \code{\link{par}} ("cex") yields the final character size;
\code{NULL} and \code{NA} are equivalent to \code{1.0}}

\item{srt}{the string rotation in degrees; see \code{\link{par}}}

\item{trans}{color transparency; see \code{\link{tcol}}}

\item{family}{the name of a font or drawing text; see \code{\link{Hershey}}}

\item{dev}{logical; if \code{TRUE}, will plot on a platform-specific
graphics device; see details}

\item{...}{further graphical parameters passed to \code{\link{text}} (or
\code{\link{par}})}
}
\description{
Allows you to add text and expressions anywhere in a plot (including
margins) with mouse click(s).
}
\details{
When \code{dev = TRUE}, a platform-specific graphics device is used to
displace the plot. Defaults are \code{\link{quartz}}, \code{\link{windows}},
and \code{\link{x11}} for apple, windows, and unix platforms, respectively.
However, any device can be used by setting \code{dev = FALSE} and opening
a device before running \code{kmplot}. If \code{dev} is \code{FALSE}, the
plot will open in the current device; see \code{\link{.Device}}. Close the
device pane or use \code{dev.off()} to turn off the current device.
}
\examples{
\dontrun{
plot.new()
click.text('hello', col = 'red', cex = .5)
click.text('goodbye', family = 'HersheyScript', cex = 3)
click.text(expression(sum(x ^ 2) == 5 ^ hat(x)), srt = 45)
}
}
\seealso{
\code{\link{click.shape}}; \code{\link{plotmath}} for help with
mathematical expressions
}

