% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{click_text}
\alias{click_text}
\title{Add text interactively in base \code{R} graphics}
\usage{
click_text(expr, ...)
}
\arguments{
\item{expr}{a character string of text or an \code{\link{expression}}}

\item{...}{additional graphical parameters passed to \code{\link{text}} (or 
\code{\link{par}}) such as \code{col}, \code{srt}, \code{family}, etc}
}
\value{
A vector of length two with the x- and y-coordinates of the text position.
}
\description{
Add text and expressions anywhere in a plot (including margins) with mouse
click(s).
}
\examples{
\dontrun{
plot.new()
click_text('hello', col = 'red', cex = .5)
click_text('goodbye', family = 'HersheyScript', cex = 3)
click_text(expression(sum(x ^ 2) == 5 ^ hat(x)), srt = 45)
}

}
\seealso{
\code{\link{click_shape}}; \code{\link{plotmath}} for help with plotting
mathematical expressions; \code{link{tcol}}
}

