% Generated by roxygen2 (4.0.0): do not edit by hand
\name{color.bar}
\alias{color.bar}
\title{Color legend}
\usage{
color.bar(cols, x = NULL, y = x, labels = NULL,
          at.x = par('usr')[2], at.y = par('usr')[3],
          cex.x = 1, cex.y = 1, ...)
}
\arguments{
\item{cols}{vector of color names (or hexadecimal) from low to high}

\item{x}{numeric vector of data}

\item{y}{optional numeric vector of data; to color by a variable other than
\code{x}, \code{y} must be specified for \code{labels} coordinates to work
properly; see examples}

\item{labels}{optional labels for color bar}

\item{at.x}{x-coordinate of lower-left corner of bar}

\item{at.y}{y-coordinate of lower-left corner of bar; note that the bar can
extend ouside the plotting region, so use \code{cex.x} and \code{cex.y} to
scale the bar down (or up)}

\item{cex.x,cex.y}{scaling parameters}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\value{
An \code{\link{invisible}} vector of colors (hexadecimal format)
used for the bar. Note that these colors may be mapped to \code{x} so that
only one \code{color.bar} needs to be called in the \code{col} argument for
both the coloring of points and the generation of the legend; see examples.
}
\description{
Continuous color bar legend
}
\examples{
plot.new()
color.bar(c('black','red','white','blue'), at.x = 0, at.y = 0)
text(x = 0, y = seq(0, par('usr')[4], length = 6), pos = 4, xpd = TRUE,
     labels = pretty(seq(0, par('usr')[4]), n = 6), cex = .8, offset = .75)

\dontrun{
## calling color.bar once in the col argument will color x accordingly
# and plot the legend simultaneously

## compare:
png('/users/rawr/desktop/tmp.png')
par(mfrow = c(2,1))
plot(mtcars$mpg, pch = 19, cex = 2,
     col = colorRampPalette(c('yellow','red'))(1000)[rescaler(mtcars$mpg, c(1, 1000))])
plot(mtcars$mpg, pch = 19, cex = 2, bty = 'l',
     col = color.bar(c('yellow','red'), mtcars$mpg, labels = mtcars$mpg))
dev.off()

## plot a color.bar legend by a variable other than x
# use y argument to match the variable in the original plot call
# and x as the variable to color and label

## compare:
png('/users/rawr/desktop/tmp.png')
par(mfrow = c(2,1))
with(mtcars,
     plot(mpg, pch = 19, cex = 2, main = 'color by weight (red = heavier)',
     col = colorRampPalette(c('yellow','red'))(1000)[rescaler(wt, c(1, 1000))]))
with(mtcars,
     plot(mpg, pch = 19, cex = 2, bty = 'l', main = 'color by weight (red = heavier)',
     col = color.bar(c('yellow','red'), x = wt, y = mpg, labels = wt)))
dev.off()
}
}

