% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{ctext}
\alias{cmtext}
\alias{ctext}
\alias{ctitle}
\title{Color text}
\usage{
ctext(text, cols, space = TRUE, ...)

cmtext(text, cols, space = TRUE, ...)

ctitle(main = NULL, sub = NULL, xlab = NULL, ylab = NULL, cols,
  space = TRUE, ...)
}
\arguments{
\item{text}{vector of text}

\item{cols}{vector of colors; should be the same lenght as \code{text} or
will me recycled with a warning}

\item{space}{logical; if \code{TRUE}, adds space between \code{text}}

\item{...}{additional parameters passed to \code{text} or \code{mtext}}

\item{main, sub, xlab, ylab}{vector(s) of text for specific labels}
}
\description{
Add color to individual words in text functions. \code{ctext},
\code{cmtext}, and \code{ctitle} are analogous to \code{\link{text}},
\code{\link{mtext}}, and \code{\link{title}}, respectively. Note that
\code{title} accepts some graphical parameters specific to the label type,
e.g., \code{col.main}, but this is not implemented in \code{ctitle}--colors
will be recycled if more than one label type is given. Similarly, further
graphical parameters such as \code{cex} or \code{line} will be passed to
all label types; see examples.
}
\examples{
plot(1, ann = FALSE)
ctext(x = 1, y = 1, text = c('hello','little','point'), cols = 1:3, pos = 1)
cmtext(c('a','side','label'), 1:2, space = FALSE, side = 4, cex = 3)

## note that line, cex, font, etc will be recycled
ctitle(main = c('the','main','label'), xlab = c('x','label'),
       ylab = c('y','label'), sub = c('sub', 'label'), col = 3:5)
ctitle(xlab = c('another','label'), ylab = c('another','label'),
       font = 3, col = 1:2, line = 2, cex = 1.5)

}

