% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{cum_reset}
\alias{cum_reset}
\title{Cumulative reset}
\usage{
cum_reset(x, value = 0L, FUN = cumsum)
}
\arguments{
\item{x}{a vector}

\item{value}{a value of \code{x} which signals the end of a group and
resets \code{FUN}}

\item{FUN}{function to apply to each group}
}
\value{
A vector having the same length as \code{x} with \code{FUN} applied to
each group defined by positions of \code{value}.
}
\description{
Reset a cumulative function if a value is encountered.
}
\examples{
x <- 1:10
cum_reset(x, 5, cummin)

x[x \%\% 4 == 0] <- 0
cum_reset(x)
cum_reset(x, FUN = sum)

set.seed(1)
data.frame(x = x <- rpois(15, 1),
           y = cum_reset(x),
           z = cum_reset(x, 0, function(x) ave(x, FUN = sum)))

## x need not be numeric if FUN is appropriate for typeof(x)
cum_reset(letters[1:10], c('d','g'), function(x) letters[as.numeric(factor(x))])

}
\seealso{
\code{\link{cummin}}, \code{\link{cummax}}, \code{\link{cumsum}},
\code{\link{cumprod}}, \code{\link{ave}}
}

