% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dapply}
\alias{dapply}
\title{Diagonal apply}
\usage{
dapply(X, MARGIN, FUN, ..., SIMPLIFY = TRUE)
}
\arguments{
\item{X}{a matrix or an object to be coerced}

\item{MARGIN}{the corner to slice from to the opposite corner: \code{1} for
lower left, \code{2} for top left, \code{3} for top right, or \code{4} for
lower right}

\item{FUN}{a function to apply to each slice}

\item{...}{additional arguments passed to \code{FUN}}

\item{SIMPLIFY}{logical or character string; attempt to reduce the result to
a vector or matrix; see the \code{simplify} argument of \code{\link{sapply}}}
}
\description{
Apply a function on all diagonal slices of a matrix starting from one
of the corners.
}
\examples{
mat <- matrix(1:12, 3)
dapply(mat, 2, identity)

dapply(mat, 2, range)
dapply(mat, 2, range, SIMPLIFY = FALSE)

}
\seealso{
\code{\link{apply}}; \code{\link{simplify2array}}
}
