% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{date_cutoff}
\alias{date_cutoff}
\title{Apply cutoff date}
\usage{
date_cutoff(
  data,
  date = NULL,
  exclude = NULL,
  verbose = FALSE,
  FUN = match.fun(">")
)
}
\arguments{
\item{data}{a data frame}

\item{date}{date of cutoff}

\item{exclude}{optional character string to match date variable names
to be excluded in the comparison}

\item{verbose}{logical; if \code{TRUE}, print summary of pre/post data with
the min/max dates pre and post and the count and number missing for each
date variable}

\item{FUN}{the function used in the comparison to \code{date}; default is
\code{>}, but any \code{\link{Comparison}} operator would work}
}
\description{
Drop rows where any dates occur after a cutoff date. By default, all date
variables in \code{data} will be compared to \code{date}
}
\examples{
dd <- data.frame(id = 1:3, d1 = 1:3, d2 = c(3, NA, 5))
dd[, -1] <- lapply(dd[, -1], as.Date)
date_cutoff(dd, '1970-01-05')
date_cutoff(dd, '1970-01-05', exclude = 'd2')
date_cutoff(dd, '1970-01-03', verbose = TRUE, FUN = `<=`)

}
