% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{describeDate}
\alias{describeDate}
\alias{describeDateBy}
\title{Describe a date}
\usage{
describeDate(x, format = "\%b \%d, \%Y", copula = " to ",
  FUN = range)

describeDateBy(x, by, format = "\%b \%d, \%Y", copula = " to ",
  FUN = range, add_total_col = TRUE, useNA = c("ifany", "no",
  "always"), useNA.digits = 0L, percentage_sign = FALSE)
}
\arguments{
\item{x, by}{vectors of dates and the variable to split \code{x} by}

\item{format}{the date format; see \code{\link{strptime}}}

\item{copula}{character string to separate the range}

\item{FUN}{a function to summarize \code{x}, usually \code{range} or
\code{min}}

\item{add_total_col}{logical, \code{"first"}, or \code{"last"}; adds
the total column to the output}

\item{useNA}{how to handle missing values, one of \code{"ifany"} (default),
\code{"no"}, or \code{"always"}}

\item{useNA.digits}{number of digits to use for missing percentages}

\item{percentage_sign}{logical; if \code{TRUE}, percent signs are added}
}
\description{
A function that returns the range of a date object.
}
\examples{
x <- c(1:4, NA, 5)
d <- as.Date(x, origin = Sys.Date())
y <- c(1, 1, 2, 2, 1, 3)
describeDate(x)
describeDateBy(d, y, copula = ' - ', format = '\%d \%B')
describeDateBy(d, y, percentage_sign = TRUE, useNA.digits = 2, FUN = min)

dd <- data.frame(x, y, d)
tabler_stat(dd, 'd', 'y')
tabler_stat2(dd, c(Integer = 'x', Date = 'd'), c(Byvar = 'y'))

}
\seealso{
\code{\link[Gmisc]{getDescriptionStatsBy}}
}
