% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{describeFuns}
\alias{describeFuns}
\title{Describe multiple functions}
\usage{
describeFuns(x, FUN, ...)
}
\arguments{
\item{x}{a vector to be described by each \code{FUN}}

\item{FUN}{a list of functions that each return a \emph{named} string; if
\code{FUN} is a named list, all non zero length names will take
precedence; if neither are available, names will be guessed based on
the list}

\item{...}{ignored}
}
\description{
Concatenate one or more functions for output in \code{\link{tabler_stat}}.
}
\examples{
describeFuns(1:10, FUN = list(mean = mean, median = median))
describeFuns(1:10, FUN = list(Gmisc::describeMean, Gmisc::describeMedian))
describeFuns(1:10, FUN = list(function(...) Gmisc::describeMedian(..., iqr = FALSE)))


## typical usage
ff <- list(
  'Median (full range)' = function(...) Gmisc::describeMedian(..., iqr = FALSE),
  Gmisc::describeMean,
  Min = min, Max = max
)
  
tabler_stat2(
  mtcars, c('mpg', 'wt'), 'gear', FUN = NA,
  statArgs = list(continuous_fn = function(x, ...) describeFuns(x, FUN = ff))
)

}
