% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat.R
\name{dlt_table}
\alias{dlt_table}
\title{DLT table}
\usage{
dlt_table(low, high, delta = 10)
}
\arguments{
\item{low}{lowest true DLT rate, percent}

\item{high}{highest true DLT rate, percent}

\item{delta}{interval of DLT rate sequence; default is 10}
}
\description{
Creates a standard 3x3 dose-limiting toxicity table with probabilities of
dose-escalation.
}
\examples{
dlt_table(10, 50)
dlt_table(10, 50, 5)
}

