% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils2.R
\name{dmy}
\alias{dmy}
\title{Date parse}
\usage{
dmy(d, m, y, origin = c(1, 1, 1900))
}
\arguments{
\item{d,m,y}{day, month, year as single integers or vectors}

\item{origin}{a vector of length three giving the origins for \code{d},
\code{m}, and \code{y}, respectively; see details}
}
\description{
Parses day, month, year columns to the standard date format.
}
\details{
For two-digit years, the \code{origin} year should be specified; otherwise,
the default of 1900 will be used. For \code{NA} year, month, or day,
\code{origin} is used for defaults, i.e., \code{origin = c(15, 6, 2000)}
will convert missing days to day 15, missing months to June, and missing
years to 2000.
}
\examples{
dmy(25, 7, 13)

## this function is vectorized:
dmy(NA, NA, 2000:2009)

set.seed(1)
dd <- data.frame(id = 1:10,
                 day = sample(1:31, 10),
                 month = sample(1:12, 10),
                 year = sample(1000:2500, 10))

cbind(dd, dt = with(dd, dmy(day, month, year)))

dd1 <- within(dd, {
  day[sample(1:10, 5)] <- NA
  month[sample(1:10, 2)] <- NA
  year[sample(1:10, 2)] <- NA
})

cbind(dd1,
      new_date = with(dd1, dmy(day, month, year, origin = c(15, 6, 2000))))
}

