% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_extra.R
\name{dodge}
\alias{dodge}
\alias{dodge.formula}
\alias{dodge.default}
\title{Point dodge}
\usage{
dodge(x, ...)

\method{dodge}{formula}(formula, data = NULL, ...)

\method{dodge}{default}(x, y, dist, jit, ...)
}
\arguments{
\item{x}{grouping variables or, equivalently, positions along x-axis}

\item{...}{additional arguments passed to other methods}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ group}, where
\code{y} is a numeric vector of data values to be split into groups
according to the grouping variable, \code{group}}

\item{data}{optional matrix or data frame containing the variables in
\code{formula}; by default, the variables are taken from
\code{environment(formula)}}

\item{y}{a numeric vector of data, y-values}

\item{jit, dist}{jittering parameters; \code{jit} describes the spread of
close points, and \code{dist} defines a range to consider points "close"}
}
\description{
Dodge and center overlapping points by group. Spreads scattered points
similar to \code{jitter} but symmetrically. Although the default method
can be used, it is recommended to use the formula method for ease of use
and to set useful defaults for \code{jit} and \code{dist}.
}
\examples{
## these are equivalent ways to call dodge:
dodge(mpg ~ gear + vs, mtcars)
with(mtcars, dodge(list(gear, vs), mpg))
dodge(mtcars[, c('gear', 'vs')], mtcars$mpg)


## compare to overlapping points and jittering
op <- par(no.readonly = TRUE)
sp <- split(mtcars$mpg, do.call(interaction, mtcars[, c('gear','vs')]))
plot.new()
par(mar = c(0,0,0,0), cex = 2)
plot.window(c(.5,6.5),c(10,35))
box()
for (ii in seq_along(sp))
  points(rep(ii, length(sp[[ii]])), sp[[ii]])
for (ii in seq_along(sp))
  points(jitter(rep(ii, length(sp[[ii]]))), sp[[ii]], col = 4, pch = 1)
points(dodge(mpg ~ gear + vs, mtcars), col = 2, pch = 4)
legend('topleft', pch = c(1,1,4), col = c(1,4,2), cex = .8,
       legend = c('overlapping','random jitter','dodging'))
par(op)


## practical use
boxplot(mpg ~ vs + gear, data = mtcars)
points(dodge(mpg ~ vs + gear, data = mtcars), col = 2, pch = 19)

}
\seealso{
\code{\link{jitter}}, \code{\link{tplot}}
}
