% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{dodge2}
\alias{dodge2}
\alias{dodge2.default}
\alias{dodge2.formula}
\title{Point dodge}
\usage{
dodge2(x, ...)

\method{dodge2}{default}(x, y, jit = NULL, dist = NULL, ...)

\method{dodge2}{formula}(formula, data = NULL, ...)
}
\arguments{
\item{x}{grouping variables or, equivalently, positions along x-axis}

\item{...}{additional arguments passed to or from other methods}

\item{y}{a numeric vector of data, y-values}

\item{jit, dist}{jittering parameters; \code{jit} describes the spread of
close points, and \code{dist} defines a range to consider points "close";
both may be specified for each group and recycled as needed}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ group}, where
\code{y} is a numeric vector of data values to be split into groups
according to the grouping variable, \code{group}}

\item{data}{optional matrix or data frame containing the variables in
\code{formula}; by default, the variables are taken from
\code{environment(formula)}}
}
\description{
Dodge and center overlapping points by group. Spreads scattered points
similar to \code{jitter} but symmetrically. Although the default method
can be used, it is recommended to use the formula method for ease of use
and to set useful defaults for \code{jit} and \code{dist}.
}
\examples{
## these are equivalent ways to call dodge2:
dodge2(mpg ~ gear + vs, mtcars)
with(mtcars, dodge2(list(gear, vs), mpg))
dodge2(mtcars[, c('gear', 'vs')], mtcars$mpg)


## compare to overlapping points and jittering
sp <- split(mtcars$mpg, do.call(interaction, mtcars[, c('gear','vs')]))
plot.new()
op <- par(mar = c(0,0,0,0), cex = 2)
plot.window(c(.5,6.5),c(10,35))
box()
for (ii in seq_along(sp))
  points(rep(ii, length(sp[[ii]])), sp[[ii]])
for (ii in seq_along(sp))
  points(jitter(rep(ii, length(sp[[ii]]))), sp[[ii]], col = 4, pch = 1)
points(dodge2(mpg ~ gear + vs, mtcars), col = 2, pch = 4)
legend('topleft', pch = c(1,1,4), col = c(1,4,2), cex = .8,
       legend = c('overlapping','random jitter','dodging'))
par(op)


## practical use
boxplot(mpg ~ vs + gear, data = mtcars)
points(dodge2(mpg ~ vs + gear, data = mtcars), col = 2, pch = 19)

}
\seealso{
\code{\link{jitter}}; \code{\link{tplot}}; \code{\link{dodge}}
}
