% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{facet_adjust}
\alias{facet_adjust}
\title{Facet labeling}
\usage{
facet_adjust(x, pos = c("up", "down"), newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{x}{\code{\link[ggplot2]{ggplot}} object}

\item{pos}{position of labels}

\item{newpage}{draw new (empty) page first; see 
\code{\link[ggplot2]{print.ggplot}}}

\item{vp}{viewport to draw plot in; see \code{\link[ggplot2]{print.ggplot}}}
}
\value{
\code{facet_adjust} object that inherits \code{\link[gtable]{gtable}} class
}
\description{
Adjusts labels on x-axes when using \code{\link[ggplot2]{facet_wrap}}.
}
\examples{
library('ggplot2')
## missing some labels 
(tmp <- ggplot(diamonds[1:100, ], aes(carat, price, colour = clarity)) + 
  geom_point() + facet_wrap( ~ cut))
facet_adjust(tmp)
facet_adjust(tmp, pos = 'down')

}

