% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor2}
\alias{factor2}
\title{Combine values to factor levels}
\usage{
factor2(
  x = character(0L),
  levels = NULL,
  exclude = NA,
  ordered = is.ordered(x),
  regex = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct values}

\item{levels}{a \emph{named} list of unique values or regular expressions
  to group elements of \code{x}

  a \code{NULL} group will collapse all non matches to a single category;
  however, if there is no \code{NULL} group, non matches will be unchanged

  if \code{regex = TRUE} or flagged with \code{(?r)}, matching elements of
  \code{x} will be combined}

\item{exclude, ordered}{passed to \code{\link{factor}}}

\item{regex}{logical; if \code{TRUE}, \code{levels} is treated as a list
of regular expressions or otherwise matched exactly (unless \code{levels}
is flagged with \code{(?r)})}

\item{...}{additional arguments passed to \code{\link{grep}} when using
regular expressions}
}
\description{
Convenience function to combine level(s) of a vector into a new or existing
factor level(s) by unique value and/or regular expression.

Requires \code{R >= 3.5.0} for duplicated labels in \code{\link{factor}}s.
}
\examples{
## combine numeric, character, or factor
x <- rep(1:5, each = 2)
factor2(x, list('1' = 1:2))
factor2(x, list(a = 1:2, b = 3))    ## all non matches unchanged
factor2(x, list(a = 1:2, b = NULL)) ## all non matches grouped to b

## levels may be swapped without losing original distinction
factor2(x, list('3' = 1:2, '1' = 3))
factor2(x, list('1' = 3, '3' = 1:2))

## combine by regular expression
x <- letters[1:5]
factor2(x, list(a = c('a', 'b')), regex = TRUE)
factor2(x, list(a = c('a|b')), regex = TRUE) ## same
factor2(x, list(a = 'a', b = 'b|c|e'), regex = TRUE)

## use regex only when flagged with (?r)
## note: must use regex = FALSE, otherwise all are treated as regex
x <- c('a', 'aaa', 'abc', 'def')
factor2(x, list(a = 'a', b = '(?r)a.', d = '(?r)d'))
factor2(x, list(a = 'a', b = '(?r)a.', d = 'd'))

}
\seealso{
\code{\link{combine_levels}}; \code{\link{combine_regex}}
}
