% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{factors}
\alias{factors}
\title{Find factors}
\usage{
factors(...)
}
\arguments{
\item{...}{integers}
}
\description{
Find common factors of two or more integers.
}
\examples{
factors(21)
factors(3 * 2 ^ 20)
factors(64, 128, 58)

}
