% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{fakeglm}
\alias{fakeglm}
\title{Fake GLM}
\usage{
fakeglm(formula, ..., family, data = NULL)
}
\arguments{
\item{formula}{a \code{\link{formula}}}

\item{...}{coefficients for new model; see details}

\item{family}{a description of the error distribution and link to be used
in the model; this can be a character string naming a family function or
the result of a clall to a family function; see \code{\link{family}}}

\item{data}{a data frame}
}
\description{
Creates a "fake" \code{\link{glm}} object using specific coefficients
which can be used with \code{predict.glm} without fitting a model first.
}
\details{
Using \code{data} a \code{glm} object will be created with a series of the
desired coefficients passed in \dots in the order that they would appear
in \code{names(coef(glm(...)))}. An unnamed value will be treated as
the intercept and all other coefficients must be named in the order given
in \code{formula}, eg, \code{1, x1 = 1, x2 = 3}, etc.

Also note that factor variables must already be factors in \code{data},
that is, it is not possible to use \code{y ~ factor(x)} currently in
\code{fakeglm}.
}
\examples{
f1 <- glm(vs ~ mpg + wt + disp, data = mtcars, family = 'binomial')
p1 <- predict(f1, type = 'response')

f2 <- fakeglm(vs ~ mpg + wt + disp, data = mtcars, family = 'binomial',
              -21.9020, mpg = 0.6470, wt = 5.3315, disp = -0.0403)
p2 <- predict(f2, newdata = mtcars, type = 'response')

all.equal(p1, p2, tolerance = .0001) ## TRUE

dat <- within(mtcars, gear <- factor(gear))
fakeglm(vs ~ mpg + gear, data = dat, family= 'binomial', 0, mpg = 1,
        gear4 = 3, gear5 = 1)

}
\references{
\url{https://gist.github.com/MrFlick/ae299d8f3760f02de6bf}
}
