% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{fcols}
\alias{fcols}
\title{Find columns by pattern}
\usage{
fcols(x, pattern, keep, ...)
}
\arguments{
\item{x}{a matrix or data frame}

\item{pattern}{pattern to match}

\item{keep}{optional vector of names of other columns to keep}

\item{...}{additional parameters passed to \code{\link{grep}}}
}
\description{
Quickly selects and returns columns from a matrix or data frame by
\code{\link{grep}}'ing for a desired \code{pattern}.
}
\examples{
fcols(iris, 'Petal')
fcols(iris, '\\\\.')
fcols(mtcars, '^[\\\\w]{2}$')
}

