% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{fill_df}
\alias{fill_df}
\title{Fill data frame}
\usage{
fill_df(data, key, ids, fill, values)
}
\arguments{
\item{data}{a data frame to recode}

\item{key}{a look-up table data frame}

\item{ids}{columns treated as id variables, as character strings or indices}

\item{fill}{columns to recode, as character strings or column indices}

\item{values}{optional vector of values to recode with \code{fill}; if
missing (default), \code{fill_df} only looks for \code{NA}s in \code{data};
otherwise, all occurrences of \code{values} will be replaced with
\code{NA}, and \code{fill_df} will procede normally}
}
\value{
A data frame with \code{NA}s from \code{fill}-columns recoded to match
the values from \code{key}.
}
\description{
Fills data frame, \code{data}, containing \code{NA} values using a look-up
table, \code{key}. \code{ids} and \code{fill} columns must be in both
\code{data} and \code{key}. If neither are given, \code{fill_df} will
smartly try to guess which columns need to be filled with the values from
the look-up table.
}
\examples{
dd <- mtcars
dd[matrix(sample(c(TRUE, FALSE), 32 * 11, replace = TRUE), 32)] <- NA
identical(mtcars, fill_df(dd, mtcars))  ## TRUE

## recode other variables instead of NAs
nn <- sum(is.na(dd))
dd[is.na(dd)] <- sample(-10:-1, nn, replace = TRUE)
identical(mtcars, fill_df(dd, mtcars, values = -1:-10)) ## TRUE

f <- function(x, n = 20) sample(x, size = n, replace = TRUE)
set.seed(1)
key_df <- data.frame(id = c(1,2,1,2), group = c(3,3,4,4),
                     x = c(100, 200, 300, 400), y = I(LETTERS[1:4]))
na_df <- data.frame(id = f(1:2), group = f(3:4),
                    x = f(c(0, NA)), y = I(f(c('', NA))), z = 1)

## auto: all cols with no NAs == ids; cols with any NAs = fill
fill_df(na_df, key_df)

## select which to be filled and returned
fill_df(na_df, key_df, ids = 1:2, fill = 'x')
fill_df(na_df, key_df, ids = 1:2, fill = 4)

}
\seealso{
\code{\link{recoder}}; \code{\link{locf}}
}
