% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{Flatten lists}
\usage{
flatten(l)
}
\arguments{
\item{l}{a list}
}
\description{
Flattens lists and nested lists of vectors, matrices, and/or data frames.
}
\examples{
(l <- list(matrix(1:3), list(1:3, 'foo'), TRUE, 'hi',
           list(head(mtcars), list(tail(mtcars)))))
flatten(l)

}
\references{
\url{https://stackoverflow.com/questions/8139677/how-to-flatten-a-list-to-a-list-without-coercion}
}

