% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{ggcaterpillar}
\alias{ggcaterpillar}
\title{Caterpillar plot}
\usage{
ggcaterpillar(re, qq = TRUE, likeDotplot = TRUE)
}
\arguments{
\item{re}{random effects from lmer object}

\item{qq}{if \code{TRUE}, returns normal q/q plot; else returns caterpillar
dotplot}

\item{likeDotplot}{if \code{TRUE}, uses different scales for random effects,
i.e., \code{\link[ggplot2]{facet_wrap}}}
}
\description{
Caterpillar plots for random effects models using \code{\link{ggplot}}.
}
\details{
Behaves like \code{\link[lattice]{qqmath}} and 
\code{\link[lattice]{dotplot}} from the lattice package; also handles
models with multiple correlated random effects
}
\examples{
\donttest{
library('lme4')
fit <- lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy)

ggcaterpillar(ranef(fit, condVar = TRUE))

## compare (requires lattice package)
lattice::qqmath(ranef(fit, condVar = TRUE))
}
}

