% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggheat}
\alias{ggheat}
\title{ggheat}
\usage{
ggheat(cors = NULL, data = NULL, limits = c(-1, 1),
  gradn = rev(heat.colors(10)), gradc = c("white", "steelblue"))
}
\arguments{
\item{cors}{matrix (or matrix-like object) of correlations}

\item{data}{data frame of data (in progress)}

\item{limits}{range of color scale in legend; see 
\code{\link[ggplot2]{discrete_scale}}}

\item{gradn}{vector of \code{n} colors to use, from low to high; see details}

\item{gradc}{vector of length two, low color and high color; see details

\code{gradc} takes a low color and a high color, respectively, and generates
a continuous scale between those colors; see 
\code{\link[ggplot2]{scale_fill_gradient}}}
}
\description{
Function to plot a heat map using \code{\link{ggplot}}.
}
\details{
\code{gradn} takes a vector of \code{n} colors: either a list of names, a 
list of hexadecimal values, an existing color palette (i.e., 
\code{\link{heat.colors}}, \code{\link{rainbow}}, etc); see also 
\code{\link{palette}}, \code{\link{colors}}, the \code{RColorBrewer} 
package, etc.
}
\examples{
library('ggplot2')
tmp <- rescaler(matrix(1:25, 5))
diag(tmp) <- 1
colnames(tmp) <- rownames(tmp) <- LETTERS[1:5]
ggheat(cors = tmp, limits = c(0, 1))
ggheat(cors = tmp, gradn = NULL, gradc = c('white','red'))

}
\seealso{
\code{\link{cor}}, \code{\link{ggheat2}}, \code{\link[iplotr]{icorr}}
}

