% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggheat2}
\alias{ggheat2}
\title{ggheat2}
\usage{
ggheat2(data, corr = cor(data, use = "pairwise.complete"), cluster = TRUE,
  nbreaks = NULL, palette = if (is.null(nbreaks)) c("blue", "white", "red")
  else 1, legend_name = expression(rho), pch, cex = c(2, 6),
  label = FALSE, label_alpha = FALSE, label_color = "black",
  label_digits = 2, midpoint = 0, clim = c(-1, 1), ...)
}
\arguments{
\item{data}{a data frame or matrix (observations x variables) of numeric
values}

\item{corr}{a correlation matrix}

\item{cluster}{logical or function; if \code{TRUE}, the variables will be
clustered and reordered; if \code{FALSE}, no reordering will be done;
otherwise, a custom clustering function may be given; see details}

\item{nbreaks}{number of breaks to categorize the correlations (default is
\code{NULL}, ie, a continuous scale)}

\item{palette}{for a continuous scale, a vector of length three giving the
low, mid, and high colors of the gradient (default is
\code{c('blue','white',''red)}, see \code{\link{scale_colour_gradient2}});
for a discrete scale, a character string of the palette name or an integer
giving the index of the palette, see \code{\link{scale_colour_brewer}}}

\item{legend_name}{the legend name; see \code{\link{discrete_scale}}}

\item{pch}{(optional) plotting character; if \code{missing},
\code{\link{geom_tile}} is used instead of plotting characters}

\item{cex}{size of \code{pch}; a vector of length one (all \code{pch} will
be the same size) or two (size will be proportional to the correlation);
default is c(2,6); see\code{\link{scale_size_identity}}}

\item{label}{logical; if \code{TRUE}, adds correlation coefficients on top
of each \code{pch}}

\item{label_alpha}{logical, if \code{TRUE}, adds alpha transparency when
\code{label} is \code{TRUE} (correlations closer to 0 will be less visible)}

\item{label_color}{color of correlations (default is \code{'black'})}

\item{label_digits}{number of digits in correlation labels}

\item{midpoint}{the midpoint value for continuous scaling of correlations
(default is \code{0})}

\item{clim}{vector of length two giving the limits of correlation
coefficients (default is \code{-1,1})}

\item{...}{additional arguments passed to \code{\link{geom_text}} for the
diagonal labels}
}
\description{
Function to plot a heat map using \code{\link{ggplot}}.
}
\details{
Default cluster method is \code{stats::hclust(dist(x), method = 'average')}
which will return a list containing a named vector, "order", which is used
to reorder the variables.

In order to pass a custom clustering function to \code{cluster}, the
function must take a single input (a correlation matrix) and return either
a vector or a list with a named vector, \code{"order"}.
}
\examples{
library('ggplot2')
ggheat2(mtcars)

ggheat2(mtcars, label = TRUE, label_alpha = TRUE, cluster = FALSE,
        ## additional args passed to diagonal labels
        colour = 'red', angle = 45, size = 7)

ggheat2(mtcars, pch = 19, nbreaks = 6, cex = c(2,10),
        palette = 'PuOr',         ## colorblind palette
        size = 5, hjust = 0.75) + ## passed to diag text
    labs(title = 'Correlation Matrix')

## custom clustering function
ggheat2(data = NULL, corr = cor(mtcars, use = 'pairwise'),
        nbreaks = 5, palette = 'Blues',
        cluster = function(...) sample(ncol(mtcars)))

}
\seealso{
\code{\link{cor}}, \code{\link{ggheat}}, \code{\link[iplotr]{icorr}},
\code{\link[arm]{corrplot}}, \url{https://github.com/briatte/ggcorr}
}

