% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggmultiplot}
\alias{ggmultiplot}
\title{Draw multiple ggplot objects in a single layout}
\usage{
ggmultiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{list of ggplot objects (optional)}

\item{cols}{number of columns in layout}

\item{layout}{matrix specifying the layout; if present, \code{cols} is
ignored; if layout is \code{matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE)}
for example, then plot 1 will go in the upper left, 2 will go in the upper
right, and 3 will go all the way across the bottom.}
}
\description{
Uses functions in \pkg{grid} to arrange one or more
\code{\link[ggplot2]{ggplot}} objects into a single layout
}
\examples{
data(mtcars)
library(ggplot2)

tmp <- ggplot(mtcars, aes(factor(1), fill = factor(cyl))) +
  geom_bar(width = 1)
tmp1 <- tmp + coord_polar()
tmp2 <- tmp + coord_polar(theta = 'y')
tmp3 <- ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar(width = 1) + coord_polar()

ggmultiplot(tmp1, tmp2, tmp3, cols = 2)
ggmultiplot(tmp1, tmp2, tmp3,
  layout = matrix(c(1, 2, 3, 3, 3, 3), ncol = 2, byrow = TRUE))
}

