% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist_url_to_raw.R
\name{gist_url_to_raw}
\alias{gist_url_to_raw}
\title{Convert github gist url into equilvalent url for raw code}
\usage{
gist_url_to_raw(url)
}
\arguments{
\item{url}{Link to a github gist}
}
\value{
A character vector of length 1 containing url for the raw version of the github gist
}
\description{
Convert github gist url into equilvalent url for raw code
}
\examples{

gist_url_to_raw("https://gist.github.com/korkridake/ea96f6569f5a4f9a1a552145e6151235")

}
