% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{h2c}
\alias{h2c}
\alias{html2char}
\alias{c2h}
\title{html2char}
\usage{
h2c(x, to_html = FALSE)

c2h(x)
}
\arguments{
\item{x}{a character or factor vector}

\item{to_html}{logical; if \code{TRUE}, convert to html entities}
}
\description{
Convert html entities to character or vice versa.
}
\examples{
x <- c('<', '>', '≤', '+-', '+/-')
c2h(x)
c2h(factor(x))
h2c(c2h(x)) == x

h2c(c('&Beta;', '&beta;'))

}
