% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{helpExtract}
\alias{helpExtract}
\title{Extract R help files}
\usage{
helpExtract(FUN, show.sections = FALSE, section = "Usage",
  type = c("text", "md_code", "md_text", "sw_code", "sw_text"), ...)
}
\arguments{
\item{FUN}{a function as name or character string}

\item{show.sections}{logical; if \code{TRUE}, returns \code{section} options
for \code{FUN}}

\item{section}{section to extract (default is \code{"Usage"}}

\item{type}{type of character vector you want returned; default is 
\code{"m_code"}, see details}

\item{...}{additional arguments passed to \code{\link[utils]{help}}}
}
\value{
A character vector to be used in a Sweave or Rmarkdown document.
}
\description{
Extracts specified portions of R help files (from \emph{loaded} libraries)
for use in Sweave or R-markdown documents.
}
\details{
The \code{type} argument accepts:
\itemize{ 
\item \code{"m_code"}: Markdown code chunks; for use with markdown documents
when highlighted code is expected.
\item \code{"m_text"}: Markdown plain text; for use with markdown documents 
where regular text is expected.
\item \code{"s_code"}: Sweave code chunks; for use with Sweave documents 
where highlighted code is expected.
\item \code{"s_text"}: Sweave plain text; for use with Sweave documents 
where regular text is expected.
}

To see the results in the console:

\code{cat(helpExtract(print, type = 'm_text'))}

To insert a (highlighted) chunk into a markdown document:

\verb{
```{r, results='asis'}
cat(helpExtract(print), sep ='\n')
```
}

To insert a (highlighted) chunk into a Sweave document:

\verb{
\\Sexpr{knit_child(textConnection(helpExtract(print, type = 's_code')),
     options = list(tidy = FALSE, eval = FALSE))}
}
}
\examples{
helpExtract(print, type = 'text')

cat(helpExtract(print), sep = '\\n')

cat(helpExtract(print, type = 'md_text', section = 'description'))

## selecting multiple sections prints section names
cat(helpExtract(print, section = c('references', 'see also')), sep = '\\n')

}

