% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{hl_est}
\alias{hl_est}
\title{Hodges-Lehmann estimator}
\usage{
hl_est(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric or logical vector}

\item{na.rm}{logical; if \code{TRUE}, \code{NA} values are removed}
}
\description{
A function to calculate a nonparametric estimation of a population's
location parameter for one sample described by Hodges and Lehmann (1963).
}
\details{
For a numeric vector \code{x} with length \code{n}, the median of the
means of \code{n(n+1)/2} pairs (i.e., including the vector itself)
are calculated.

For symmetric distributions, the Hodges-Lehmann statistic estimates the
population's median and has greater efficiency than the sample median.
For the normal distribution, the Hodges-Lehmann statistic is nearly as
efficient as the sample mean and better when estimating mixtures of
normal distributions.

For non-symmetric distributions, the Hodges-Lehmann statistic estimates
the population's "pseudo-median" (see \code{\link[stats]{wilcox.test}}), a
location parameter that is closely related to the median. The pseudo-median
is well defined for all distributions of random variables having dimension
two or greater, and like the median, the pseudo-median is defined for even
heavy-tailed distributions that lack any finite mean.
}
\examples{
set.seed(1)
x <- rnorm(100)
hl <- hl_est(x)

## compare
hl2 <- wilcox.test(x, exact = TRUE, conf.int = TRUE)
identical(hl, unname(hl2$estimate))

}
\references{
Hettmansperger, T. P.; McKean, J. W. (1998). \emph{Robust nonparametric
statistical methods}. Kendall's Library of Statistics. \strong{5}.

Hodges, J. L.; Lehmann, E. L. (1963). "Estimation of location based on
ranks." \emph{Annals of Mathematical Statistics}. \strong{34} (2): 598-611.
}
\seealso{
\code{\link{wilcox.test}}; \code{ICSNP::hl.loc}
}
