% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{html_align}
\alias{html_align}
\title{Align html}
\usage{
html_align(x, sep = "&nbsp;", where = "&&", min_width = "35px")
}
\arguments{
\item{x}{an object of class \code{\link[htmlTable]{htmlTable}}}

\item{sep}{a character string used as the center of alignment}

\item{where}{a character string or regular expression (see examples)
defining where strings should be aligned; the easiest method is to
use \code{"&&"} at the desired alignment point}

\item{min_width}{minimum width of the span tag; too narrow will not
align strings but too wide adds whitespace}
}
\description{
Align text in columns of an \code{\link[htmlTable]{htmlTable}} at a
specific location (similar to \code{align} or \code{eqnarray} environments
in latex).
}
\examples{
library('htmlTable')
tmp <- within(cars, {
  align2 <- sprintf('\%s&&(\%s)', speed, dist)
  align1 <- sprintf('\%s (\%s)', speed, dist)
  raw    <- sprintf('\%s - (\%s)', speed, dist)
})

ht <- htmlTable(
  rawr::ht(tmp), n.cgroup = 2:3, cgroup = c('raw', 'align'),
  caption = 'caption', rnames = FALSE
)

## default
structure(ht, class = 'htmlTable')

## align at '&&'
structure(html_align(ht), class = 'htmlTable')
structure(html_align(ht, ' --- '), class = 'htmlTable')

## align at '&&' or ' '
## the regex should capture the left text in group 1, use non-capture
## for separating text, andn capture the right text in group 2
structure(
  html_align(ht, '&nbsp;', '(\\\\d+)(?: |&&)([()0-9]+)'),
  class = 'htmlTable'
)

}
