% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{html_test}
\alias{html_test}
\title{html test}
\usage{
html_test(x)
}
\arguments{
\item{x}{character string of html code}
}
\description{
Render html in rstudio viewer
}
\examples{
\dontrun{
html_test("
<div align = center><h1>A heading<sup>&dagger;</sup><h1></div>
<font size = 1><sup>&dagger;</sup>That was the heading</font>
")

library(htmlTable)
html_test(htmlTable(mtcars, output = FALSE))
}
}

