% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{icols}
\alias{icols}
\title{Index columns by pattern}
\usage{
icols(x, pattern, keep, ...)
}
\arguments{
\item{x}{a matrix or data frame}

\item{pattern}{pattern to match}

\item{keep}{optional vector of names of other columns to keep}

\item{...}{additional parameters passed to \code{\link{grep}}}
}
\description{
Quickly selects and returns columns from a matrix or data frame by
\code{\link{grep}}'ing for a desired \code{pattern}.
}
\examples{
icols(iris, 'Petal')
icols(iris, '\\\\.')
icols(mtcars, '^[\\\\w]{2}$')

}
