% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{identical2}
\alias{identical2}
\title{Test two or more objects for exact equality}
\usage{
identical2(
  ...,
  num.eq = TRUE,
  single.NA = TRUE,
  attrib.as.set = TRUE,
  ignore.bytecode = TRUE,
  ignore.environment = FALSE
)
}
\arguments{
\item{...}{any \code{R} objects}

\item{num.eq}{logical indicating if (\code{\link{double}} and
\code{\link{complex}} non-\code{\link{NA}}) numbers should be compared
using \code{\link{==}} ("equal"), or by bitwise comparison. The latter
(non-default) differentiates between -0 and +0}

\item{single.NA}{logical indicating if there is conceptually just one
numeric \code{NA} and one \code{\link{NaN}}; \code{single.NA = FALSE}
differentiates bit patterns}

\item{attrib.as.set}{logical indicating if \code{\link{attributes}} of
\code{...} should be treated as \emph{unordered} tagged pairlists
("sets"); this currently also applies to \code{\link{slot}}s of S4
objects. It may well be too strict to set \code{attrib.as.set = FALSE}}

\item{ignore.bytecode}{logical indicating if byte code should be ignored
when comparing \code{\link{closure}}s.}

\item{ignore.environment}{logical indicating if their environments should
be ignored when comparing \code{closure}s.}
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}, never \code{NA}
and never anything other than a single value.
}
\description{
The safe and reliable way to test two or more objects for being exactly
equal; returns \code{TRUE} in this case, \code{FALSE} in every other case.
}
\examples{
identical2(1, 1.)
identical2(1, 1., 1L)

## for unusual R objects:
identical2(.GlobalEnv, environment(), globalenv(), as.environment(1))

identical2(0., 0, -0.) ## not differentiated
identical2(0., 0, -0., num.eq = FALSE)

identical2(NaN, -NaN)
identical2(NaN, -NaN, single.NA = FALSE) ## differ on bit-level

## for functions
f <- function(x) x
g <- compiler::cmpfun(f)
identical2(f, g)
identical2(f, g, ignore.bytecode = FALSE)

}
\seealso{
\code{\link{identical}}; \code{\link{all.equal}} for descriptions of how
two objects differ; \code{\link{Comparison}} for operators that generate
element-wise comparisons; \code{\link{isTRUE}} is a simple wrapper based
on \code{identical}; \code{\link{all_equal2}}
}
