% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_domain.R
\name{identify_domain}
\alias{identify_domain}
\title{Identify the domain of the url}
\usage{
identify_domain(urls)
}
\arguments{
\item{urls}{Vector of urls}
}
\description{
Identify the domain of the url
}
\examples{

test_domains <- c("https://github.com/hadley/vis-eda/blob/master/travel.R",
"https://www.datacamp.com/community/tutorials/sentiment-analysis-R",
"https://www.tidytextmining.com/sentiment.html",
"https://www.kaggle.com/vrtjso/mercari-eda-more-info-than-you-can-imagine")

identify_domain(test_domains)



}
