% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{imgpal}
\alias{imgpal}
\title{Image palettes}
\usage{
imgpal(path, n = 10L)
}
\arguments{
\item{path}{full file path to image}

\item{n}{maximum number of colors to extract, result will be <= \code{n},
and the calculated number of unique colors will also be provided}

\item{x}{an object of class \code{"imgpal"}}

\item{fullrange}{logical; if \code{TRUE}, all <=\code{n} colors will be
displayed}

\item{...}{ignored}
}
\value{
A list of class \code{"imgpal"} with the following elements:
\item{filename}{the image file name}
\item{n_unique}{the calculated number of unique colors}
\item{col}{a vector of colors}
\item{counts}{frequency counts for each \code{col}}
\item{call}{the call made to \code{convert}}
\item{convert}{the result of \code{call}}
}
\description{
Extract unique and most commonly-used unique colors from an image file
(requires \href{https://imagemagick.org/index.php}{ImageMagick}).
}
\examples{
\dontrun{
ip <- imgpal('https://www.google.com/images/branding/googlelogo/2x/googlelogo_color_272x92dp.png')
show_pal(ip)

img <- system.file('fig', package = 'rawr')
img <- list.files(img, full.names = TRUE, pattern = 'g$')
op <- par(mfrow = n2mfrow(length(img)))
sapply(img, function(x) show_pal(imgpal(x), fullrange = TRUE))
par(op)
}

}
