% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_matrix}
\alias{insert_matrix}
\title{Insert rows or columns}
\usage{
insert_matrix(m, rowsep, colsep, rowrep = NA, colrep = rowrep)
}
\arguments{
\item{m}{a matrix}

\item{rowsep, colsep}{index of row or column to shift right or down,
respectively}

\item{rowrep, colrep}{row and column replacement values, recycled if needed;
note that rows are replaced first followed by columns, so for the opposite
behavior, transpose \code{m} before and after using \code{insert_matrix}
and swap \code{rowrep} and \code{colrep}; see examples}
}
\description{
Insert rows and/or columns into a matrix.
}
\examples{
m <- col(matrix(0, 5, 5))
insert_matrix(m, 2, c(2,4,4,4))

## anticipate number of values needed for replacement(s)
insert_matrix(m, 4, 4:5, colrep = 1:6)

## these are _almost_ identical
insert_matrix(m, 5, 5, 0, 1) == t(insert_matrix(t(m), 5, 5, 1, 0))

}

