% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{inset}
\alias{inset}
\title{Inset plots}
\usage{
inset(x, y = NULL, pct = 0.25, ...)
}
\arguments{
\item{x}{a keyword (see details) or a vector of length two giving the 
positions on the current plot at which to draw the inset plot along the
x-axis}

\item{y}{ignored if \code{x} is a keyword or a vector of length two giving
the positions on the current plot at which to draw the inset plot along the
y-axis}

\item{pct}{inset plot scaling (only used if \code{x} is a keyword)}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\description{
Inset new plots in existing plot windows.
}
\details{
\code{x} should be of the form \code{x0, x1} and similar for \code{y} 
giving the starting and ending coordinates to draw the inset plot and must
be in the range defined in the current plotting area.

Alternatively, \code{x} can be a keyword ("bottomright", "bottom",
"bottomleft", "left", "topleft", "top", "topright" "right", or "center")
giving the approximate location of the inset plot. \code{pct} is used to
adjust the size.
}
\examples{
op <- par(no.readonly = TRUE)
plot(mpg ~ wt, data = mtcars, col = 'blue')
abline(lm(mpg ~ wt, data = mtcars), col = 'red')
inset('topright', pct = .4)
hist(mtcars$mpg, ann = FALSE, panel.last = box(),
     col = 'dodgerblue2', las = 1)

par(op)
plot(1:10)
op <- par(no.readonly = TRUE)
Map(function(x) {
 inset(x, las = 1, col = 'red', pct = 1/3)
 plot(rnorm(10), ann = FALSE, axes = FALSE, panel.last = box())
 par(op)
 Sys.sleep(.5)
 }, c("bottomright", "bottom", "bottomleft", "left",
      "topleft", "top", "topright", "right", "center"))

}

