% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{install.bioc}
\alias{install.bioc}
\title{Install bioconductor}
\usage{
install.bioc(pkgs, upgrade = FALSE)
}
\arguments{
\item{pkgs}{character vector of package names to install}

\item{upgrade}{logical; if \code{TRUE} updates installed packages}
}
\description{
Installs bioconductor base, packages, or upgrades.
}
\examples{
\dontrun{
install.bioc()
install.bioc('Biostrings')
}

}
