% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{install_temp}
\alias{install_temp}
\title{Install packages temporarily}
\usage{
install_temp(pkgs, lib, ...)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose current
versions should be downloaded from the repositories}

\item{lib}{character vector giving the library directories where to install
\code{pkgs}; recycled as needed; if missing (default), a
\code{\link{tempdir}} will be created}

\item{...}{additional arguments passed to
\code{\link[utils]{install.packages}}}
}
\description{
This function will create a temporary \code{.libPath}, install, and load
packages for use in a single \code{R} session. \cr \cr To install a repo
from github temporarily, use \code{\link[withr]{with_libpaths}}.
}
\examples{
\dontrun{
install_temp(c('devtools', 'testthat'))
}

}
