% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{installRawfileReaderDLLs}
\alias{installRawfileReaderDLLs}
\alias{Thermo}
\alias{ThermoFisher}
\alias{ThermoFisherScientific}
\title{Install the New RawFileReader from Thermo Fisher Scientific}
\usage{
installRawfileReaderDLLs(
 
    sourceUrl = "https://github.com/compomics/ThermoRawFileParser/raw/master/packages/mzLib.1.0.450/lib/netstandard2.0/",
  ...
)
}
\arguments{
\item{sourceUrl}{url of New RawFileReader from Thermo Fisher Scientific
assemblies.}

\item{...}{other parameter for \code{download.file}}
}
\value{
An (invisible) vector of integer code, 0 for success and non-zero for
failure. For the "wget" and "curl" methods this is the status code returned
by the external program.
}
\description{
Install the New RawFileReader from Thermo Fisher Scientific
}
\seealso{
\url{https://planetorbitrap.com/rawfilereader}
}
