% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interleave}
\alias{interleave}
\title{Interleave rows or columns}
\usage{
interleave(..., which)
}
\arguments{
\item{...}{vectors, matrices, or data frames}

\item{which}{joining method; \code{'rbind'} or \code{'cbind'}}
}
\description{
Interleave rows (or columns) of vectors, matrices, or data frames.
}
\examples{
interleave(matrix(1:9, 3, 3), matrix(letters[1:9], 3, 3), which = 'rbind')
interleave(matrix(1:9, 3, 3), matrix(letters[1:9], 3, 3), which = 'cbind')
}

