% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{intr}
\alias{intr}
\title{Interval formatter}
\usage{
intr(..., fun = median, conf = NULL, digits = 0L, na.rm = FALSE)
}
\arguments{
\item{...}{numeric vector or string of numeric vectors}

\item{fun}{summary stat function, usually \code{\link{mean}} or
\code{\link{median}}}

\item{conf}{width of confidence interval in \code{[0,1]}; if \code{NULL}
(default), returns min and max of \code{...}}

\item{digits}{number of digits (includes trailing 0s)}

\item{na.rm}{logical; if \code{TRUE}, any \code{\link{NA}} and \code{NaN}
are removed from \code{...} before \code{FUN} and \code{\link{quantile}}
are computed}
}
\description{
Calculate summary statistic with range or confidence interval.
}
\examples{
intr(1:10)
intr(1:10, conf = 0.95)

## inner quartile range
cbind(
  'median (IQR)' = lapply(mtcars, intr, conf = 0.5, digits = 2),
  'mean (range)' = lapply(mtcars, intr, fun = mean, digits = 2)
)

## compare to
summary(mtcars)

}
\seealso{
\code{\link[rawr]{roundr}}
}
