% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{iprint}
\alias{iprint}
\title{In-line printing}
\usage{
iprint(..., wrap = "", sep = ", ", copula, digits = 2)
}
\arguments{
\item{...}{one or more numeric or character elements to be converted into
character vectors}

\item{wrap}{character string to wrap each term}

\item{sep}{character string to separate the terms}

\item{copula}{character string to separate last two terms}

\item{digits}{number of digits past the decimal point to keep; see
\code{\link{roundr}}}
}
\description{
Modified \code{\link[pander]{p}} function from the \pkg{pander} package.
}
\examples{
iprint('fee','fi','fo','fum')
iprint(rnorm(2))
iprint(-0.000, 0.100)
iprint(LETTERS[1:5], copula = ", and the letter ")
iprint("Thelma", "Louise", copula = " & ")

}
\seealso{
\code{\link{roundr}}, \code{\link[pander]{p}}
}

