% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{jt.test}
\alias{jt.test}
\title{Jonckheere-Terpstra test}
\usage{
jt.test(x, y = NULL)
}
\arguments{
\item{x}{either a two-dimensional contingency table in matrix form or a
factor object}

\item{y}{a factor object; ignored if x is a matrix}
}
\value{
A list with class "\code{htest}" containing the following components:

\item{statistic}{the value of the test statistic with a name describing it}
\item{p.value}{the p-value for the test (two-sided, corrected for ties)}
\item{method}{a character string describing the test used}
\item{data.name}{a character string giving the names of the data}
}
\description{
An implementation of Jonckheere's trend test for \emph{ordered} independent
samples.
}
\examples{
(tbl <- table(mtcars$gear, mtcars$cyl))
jt.test(tbl)
cor.test(mtcars$gear, mtcars$cyl, method = 'kendall')
# clinfun::jonckheere.test(mtcars$gear, mtcars$cyl)

## from stats::cor.test
x <- c(44.4, 45.9, 41.9, 53.3, 44.7, 44.1, 50.7, 45.2, 60.1)
y <- c(2.6, 3.1, 2.5, 5.0, 3.6, 4.0, 5.2, 2.8, 3.8)

cor.test(x, y, method = 'kendall', exact = FALSE)
jt.test(x, y)

}
\references{
Jonckheere, A. R. (1954). A distribution-free k-sample test again ordered
alternatives. \emph{Biometrika} \strong{41}:133-145.

Terpstra, T. J. (1952). The asymptotic normality and consistency of
Kendall's test against trend, when ties are present in one ranking.
\emph{Indagationes Mathematicae} \strong{14}:327-333.
}
\seealso{
\code{\link[stats]{cor.test}}; \code{clinfin::jonckheere.test}
}
