% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{justify}
\alias{justify}
\title{Justify text}
\usage{
justify(string, width = getOption("width")/2, fill = c("random", "right",
  "left"))
}
\arguments{
\item{string}{a character string}

\item{width}{desired width text in characters given as a positive integer}

\item{fill}{method of adding whitespace}
}
\description{
Add whitespace to (monospaced) text for justified or block-style spacing.
}
\examples{
x <- paste(rownames(mtcars), collapse = ' ')
cat(justify(x))

plot.new()
mtext(justify(x), line = -10, family = 'mono')
mtext(justify(x, fill = 'right'), line = -10, family = 'mono', col = 2)
mtext(justify(x, fill = 'left'), line = -10, family = 'mono', col = 3)

}
\references{
Adapted from http://stackoverflow.com/questions/34710597/justify-text-in-r
}
\seealso{
\code{\link{strwrap}}
}

