% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{kmer}
\alias{kmer}
\title{k-mers}
\usage{
kmer(x, k)
}
\arguments{
\item{x}{a vector of character string(s)}

\item{k}{length(s) of sub-sequences}
}
\description{
Generate all k-mers from a sequence of characters.
}
\examples{
x <- 'AATTGCGCGCTGT'
kmer(x, 3)
kmer(x, 3:4)

}
