% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot}
\alias{kmplot}
\title{Survival curves}
\usage{
kmplot(
  object,
  data = NULL,
  lty.surv = par("lty"),
  lwd.surv = par("lwd"),
  col.surv = seq_along(object$n),
  mark = 3L,
  lwd.mark = lwd.surv,
  lty.ci = 0,
  lwd.ci = lwd.surv,
  col.ci = col.surv,
  col.band = FALSE,
  alpha.band = 0.5,
  kmdiff = FALSE,
  atrisk.table = TRUE,
  atrisk.lab = NULL,
  atrisk.pad = 0.5,
  atrisk.type = c("atrisk", "events", "atrisk-events", "survival", "survival-ci",
    "percent", "percent-ci", "cuminc", "percent-cuminc", "cuminc-ci",
    "percent-cuminc-ci"),
  atrisk.digits = (!grepl("percent", atrisk.type)) * 2,
  atrisk.lines = TRUE,
  atrisk.lines.adj = NA,
  atrisk.col = !atrisk.lines,
  atrisk.min = NULL,
  strata.lab = NULL,
  strata.expr = NULL,
  strata.order = seq_along(object$n),
  extra.margin = 5,
  mar = NULL,
  median = FALSE,
  digits.median = 0L,
  ci.median = TRUE,
  args.median = list(),
  xaxs = "s",
  xlim = NULL,
  ylim = NULL,
  xaxis.at = pretty(xlim),
  xaxis.lab = xaxis.at,
  atrisk.at = xaxis.at,
  yaxis.at = pretty(0:1),
  yaxis.lab = yaxis.at,
  xlab = "Time",
  ylab = "Probability",
  main = NULL,
  line.xlab = NA,
  cex.axis = par("cex.axis"),
  cex.atrisk = cex.axis,
  cex.lab = par("cex.lab"),
  cex.main = par("cex.main"),
  legend = !atrisk.table && !is.null(object$strata),
  args.legend = list(),
  median.legend = FALSE,
  lr_test = TRUE,
  tt_test = FALSE,
  test_details = TRUE,
  args.test = list(),
  hr_text = FALSE,
  args.hr = list(),
  events = FALSE,
  pw_test = FALSE,
  args.pw = list(),
  format_pval = TRUE,
  stratify = NULL,
  fun = c("S", "F"),
  times = NULL,
  times.lab = NULL,
  times.type = c("percent", "atrisk", "events", "atrisk-events", "survival",
    "survival-ci", "percent-ci", "cuminc", "percent-cuminc", "cuminc-ci",
    "percent-cuminc-ci"),
  times.digits = (!grepl("percent", times.type)) * 2,
  args.times = list(),
  add = FALSE,
  panel.first = NULL,
  panel.last = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link[survival]{survfit}} or
\code{\link[survival]{survfit.coxph}}; note that for the latter some
features are not available, e.g., \code{kmdiff}, \code{lr_test},
\code{tt_test}, \code{hr_text}, \code{atrisk.type} and/or
\code{times.type \%in\% c("atrisk", "events", "atrisk-events")}, etc.}

\item{data}{(optional) the data frame used to create \code{s}, used to
obtain some data not available in \code{survfit} objects; if not given,
\code{object$call$data} will be searched for in the \code{\link{sys.frames}}}

\item{lty.surv, lwd.surv, col.surv}{line type, width, and color for survival
  curve(s); colors may be either numeric, color names as character
  string(s), or hexadecimal string(s)

  colors will be assigned to curves in order of strata or may be mapped
  using a named vector (the names must match the ones given to
  \code{strata.lab} or those that are created by
  \code{\link[survival]{strata}}); this can be useful to set colors for
  specific strata levels in multiple plots if some levels are missing; see
  examples in \code{\link{kmplot_by}}}

\item{mark, lwd.mark}{numeric plotting character (\code{\link{pch}}) or
character string, e.g., \code{''}, \code{'|'}; if \code{mark = 'bump'}, a
mark will be drawn only above the curve; \code{lwd.mark} controls the line
width when a \code{pch} or \code{"bump"} is used}

\item{lty.ci, lwd.ci, col.ci}{line type, width, and color for confidence
interval(s); not plotted (i.e., \code{= 0}) by default}

\item{col.band}{color for confidence band(s); either as numeric, color
string(s), or hexadecimal string(s); if \code{TRUE}, \code{col.surv}
values will be used; if \code{FALSE}, \code{NA}, or \code{NULL}, no
bands will be plotted; also note that this is not a true confidence band;
see details}

\item{alpha.band}{alpha transparency (in \code{[0, 1]}) of the band}

\item{kmdiff}{logical; if \code{TRUE}, a confidence band for the difference
between two curves is shown; see \code{\link{kmdiff}} (note that the color
can be passed via \code{col.ci} and the confidence interval is taken from
\code{object$conf.int})}

\item{atrisk.table}{logical; if \code{TRUE} (default), draws at-risk table}

\item{atrisk.lab}{heading for at-risk table}

\item{atrisk.pad}{extra padding between plot and at-risk table; alternatively,
a vector of padding for each line in the at-risk table, recycled as needed}

\item{atrisk.type}{a character string giving the type of "at-risk" table to
show; one of \code{"atrisk"} (number at-risk), \code{"events"} (cumulative
number of events), \code{"atrisk-events"} (both), \code{"survival"}
(survival probability), \code{"percent"} (survival percent), or
\code{"cuminc"} (cumulative incidence); \code{"survival-ci"},
\code{"percent-ci"}, \code{"percent-cuminc"}, \code{"cuminc-ci"}, and
\code{"percent-cuminc-ci"} are similar but add confidence intervals}

\item{atrisk.digits}{when survival probabilities are shown in at-risk table
(see \code{atrisk.type}), number of digits past the decimal to keep}

\item{atrisk.lines}{logical; draw lines next to strata in at-risk table}

\item{atrisk.lines.adj}{a vector of length 2 giving the left and right
adjustment for the at-risk lines in normalized units, e.g.,
\code{c(0, 1)} and \code{c(-0.5, 0.5)} doubles the length by extending
right-only and equally left and right, respectively}

\item{atrisk.col}{logical or a vector with colors for at-risk table text;
if \code{TRUE}, \code{col.surv} will be used}

\item{atrisk.min}{optional integer to replace any at-risk counts
\code{< atrisk.min} with \code{"---"} by default; note that if
\code{atrisk.min} is \emph{named}, then \code{names(atrisk.min)} will be
the replacement string}

\item{strata.lab}{labels used in legend and at-risk table for strata; if
\code{NULL} (default), labels created in \code{survfit} are used; if only
one strata is present, "All" is used by default; if \code{FALSE}, labels
are not used; if \code{TRUE}, labels will be stripped of variable names}

\item{strata.expr}{an alternative to \code{strata.lab} which allows for
\code{\link{bquote}} or \code{\link{expression}} to be passed to labels
for at-risk table; note that \code{strata.expr} trumps \code{strata.lab}}

\item{strata.order}{order of strata in legend and at-risk table}

\item{extra.margin}{increase left margin when strata labels in at-risk
table are long (note that this will be overridden by \code{mar})}

\item{mar}{margins; see \code{mar} section in \code{\link{par}}}

\item{median}{logical or numeric; if \code{TRUE}, median and confidence
interval for each curve is added to at-risk table at a calculated
position; for more control, use a specific x-coordinate}

\item{digits.median}{number of digits past the decimal point to keep for
median(s)}

\item{ci.median}{logical; if \code{TRUE}, confidence interval for medians
are shown}

\item{args.median}{an optional \emph{named} list of \code{\link{mtext}}
  arguments controlling the \code{median} text

  additionally, \code{list(x = , y = )} may be given for exact placement
  (translated to \code{at} and \code{line}, respectively)}

\item{xaxs}{style of axis; see details or \code{\link{par}}}

\item{xlim, ylim}{x- and y-axis limits}

\item{xaxis.at, yaxis.at}{positions for x- and y-axis labels and ticks}

\item{xaxis.lab, yaxis.lab}{x- and y-axis tick labels}

\item{atrisk.at}{x-coordinates to show at-risk table (default is
\code{xaxis.at})}

\item{xlab, ylab}{x- and y-axis labels}

\item{main}{title of plot}

\item{line.xlab}{line position for \code{xlab}, passed to \code{\link{title}}}

\item{cex.axis, cex.atrisk}{text size for axis tick labels and at-risk table}

\item{cex.lab, cex.main}{text size for axis and main titles}

\item{legend}{logical, a vector of x/y coordinates, or a keyword (see
\code{\link{legend}}); if \code{TRUE}, the default position is
\code{"bottomleft"}}

\item{args.legend}{an optional \emph{named} list of \code{\link{legend}}
arguments controlling the \code{legend}}

\item{median.legend}{logical; if \code{TRUE}, medians (and optionally
confidence intervals if \code{ci.median = TRUE}) are added to the legend,
useful for adding a legend and median without at-risk table}

\item{lr_test}{logical or numeric; if \code{TRUE}, a log-rank test will be
performed and the results added to the top-right corner of the plot; if
numeric, the value is passed as \code{rho} controlling the type of test
performed; see \code{\link{survdiff}}}

\item{tt_test}{logical; if \code{TRUE}, Tarone's trend test will be
performed and the resultls added to the top-right corner of the plot;
note that this will override \code{lr_test}}

\item{test_details}{logical; if \code{TRUE} (default), all test details
(test statistic, degrees of freedom, p-value) are shown; if \code{FALSE},
only the p-value is shown}

\item{args.test}{an optional \emph{named} list of \code{\link{mtext}}
arguments controlling the \code{*_test} text; additional text can be
appended with \code{list(.prefix = '')} or \code{list(.suffix = '')}}

\item{hr_text}{logical; if \code{TRUE}, a \code{\link{coxph}} model is fit,
and a summary (hazard ratios, confidence intervals, and Wald p-values)
is shown}

\item{args.hr}{an optional \emph{named} list of \code{\link{legend}}
arguments controlling the \code{hr_text} legend}

\item{events}{logical; if \code{TRUE} and \code{hr_text = TRUE}, the total
events by group is shown}

\item{pw_test}{logical; if \code{TRUE}, all pairwise tests of survival
curves are performed, and p-valuees are shown}

\item{args.pw}{an optional \emph{named} list of \code{\link{legend}}
arguments controlling the \code{pw_text} legend}

\item{format_pval}{logical; if \code{TRUE}, p-values are formatted with
\code{\link{pvalr}}; if \code{FALSE}, no formatting is performed;
alternatively, a function can be passed which should take a numeric value
and return a character string (or a value to be coerced) for printing}

\item{stratify}{(dev) \code{\link[survival]{strata}} variables}

\item{fun}{survival curve transformation, one of \code{"S"} or \code{"F"}
for the usual survival curve or empirical CDF, respectively}

\item{times}{time points to show additional summaries in a table added to
the plot (requires \href{https://github.com/raredd/plotr}{\pkg{plotr}})}

\item{times.lab}{heading for the \code{times} table}

\item{times.type}{see \code{atrisk.type}}

\item{times.digits}{when survival probabilities are shown in times table
(see \code{times.type}), number of digits past the decimal to keep}

\item{args.times}{an optional \emph{named} list of \code{plotr::tableplot}
arguments controlling the table appearance and/or location}

\item{add}{logical; if \code{TRUE}, \code{par}s are not reset; allows for
multiple panels, e.g., when using \code{par(mfrow = c(1, 2))}}

\item{panel.first}{an expression to be evaluated after the plot axes are
set up but before any plotting takes place}

\item{panel.last}{an expression to be evaluated after plotting but before
returning from the function}

\item{...}{additional parameters (\code{font}, \code{mfrow}, \code{bty},
\code{tcl}, etc.) passed to \code{par}}
}
\description{
Plot Kaplan-Meier or Cox regression models with optional at-risk table,
median survival summary, confidence bands/intervals, pairwise tests,
hazard ratios, cumulative incidences, and other features.
}
\details{
Line specifications (e.g., \code{lty.surv}, \code{lwd.surv}, etc.) will be
recycled as needed.

If \code{col.band} is not \code{NULL}, \code{NA}, or \code{FALSE}, a
confidence band is plotted; however, this is not a confidence band in the
statistical sense, i.e., a xx-percent chance of containing the entire
population of the survival curve which are wider than the point-wise
confidence limits. Rather, it refers to a band of color plotted between
the confidence limits calculated in the survfit object. That is, the xx-\%
confidence interval (plotted when \code{lty.ci != 0}) and the confidence
bands are identical--just two ways of plotting the same invervals.

\code{xaxs} is the style of the x-axis; see \code{\link{par}}. The default
for \code{kmplot} is \code{"S"} which is equivalent to \code{xaxs = "i"}
but with the maximum \code{xlim} value increased by 4\%. Other styles for
\code{xaxs} currently implemented in \code{R} are \code{"r"} (default for
plotting and the previous value for \code{kmplot}) and \code{"i"} which
will \emph{not} add padding to the ends of the axes.

When saving plots, it is highly recommended to use \code{\link{png}},
\code{\link{svg}}, \code{\link{pdf}}, etc. instead of exporting directly
from the \code{R} graphics device. Doing so may cause the at-risk table or
legend to be mis-aligned.
}
\examples{
library('survival')
km1 <- survfit(Surv(time, status) ~ sex, data = colon)
km2 <- survfit(Surv(time, status) ~ I(rx == "Obs") + adhere, data = colon)


## basic usage
kmplot(km1)
kmplot(km1, args.median = list(x = 500, y = 0.2))
kmplot(km1, atrisk.table = FALSE, median.legend = TRUE)
kmplot(km1, fun = 'F')
kmplot(km1, atrisk.col = c('grey50', 'tomato'), test_details = FALSE,
       args.test = list(col = 'red', cex = 1.5, .prefix = 'Log-rank: '))
kmplot(km1, mark = 'bump', atrisk.lines = FALSE, median = TRUE)
kmplot(km1, mark = 'bump', atrisk.lines = FALSE, median = 3500)

## atrisk spacing adjustment
kmplot(km2, atrisk.pad = 3)
kmplot(km2, atrisk.pad = 1:4 / 4)

kmplot(km2, atrisk.table = FALSE, lwd.surv = 2, lwd.mark = 0.5,
       col.surv = 1:4, col.band = c(1, 0, 0, 4))


## table of additional stats
kmplot(km1, times = 1:3 * 1000, times.type = 'percent-ci')
kmplot(km1, times = 1:3 * 1000, times.type = 'percent-ci',
       atrisk.table = FALSE, args.times = list(x = 'bottomright'))
kmplot(km1, times = 0:10 * 250, times.type = 'survival',
       times.digits = 3, times.lab = 'Prob. of survival',
       args.times = list(x = mean(par('usr')[1]), y = 0))


## use stratified models
kmplot(km1, stratify = 'differ')
kmplot(km1, stratify = c('differ', 'adhere'))


## for hazard ratios, use factors to fit the proper cox model
kmplot(survfit(Surv(time, status) ~ factor(sex), data = colon),
       hr_text = TRUE, strata.lab = TRUE)
kmplot(survfit(Surv(time, status) ~ interaction(sex, rx), data = colon),
       hr_text = TRUE, atrisk.table = FALSE, legend = FALSE)
kmplot(survfit(Surv(time, status) ~ interaction(sex, rx), data = colon),
       hr_text = TRUE, atrisk.table = FALSE, legend = FALSE,
       format_pval = function(x) format(x, digits = 2, scipen = 0))


## at-risk and p-value options
kmplot(km2, tt_test = TRUE, test_details = FALSE)
kmplot(km2, tt_test = TRUE, format_pval = format.pval)
kmplot(km1, atrisk.type = 'survival', atrisk.digits = 3L)
kmplot(km1, atrisk.type = 'atrisk-events')
kmplot(survfit(Surv(time, status) ~ rx, data = colon),
       pw_test = TRUE, args.pw = list(text.col = 1:3, x = 'bottomleft'))

## expressions in at-risk table (strata.expr takes precedence)
kmplot(km1, strata.lab = c('\u2640', '\u2642'))
kmplot(km1, strata.lab = c('\u2640', '\u2642'),
               strata.expr = expression(widetilde(ring(Female)),
                                        phantom() >= Male))

## character vectors passed to strata.expr will be parsed
kmplot(km1, strata.expr = c('Sex[Female]', 'Sex[Male]'))
## but not those passed to strata.lab
kmplot(km1, strata.lab  = c('Sex[Female]', 'Sex[Male]'))


## when using mfrow options, use add = TRUE and mar to align axes
mar <- c(9, 6, 2, 2)
op <- par(mfrow = c(1, 2))
kmplot(km1, add = TRUE, mar = mar)
kmplot(km2, add = TRUE, strata.lab = TRUE, mar = mar)
par(mfrow = op)


## survfit.coxph object
cx1 <- coxph(Surv(time, status) ~ sex, data = colon)
cx2 <- coxph(Surv(time, status) ~ sex + adhere, data = colon)
cx3 <- coxph(Surv(time, status) ~ I(rx == "Obs") + adhere + age, data = colon)

kmplot(survfit(cx1))
kmplot(survfit(cx1, data.frame(sex = 0:1)), strata.lab = c('F', 'M'))

nd <- expand.grid(adhere = 0:1, sex = 0:1)
kmplot(survfit(cx2, nd), col.surv = 1:4, strata.lab = do.call(paste, nd))
## compare
kmplot(survfit(Surv(time, status) ~ sex + adhere, data = colon),
       atrisk.type = 'survival')

nd <- data.frame(rx = 'Obs', adhere = 0:1, age = c(30, 30, 70, 70))
kmplot(survfit(cx3, nd), col.surv = 1:4, strata.lab = do.call(paste, nd))

}
\references{
Adapted from \url{http://biostat.mc.vanderbilt.edu/wiki/Main/TatsukiRcode}
}
\seealso{
\code{\link{kmplot_by}}; \code{survival:::plot.survfit};
\code{\link{kmplot_data_}}; \code{\link{atrisk_data_}};
\code{\link{lr_text}}; \code{\link{lr_pval}}; \code{\link{points.kmplot}};
\code{\link{kmdiff}}; \code{\link[plotr]{ggsurv}}
}
