% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot_by}
\alias{kmplot_by}
\title{kmplot_by}
\usage{
kmplot_by(strata, event = "pfs", dat, by, pval = TRUE, ..., cols,
  mar = c(8, 8, 3, 1), main, labs, fig)
}
\arguments{
\item{strata}{character string of the strata variable}

\item{event}{character string indicating the event (pfs, os, ttp, etc);
see details}

\item{dat}{data frame to use}

\item{by}{optional character string of stratification variable}

\item{pval}{logical; if \code{TRUE}, a log-rank test will be performed
and the results added to the top-right corner of the plot}

\item{...}{additional arguments passed to \code{\link{kmplot}} or
graphical parameters subsequently passed to \code{\link{par}}}

\item{cols}{vector of colors for curves}

\item{mar}{plotting margins}

\item{main}{title for figure}

\item{labs}{at risk table labels}

\item{fig}{figure panel label}
}
\value{
Invisibly returns a list of data frames stratified by the \code{by}
variable.
}
\description{
This function helps create stratified \code{\link{kmplot}}s quickly with
panel labels and log-rank tests for the subsets.
}
\details{
The data used should have at least three variables: \code{strata},
\code{*_time}, and \code{*_ind} where \code{*} is \code{event}. For
example, to use progression-free survival, \code{dat} should have columns
\code{"pfs_time"} and \code{"pfs_ind"} and optionally the \code{strata}
column unless \code{strata = "1"}.
}
\examples{
library('survival')
data(cancer)
colon <- within(colon[duplicated(colon$id), ], {
  pfs_time <- time
  pfs_ind <- status
  sex <- c('Female','Male')[sex + 1]
})

kmplot_by('1', dat = colon)
kmplot_by('sex', dat = colon, fig = 'B', labs = c('F','M'))
kmplot_by('rx', dat = colon, fig = '', labs = FALSE)
kmplot_by('rx + sex', dat = colon)
kmplot_by('rx', dat = colon, by = 'sex')
}
\seealso{
\code{\link{kmplot}}, \code{\link{survdiff}}
}

