% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{local_coxph_test}
\alias{local_coxph_test}
\title{Compute local p-value from coxph}
\usage{
local_coxph_test(s, pos, C = NULL, d = NULL, digits = 3)
}
\arguments{
\item{s}{survival object of class \code{\link[survival]{coxph}}}

\item{pos}{vector of positions of \code{\link{coefficients}} of interest 
from \code{summary(coxph)}; defaults to \code{1:length(coef(s))}}

\item{C, d}{\code{C}, a q-by-p matrix, and \code{d}, a q-by-1 matrix, define
the null hypothesis being checked; default is a global test on the variables
in \code{pos}, i.e., \code{C} is the identity matrix, and \code{d} is a
vector of zeros}

\item{digits}{number of significant figures in output}
}
\description{
Checks the null hypothesis: C * beta.hat = c, i.e., the local
p-value of one or more factors in a model; can also be used to test more
comlex hypotheses.
}
\examples{
library(survival)
fit <- coxph(Surv(time, status) ~ sex + ph.ecog, data = cancer)

## compare to summary(fit)
local_coxph_test(fit)
local_coxph_test(fit, 2)

}
\references{
\url{http://www.ddiez.com/teac/surv/}
}

