% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{lr_pval}
\alias{lr_pval}
\title{Extract values from \code{survdiff} call}
\usage{
lr_pval(s, details = FALSE)
}
\arguments{
\item{s}{a \code{\link{survfit}} or \code{\link{survdiff}} object}

\item{details}{logical; \code{TRUE} returns statistic, degrees of freedom,
and p-value where \code{FALSE} returns only a pvalue}
}
\description{
Extract values from \code{survdiff} call
}
\seealso{
\code{\link{lr_text}}
}

