% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{lr_text}
\alias{lr_text}
\title{Format results from \code{survdiff} for plotting}
\usage{
lr_text(formula, data, rho = 0, ...)
}
\arguments{
\item{formula, data, rho, ...}{passed to \code{\link{survdiff}}}
}
\description{
Format results from \code{survdiff} for plotting
}
\seealso{
\code{\link{lr_pval}}
}
