% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{lsf}
\alias{lsf}
\title{List package files}
\usage{
lsf(package, file = "DESCRIPTION")
}
\arguments{
\item{package}{package name, as a \code{\link{name}} or literal character
string}

\item{file}{file to return as a character string; usual options are
\code{'DESCRIPTION'}, \code{'NEWS'}, \code{'INDEX'}, \code{'NAMESPACE'};
partial matching is supported and case is ignored}
}
\value{
Returns (invisibly) the contents of \code{file} as a vector of character
strings.
}
\description{
Prints package files (\code{DESCRIPTION}, \code{NEWS}, \code{INDEX},
\code{NAMESPACE}, etc.) to console.
}
\examples{
## these are equivalent
lsf(rawr)
lsf(rawr, 'DESCRIPTION')
lsf('rawr', 'des')
}
\seealso{
\code{\link{lsp}}
}

