% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lsp}
\alias{lsp}
\title{list package}
\usage{
lsp(package, what = 'all')
}
\arguments{
\item{package}{package name as a character string}

\item{what}{what to get; \code{'all'} is default which returns all exported
and non exported functions in \code{package}; see description for more}
}
\description{
list all exported and/or non exported objects in a package.

This is a helper/wrapper function to easily list exported (\code{?'::'}) and
non exported (\code{?':::'}) functions (and other features from a package's
\code{NAMESPACE} file). Note that \code{base} and older packages do not have
a \code{'NAMESPACE'} file in which case, for \code{base} packages,
\code{lsp} returns \code{ls(.BaseNamespaceEnv, all.names = TRUE)}, and for
older packages with no \code{NAMESPACE} will throw an error.

Possible values for \code{what} are \code{'all'} (default), \code{NULL},
\code{'exports'}, \code{'imports'}, \code{'dynlibs'}, \code{'lazydata'},
\code{'path'}, \code{'S3methods'}, \code{'spec'}, and others depending on
the package.

\code{lsp(packagename, '?')} to see options for a specific package.

\code{lsp(packagename, NULL)} to return all information.
}
\examples{
lsp('rawr')

## for "what" options
lsp('rawr', '?')

## library path
lsp('rawr', 'path')

## to return everything
lsp('rawr', NULL)

## data sets
lsp('ggplot2', 'lazydata')
}

