% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{lunique}
\alias{lunique}
\title{Number of unique values}
\usage{
lunique(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical; if \code{TRUE}, \code{NA} will not be counted as a
unique level (default is to include \code{NA}s)}
}
\description{
Number of unique values
}
\examples{
x <- c(1:5, NA)
lunique(factor(x))
lunique(x, TRUE)

}
