% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{match_ctc}
\alias{match_ctc}
\title{Match CTCAE codes}
\usage{
match_ctc(..., version = 4L)
}
\arguments{
\item{...}{character string(s) of toxicity codes (usually of the form
\code{AB123} but can handle \code{AB-123} or \code{AB 123}) or keyword(s)
to be matched in the toxicity description}

\item{version}{version number; default is CTCAE v4}
}
\value{
A list containing:
\item{\code{matches}}{a data frame of matches with toxicity codes and their
respective descriptions and categories}
\item{\code{version}}{CTCAE version used}
}
\description{
Convenience function to convert CTCAE (version 3 or 4) toxicity codes or
descriptions with their appropriate matches. Especially useful in data from
paper trials where only the toxicity codes are reported excluding (the more
meaningful) descriptions.
}
\examples{
codes <- sample(rawr::ctcae_v4$tox_code, 10)
match_ctc(codes)

match_ctc('injury', version = 3L)
match_ctc('aortic', 'arterial')

}
