% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medium.R
\name{medium}
\alias{medium}
\title{Retrieve r code from Medium article or Towards Data Science article}
\usage{
medium(url, method, padding)
}
\arguments{
\item{url}{Link to a medium or towards data science article}

\item{method}{Not all websites are formatted consistently. To overcome this, try a different
method by setting the method parameter to integers 2 and greater to try
other available methods}

\item{padding}{Specify what goes between the last character of one code block and the
first character of the next code block. Default is a two new lines, which appears
visually as one new line between code blocks.}
}
\value{
A character vector of length 1 containing the R code from the target url. All code
   at the target url (including javascript, ruby, python) will be returned.
}
\description{
Retrieve r code from Medium article or Towards Data Science article
}
\examples{
library(dplyr)
url <-"https://medium.com/@data_datum/ggplot2-or-how-to-make-fancy-graphics-with-r-fabae7fc16d4"
medium(url)

}
