% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge2}
\alias{merge2}
\title{Recursively merge a list of data frames}
\usage{
merge2(l, ...)
}
\arguments{
\item{l}{list of data frames or objects to be coerced}

\item{...}{additional arguments passed to \code{merge} (eg, \code{by},
\code{all}, etc)}
}
\description{
Use \code{\link{merge}} to join \code{n} data frames
}
\examples{
a <- data.frame(id = 1:10, a = rnorm(10))
b <- data.frame(id = 4:6, b = rnorm(3))
c <- data.frame(id = 4:14, c = rpois(11, 1))
d <- matrix(c(1:5, rnorm(5)), nrow = 5, dimnames = list(NULL, c('id', 'd')))

merge2(list(a, b, c))
merge2(list(a, b, c), all = TRUE)
merge2(list(a, b, c, d), all = TRUE)
}
\seealso{
\code{\link[plyr]{join_all}}
}

