% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{mgrep}
\alias{mgrep}
\alias{mgrepl}
\alias{msub}
\alias{mgsub}
\title{Multiple pattern matching and replacement}
\usage{
mgrepl(pattern, x, ..., parallel = length(pattern) > 10000L)

mgrep(pattern, x, ..., parallel = length(pattern) > 10000L)

msub(pattern, replacement, x, ...)

mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{for substituting, a vector of length two for a single
replacement or a \emph{list} of length two vectors for multiple
replacements where each vector is \code{c(pattern,replacement)}; or for
grepping, a vector of character strings containing regular expressions
to be matched in \code{x}}

\item{x}{a character vector where matches are sought}

\item{...}{additional parameters passed onto other methods}

\item{parallel}{logical; if \code{TRUE}, grepping will be performed in
\pkg{\link{parallel}}; also, if \code{pattern} is a vector greater than
\code{1e4} elements in length, \code{parallel} defaults to \code{TRUE}}

\item{replacement}{optional; if given, both \code{pattern} and
\code{replacement} should be character vectors of equal length
(\code{replacement} will be recycled if needed)}
}
\description{
Perform multiple pattern matching and replacement.
}
\examples{
## grepping
mgrep(letters[1:5], letters[1:5])
mgrepl(letters[1:5], letters[1:5])

## subbing
s1 <- 'thiS iS SooD'

## if replacement is given, acts like gsub
mgsub(c('hi', 'oo'), c('HI', '00'), s1)
mgsub(c('\\\\bS','$','i'), '_', rep(s1, 3))

## pattern can also be a list of c(pattern, replacement)
r1 <- c('hi','HI')
r2 <- c(list(r1), list(c('oo', '00')))
r3 <- c(r2, list(c('i', '1'), c('\\\\b(\\\\w)', '\\\\U\\\\1')))

mgsub(r1, x = s1, ignore.case = TRUE)
mgsub(r2, x = s1)
mgsub(r3, x = s1, perl = TRUE)

}
\seealso{
\code{\link[base]{grep}}; \code{\link{vgrep}}
}
