% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgsub}
\alias{mgsub}
\title{Multiple gsub}
\usage{
mgsub(pattern, x, ...)
}
\arguments{
\item{pattern}{a vector of length two for a single replacement or a
\emph{list} of length two vectors for multiple replacements where each
vector is \code{c("pattern","replacement")}}

\item{x}{a character string}

\item{...}{additional parameters passed to \code{\link{gsub}}}
}
\description{
Perform multiple pattern replacements with \code{\link{gsub}}.
}
\examples{
s1 <- 'thiS iS gooD'

r1 <- c('s','5')
r2 <- c(list(r1), list(c('o', '0'), c('i', '1')))
r3 <- c(r2, list(c('\\\\b(\\\\w)', '\\\\U\\\\1')))

mgsub(r1, s1, ignore.case = TRUE)
mgsub(r2, s1)
mgsub(r3, s1, perl = TRUE)


}

