% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{mlegend}
\alias{mlegend}
\title{mlegend}
\usage{
mlegend(x, y = NULL, legend, horizontal = FALSE, ...)
}
\arguments{
\item{x, y}{the x- and y-coordinates to be used or a keyword, see
\code{\link[grDevices]{xy.coords}}; note that this is only used for
the first legend, and the coordinates for subsequent legends are
automatically calculated from the previous legend}

\item{legend}{a list or list of lists of arguments passed to
\code{\link[graphics]{legend}} with each list corresponding to a single
legend}

\item{horizontal}{logical; if \code{TRUE}, stack legends horizontally}

\item{...}{additional arguments passed to \emph{all} legends, e.g., to set
defaults; these will be over-ridden in individual legends by passing a
list of arguments to \code{legend}}
}
\description{
Stack multiple legends.
}
\examples{
op <- par(mfrow = c(1, 2))
plot(1)
l <- list(
  legend = 'first\nlegend', fill = 'red'
)
mlegend('topleft', legend = l)

l <- list(
  l,
  list(
    legend = 'second legend', fill = 'blue'
  ),
  list(
    legend = letters[1:5], pch = 16, col = 1:5, bty = 'n'
  ),
  list(
    legend = LETTERS[1:3], horiz = TRUE, lty = 1, col = 1:3
  ),
  list(
    legend = 'final legend'
  )
)

mlegend('left', legend = l, horizontal = TRUE)

## dot arguments will be applied to all legends
mlegend('topright', legend = l, bty = 'n', inset = c(-0.5, 0), xpd = NA)
par(op)

}
