% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{moods_test}
\alias{moods_test}
\title{Mood's median test}
\usage{
moods_test(X, ...)
}
\arguments{
\item{X}{a list of two or more numeric vectors}

\item{...}{additional parameters passed to \code{\link{fisher.test}}}
}
\description{
A nonparametric test used to test the equality of medians from two or more
populations.
}
\details{
Mood's median test is an exact test which dichotomizes values in each group
above or below the median of the pooled data. The test is conservative in 
small sample sizes.

Note that this test is different than \code{\link{mood.test}} in the stats
package which tests for a difference in \emph{scale parameters} rather than 
medians.
}
\examples{
set.seed(1)
X <- list(rnorm(10), rnorm(10, 1), rnorm(20, 2))
moods_test(X)

plot(density(X[[1]]), xlim = range(unlist(X)), ylim = c(0, .5))
for (x in 2:3)
  lines(density(X[[x]]), col = x)

}
\seealso{
\code{\link{mood.test}}; \code{\link{kruskal.test}};
\code{\link{wilcox.test}}
}

