% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{na.points}
\alias{na.points}
\title{na.points}
\usage{
na.points(
  x = NULL,
  y = NULL,
  xat = NULL,
  yat = NULL,
  label = "NA",
  args.beeswarm = list(),
  ...
)
}
\arguments{
\item{x, y}{vectors of data to be plotted; if only one is given, all points
will be plotted on the corresponding axis; if both are given, only \code{x}
values where \code{y} is \code{NA} are plotted and vice versa}

\item{xat, yat}{the x- and y-axis positions for points}

\item{label}{axis label for points; use \code{FALSE} or \code{''} to omit}

\item{args.beeswarm}{a \emph{named} list of additional arguments passed to
\code{\link[beeswarm]{beeswarm}} such as \code{corral} or \code{corralWidth}}

\item{...}{additional arguments passed to \code{\link{tplot}} or further
to \code{par} such as \code{col}, \code{cex}, or \code{pch}}
}
\description{
Add 1-D scatter plot to plot margins for missing values.
}
\examples{
op <- par(mar = c(5, 5, 5, 5))
plot(mpg ~ wt, mtcars)
na.points(y = mtcars$mpg, col = 'red')
na.points(x = mtcars$wt, pch = 16, cex = 1.5, col = 'blue')

plot(mpg ~ wt, mtcars, bty = 'l')
na.points(x = mtcars$wt, yat = par('usr')[4], label = 'Missing',
          args.beeswarm = list(corral = 'wrap', corralWidth = 1.5))

plot(mpg ~ wt, mtcars, col = rep(1:2, each = 10))
na.points(replace(mtcars$wt, 1:10, NA), replace(mtcars$mpg, 11:20, NA), col = 2)
par(op)

}
