% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{notin}
\alias{\%ni\%}
\alias{notin}
\title{not in}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{vector or \code{NULL}; the values to be matched}

\item{table}{vector or \code{NULL}; the values to be matched against}
}
\description{
Negation of \code{\link[base]{\%in\%}}
}
\examples{
1:5 \%ni\% 3:5
}

