% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{num2char}
\alias{num2char}
\title{Numeric to character string}
\usage{
num2char(x, informal = FALSE, cap = TRUE)
}
\arguments{
\item{x}{an integer to convert to words; can be negative or positive but
decimals will be rounded first}

\item{informal}{logical; if \code{TRUE}, adds "and" before tens or ones}

\item{cap}{logical; if \code{TRUE}, capitalizes the first word}
}
\description{
Convert a number to its word equivalent.
}
\details{
Whole numbers twenty-one through ninety-nine are hyphenated when they are
written out whether used alone or as part of a larger number; for example:
"twenty-one" or "one million twenty-one."

Whole numbers in this range are \emph{not} hyphenated for other orders of
magnitude; for example, 52,052 is written "\emph{fifty two} thousand fifty-
two" and not "\emph{fifty-two} thousand fifty-two." This rule applies only
to two-word numbers 21-99.

Informal and formal case differ only by the use of "and" to separate
1-99: "one hundred one" is the formal case, and "one hundred and one" is
the informal case.
}
\examples{
num2char(52052, TRUE)
num2char(52052, FALSE)

## vectorized
num2char(-1:2)

x <- c(-1000, 100, 52052, 3922, 3012, 201, -152, 1002, 91070432)
cbind(x, informal = num2char(x, TRUE), formal = num2char(x, FALSE))

}
\references{
\url{http://dictionary.reference.com/help/faq/language/g80.html}

\url{https://www.grammarbook.com/numbers/numbers.asp}
}
\seealso{
\code{\link{case}}; adapted from
\url{github.com/ateucher/useful_code/blob/master/R/numbers2words.r}
}
