% Generated by roxygen2 (4.0.0): do not edit by hand
\name{oror}
\alias{\%||\%}
\alias{oror}
\title{oror}
\usage{
e1 \%||\% e2
}
\arguments{
\item{e1,e2}{raw or logical or "number-like" vectors or objects}
}
\description{
\code{function_that_may_return_null()} \code{oror} default value
}
\examples{
\dontrun{
NULL || TRUE    # error
NULL \%||\% TRUE  # no error
}
}

