% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outer2}
\alias{outer2}
\title{Outer product of n-dimensional arrays}
\usage{
outer2(..., FUN)
}
\arguments{
\item{...}{arguments passed to \code{FUN} in the order given}

\item{FUN}{a function to use on the outer products}
}
\description{
The outer product of the arrays \code{X}, \code{Y}, ... with dimensions
\code{c(dim(X), dim(Y), ...)}; see \code{\link{outer}}.
}
\examples{
outer2(LETTERS[1:3], letters[1:3], LETTERS[24:26], FUN = paste0)
outer2(1:3, 1:3, 1:2, FUN = prod)

## three-way example in ?outer
x <- setNames(1:9, 1:9); y <- setNames(2:8, paste0(2:8, ':'))
x \%o\% x \%o\% y[1:3]
outer2(x, x, y[1:3], FUN = prod)

}
\seealso{
\code{\link{outer}}; \code{\link{Vectorize}}
}
