% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_to_file.R
\name{output_to_file}
\alias{output_to_file}
\title{Open contents in a new tempfile or file}
\usage{
output_to_file(content, file_path)
}
\arguments{
\item{content}{Some (text) content to go into the new file}

\item{file_path}{Specifies a file path and file name.}
}
\description{
Open contents in a new tempfile or file
}
\examples{

\dontrun{
code_sample <- "#Sample code\\nx <- 6\\n2 * 2"
output_to_file(code_sample)
}




}
