% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{path_extract}
\alias{path_extract}
\alias{fname}
\alias{file_name}
\alias{file_ext}
\alias{rm_ext}
\title{Extract parts of file path}
\usage{
path_extract(path)

fname(path)

file_name(path)

file_ext(path)

rm_ext(path)
}
\arguments{
\item{path}{file path as character string}
}
\description{
These functions will extract the directory, file name, and file extension
of some common types of files. Additionally, \code{path_extract} will
check its results by recreating \code{path} and will give warnings if
the results fail to match the input.
}
\details{
\code{fname} and \code{path_extract} do the text processing; 
\code{file_name} and \code{file_ext} are convenience functions that only
return the file name or file extension, respectively.
}
\note{
Known examples where this function fails:
\itemize{
 \item{\code{.tar.gz} }{files with compound file extensions}
}
}
\examples{
l <- list(
  '~/desktop/tmp.csv',               ## normal file with directory
  '.dotfile.txt',                    ## dotfile with extension
  '.vimrc',                          ## dotfile with no extension
  '~/file.',                         ## file name ending in .
  '~/DESCRIPTION',                   ## no extension
  '~/desktop/tmp/a.filename.tar.gz'  ## compound extension fails
)

setNames(lapply(l, path_extract), l)
setNames(lapply(l, fname), l)
setNames(lapply(l, file_name), l)
setNames(lapply(l, file_ext), l)

}
\seealso{
\code{\link[rawr]{regcaptures}}; \code{\link{basename}};
\code{\link{dirname}}
}
