% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{perm.t.test}
\alias{perm.t.test}
\alias{perm.t.test.default}
\alias{perm.t.test.formula}
\title{Permutation t-test}
\usage{
perm.t.test(x, ...)

\method{perm.t.test}{default}(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95,
  midp = TRUE,
  B = 10000L,
  ...
)

\method{perm.t.test}{formula}(formula, data, ...)
}
\arguments{
\item{x, y}{vectors of numeric data}

\item{...}{additional arguments passed to or from methods}

\item{alternative}{a character string specifying the alternative hypothesis,
one of \code{"two.sided"} (default), \code{"greater"}, or \code{"less"}}

\item{paired}{logical; if \code{TRUE}, \code{x} and \code{y} are assumed
to be paired (\code{NA}s will remove data in pairs); if \code{FALSE},
\code{x} and \code{y} are assumed to be independent samples (\code{NA}s
are removed from each sample independently)}

\item{var.equal}{logical; if \code{TRUE}, two-sample variances are treated
equally; otherwise, variances are estimated separately for both samples
(default)}

\item{conf.level}{confidence level in \code{(0, 1)} for the p-value}

\item{midp}{logical; if \code{TRUE} (default), the mid p-value is used,
i.e., half the conditional probability of the observed statistic plus the
conditional probability of more extreme values}

\item{B}{an integer specifying the number of permutations}
}
\description{
Performs paired and two-sample t-tests based on Monte Carlo permutation.
}
\examples{
set.seed(1)
x <- rnorm(10, 0, 0.5)
y <- rnorm(10, 0.5, 1)
dat <- data.frame(value = c(x, y), group = rep(1:2, each = 10))

## two-sample data
t.test(x, y)

## equivalent ways to call perm.t.test for two-sample data
perm.t.test(x, y)
perm.t.test(value ~ group, dat)

## paired data
t.test(x - y)
t.test(x, y, paired = TRUE)

## equivalent ways to call perm.t.test for paired data
perm.t.test(x, y, paired = TRUE)
perm.t.test(value ~ group, dat, paired = TRUE)
perm.t.test(x - y)

}
\seealso{
\code{\link{t.test}}

\code{perm.t.test} from the \pkg{Deducer} package

\code{independence_test} and \code{symmetry_test} from the \pkg{coin} package

\code{paired.perm.test} and \code{perm.test} from the \pkg{broman} package
}
