% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{pickcol}
\alias{pickcol}
\title{Pick elements from columns}
\usage{
pickcol(data, ind = 1L, value = FALSE)
}
\arguments{
\item{data}{a data frame or matrix}

\item{ind}{if \code{value = FALSE}, a vector (usually a single value)
which, for each row, will return the corresponding column name; if
\code{value = FALSE} (default), a vector of values to be \emph{ignored}}

\item{value}{logical; if \code{TRUE}, returns column value; otherwise,
returns column names (default)}
}
\value{
If \code{value} is \code{FALSE} (default), the column names of \code{data}
for which each row of \code{data} contained \code{ind}.

If \code{value} is \code{TRUE}, the column values of \code{data} which are
\emph{not} values of \code{ind}.
}
\description{
This function will return \code{\link{colnames}} or column values (if
\code{value = TRUE}) for "indicator-like" matrices or data frames.
}
\examples{
set.seed(1)
ss <- sample(10)
dd <- as.matrix(ftable(1:10, ss))

all(pickcol(dd) == ss)

rn <- rnorm(10)
dd[dd == 1] <- rn
all(pickcol(dd, value = TRUE, ind = 0) == rowSums(dd))


dd <- data.frame(x = c(1,0,0), y = c(0,0,1), z = c(0,1,0),
                 a = c('one','',''), b = c('','','three'), c = c('','two',''))

pickcol(dd[1:2])
pickcol(dd[1:2], 0)
pickcol(dd[1:3] + 1, ind = 2)

pickcol(dd[4:6], value = TRUE, ind = '')
pickcol(dd, value = TRUE, ind = c('', 0:1))

dd[dd == ''] <- NA
pickcol(dd[4:6], value = TRUE)

}
