% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{pickwin}
\alias{pickwin}
\title{Pick-the-winner design}
\usage{
pickwin(
  n = NULL,
  k = 2L,
  p0 = NULL,
  p1 = NULL,
  power = NULL,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{n}{sample size per arm}

\item{k}{number of arms}

\item{p0, p1}{null and alternative response rates, respectively}

\item{power}{probability of choosing the winning arm}

\item{tol}{the desired accuracy (convergence tolerance); see
\code{\link[stats]{uniroot}}}
}
\value{
Ab object of class \code{power.htest} with the following elements:

\item{n}{sample size per arm}
\item{arm}{number of arms}
\item{n.total}{total sample size for the study}
}
\description{
Calculate sample size of Simon's pick winner design or determine the
probability of alternative hypothesis to obtain target power.
}
\details{
These calculations use method proposed by R. Simon for pick-the-winner design
for randomized phase II clinical trials.

Exactly one of the parameters \code{n}, \code{power}, and \code{p1} must
be \code{NULL}.

This design is not a hypothesis test, nor is there a formal comparison
between arms. Therefore, the type I error rate does not apply.
}
\examples{
pickwin(20, 2, 0.25, 0.5)
pickwin(NULL, 2, 0.25, 0.5, 0.95)

pickwin(20, 3, 0.25, 0.5)$power
pickwin(20, 4, 0.25, 0.5)$power

}
\references{
Adapted from \code{power.ctepd::n.pick.winner}, Bingshu E. Chen

Simon, R., R.E. Wittes, S.S. Ellenberg. Randomized phase II clinical
trials. \emph{Cancer Treat Rep}. \strong{69}:1375-1381, 1985.
}
\seealso{
\code{power.ctepd::n.pick.winner}; \code{\link[clinfun]{pselect}};
\code{\link[desmon]{pickwin}};
}
