% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{plothc}
\alias{plothc}
\title{Color \code{plot.hclust}}
\usage{
plothc(hc, labels = hc$labels, col = as.factor(labels), hang = 0.1, ...)
}
\arguments{
\item{hc}{an \code{\link{hclust}} object}

\item{labels}{a character vector of labels for the leaves of the tree}

\item{col}{a vector of valid colors for \code{labels}}

\item{hang}{the fraction of the plot height by which \code{labels} should
hang below the rest of the plot; a negative value will cause \code{labels}
to hang down from 0}

\item{...}{additional arguments passed to \code{\link{plot.hclust}}}
}
\description{
Plot an \code{\link{hclust}} object with colored labels.
}
\examples{
hc <- hclust(dist(iris[, 1:4]))
plothc(hc)
plothc(hc, labels = iris$Species, col = as.numeric(iris$Species))


## add features below leaves
plothc(hc, col = iris$Species, xlab = '', sub = '')

set.seed(1)
mat <- matrix(rep(as.numeric(iris$Species), 3), 3)
mat[sample(length(mat), length(mat) / 2)] <- NA
mat <- sort_matrix(mat)[3:1, ]
shift <- c(.5, 6)
rect(col(mat) - shift[1], row(mat) - shift[2], col(mat) + shift[1],
     row(mat) - shift[2] + .5, col = mat, xpd = NA, border = 'white')

}
\author{
Eva KF Chan, \url{https://github.com/ekfchan/evachan.org-Rscripts},
modifications by Robert Redd
}

