% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pplot}
\alias{pplot}
\alias{pplot.formula}
\alias{pplot.default}
\title{pplot}
\usage{
pplot(x, ...)

\method{pplot}{formula}(formula, data, ...)

\method{pplot}{default}(
  x,
  at = seq_along(x),
  pad = 0.05,
  test = wilcox.test,
  args.test = list(),
  col = 1:2,
  pch = par("pch"),
  cex = par("cex"),
  names,
  quantiles = NULL,
  pch.quantiles = rep_len(16L, length(quantiles)),
  xlab = NULL,
  ylab = NULL,
  legend.text = NULL,
  ...
)
}
\arguments{
\item{x}{for the default method, a list of groups each having a list of
length 2 of data values}

\item{...}{for the \code{formula} method, named arguments to be passed to
the default method

for the default method, additional arguments passed to \code{\link{tplot}}
or graphical parameters passed to \code{\link{par}}}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ group1 + group2},
where y is a numeric vector of data values to be split into n groups
(\code{group1}), each group split into at most two groups (\code{group2})}

\item{data}{a data frame (or list) from which the variables in
\code{formula} should be taken}

\item{at}{the x-axis group positions}

\item{pad}{the padding between \code{at} for each group and the points}

\item{test}{logical or function; if \code{TRUE}, a \code{\link{wilcox.test}}
p-value is added for each group

alternatively, a function (or function name as a character string) can be
used, e.g., \code{test = t.test} or \code{function(x, y)
t.test(x, y)}; note that if \code{test} is a function, it must have
at least two arguments with the two groups of data; see examples}

\item{args.test}{a \emph{named} list of arguments passed to \code{test}}

\item{col, pch, cex}{color, plotting character, and character expansion
vectors, each may be length 1, 2, or total number of groups to be plotted,
recycled as needed}

\item{names}{group labels having the same length as \code{l}}

\item{quantiles}{optional probabilities passed to \code{\link{quantile}}
to show for each group}

\item{pch.quantiles}{plotting character for each quantile}

\item{xlab, ylab}{x- and y-axis labels}

\item{legend.text}{group labels of length 2}
}
\description{
A grouped paired plot for continuous data.
}
\examples{
set.seed(1)
x <- lapply(1:5, function(x) list(rnorm(20, 0, x), rnorm(20, 0, x)))

pplot(x)
pplot(x[1], ylim = c(-3, 3))
pplot(x, test = t.test, args.test = list(paired = TRUE))
pplot(x, col = rep(2:6, each = 2), pch = c(1, 16), legend.text = c('+', '-'))
pplot(x, quantiles = 0.5, legend.text = c('+', '-'))


## pplot returns the object returned by test
pv <- pplot(
  x, quantiles = c(0.25, 0.5, 0.75), pch.quantiles = c(5, 18, 5),
  legend.text = c('+', '-'), pad = 0.1, pch = 16, bty = 'l',
  names = paste('group', 1:5), las = 1
)
sapply(pv, function(x) x$p.value)


## use a custom test function
pplot(x, test = function(x, y) cor.test(x, y, method = 'spearman'))

moods <- function(x, y) rawr::moods.test(list(x, y))
sapply(x, function(y) do.call('moods', y)$p.value)
pplot(x, test = moods)


## formula method
pplot(mpg ~ gear + am, mtcars)
pplot(mpg ~ gear + vs, mtcars, xlab = 'Gears', legend.text = c('V', 'S'))

}
