% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{pr_tables}
\alias{pr_tables}
\alias{dlt_table}
\alias{pr_table}
\title{Probability tables}
\usage{
dlt_table(probs, digits = NULL)

pr_table(probs, n, crit, greater = FALSE, digits = NULL)
}
\arguments{
\item{probs}{a vector of probabilities}

\item{digits}{number of digits past the decimal point to keep; if
\code{NULL}, then no rounding is done}

\item{n, crit}{the total sample size and critical number of events; note
that this value will \emph{always} be included in the outcome regardless
of the value of \code{greater}}

\item{greater}{logical; the direction of \code{crit}: if \code{FALSE},
the probabilities are calculated for \code{crit} \emph{or fewer} events;
if \code{TRUE}, the probabilities are calculated for \code{crit} \emph{or
greater} events}
}
\description{
\code{dlt_table} creates a standard 3+3 dose-limiting toxicity table with
probabilities of dose-escalation for true but unknown probabilities of
experiencing toxicity.

\code{pr_table} is a general single-stage (as opposed to \code{dlt_table})
function for calculating the probabilities of \code{crit} or fewer (or
greater if \code{greater = TRUE}) events for true but unknown probabilities
of the event occurring.
}
\examples{
probs <- c(1, 5, 1:5 * 10) / 100

## standard 3+3 escalation design for true but unknown
## probabilities of experiencing event
dlt_table(probs)
t(dlt_table(probs, 3))


## probabilities of crit or fewer events given true but unknown probs
pr_table(probs, n = 15, crit = 3, greater = FALSE)


## probabilities of crit or more events given true but unknown probs
pr_table(probs, n = 15, crit = 3, greater = TRUE)

## compare
1 - pr_table(probs, n = 15, crit = 2, greater = FALSE)[2L, ]

}
