% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_extra.R
\name{pretty_sci}
\alias{oom}
\alias{parse_sci}
\alias{pretty_sci}
\title{Print scientific numbers}
\usage{
pretty_sci(x, digits = 0, limit = 1000)

oom(x)

parse_sci(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places to be used}

\item{limit}{a numeric value whose order of magnitude will set a limit
beyond which values of \code{x} will be displayed in scientific notation;
default is to display numbers beyond a magnitude of 3; to display
scientific notation always, use a negative value}
}
\value{
For \code{oom} a vector of order of magnitudes. For \code{parse_sci} an
expression of values in scientific notation. For \code{pretty_sci} an
expression of values in standard or scientific notation depending on the
value of \code{limit}.
}
\description{
Functions to parse numeric vectors in scientific notation and return an
\code{\link{expression}} for a pretty display.
}
\examples{
x <- 10^(1:5) / 10
oom(x)

parse_sci(x)

par(xpd = NA, mar = c(6,4,4,2) + .1)
plot(1:5, type = 'n', axes = FALSE, ann = FALSE)
axis(2, at = 1:5, labels = pretty_sci(x), las = 1, lwd.ticks = 0)

text(1:5, 0, pretty_sci(1 / x ^ 10))
text(1:5, 1, pretty_sci(1 / x, digits = 3))
text(1:5, 2, pretty_sci(1 / x, digits = 2, limit = 1e2))
text(1:5, 3, x)
text(1:5, 4, pretty_sci(x, limit = 1e2))
text(1:5, 5, pretty_sci(x, digits = 1))
text(1:5, 6, pretty_sci(x ^ 10))

text(1:5, -1, pretty_sci(1 / x, limit = -1))

}
\seealso{
\code{\link{roundr}}; \code{\link{format}}; \code{\link{sprintf}},
\code{\link{rawr_parse}}
}

