% Generated by roxygen2 (4.0.0): do not edit by hand
\name{prettybars}
\alias{prettybars}
\title{prettybars}
\usage{
prettybars(x, y = names(x), emph = NULL,

           ## aesthetics
           col.bar = grey(.9), col.emph = 'magenta1', col.y = 'black',
           col.bg = 'lightblue', cex.y = 0.5,

           ## summary line(s)
           FUN = mean, ..., fun.lab = 'overall mean', digits = 2,

           ## labels and text
           title = paste0('prettybar of ', m$x), sub = NULL, note = NULL,
           col.note = col.emph, subnote = 'source: github.com/raredd/rawr',

           ## etc
           extra.margin = 0, pars = NULL)
}
\arguments{
\item{x}{numeric vector; values of bar heights}

\item{y}{rownames of bars (or extracted from a named vector)}

\item{emph}{vector of bars to emphasize (must be in \code{y})}

\item{col.bar}{color of bars}

\item{col.emph}{color of emphasis bars}

\item{col.y}{color of y labels}

\item{col.bg}{background color bars}

\item{cex.y}{size of y labels}

\item{FUN}{summary function to apply to \code{x}}

\item{...}{additional arguments passed to \code{FUN}}

\item{fun.lab}{text label for \code{FUN}}

\item{digits}{numeric; digits after decimal after applying \code{FUN}}

\item{title}{main title}

\item{sub}{subtitle}

\item{note}{optional note (top right)}

\item{col.note}{color of note}

\item{subnote}{optional subnote (bottom right)}

\item{extra.margin}{extra left spacing for long \code{y} labels}

\item{pars}{additional list of graphical parameters passed to
\code{\link{par}}}
}
\description{
A barplot
}
\examples{
set.seed(1618)
myrandomdata <- setNames(round(runif(10), 2) * 100, LETTERS[1:10])
prettybars(myrandomdata, FUN = NULL, fun.lab = NULL, cex.y = 1.2,
           bg = 'white', emph = 'F', digits = 2)

x <- setNames(mtcars$mpg, rownames(mtcars))
prettybars(x, emph = rownames(mtcars)[mtcars$hp < 100], extra.margin = 1,
           FUN = median, fun.lab = 'overall median mpg',
           title = 'motor trend cars', sub = '   (miles per gallon)',
           note = 'vehicles with less than 100 hp in bold')

prettybars(mtcars$mpg, y = rownames(mtcars), col.bg = 'snow',
           emph = rownames(mtcars)[grepl('Merc', rownames(mtcars))],
           extra.margin = 1, col.emph = 'cyan2',
           FUN = quantile, probs = c(.25, .5, .75), na.rm = TRUE,
           fun.lab = c('lower quartile','median','upper quartile'),
           note = "if you buy a Mercedes,\\nget ready to pay for lots of gas",
           title = 'motor trend cars', sub = '   (miles per gallon)')
}
\seealso{
\code{\link{prettybars2}}
}

