% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prettybars2}
\alias{prettybars2}
\title{prettybars2}
\usage{
prettybars2(x, lab.y = colnames(x), n.y = ncol(x), lab.group = rownames(x),
            n.group = nrow(x), col.group = tcol(1:n.group, 100),
            col.line = 'skyblue3', extra.margin = 0, col.bg = 'snow',
            cex.y = 0.7, title = paste0('prettybar of ', m$x),
            sub = paste0('     n = ', sum(x)),  note = NULL,
            subnote = 'source: github.com/raredd/rawr', legend = 1,
            notext = FALSE, ...)
}
\arguments{
\item{x}{a table of data, typically from \code{\link{table}} or
\code{\link{xtabs}}}

\item{lab.y}{y-axis labels}

\item{n.y}{number of y-groups (calculated from \code{x})}

\item{lab.group}{group labels}

\item{n.group}{number of groups (calculated from \code{x})}

\item{col.group}{color for each group level}

\item{col.line}{color of origin line}

\item{extra.margin}{extra left spacing for long \code{lab.y} labels}

\item{col.bg}{background color}

\item{cex.y}{size of \code{lab.y}}

\item{title}{main title}

\item{sub}{subtitle}

\item{note}{optional note (top right)}

\item{subnote}{optional subnote (bottom right)}

\item{legend}{numeric (0, 1, 2); \code{0} for no legend; \code{1} default
legend; or \code{2} for a default \code{\link{legend}}}

\item{notext}{logical; if \code{TRUE}, suppresses all plot text, labels, and
axes for user input}

\item{...}{additional graphical parameters passed to \code{\link{par}}}
}
\description{
A barplot
}
\examples{
set.seed(1618)
f <- function(...) sample(1:5, 100, replace = TRUE, prob = c(...))
dat <- data.frame(q1 = f(.1, .2, .3, .3, .1),
                  q2 = f(.1, .4, .1, .3, .1),
                  q3 = f(.1, .2, .3, .3, .1),
                  q4 = f(.1, .1, .3, .3, .1),
                  q5 = f(.2, .1, .2, .3, .2),
                  q6 = f(.1, .3, .3, .2, .1),
                  q7 = f(.1, .4, .1, .1, .3))
dat <- stack(dat)
dat <- within(dat, {
  values <- factor(values, levels = 1:5, labels = c('NA','SA','A','D','SD'))})

mydata <- table(dat)
cols <- c(grey(.9), tcol(c('lightblue','lightblue','magenta1','magenta1'),
               c(200, 100, 100, 200)))

## compare:
barplot(mydata, horiz = TRUE, las = 1, col = cols, border = NA)
prettybars2(mydata, lab.y = paste('Question #', 1:7), extra.margin = 3,
            col.group = cols)
}
\seealso{
\code{\link{prettybars}}
}

