% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psum}
\alias{psum}
\title{Pairwise sum}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric vectors equal in length}

\item{na.rm}{logical; if \code{TRUE}, omits missing values (including
\code{\link{NaN}}) from calculations}
}
\value{
A single vector of element-wise sums.
}
\description{
Compute the pairwise sum of two or more vectors.
}
\examples{
x <- c(-1, NA, 4, 15)
y <- c(NA, NA, 6, -1)
psum(x, y)
psum(x, y, na.rm = TRUE)

}
\seealso{
Adapted from \url{https://stackoverflow.com/a/13123779/2994949};
\code{\link{pmin}}; \code{\link{pmax}}
}
