% Generated by roxygen2 (4.0.1): do not edit by hand
\name{r2sas}
\alias{r2sas}
\title{r2sas}
\usage{
r2sas(code, saspath, force = FALSE, out = getwd())
}
\arguments{
\item{code}{character string of valid \code{SAS} code}

\item{saspath}{character string of directory of \code{sas.exe}; usually for
windows, the path \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory,
and \code{sas.mget} defaults to this (the most recent version of \code{SAS})}

\item{force}{logical; by default, user must interactively allow
\code{r2sas} to continue running \code{code}; set to \code{TRUE} to ignore
this or for non-interactive \code{R}}

\item{out}{either \code{FALSE}, directory (as character string) to dump
\code{.sas} program file, \code{.log} file, and any \code{SAS} output, i.e.,
\code{.lst} file(s)}
}
\description{
Write and run \code{SAS} code in \code{R}. To run an existing \code{.sas}
file, see \code{\link{source_sas}}.
}
\examples{
\dontrun{
code <- "
* this is a sas program file : ;

options nodate nocenter nonumber ;

x 'cd ./newfolder' ;

libname lib './newfolder' ;

data data ;
  set data ;
if x = 1 then delete ;
run ;
"

r2sas(code)

## * this is a sas program file : ;
##
## options nodate nocenter nonumber ;
##
## x 'cd ./newfolder' ;
##
## libname lib './newfolder' ;
##
## data data ;
##   set data ;
## if x = 1 then delete ;
## run ;
##
##
##
## ... will be run. Continue? (y/n):
}
}
\seealso{
\code{\link{rmacro}}, \code{\link{get_margs}}, \code{\link{sas.mget}},
\code{\link{source_sas}}
}

