% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawr.R
\docType{package}
\name{rawr-package}
\alias{rawr-package}
\title{Retreive Raw R Code from Popular Tutorials and Websites}
\description{
Retreive Raw R Code from Popular Tutorials and Websites
}
\details{
It has the goal of providing a very simple and easy way of quickly retrieving the raw R code from
    popular coding websites, including github, kaggle, and data camp.
}
\author{
Steve Condylios \email{steve.condylios@gmail.com}
}
