% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawr_ls}
\alias{rawr_ls}
\alias{lss}
\alias{lsf}
\alias{lsp}
\title{List utilities}
\usage{
lss(
  pos = 1L,
  pattern,
  by = NULL,
  all.names = FALSE,
  decreasing = TRUE,
  n = Inf
)

lsf(package, file = "DESCRIPTION")

lsp(package, what, pattern)
}
\arguments{
\item{pos}{argument specifying the environment as a position in search list
or a "list-like" object such as a data frame or list of objects}

\item{pattern}{optional \code{\link{regex}}; for \code{lss} only names
matching \code{pattern} are returned; \code{\link{glob2rx}} can be used
to convert wildcard patterns to regular expressions; for \code{lsp} a text
pattern or regular expression passed to \code{\link{grep}} to filter the
results}

\item{by}{variable to order output ("type", "size" (default), "sizef",
"nrow", or "ncol")}

\item{all.names}{logical; if \code{TRUE}, all object names are returned; if
\code{FALSE}, names which begin with a \code{.} are omitted}

\item{decreasing}{logical; if \code{TRUE}, displays output in decreasing
order}

\item{n}{number of objects to displace if \code{head} is \code{TRUE}}

\item{package}{package name, as a \code{\link{name}} or literal character
string}

\item{file}{file to return as a character string; usual options are
\code{"DESCRIPTION"}, \code{"NEWS"}, \code{"INDEX"}, \code{"NAMESPACE"};
partial matching is supported and case is ignored}

\item{what}{what to get; \code{"all"} is default which returns all exported
and non exported functions in \code{package}; see details for more}
}
\value{
\code{lss} returns a data frame with each object from the workspace with
type, object size, and dimensions for each. \code{lsf} (invisibly) returns
the contents of \code{file} after being printed to the console. \code{lsp}
returns a vector of character strings.
}
\description{
These functions begin with \code{ls} and list different things. \code{lss}
is an "improved" \code{\link{ls}} which gives more details of objects in
the workspace such as type, size, and dimensions; \code{lsp} lists
\code{p}ackage contents, i.e., exported and/or non exported objects,
methods, lazy data, etc.; and \code{lsf} lists package \code{f}iles. See
details and examples.
}
\details{
\code{lsf} prints package files (e.g., \code{DESCRIPTION}, \code{NEWS},
\code{INDEX}, \code{NAMESPACE}, etc.) to console and returns (invisibly)
the files parsed for easier use.

\code{lsp} is a helper function to list exported (\code{?"::"}) and non
exported (\code{?":::"}) functions (and other features from a package's
\code{NAMESPACE} file).

Note that \code{base} and older packages do not have a \code{NAMESPACE}
file in which case, for \code{base} packages, \code{lsp} returns
\code{ls(.BaseNamespaceEnv, all.names = TRUE)}, and throws an error
otherwise.

Possible values for \code{what} are "\code{all}" (default), \code{NULL},
"\code{exports}", "\code{imports}", "\code{dynlibs}", "\code{lazydata}",
"\code{path}", "\code{S3methods}", "\code{spec}", and others depending on
the package.

\code{lsp(packagename, "?")} to see options for a specific package.

\code{lsp(packagename, NULL)} to return all information in a list.
}
\examples{
## lss(): use like ls()
lss()
a <- rnorm(100000)
b <- matrix(1, 1000, 100)
lss()

## lsf(): these are equivalent ways to use
lsf(rawr)
lsf(rawr, 'DESCRIPTION')
lsf('rawr', 'des')

## lsp: see the contents of a package
lsp(rawr)

## return all one- or two-character functions from base package
lsp(base, pat = '^.{1,2}$')

## for "what" options
lsp('rawr', '?')

## library path
lsp('rawr', 'path')
system.file(package = 'rawr')

## to return everything
lsp('rawr')

## data sets
lsp('rawr', 'lazydata')

}
\seealso{
\code{\link{ls}}; \code{\link{search}}; \code{\link[rawr2]{rawr_parse}}
}
