% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawr_ops}
\alias{rawr_ops}
\alias{oror}
\alias{notin}
\alias{inside}
\alias{\%ni\%}
\alias{\%inside\%}
\alias{\%==\%}
\alias{\%||\%}
\alias{\%:\%}
\title{rawr operators}
\usage{
x \%ni\% table

x \%inside\% interval

a \%==\% b

a \%||\% b

object \%:\% range
}
\arguments{
\item{x}{vector or \code{NULL}; the values to be matched}

\item{table}{vector or \code{NULL}; the values to be matched against}

\item{interval}{numeric vector of length two representing the interval}

\item{a, b}{raw, logical, "number-like" vectors or objects}

\item{object}{a \emph{named} vector or list, a matrix or data frame}

\item{range}{a numeric or character vector of length two with the indices
or names from \code{object}, generally of the structure \code{c(from, to)}}
}
\description{
Some useful binary operators.
}
\details{
\code{\%ni\%} is the negation of \code{\link[base]{\%in\%}}.

\code{\%inside\%} returns a logical vector indicating if \code{x} is inside
the \code{interval} (inclusive).

\code{\%=\%} is an operator combining the qualities of \code{\link{==}} and
\code{\link{\%in\%}} to compare vectors in a pairwise manner which may
include \code{\link{NA}}s.

\code{\%||\%} is useful for a function, \code{f}, that may return a value
or \code{NULL}; however if \code{NULL} is the return value of \code{f}, it
is desirable to return some other default value.

\code{\%:\%} is useful for obtaining a range of a vector (usually
\code{colnames} or \code{names} of a matrix or data frame) by literal
character strings rather than by index.
}
\examples{
1:5 \%ni\% 3:5

c(0,4) \%inside\% c(0, 4)
-5:5 \%inside\% c(0,5)
-5:5 \%inside\% c(5,0)

a <- c(1, NA, 2)
b <- c(2, NA, 1)
## not desired
a == b     # FALSE NA   FALSE
a \%in\% b   # TRUE  TRUE TRUE
## desired results
a \%==\% b   # FALSE TRUE FALSE

# NULL || TRUE   # error
NULL \%||\% TRUE   # TRUE

1:5 \%:\% c(3,5)
letters \%:\% c('e', 'n')

## these are equivalent
mtcars \%:\% c('hp','vs')
mtcars \%:\% c(4, 8)
names(mtcars[, 4:8])

}
\seealso{
\code{\link{==}}, \code{\link{\%in\%}}, \code{\link{||}}
}
