% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{rawr_palettes}
\alias{rawr_palettes}
\alias{rawr_pal}
\alias{show_pal}
\title{rawr palettes}
\usage{
rawr_palettes

rawr_pal(
  name,
  n = NULL,
  z = n,
  type = c("discrete", "continuous"),
  rev = FALSE
)

show_pal(x, n = Inf, fullrange = FALSE, counts = inherits(x, "imgpal"))
}
\arguments{
\item{name}{the palette name, one of \code{names(rawr_palettes)} or an
unambiguous abbreviation}

\item{n}{the first \code{n} colors from the palette to use}

\item{z}{for \code{type = 'continuous'}, the number of colors to
interpolate from the sequence of \code{n} palette colors}

\item{type}{return a discrete or continuous (gradient) of colors}

\item{rev}{logical; if \code{TRUE}, the palette is reversed}

\item{x}{one of 1) a \code{rawr_palette} name; 2) a vector of two or more
colors; 3) an \code{\link{imgpal}} object}

\item{fullrange}{logical; for \code{\link{imgpal}} objects, if \code{TRUE},
the entire palette is shown; otherwise, only the unique colors (estimated
from ImageMagick) are shown}

\item{counts}{logical; for \code{\link{imgpal}} objects, if \code{TRUE},
the frequencies are shown for each color}
}
\description{
Functions for generating and viewing color palettes.
}
\examples{
## some built-in palettes
rawr_palettes

## use or generate new palettes from existing
show_pal(rawr_pal('dfci'))
show_pal(rawr_pal('dfci', 4))
show_pal(rawr_pal('dfci', 4, 100, type = 'continuous'))

## view palettes from other sources
# show_pal(nord::nord_palettes$afternoon_prarie)
show_pal(rainbow(8))

\dontrun{
filled.contour(volcano, col = rawr_pal('dfci', 4, 21, type = 'c'))
filled.contour(volcano, col = rawr_pal('dfci', z = 21, type = 'c'))
filled.contour(volcano, col = rawr_pal('pokrie', 4, 21, type = 'c'))
}

}
\seealso{
\code{\link{imgpal}}; \code{\link{palette}}; \code{\link{colorRampPalette}};
\code{wesanderson::wes_palettes}; \code{nord::nord_palettes};
\code{faulkner::faulkners}
}
