% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawr_parse}
\alias{rawr_parse}
\alias{parse_yaml}
\alias{parse_index}
\alias{parse_news}
\alias{parse_namespace}
\title{Parsers}
\usage{
parse_yaml(x)

parse_index(x)

parse_news(x)

parse_namespace(x, wh = c("import", "export", "exportPattern",
  "importClass", "importMethod", "exportClass", "exportMethod",
  "exportClassPattern", "useDynLib", "nativeRoutine", "S3method"))
}
\arguments{
\item{x}{a vector of character strings}

\item{wh}{for \code{parse_namespace}, which types to parse and return}
}
\value{
A named list for each section type.
}
\description{
Some (mostly internal) simple parsers. \code{parse_yaml} can process
simple, single-level yaml-like files such as the \code{DESCRIPTION} files
of \code{R} packages; \code{parse_index}, \code{parse_news}, and
\code{parse_namespace} process their respective files. These have not
been tested for every case, particularly \code{parse_news} since there is
not a requirememnt on the structure of \code{NEWS} files.
}
\examples{
parse_yaml(lsf(rawr, 'desc'))
parse_index(lsf(rawr, 'index'))
parse_news(lsf(rawr, 'news'))
parse_namespace(lsf(rawr, 'name'))

}
\seealso{
\code{\link{parseNamespaceFile}}, \code{\link{lss}}, \code{\link{lsf}},
\code{\link{lsp}}
}
