% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawr_view}
\alias{rawr_view}
\alias{View2}
\alias{view}
\title{View data}
\usage{
View2(x, title, ...)

view(x, use_viewer = FALSE, ...)
}
\arguments{
\item{x}{an \code{R} object which can be coerced to a data frame with
non-zero numbers of rows and columns or an \code{htmlwidget} object}

\item{title}{title for viewer window, defaults to name of \code{x}
prefixed by \code{Data:}}

\item{...}{additional arguments passed to \code{\link[DT]{datatable}}}

\item{use_viewer}{logical; if \code{TRUE}, opens in the
\code{\link[rstudioapi]{viewer}} if available; otherwise, opens in the
default browser}
}
\description{
Convenience functons to use the base \code{R} data viewer (\code{View2}
always invokes \code{\link[utils]{View}} instead of the rstudio viewer)
or the default browser (\code{view} which can open html or widgets in
the browser or to view data frame- or matrix-like objects using
\code{\link[DT]{datatable}}.
}
\examples{
View2(mtcars)
view(mtcars)
view(htmlTable::htmlTable(mtcars))

\dontrun{
view(qtlcharts::iplot(1:5, 1:5))
}

}
