% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{rawrops}
\alias{\%=\%}
\alias{\%inside\%}
\alias{\%ni\%}
\alias{\%||\%}
\alias{inside}
\alias{notin}
\alias{oror}
\alias{rawrops}
\title{rawr operators}
\usage{
x \%ni\% table

x \%inside\% interval

a \%=\% b

a \%||\% b
}
\arguments{
\item{x}{vector or \code{NULL}; the values to be matched}

\item{table}{vector or \code{NULL}; the values to be matched against}

\item{interval}{numeric vector of length two representing the interval}

\item{a,b}{raw, logical, "number-like" vectors or objects}
}
\description{
Some useful binary operators.
}
\details{
\code{\%ni\%} is the negation of \code{\link[base]{\%in\%}}.

\code{\%inside\%} returns a logical vector indicating if \code{x} is inside
the \code{interval}.

\code{\%=\%} is an operator combining the qualities of \code{\link{==}} and
\code{\link{\%in\%}} to compare vectors in a pairwise manner which may
include \code{\link{NA}}s.

\code{\%||\%} is useful for a function, \code{f}, that may return a value
or \code{NULL}, but if \code{NULL} is the result of \code{f}, it is
desirable to return some other default value without errors.
}
\examples{
\dontrun{
1:5 \%ni\% 3:5

-5:5 \%inside\% c(0,5)

a <- c(1, NA, 2)
b <- c(2, NA, 1)
## not desired
a == b     # FALSE NA   FALSE
a \%in\% b   # TRUE  TRUE TRUE
## desired results
a \%=\% b    # FALSE TRUE FALSE

f <- function(x0 = TRUE) NULL || x0
f() # error
f <- function(x0 = TRUE) NULL \%||\% x0
f() # TRUE
}
}
\seealso{
\code{\link{==}}, \code{\link{\%in\%}}, \code{\link{||}}
}

