% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rcor}
\alias{rcor}
\title{Correlate variables}
\usage{
rcor(y, x = NULL, rho = 0.5)
}
\arguments{
\item{y}{a numeric vector or matrix used to set correlation(s)}

\item{x}{a vector to be correlated with \code{y}; if not given, a random,
normally distributed vector is used}

\item{rho}{the desired correlation(s); should be equal in length to the
number of columns of \code{y}, recycled as needed (or length 1 if \code{y}
is a vector)}
}
\description{
Generate a (random) vector with a desired correlation to one or more other
variables.
}
\examples{
set.seed(1)
y <- matrix(rnorm(150), 50)
x <- rnorm(50)

cor(y[, 1], rcor(y[, 1]))
cor(y[, 1], rcor(y[, 1], x, rho = 0.8))

cor(y, rcor(y))
cor(y, rcor(y, rho = c(.75, -.5, .25)))

pairs(
  cbind(rcor(y, x, rho = c(.75, -.5, .25)), y),
  upper.panel = function(x, y, ...) {
    points(x, y)
    abline(lm(y ~ x))
    mtext(parse(text = sprintf('rho==\%.2f', cor(x, y))), cex = 2, col = 2)
  }
)

}
\seealso{
\url{https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables}
}
