% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{read.clip}
\alias{read.clip}
\alias{read.clip.csv}
\alias{read.clip.fwf}
\alias{read.clip.tab}
\title{Read data from clipboard}
\usage{
read.clip(header = TRUE, ...)

read.clip.csv(header = TRUE, sep = ",", ...)

read.clip.tab(header = TRUE, sep = "\\t", ...)

read.clip.fwf(header = TRUE, widths, ...)
}
\arguments{
\item{header}{logical; indicates if variable names are in first line}

\item{...}{additional arguments passed to \code{\link{read.table}}}

\item{sep}{separator as a character string}

\item{widths}{a vector of widths of the fixed-width fields or a list of
vectors giving the widths for multiple lines}
}
\description{
Reads data (comma-, tab-, or fixed-width separated) data from clipboard and
returns as a data frame.
}
\seealso{
\code{\link[psych]{read.clipboard}}, \code{\link{read.table}},
\code{\link{read.fwf}}
}

