% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{read_clip}
\alias{read_clip}
\alias{read_clip.csv}
\alias{read_clip.tab}
\alias{read_clip.fwf}
\title{Read data from clipboard}
\usage{
read_clip(header = TRUE, ...)

read_clip.csv(header = TRUE, sep = ",", ...)

read_clip.tab(header = TRUE, sep = "\\t", ...)

read_clip.fwf(header = TRUE, widths, ...)
}
\arguments{
\item{header}{logical; indicates if variable names are in first line}

\item{...}{additional arguments passed to \code{\link{read.table}}}

\item{sep}{separator as a character string}

\item{widths}{a vector of widths of the fixed-width fields or a list of
vectors giving the widths for multiple lines}
}
\description{
Reads data (comma-, tab-, or fixed-width separated) data from clipboard and
returns as a data frame.
}
\seealso{
\code{\link{read.table}}; \code{\link{read.fwf}}
}
