% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{reload}
\alias{reload}
\title{Re-load a package}
\usage{
reload(package)
}
\arguments{
\item{package}{package name, as a \code{\link{name}} or literal character
string}
}
\description{
Detach and re-load a package.
}
