% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Reshape}
\alias{Reshape}
\alias{cast}
\alias{melt}
\title{Reshape data}
\usage{
cast(data, idvar = list(1), timevar = list(2),
  v.names = list(3:ncol(data)), ...)

melt(data, varying = list(1:ncol(data)), ...)
}
\arguments{
\item{data}{a data frame}

\item{idvar}{id variable(s)}

\item{timevar}{time point variable(s)}

\item{v.names}{long variable(s) to be flattened}

\item{...}{additional arguments passed to \code{reshape}}

\item{varying}{variable(s) to melt into column(s)}
}
\value{
The reshaped data frame with added attributes to simplify reshaping back
to the original form.
}
\description{
Convenience wrappers for \code{\link{reshape}} where \code{direction}
is \code{"long"} (\code{melt}) or \code{"wide"} (\code{cast}).
}
\details{
These functions set defaults for reshaping data; any defaults may be
overwritten by simply passing arguments to \dots (names must match exactly
and no partial matching is allowed or they will be ignored).

By default, \code{cast} assumes \code{data} is a data frame with at least
three columns representing id, time point, and value variables (any
additional columns will be considered value variables as well). \code{melt}
by default assumes no id variables and will melt all columns.

\code{idvar}, \code{timevar}, \code{v.names}, and \code{varying} can be
passed as a \emph{vector} of column indices or character strings of
variable names or as a \emph{list} of indices or character strings names
for convenience.
}
\examples{
## default of melt is no id variable so all are melted
melt(mtcars)

## compare to
reshape(mtcars, dir = 'long', varying = list(names(mtcars)))


dat <- data.frame(id = rep(1:4, rep(2, 4)),
                  visit = I(rep(c('Before', 'After'), 4)),
                  x = abs(rpois(4, 10)), y = -runif(4))

## equivalent ways to give `varying`
melt(dat, list(3:4))
melt(dat, 3:4)
(l <- melt(dat, c('x','y')))

## equivalent ways to give `idvar`, `timevar`, `v.names`
cast(l, c('id','visit'), 'variable', 'value')
cast(l, list('id','visit'), 3, 4)
cast(l, 1:2, 3, list('value'))

## cast assumes columns are id, time, followed by all varying:
cast(dat)


dat <- within(dat, {
  xx <- abs(rpois(4, 10))
  yy <- -runif(4, 1, 2)
})

## melting multiple variables simultaneously
melt(dat, list(c(3,6), c(4:5)))

## note that a vector cannot be used here for the same results
melt(dat, c('x','xx','y','yy'))

## use a list instead
melt(dat, list(c('x','xx'), c('y','yy')))


## melt or cast can also be undone as other reshape objects
w1 <- reshape(Indometh, v.names = "conc", idvar = "Subject",
              timevar = "time", direction = "wide", new.row.names = 1:6)
w2 <- cast(Indometh, v.names = "conc", idvar = "Subject",
           timevar = "time", direction = "wide")

identical(w1, w2) ## TRUE
identical(reshape(w1), reshape(w2)) ## TRUE

}
\seealso{
\code{\link{reshape}}
}

