% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{Restart}
\alias{Restart}
\title{Restart}
\usage{
Restart(afterRestartCommand = "")
}
\arguments{
\item{afterRestartCommand}{character string of command(s) to be
executed after restarting}
}
\description{
Ends current and restarts a clean \code{R} session.
}
\examples{
\donttest{
Restart("clear(); cat('Here is a clean session just for you')")
}
}

