% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{rgene}
\alias{rgene}
\title{Generate random gene names}
\usage{
rgene(
  n = 1L,
  alpha = LETTERS[1:5],
  nalpha = 2:5,
  num = 0:9,
  nnum = 1:5,
  sep = "-"
)
}
\arguments{
\item{n}{number of gene names to return}

\item{alpha}{vector of letters to select from}

\item{nalpha}{range of possible number of \code{alpha} to select}

\item{num}{numerics to select from}

\item{nnum}{range of possible number of \code{num} to select}

\item{sep}{character to separate \code{alpha} and \code{num}}
}
\description{
Generate random character strings from pools of letters and digits.
}
\examples{
rgene()
rgene(5, alpha = 'ABCD', nalpha = 1, nnum = 5:6)
rgene(5, alpha = c('A','T','C','G'), num = '', sep = '')

}
