% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmacro}
\alias{rmacro}
\title{Call SAS macros}
\usage{
rmacro(mpath, mname, args, saspath, show.args, log.file, moreArgs)
}
\arguments{
\item{mpath}{character string of path to \code{macro.sas} file}

\item{mname}{macro name; if missing, \code{\link{get_margs}} will search
\code{mpath} for macro names; if missing and \code{get_margs} finds more
than one macro in \code{mpath}, will throw an error}

\item{args}{arguments passed to the macro; if unsure of arguments, simply
run \code{rmacro} with \code{show.args = TRUE}}

\item{saspath}{character string of directory of \code{sas.exe}; usually for
windows, the path \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory,
and \code{mpath} defaults to this (the most recent version of \code{SAS})}

\item{show.args}{logical; return macro names and arguments in \code{mpath}}

\item{log.file}{character string of directory to create \code{_temp_.log},
the log file for the \code{SAS} call; defaults to the \code{mpath} directory}

\item{moreArgs}{character string of additional \code{SAS} (legal) commands
to use separated by semicolons; for example, \code{'options <OPTIONS> ;'} or
\code{"x 'cd c:/path/to/new/directory ;'"}}
}
\description{
\code{rmacro} runs \code{SAS} macros in \code{*.sas} files from \code{R};
currently only for windows platforms.
}

