% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Round}
\alias{Round}
\title{Round to specified target}
\usage{
Round(x, target)
}
\arguments{
\item{x}{numeric values}

\item{target}{desired sum of code{x} after rounding}
}
\description{
Rounds a numeric vector constrained to sum to a \code{target} value.
}
\examples{
pcts <- data.frame(pct1 = c(33.3, 21.5, 45.51),
                   pct2 = c(33.3,33.3,33.3))

Map(round, pcts); sapply(.Last.value, sum)
Map(Round, pcts, 100); sapply(.Last.value, sum)
}

