% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{round2}
\alias{round2}
\title{Round to nearest}
\usage{
round2(x, to = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{to}{nearest fraction or integer}
}
\description{
Round numerics to nearest multiple of \code{to}.
}
\examples{
round2(mtcars$mpg, 5)
round2(mtcars$mpg, .5)
}

