% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rpart_utils}
\alias{rpart_nodes}
\alias{rpart_parent}
\alias{rpart_subset}
\alias{rpart_utils}
\title{\code{rpart} utilities}
\usage{
rpart_parent(node = 1L)

rpart_subset(tree, node = 1L)

rpart_nodes(tree, node_labels = FALSE, droplevels = TRUE)
}
\arguments{
\item{node}{an integer representing the node number}

\item{tree}{an object returned from \code{rpart}}

\item{node_labels}{a vector of labels having the same length as the number
of terminal nodes or total nodes}

\item{droplevels}{logical; if \code{TRUE}, only node labels with at least
one observation are used (i.e., only terminal node labels are used)}
}
\value{
\code{rpart_parent} returns a vector representing the path from the root to
\code{node}.

\code{rpart_subset} returns the data frame of observations in \code{node}.
For any \code{tree}, the possibilities 

\code{rpart_nodes} returns a factor variable
}
\description{
Utilities for the \pkg{\link{rpart}} package:

\code{rpart_parent} returns all parent nodes of \code{node}.

\code{rpart_subset} and \code{rpart_subset2} (in examples) return a
subset of the data used in \code{rpart} for any intermediate or terminal
\code{node}.

\code{rpart_nodes} returns the terminal node label for each observation
in the original data frame used for \code{tree}.
}
\examples{
rpart_parent(116)
rpart_parent(29)

\dontrun{
library('rpart')
(fit <- rpart(Kyphosis ~ Age + Number + Start, kyphosis, minsplit = 5))

## children nodes should have identical paths
identical(head(rpart_parent(28), -1L), head(rpart_parent(29), -1L))

## terminal nodes should combine to original data
nodes <- as.integer(rownames(fit$frame[fit$frame$var \%in\% '<leaf>', ]))
sum(sapply(nodes, function(x) nrow(rpart_subset(fit, x))))
nrow(kyphosis)

## all nodes
nodes <- as.integer(rownames(fit$frame))
sapply(nodes, function(x) nrow(rpart_subset(fit, x)))


rpart_subset2 <- function(tree, node = 1L) {
  require('partykit')
  ptree <- as.party(tree)
  ptree$data <- model.frame(eval(tree$call$data, parent.frame(1L)))
  ## retain transformed variables but drop those not in formula
  ## http://stackoverflow.com/a/36816883/2994949
  # ptree$data <- model.frame(tree)
  data_party(ptree, node)[, seq_along(ptree$data)]
}

## note differences in nodes labels in party vs rpart
dim(rpart_subset(fit, 4))
dim(rpart_subset2(fit, 3))


rpart_nodes(fit)
rpart_nodes(fit, TRUE)

table(rpart_nodes(fit, letters[1:10]),
      rpart_nodes(fit, letters[1:19]))

## subset an rpart object by node id which should only include
## observations found in children of the node id(s) selected
identical(kyphosis, rpart_subset(fit, unique(rpart_nodes(fit))))

kyphosis$node <- rpart_nodes(fit)
rpart_subset(fit, 14:15)
}

}
\seealso{
\url{http://stackoverflow.com/questions/36086990/how-to-climb-the-tree-structure-of-rpart-object-using-path-in-order-to-purge-man}

\url{http://stackoverflow.com/questions/36748531/getting-the-observations-in-a-rparts-node-i-e-cart}
}

