% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rsum}
\alias{rsum}
\title{Random sum}
\usage{
rsum(a, b, n, k, unique = FALSE, iterations = 100L)
}
\arguments{
\item{a, b, n, k}{\code{n} integers in the range \code{[a,b]} which sum to
\code{k}}

\item{unique}{logical; if \code{TRUE}, the returned vector of integers
will be unique if possible}

\item{iterations}{maximum number of iterations to repeat the algorithm
until a solution is found}
}
\description{
Returns a vector of length \code{n} from \code{a} to \code{b} which sum to
\code{k}.
}
\details{
To find the \code{n} integers in \code{[a, b]} which sum to \code{k}, the
range \code{1:k} is broken into \code{n} pieces requiring \code{n - 1}
cutpoints, and the distance between cutpoints is summed and shifted by
\code{a}.

Note that the solution may not contain a unique set of integers unless
\code{unique = TRUE}. Additionally, this algorithm is not guaranteed to
find a solution even if one exists regardless of many \code{iterations}.
Finally, depending on the input parameters, the algorithm may complete
after one or require thousands of iterations.
}
\examples{
set.seed(1)
rsum(1, 100, 5, 98)
rsum(1, 100, 5, 120, TRUE)
sum(rsum(1, 100, 10, 98))

}
\seealso{
Adapted from \url{https://stackoverflow.com/a/49016614/2994949};
\code{Surrogate::RandVec}
}
