% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{sample_each}
\alias{sample_each}
\title{Sample each}
\usage{
sample_each(x, n = 1L)
}
\arguments{
\item{x}{a vector}

\item{n}{number to sample from each unique group in order; if \code{x} is
a factor, \code{n} should correspond to the order of \code{levels(x)};
otherwise, \code{n} will be matched with the sorted unique groups}
}
\value{
A logical vector the same length as \code{x} identifying selected indices.
}
\description{
Sample each unique value of a vector.
}
\examples{
x <- mtcars$gear
X <- factor(x, 5:3)

sample_each(x)

mtcars[sample_each(x), ]

## compare numeric vs factor vectors
mtcars[sample_each(x, c(3,4,5)), ]
mtcars[sample_each(X, c(3,4,5)), ]

}
\seealso{
\code{\link{sample}}; \code{\link{kinda_sort}}
}

