% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sas.mget}
\alias{sas.mget}
\title{Convert multiple SAS data sets to R data frame}
\usage{
sas.mget(libpath, dsn, saspath, fmtpath, log.file, ..., force = FALSE)
}
\arguments{
\item{libpath}{character string of directory of data set(s)}

\item{dsn}{data set name(s); either \code{data1} or \code{data1.sas7bdat}
will work; if missing or \code{NULL}, will get all \code{.sas7bdat} files}

\item{saspath}{character string of directory of \code{sas.exe}; usually for
windows, the path \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory,
and \code{sas.mget} defaults to this (the most recent version of \code{SAS})}

\item{fmtpath}{character string of path to a format \code{.sas} file;
\code{sas} throws an error if the host used to make the format catalog does
not use the same platform; this creates a new format catalog either by
running a \code{proc format} macro or by making a copy of the non native
\code{.sas7bcat} file}

\item{log.file}{character string of directory to create \code{_temp_.log},
the log file for the \code{SAS} call; defaults to \code{libpath} directory}

\item{...}{additional arguments passed to \code{\link[Hmisc]{sas.get}}}

\item{force}{logical; by default, user must interactively allow
\code{sas.mget} to continue reading all data sets; set to \code{TRUE} to
ignore this or for non-interactive use}
}
\value{
A list of data frames resembling the \code{SAS} data sets.
}
\description{
User-friendly wrapper of \code{\link[Hmisc]{sas.get}} to convert one or more
\code{SAS} data sets (\code{.sas7bdat} files) into a list of \code{R} data
frames.
}

