% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_catalog}
\alias{sas_catalog}
\title{\code{SAS} catalog}
\usage{
sas_catalog(path, libpath = dirname(path), saspath = sas_path(),
  log = "_temp_formats_.log")
}
\arguments{
\item{path}{file path to a \code{.sas} file}

\item{libpath}{directory to data set(s) as character string; if missing,
searches the current working directory}

\item{saspath}{file path to \code{sas.exe} as character string}

\item{log}{name of log file as character string}
}
\description{
Create a \code{SAS} format catalog from a \code{.sas} file or macro.
}
\examples{
\dontrun{
p <- system.file('testfiles', 'formats.sas', package = 'rawr')
cat(readLines(p), sep = '\\n')
sas_catalog(p, getwd())
}

}
\seealso{
\code{\link{sas_path}}, \code{\link{r2sas}}, \code{\link{rmacro}},
\code{\link{get_margs}}, \code{\link{sas_mget}}, \code{\link{source_sas}},
\code{\link{parse_formats}}
}

