% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_mget}
\alias{sas_mget}
\title{Convert multiple SAS data sets to R data frame}
\usage{
sas_mget(libpath, dsn, saspath, fmtpath, catalog = FALSE, log.file, ...,
  force = FALSE)
}
\arguments{
\item{libpath}{directory to data set(s) as character string; if missing,
searches the current working directory}

\item{dsn}{data set name(s); either \code{data1} or \code{data1.sas7bdat}
will work; if missing or \code{NULL}, all \code{.sas7bdat} files found in
\code{libpath} will be read}

\item{saspath}{file path to \code{sas.exe} as character string

Usually \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory

Note that specifying \code{saspath} is not required if v9.2 is installed in
the default directory (i.e.,
\code{c:/program files/sasfoundation/9.2/sas.exe});

For previous versions or if\code{SAS} is not installed in the default
directory, it will be necessary to supply \code{saspath} to \code{sas.exe}}

\item{fmtpath}{path to a format \code{.sas} file as character string;
\code{SAS} throws an error if the host used to make the format catalog was
not on the same platform, so we create a new format catalog either by
running a \code{proc format} macro or by making a copy of the non-native
\code{.sas7bcat} file}

\item{catalog}{logical; use the format catalog in \code{libpath}; by default
this is \code{FALSE} since using cross-platform catalogs throws errors, so
\code{sas_mget} will either ignore the catalog or create a new one if
\code{fmtpath} is given; set \code{catalog = TRUE} if you are sure that
the format catalog is native to windows}

\item{log.file}{name of \code{SAS} log file; default value will create
\code{_temp_.log} in the \code{libpath} directory}

\item{...}{additional arguments passed to \code{\link[Hmisc]{sas.get}}}

\item{force}{logical; by default, user must interactively allow
\code{sas_mget} to continue reading all data sets; set to \code{TRUE} to
ignore this or for non-interactive \code{R}}
}
\value{
A list of data frames resembling the \code{SAS} data sets.
}
\description{
User-friendly wrapper of \code{\link[Hmisc]{sas.get}} to convert one or
more \code{SAS} data sets (\code{.sas7bdat} files) into a list of \code{R}
data frames.
}
\examples{
\dontrun{
data.list <- sas_mget('./tests/testfiles/')

## !!! Two data set(s) will be read !!!
##
##  Size in Kb:
##
##       size (Kb)
## data1         0
## data2         0
##
##
##
##
##
## do you want to continue (y/n)?
}
}
\seealso{
\code{\link{rmacro}}, \code{\link{get_margs}}, \code{\link{r2sas}},
\code{\link{source_sas}}
}

