% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{search_hist}
\alias{search_hist}
\title{Search history}
\usage{
search_hist(x, ...)
}
\arguments{
\item{x}{numeric or character; if numeric, shows the most recent \code{n}
lines in \code{.Rhistory}; if character, searches for pattern matches}

\item{...}{additional arguments passed to \code{\link{grep}}}
}
\value{
A list of recent commands that match \code{pattern}.
}
\description{
Searches \code{.Rhistory} file for pattern matches.
}
\examples{
search_hist()
search_hist(25)
search_hist('?')
search_hist('?', fixed = TRUE)
search_hist('\\\\?')

}
