% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{show_math}
\alias{show_math}
\title{Show math equations}
\usage{
show_math(..., css = "", use_viewer = !is.null(getOption("viewer")))
}
\arguments{
\item{...}{one or more character strings}

\item{css}{optional css formatting}

\item{use_viewer}{logical; if \code{TRUE}, attempts to use
\code{rstudioapi::viewer} or opens in default browser on error}
}
\value{
The html code (invisibly) as a character string.
}
\description{
Displays math equations in \code{rstudioapi::viewer} or browser using the
\href{http://www.mathjax.org}{MathJax} javascript engine.
}
\examples{
\dontrun{
form1 <- '$$A=\\\\frac{B}{C}$$'

form2 <- '$$
  \\\\frac{1}{\\\\displaystyle 1+
  \\\\frac{1}{\\\\displaystyle 2+
  \\\\frac{1}{\\\\displaystyle 3+x}}} +
  \\\\frac{1}{1+\\\\frac{1}{2+\\\\frac{1}{3+x}}}
$$'

form3 <- '\\\\frac{d}{dx}\\\\left( \\\\int_{0}^{x} f(u)\\\\,du\\\\right)=f(x)'

form4 <- "
  \\\\forall a,b,c \\\\in \\\\mathbb{R} \\\\\\\\
  \\\\begin{align}
                        a + b &= c \\\\\\\\
               (a + b)(a - b) &= c(a - b) \\\\\\\\
                    a^2 - b^2 &= ca - cb \\\\\\\\
                     a^2 - ca &= b^2 - cb \\\\\\\\
    a^2 - ca + \\\\frac{c^2}{4} &= b^2 - cb + \\\\frac{c^2}{4} \\\\\\\\
         (a - \\\\frac{c}{2})^2 &= (b - \\\\frac{c}{2})^2 \\\\\\\\
             a - \\\\frac{c}{2} &= b - \\\\frac{c}{2} \\\\\\\\
                            a &= b \\\\qquad \\\\qquad \\\\blacksquare \\\\\\\\
   \\\\end{align}
"

show_math(form1)
cat(show_math(form4))

## use default browser
show_math(form2, use_viewer = FALSE)

## concatenate multiple expressions
show_math(form1, form2, form3, css = 'color: red; font-size: 15px;')
}

}
\seealso{
\code{\link{show_html}}, \code{\link{show_markdown}},
\href{http://detexify.kirelabs.org/classify.html}{draw math symbols},
\url{https://stackoverflow.com/q/31193843/2994949}
}
