% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{show_pch}
\alias{show_pch}
\title{Show plotting characters}
\usage{
show_pch()
}
\description{
In \code{R}, there are 26 numeric plotting characters. This function shows
these options and their respective numbers. Note that \code{col} specifies
both the border and fill color (if applicable) for \code{0:20}; \code{pch}s
\code{21:25} can be filled with \code{bg}.
}
\examples{
show_pch()
}
\seealso{
\code{\link{show_colors}}
}

