% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort2}
\alias{sort2}
\title{Sorting or ordering of vectors}
\usage{
sort2(x, decreasing = FALSE, index.return = FALSE, method = c("shell",
  "radix"))
}
\arguments{
\item{x}{a numeric, complex, or logical vector}

\item{decreasing}{logical; if \code{TRUE}, \code{x} is sorted in decreasing
order}

\item{index.return}{logical; if \code{TRUE}, an integer vector with the
ordering index is returned}

\item{method}{the sorting method used, \code{"shell"} (default) or
\code{"radix"}; see \code{\link{order}}}
}
\description{
Sort (or order) a vector or factor (partially) into ascending or descending
order keeping \code{NA} values in place.
}
\examples{
set.seed(1)
x <- sample(10)
x[4:6] <- NA

sort2(x)

## compare
sort(x, na.last = NA)
sort(x, na.last = TRUE)
sort(x, na.last = FALSE)

}
