% Generated by roxygen2 (4.0.1): do not edit by hand
\name{source_sas}
\alias{source_sas}
\title{Source SAS code}
\usage{
source_sas(path, ...)
}
\arguments{
\item{path}{character string path to \code{.sas} file}

\item{...}{additional parameters passed to \code{\link{r2sas}}}
}
\description{
\code{\link{source}}-esque function for \code{.sas} files
}
\examples{
\dontrun{
source_sas('./tests/testfiles/onemacro.sas')

## \%macro macro1(arg1, arg2);
## data data;
##   set data;
## run;
## \%mend macro1;
##
##
##
## ... will be run. Continue? (y/n):
}
}

