% Generated by roxygen2 (4.0.0): do not edit by hand
\name{spark}
\alias{spark}
\title{SparkBar generator}
\usage{
spark(...)
}
\arguments{
\item{...}{numeric vector}
}
\description{
spark(...)
}
\examples{
spark(30, 31, 32, 33)
spark(runif(20))
}
\seealso{
\url{https://gist.github.com/ramnathv/7793167}
}

