% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{sparkDT}
\alias{sparkDT}
\title{\code{datatable}s with sparklines}
\usage{
sparkDT(
  data,
  spark,
  type = c("line", "bar", "box1", "box2"),
  spark_range,
  options = list(),
  ...
)
}
\arguments{
\item{data}{a data frame or matrix}

\item{spark}{a \emph{named} list of lists for each column of \code{data}
for which an interactive sparkline will replace each row cell

each named list of \code{spark} should have length \code{nrow(data)} and
contain at least one numeric value}

\item{type}{the type of sparkline, one or more of "line", "bar", "box1",
or "box2", recycled as needed; the only difference between "box1" and
"box2" is the use of \code{spark_range}}

\item{spark_range}{an optional list or vector (recycled as needed) giving
the overall range for each list of \code{spark}; if missing, the ranges
will be calculated; note this is only applicable for \code{type = "line"}
or \code{type = "box1"}}

\item{options, ...}{\code{options} or additional arguments passed to
\code{\link[DT]{datatable}}}
}
\description{
Create an HTML table widget using the JavaScript library DataTables
(\code{\link[DT]{datatable}}) with \code{\link[sparkline]{sparkline}}
columns.
}
\examples{
\dontrun{
library('DT')

## strings of data separated by commas should be passed to each row
## this data will be used to generate the sparkline
dd <- aggregate(cbind(wt, mpg) ~ gear, mtcars, function(x)
  toString(fivenum(x)))
sparkDT(dd)


## for each column, create a list of vectors for each row to be plotted
l <- sapply(c('wt', 'mpg'), function(x)
  split(mtcars[, x], mtcars$gear), simplify = FALSE, USE.NAMES = TRUE)

sparkDT(dd, l, type = 'box1')


set.seed(1)
spark <- replicate(nrow(mtcars), round(rnorm(sample(20:100, 1)), 2), FALSE)

sparkDT(mtcars, list(mpg = spark, wt = spark, disp = spark, qsec = spark))

sparkDT(mtcars, list(mpg = spark, wt = spark, disp = spark, qsec = spark),
        spark_range = list(disp = c(-5, 5), mpg = c(0, 10)))

## note difference between box1 (boxes aligned) and box2 (max size)
sparkDT(mtcars[, c('mpg', 'wt')],
        list(mpg = spark, wt = spark),
        type = c('box1', 'box2'),
        # range = c(-2, 2),
        rownames = FALSE,
        colnames = c('box1', 'box2')
)
}

}
\seealso{
Adapted from \url{leonawicz.github.io/HtmlWidgetExamples/ex_dt_sparkline.html}
}
