% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_nth}
\alias{split_nth}
\title{Split at nth pattern}
\usage{
split_nth(text, pattern, n = NULL, keep_split = FALSE, repl = "$$$",
  ...)
}
\arguments{
\item{text, pattern}{the text and pattern character strings}

\item{n}{integer value(s) for splitting at the \code{n}th occurrence of
\code{pattern}; \code{NULL} will split at each \code{pattern}}

\item{keep_split}{logical; if \code{TRUE}, \code{text} will be split at
the boundaries of \code{pattern}}

\item{repl}{a string used internally for place-keeping during splitting;
to avoid unexpected results, make sure this string does not occur in
\code{text}}

\item{...}{additional arguments passed to \code{\link{gregexpr}}}
}
\description{
Split a string at the \code{n}th occurrence(s) of a \code{pattern}.
}
\examples{
s <- 'this  is a  test string to use   for testing   purposes'
split_nth(s, '\\\\s+')
split_nth(s, '\\\\s+', 3)
split_nth(s, '\\\\s+', c(3, 5))
split_nth(s, '\\\\s+', c(3, 5), keep_split = TRUE)
split_nth(s, '\\\\s{2,}', keep_split = TRUE)
split_nth(s, '\\\\s{2,}', 2:4)

}
