% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_median}
\alias{surv_median}
\title{Extract \code{survfit} summaries}
\usage{
surv_median(x, what = "median")
}
\arguments{
\item{x}{an object of class \code{\link[survival]{survfit}}}

\item{what}{the data to return, either the index, character string(s), or
\code{NULL} (returns the entire table)}
}
\description{
Convenience function to get summary results from \code{summary(x)$table}
where \code{x} is a \code{\link[survival]{survfit}} object.
}
\examples{
library('survival')
sfit1 <- survfit(Surv(time, status) ~ 1, lung)
sfit2 <- survfit(Surv(time, status) ~ sex, lung)

surv_median(sfit1)
surv_median(sfit2)

surv_median(sfit1, NULL)
surv_median(sfit1, 2:3)
surv_median(sfit1, 'median|CL')
surv_median(sfit1, c('events', 'median'))

}
