% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_summary}
\alias{surv_summary}
\title{Summary of a survival curve}
\usage{
surv_summary(s, digits = 3L, locf = FALSE, ...)
}
\arguments{
\item{s}{a \code{\link{survfit}} object}

\item{digits}{number of digits to use in printing numbers}

\item{locf}{logical; if \code{TRUE}, any \code{NA} probabilities will be
carried forward, e.g., if no events occurred during two consecutive time
intervals}

\item{...}{additional arguments passed to
\code{\link[survival]{summary.survfit}}}
}
\value{
A list with \code{survival:::print.summary.survfit} matrices for each
strata, i.e., the output printed.
}
\description{
Prints and returns a list containing the survival curve, confidence limits
for the curve, and other information which can be more useful than the
acutal return value of \code{survival:::print.summary.survfit}.
}
\examples{
library('survival')
fit1 <- survfit(coxph(Surv(time, status) ~ strata(I(age > 60)),
                      data = cancer),
                conf.type = 'log-log')
surv_summary(fit1, times = c(0, 100, 200))

}
\seealso{
\code{\link{survfit}}; \code{\link{print.summary.survfit}}
}
