% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_table}
\alias{surv_table}
\title{Summary table}
\usage{
surv_table(s, digits = 3, times = pretty(s$time), maxtime = FALSE, ...)
}
\arguments{
\item{s}{a \code{\link[survival]{survfit}} object}

\item{digits}{number of digits to use in printing numbers}

\item{times}{vector of times}

\item{maxtime}{logical; if \code{TRUE}, adds the maximum time for which an
even occurs; if \code{FALSE}, number of events may not sum to total}

\item{...}{additional arguments passed to \code{\link{summary.survfit}}}
}
\value{
A matrix (or list of matrices) with formatted summaries for each strata; see
\code{\link{summary.survfit}}
}
\description{
Prints a formatted summary table for \code{\link{survfit}} objects.
}
\examples{
library('survival')
fit0 <- survfit(coxph(Surv(time, status == 2) ~ 1,
                      data = cancer),
                conf.type = 'log-log')
surv_table(fit0, times = 0:2 * 100, maxtime = FALSE)

## also works for list of tables
fit1 <- survfit(coxph(Surv(time, status == 2) ~ strata(I(age > 60)),
                      data = cancer),
                conf.type = 'log-log', conf.int = 0.9)
surv_table(fit1)

s <- `colnames<-`(surv_table(fit0, times = 0:8 * 100, digits = 2)[, -4],
                  c('Time','No. at risk','No. of events','Surv (95\% CI)'))
ht <- htmlTable::htmlTable(s, caption = 'Table: Overall survival.')
structure(ht, class = 'htmlTable')

}
\seealso{
\code{\link{survfit}}, \code{\link{print.summary.survfit}}
}
