% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_test}
\alias{surv_test}
\alias{lr_pval}
\alias{lr_text}
\alias{tt_pval}
\alias{tt_text}
\alias{hr_pval}
\alias{hr_text}
\title{Survival curve tests}
\usage{
lr_pval(object, details = FALSE, data = NULL, ...)

lr_text(formula, data, rho = 0, ..., details = TRUE, pFUN = NULL)

tt_pval(object, details = FALSE, data = NULL, ...)

tt_text(formula, data, ..., details = TRUE, pFUN = NULL)

hr_pval(object, details = FALSE, data = NULL, ...)

hr_text(formula, data, ..., details = TRUE, pFUN = NULL)
}
\arguments{
\item{object}{a \code{\link{survfit}}, \code{\link{survdiff}}, or
\code{\link{coxph}} object; alternatively a \code{\link[=Surv]{survival
formula}} (\code{data} must be given)}

\item{details}{logical; \code{TRUE} returns statistic, degrees of freedom,
and p-value where \code{FALSE} returns only a pvalue}

\item{formula, data, rho, ...}{passed to \code{\link{survdiff}} or
\code{\link{coxph}}}

\item{pFUN}{logical; if \code{TRUE}, p-values are formatted with
\code{\link{pvalr}}; if \code{FALSE}, no formatting is performed;
alternatively, a function can be passed which should take a numeric value
and return a character string (or a value to be coerced) for printing}
}
\description{
Internal functions for \code{\link{survdiff}} and \code{\link{survfit}}
objects.

\code{lr_pval} and \code{tt_pval} take objects or formulas and compute
log-rank and Tarone's trend test, respectively. \code{lr_text} and
\code{tt_text} format the test results for plotting.
}
\examples{
\dontrun{
library('survival')
data('larynx', package = 'KMsurv')

form <- Surv(time, delta) ~ stage
sf <- survfit(form, larynx)
sd <- survdiff(form, larynx)

kmplot(sf, lr_test = TRUE)

rawr:::lr_pval(sf)
rawr:::lr_pval(sd, TRUE)
rawr:::lr_text(Surv(time, delta) ~ stage, larynx)


rawr:::tt_pval(sf)
rawr:::tt_pval(sd, TRUE)
rawr:::tt_text(Surv(time, delta) ~ stage, larynx)

## compare
chi <- coxph(Surv(time, delta) ~ stage, larynx)$score
pchisq(chi, 1, lower.tail = FALSE)
}

}
\references{
Tarone, Robert E. Tests for Trend in Life Table Analysis. \emph{Biometrika}
\strong{62} vol. 62 (Dec 1975), 679-82.

\url{https://stat.ethz.ch/pipermail/r-help/2008-April/160209.html}
}
\seealso{
\code{\link{survdiff_pairs}}
}
