% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{survdiff_time}
\alias{survdiff_time}
\title{survdiff time}
\usage{
survdiff_time(object, time, method = c("log-log", "plain"))
}
\arguments{
\item{object}{a \code{\link{survfit}} object}

\item{time}{time at which to evaluate differences}

\item{method}{transformation, one of \code{"log-log"} (default) often
more stable and bounded on real line or \code{"plain"} for no
transformation}
}
\description{
Compare survival curves at a specific time point.
}
\examples{
library('survival')
fit1 <- survfit(Surv(time, status) ~ sex, lung)
kmplot(fit1)
survdiff_time(fit1, 200)
survdiff_time(fit1, c(200, 800))

dd <- lung[lung$ph.ecog < 3, ]
fit2 <- survfit(Surv(time, status) ~ ph.ecog, dd)
kmplot(fit2)
survdiff_time(fit2, 400)

## pairwise comparisons
sp <- split(dd, dd$ph.ecog)
xx <- combn(sp, 2, simplify = FALSE, function(x) {
  x <- do.call('rbind', x)
  survdiff_time(survfit(Surv(time, status) ~ ph.ecog, x), 400)
})
do.call('rbind', xx)

}
