% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{swimmer}
\alias{swimmer}
\alias{lines.swimmer}
\alias{points.swimmer}
\alias{legend.swimmer}
\title{Swimmer plot}
\usage{
swimmer(
  id,
  start,
  end,
  event,
  time,
  pch = NULL,
  col.pch = par("col"),
  cex.pch = par("cex"),
  bg.pch = par("bg"),
  line = c("line", "bar"),
  col.line = "grey",
  lwd.line = par("lwd"),
  xlim = NULL,
  ylim = NULL,
  xlab = "Time",
  ylab = "",
  label = TRUE,
  args.legend = list(),
  ...
)

\method{lines}{swimmer}(
  object,
  id,
  start,
  end,
  event,
  col.line = adjustcolor(par("col"), 0.25),
  lwd.line = par("lwd"),
  hadj = 0,
  vadj = 0,
  censor = FALSE,
  args.legend = list(),
  ...
)

\method{points}{swimmer}(
  object,
  id,
  time,
  event,
  pch = NULL,
  col.pch = par("col"),
  cex.pch = par("cex"),
  bg.pch = par("bg"),
  hadj = 0,
  vadj = 0,
  censor = FALSE,
  args.legend = list(),
  ...
)

legend.swimmer(object, ...)
}
\arguments{
\item{id, event, time}{vectors of ids, events, and event times; for
\code{swimmer} this should be one row per ID; \code{points.swimmer} may
have 0 or more for each ID

\code{event} and \code{time} in \code{swimmer} are intended to be single,
absorbing states such as end of follow-up, death, on-going response, etc,;
for \code{points.swimmer}, these may be events that occur multiple times
during follow-up such as response assessments or toxicity events}

\item{start, end}{vectors of start and end/censored times}

\item{pch, col.pch, cex.pch, bg.pch}{plotting characters and attributes for
each event type}

\item{line}{draw a line or bar for the follow-up time}

\item{col.line, lwd.line}{the color and line width for \code{line}}

\item{xlim, ylim}{x- and y-axis limits}

\item{xlab, ylab}{x-and y-axis labels}

\item{label}{logical; if \code{TRUE}, IDs will be shown on y-axis}

\item{args.legend}{a \emph{named} list of arguments passed to \code{\link{legend}}}

\item{...}{additional arguments passed to \code{\link{plot}} or further to
\code{\link{par}}}

\item{object}{an object of class \code{"swimmer"}, usually returned by
\code{swimmer} or \code{points.swimmer}}

\item{hadj, vadj}{horizontal and vertical adjustments to points}

\item{censor}{logical; if \code{TRUE}, events occurring after end time
defined by \code{swimmer} will be removed; otherwise, all events will be
included, potentially after the end date (default)}
}
\value{
A list of length 2 with \code{$data} and \code{$legend} elements containing
the data plotted and used for the call to \code{legend.swimmer}.
}
\description{
Functions to draw swimmer plots, points, and legends.
}
\examples{
dat <- data.frame(
  id = c('c1', 'd1', 'c2', 'c3', 'd2'),
  start = 0,
  end = 1:5 + 10,
  event = c('Ongoing', 'Death', 'Ongoing', 'Ongoing', 'Death'),
  time = 1:5 + 10
)

rsp <- data.frame(
  id = rep(dat$id, each = 3),
  time = sample(1:10, 15, replace = TRUE),
  event = sample(c('CR', 'PR', 'SD'), 15, replace = TRUE)
)

## base swimmer plot with single events
sw <- swimmer(
  dat$id, dat$start, dat$end, dat$event, dat$time,
  pch = c('X', '>'), col.pch = 2:1, cex.pch = 1.5
)
## add auto legend
legend.swimmer(sw, x = 'bottomright')

## add single time point events and legend
pt <- points(sw, rsp$id, rsp$time, rsp$event)
legend.swimmer(pt, x = 'bottomright', inset = c(0, 0.15))

## alternatively combine all legends (see rawr::mlegend)
legend.swimmer(list(pt, sw), x = 'topleft')

## add interval events and legend
lin <- data.frame(
  id = c('d1', 'd1', 'c3'),
  start = c(2, 10, 3), end = c(4, 11, 10),
  event = c('AE 1', 'AE 2', 'AE 1'),
  col = adjustcolor(c('red', 'blue', 'red'), 0.25)
)
li <- lines(sw, lin$id, lin$start, lin$end, lin$event, col = lin$col)
legend.swimmer(li, x = 'right')

}
