% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch2}
\alias{switch2}
\title{Select one of a list of alternatives}
\usage{
switch2(x, ...)
}
\arguments{
\item{x}{a vector of character strings}

\item{...}{a list of alternatives; see \code{\link{switch}}}
}
\description{
\code{\link{switch}} which accepts a vector of strings.
}
\examples{
switch('b',               a = 10, b = 2:0, c = 3, d = 1.5, e = mean)
switch2(c('d', 'a'),      a = 10, b = 2:0, c = 3, d = 1.5, e = mean)
switch2(c('b', 'a'),      a = 10, b = 2:0, c = 3, d = 1.5, e = mean)
switch2(c('b', 'a', 'e'), a = 10, b = 2:0, c = 3, d = 1.5, e = mean)

}
