% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_stat2}
\alias{tabler_stat2}
\title{\code{tabler_stat} wrappers}
\usage{
tabler_stat2(
  data,
  varname,
  byvar = NULL,
  varname_label = names(varname),
  byvar_label = names(byvar),
  digits = NULL,
  FUN = NULL,
  paired = FALSE,
  id = NULL,
  confint = FALSE,
  survmedian = TRUE,
  survtime = FALSE,
  time = 0,
  total = TRUE,
  n = NULL,
  include_na_in_prop = TRUE,
  iqr = FALSE,
  format_pval = TRUE,
  color_pval = TRUE,
  correct = FALSE,
  color_missing = TRUE,
  dagger = TRUE,
  group = NULL,
  color_cell_by = c("none", "value", "pct"),
  cell_color = palette()[1:2],
  statArgs = NULL,
  align = NULL,
  rgroup = NULL,
  cgroup = NULL,
  tfoot = NULL,
  tfoot2 = NULL,
  htmlArgs = NULL,
  zeros = "-",
  clean_daggers = FALSE
)
}
\arguments{
\item{data}{a matrix or data frame with variables \code{varname} and
\code{byvar}}

\item{varname}{one or more variables in \code{data} to calculate
statistics by \code{byvar}; the rows variable(s) of the table}

\item{byvar}{a stratification variable; the column variable of the table}

\item{varname_label, byvar_label}{optional labels for each \code{varname}
  and \code{byvar}

  note that duplicated \code{varname_label}s will be processed individually
  then grouped; this is useful if one variable is shown as continuous and
  categorical, for example; grouped variables should be consecutive, see
  examples}

\item{digits}{\code{NULL} or a vector of digits past the decimal point to
  keep for each \code{varname}; if \code{NULL}, these will be guessed using
  \code{rawr:::guess_digits}; if length 1, all will be rounded to
  \code{digits}

  alternatively, to set \code{digits} for a single \code{varname} manually,
  a \emph{named} vector of \code{digits} corresponding to one or more
  \code{varname} variables can be used; see examples}

\item{FUN}{\code{NULL} or a list of functions performing the test of
association between each \code{varname} and \code{byvar}; see
\code{\link{tabler_stat}}}

\item{paired, id}{(experimental) logical; if \code{TRUE}, paired tests are
used grouped by \code{byvar}; \code{id} should be the column name of
\code{data} giving the IDs}

\item{confint}{optional vector of \code{varname}(s) to summarize as
confidence intervals}

\item{survmedian, survtime, time}{survival object options - median survival or
survival time}

\item{total}{logical; if \code{TRUE}, total column will be shown}

\item{n}{(optional) the sample size for each column used to calculate
percents; if length 1, recycled as necessary; if length > 1, length must
be the same as the number of columns in the table including total}

\item{include_na_in_prop}{logical; if \code{TRUE} (default), the number of
missing values is included when calculating proportions for factor levels;
if \code{FALSE}, only non-missing levels count towards proportions}

\item{iqr}{logical; if \code{TRUE}, the interquartile range is used
instead of the full range (default) for continuous variables}

\item{format_pval}{logical; if \code{TRUE}, p-values will be formatted
using \code{\link{pvalr}}; alternatively, a function may by used which
will be applied to each p-value}

\item{color_pval, color_missing, dagger}{\code{NULL} or vectors, recycled as
needed for each \code{varname}; see \code{\link{tabler_stat}}}

\item{correct}{logical or one of \code{\link{p.adjust.methods}}; if p-value
correction is desired, a column is added to the table with the corrected
p-values}

\item{group}{a \emph{named} vector of variables (either index of or a
value in \code{varname}) to begin a new group (see \code{tspanner} in
\code{\link[htmlTable]{htmlTable}})}

\item{color_cell_by, cell_color}{apply a color gradient (\code{cell_color})
to each cell (for html output); one of \code{"none"}, \code{"value"}, or
\code{"pct"}; see \code{\link{tabler_stat}}}

\item{statArgs}{a named list of additional arguments passed to
\code{\link[Gmisc]{getDescriptionStatsBy}}}

\item{align, rgroup, cgroup, tfoot}{optional arguments passed to
\code{\link[htmlTable]{htmlTable}}; if \code{cgroup = FALSE}, no column
spanners will be shown}

\item{tfoot2}{optional footnote(s) appended to \code{tfoot}}

\item{htmlArgs}{a named list of additional arguments passed to
\code{\link[htmlTable]{htmlTable}}}

\item{zeros}{a character string used in place of zero cells (non-character
value keeps cells as-is)}

\item{clean_daggers}{logical or one of \code{"letters"}, \code{"numbers"};
values other than \code{FALSE} will replace default dagger characters}
}
\description{
Helper functions for using \code{\link{tabler_stat}}.
}
\examples{
sapply(mtcars[1:6], rawr:::guess_digits)
Map(rawr::roundr, mtcars[1:6], sapply(mtcars[1:6], rawr:::guess_digits))

rawr:::get_tabler_stat_n(mtcars$gear)


## typical usage
mt <- within(mtcars, {
  cyl  <- factor(cyl)
  mpg2 <- factor(+(mpg > 20), 0:1, c('&le; 20', '&gt; 20'))
})

tabler_stat2(mt, c('mpg', 'cyl', 'wt'))
tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'vs')
tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'vs', group = 'wt')
tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'vs',
  group = c(group1 = 1, 'Weight' = 3))
tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'vs',
  FUN = list('kruskal', 'fisher', FALSE))

tabler_stat2(
 mt, c('mpg', 'cyl', 'wt'), 'vs',
 cgroup = c('', 'V/S engine', ''),
 rgroup = c('Miles/gallon', 'No. cylinders', 'Weight (1000 lbs)')
)

tabler_stat2(
  mt,
  varname = c(MPG = 'mpg', MPG = 'mpg2', Cylinders = 'cyl', Weight = 'wt'),
  byvar = c('V/S engine' = 'vs'),
  confint = c('mpg2', 'cyl'),
  zeros = NULL,
  digits = c(wt = 2)
  # digits = c('3' = 2) ## equivalently
)


## multiple column variables
tabler_stat2(mt, c('mpg', 'cyl', 'wt'), c('vs', 'gear'), FUN = NA)


\dontrun{
## experimental -- row- or column-merge tables
h1 <- tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'vs')
h2 <- tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'am')
h3 <- tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'gear')

rawr:::combine_tabler_stat2(list(h1, h2), how = 'rbind')
rawr:::combine_tabler_stat2(list(table1 = h1, table2 = h2), how = 'rbind')

rawr:::combine_tabler_stat2(list(h1, h2, h3), how = 'cbind')
rawr:::combine_tabler_stat2(list(t1 = h1, t2 = h2, t3 = h3), how = 'cbind')
}

}
\seealso{
\code{\link{tabler_stat}}; \code{rawr:::get_tabler_stat_n}

\code{rawr:::guess_digits}; \code{rawr:::name_or_index}

\code{rawr:::tabler_stat_list}; \code{rawr:::tabler_stat_html}

Other tabler:
\code{\link[=tabler]{tabler()}},
\code{\link[=tabler_by]{tabler_by()}},
\code{\link[=tabler_resp]{tabler_resp()}},
\code{\link[=tabler_stat]{tabler_stat()}}
}
\concept{tabler}
