% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{tcol}
\alias{tcol}
\title{Transparent colors}
\usage{
tcol(col, alpha = 1)
}
\arguments{
\item{col}{a vector of color names, hexadecimal strings, or integers which
correspond to the current \code{\link{palette}}}

\item{alpha}{the alpha transparency in \code{[0,1]}}
}
\description{
Add alpha transparency to colors.
}
\examples{
cols <- c('red', 'green', 'blue')

## a normal plot
plot(rnorm(100), col = tcol(cols), pch = 16, cex = 4)

## more transparent
plot(x <- rnorm(100), col = tcol(cols, 0.5), pch = 16, cex = 4)

## hexadecimal strings also work
cols <- c('#FF0000', '#00FF00', '#0000FF')
plot(rnorm(100), col = tcol(cols, c(0.2, 0.4, 1)), pch = 16, cex = 4)

}
\seealso{
\code{\link{as.hexmode}}; \code{\link{col2rgb}}; \code{\link{adjustcolor}};
\code{\link{rgb}};
}
