% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytext.R
\name{tidytext}
\alias{tidytext}
\alias{tidytextmining}
\title{Retrieve raw R code tidytext tutorial}
\usage{
tidytext(url)

tidytextmining(url)
}
\arguments{
\item{url}{Link to tidytext tutorial}
}
\description{
Retrieve raw R code tidytext tutorial
}
\examples{
\dontrun{
tidytext("https://www.tidytextmining.com/sentiment.html")

# Same as above but provided to cat for easy viewing
tidytext("https://www.tidytextmining.com/sentiment.html") \%>\%
  cat

}
}
