% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tox_worst}
\alias{tox_worst}
\title{Find most severe toxicities}
\usage{
tox_worst(dat, id = "casenum", tox_desc = "tox_desc",
  tox_grade = "tox_grade")
}
\arguments{
\item{dat}{toxicity data frame}

\item{id}{column name with identifier}

\item{tox_desc}{column name with toxicity descriptions (or codes)}

\item{tox_grade}{column name with toxiticity grades; should be a factor
with the desired order to be properly sorted}
}
\value{
A list of length three with \code{tox_worst}, the filtered data frame;
\code{dat}, the input data frame sorted by \code{id}, \code{tox_desc},
and \code{tox_grade}; and \code{duplicates}, the rows which correspond to
duplicate \code{tox_desc} per \code{id} but of an equal or lesser 
\code{tox_grade}.
}
\description{
Returns a subset of input data with worst toxicity grade per toxicity
code per patient, ie, sorts and removes duplicates.
}
\examples{
oo <- options()
options(stringsAsFactors = FALSE)
set.seed(1)

f <- function(x, ...) sample(x, 100, replace = TRUE, ...)
tox <- data.frame(casenum = rep(1:10, 10), phase = 1:2,
                  tox_code = f(rawr::ctcae_v4$tox_code[1:10]),
                  tox_grade = f(1:3, prob = c(.3, .4, .3)))

tox <- cbind(tox, match_ctc(tox$tox_code)$matches[, c('tox_cat', 'tox_desc')])

tox <- within(tox, {
  tox_grade <- factor(tox_grade, levels = 1:3, 
                      labels = c('Mild','Moderate','Severe'))
  tox_cat <- factor(tox_cat)
  tox_desc <- factor(tox_desc)
})

(out <- tox_worst(tox))

## use tabler_by to summarize:
tabler_by(out$tox_worst, 'tox_desc', 'tox_grade')
options(oo)

}

