% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{tree}
\alias{tree}
\title{tree}
\usage{
tree(path = ".", full.names = FALSE, ndirs = 5, nfiles = 5)
}
\arguments{
\item{path}{file name path as character string}

\item{full.names}{logical; if \code{TRUE}, the full file path will be
returned; otherwise, only the \code{\link{basename}} is returned (default)}

\item{ndirs, nfiles}{maximum number of directories and files per directory
to print}
}
\description{
List contents of directories in a tree-like format.
}
\examples{
str(tree(system.file(package = 'rawr'), FALSE))

}
\references{
\href{http://stackoverflow.com/questions/14188197/representing-a-directory-tree-as-a-recursive-list}{SO question}
}

