% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryCatch2}
\alias{tryCatch2}
\title{Condition handling and recovery}
\usage{
tryCatch2(expr, ..., simplify = TRUE)
}
\arguments{
\item{expr}{an expression}

\item{...}{ignored}

\item{simplify}{logical; if \code{TRUE}, simplifies the returned list to
a vector if there were no exceptions}
}
\value{
If \code{expr} is evaluated with no errors, warnings, messagess, or
interruptions, then only the value of \code{expr} is returned unless
\code{simplify = FALSE} in which case a list with the return value.

If \code{expr} results in one of the above conditions, then a list with
the value of \code{expr} along with an additional element for the
condition(s) which can be accessed with \code{$warning}, \code{$error},
etc.
}
\description{
Modification of \code{\link{tryCatch}} which allows recovery from warnings
and messages while returning the value from \code{expr}. Note that the
return value from errors and interruptions is \code{NULL}.
}
\examples{
## returns value if no errors, warnings, etc
tryCatch2(1)
tryCatch2(1, simplify = FALSE)


tryCatch2(stop('halt at once!'))
tryCatch2({warning('warning'); message('message'); 1})


## compare
tryCatch({warning('this is your warning...'); 1}, warning = function(w) w)
tryCatch2({warning('this is your warning...'); 1})

tryCatch({message('calculating...'); 1}, message = function(m) m)
tryCatch2({message('calculating...'); 1})

}
\seealso{
\code{\link[base]{tryCatch}}; \code{\link[logR]{tryCatch2}};
\url{https://stackoverflow.com/q/4948361/2994949};
\href{https://github.com/jangorecki/logR}{\pkg{logR} github repo}
}
\author{
Martin Morgan, Jan Gorecki, Robert Redd (modifications)
}
