% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{tte}
\alias{tte}
\title{Time-to-event data}
\usage{
tte(start, fwup, ..., crisk = FALSE, fix = TRUE, censor = Inf, factor = 1)
}
\arguments{
\item{start, fwup}{start and last follow-up date vectors}

\item{...}{one or more event date vectors; note that event dates occurring
before \code{start} are removed}

\item{crisk}{logical; if \code{TRUE}, event dates are treated as competing
events and the order of dates in \code{...} matters, i.e., the first date
takes precedence, and indicators are assigned in order of event date
vectors}

\item{fix}{logical; if \code{TRUE} (default); censor negative times at 0
and censor the event}

\item{censor}{max time after which events/times are censored}

\item{factor}{time factor to convert days (default) to months (e.g, use
30.437), years (365.242), etc.}
}
\description{
Calculate time-to-event and competing risks times and event indicators.
}
\examples{
dd <- data.frame(
  start = 0,
  end = 10,
  death = c(5, 2, NA, 2, NA),
  prog = c(3, 2, NA, 3, 5),
  trt = c(4, 1, NA, NA, 8)
)
dd[] <- lapply(dd, as.Date, origin = '2000-01-01')

tte(dd$start, dd$end, dd$death)
tte(dd$start, dd$end, dd$death, dd$prog)
tte(dd$start, dd$end, dd$death, dd$prog, crisk = TRUE)
tte(dd$start, dd$end, dd$death, dd$prog, dd$trt, crisk = TRUE)

}
