% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{updateR}
\alias{updateR}
\title{Update \code{R}}
\usage{
updateR(update = TRUE)
}
\arguments{
\item{update}{logical; if \code{TRUE}, checks for available packages
updates, downloads, and installs}
}
\description{
Copies and updates \code{R} libraries from most recent installed version
into the current \code{\link{.libPaths}} directory. This assumes that the
user has installed a new \code{X.x} version of \code{R} but will not copy
any libraries from previous frameworks into the new library.
}
\seealso{
\code{\link{update.packages}}
}
