% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{vs.glmnet}
\alias{vs.glmnet}
\title{glmnet variable selection}
\usage{
vs.glmnet(formula, data, alpha = 1, ...)
}
\arguments{
\item{formula, data}{a formula and data frame containing the response and
all potential predictor variables}

\item{alpha}{elastic net mixing parameter; default penalty is 1 for lasso,
or any value between 0 (ridge penalty) and 1}

\item{...}{additional arguments passed to \code{\link[glmnet]{cv.glmnet}}
or further to \code{\link[glmnet]{glmnet}}}
}
\description{
Drop least relevant variables from a \code{\link[glmnet]{glmnet}}
model with optional diagnostics.
}
\examples{
vs.glmnet(iris, alpha = 1) ## lasso - default
vs.glmnet(iris, alpha = 0) ## ridge

vs.glmnet(I(Species == 'setosa') ~ ., iris)

}
\seealso{
\code{\link{vs.rfsrc}}
}
