% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{vs.rfsrc}
\alias{vs.rfsrc}
\title{rfsrc variable selection}
\usage{
vs.rfsrc(
  formula,
  data,
  nvar = -1L,
  depth = NULL,
  verbose = FALSE,
  plot = verbose,
  refit = TRUE,
  ...
)
}
\arguments{
\item{formula, data}{a formula and data frame containing the response and
all potential predictor variables}

\item{nvar}{number of variables desired in final model (if positive) or
the number of variables to drop (if negative)}

\item{depth}{logical or a numeric value (if \code{NULL}, variables will
be selected via \code{nvar}); if \code{TRUE}, variables will be selected
using first-order depths and the \code{threshold} value returned by
\code{\link[randomForestSRC]{max.subtree}}; alternatively, a numeric
value to be used as the threshold}

\item{verbose, plot}{logical; if \code{TRUE}, an updated \code{formula}
and/or variable importance figures are printed after each step}

\item{refit}{logical; if \code{TRUE} (default), model is re-fit after each
variable is dropped; otherwise, all variables to be dropped will be done
in one step which could give significant performance gains for some models
such as survival models (note the final model may be different than if
\code{refit = TRUE})}

\item{...}{additional parameters passed to
\code{\link[randomForestSRC]{rfsrc}}}
}
\value{
The \code{formula} of the final model.
}
\description{
Drop least relevant variables from a \code{\link[randomForestSRC]{rfsrc}}
model with optional diagnostics.
}
\examples{
vs.rfsrc(iris)
vs.rfsrc(I(Species == 'setosa') ~ ., iris)


## select variables based on first-order depth
f <- formula(rev(iris))
vs.rfsrc(f, iris, depth = TRUE, verbose = TRUE, plot = FALSE, ntree = 10)

## keep only most relevant 2 variables
vs.rfsrc(f, iris, nvar = 2)

## drop 2 least relevant variables
vs.rfsrc(f, iris, nvar = -2)
vs.rfsrc(f, iris, nvar = -2, refit = FALSE)


library('survival')
f <- Surv(time, status == 0) ~ rx + sex + age + obstruct + adhere + nodes
vs.rfsrc(f, colon, nvar = 4, ntree = 5)
# vs.rfsrc(f, colon, nvar = -2, ntree = 5) ## same

## for slower models, refit = FALSE may improve performance
vs.rfsrc(f, colon, nvar = 4, ntree = 5, refit = FALSE, plot = TRUE)

}
\seealso{
\code{\link{vs.rfsrc}}
}
