% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{waffle}
\alias{waffle}
\title{waffle}
\usage{
waffle(
  mat,
  xpad = 0,
  ypad = 0,
  heights = NULL,
  widths = NULL,
  colpad = 0,
  rowpad = 0,
  invert = "",
  ...,
  reset_par = TRUE,
  add = FALSE
)
}
\arguments{
\item{mat}{a matrix of integers or character strings of color names; if
\code{mat} is a matrix of integers, the colors used will correspond to
the current \code{\link{palette}}}

\item{xpad, ypad}{padding between \code{\link{rect}}s, recycled as needed;
note that these do not affect the center coordinates of the rectangles}

\item{heights, widths}{row heights and column widths, respectively, usually
in \code{[0,1]}, recycled as needed; note \code{xpad} and/or \code{ypad}
are ignored if \code{widths} and/or \code{heights} are used, respectively}

\item{colpad, rowpad}{amount of padding between columns and rows; note that
these shift the \code{\link{rect}} center coordinates rather than the
heights and widths directly}

\item{invert}{character string indicating about which axis the matrix
should be inverted; possible values are \code{"x"}, \code{"y"}, or
\code{"xy"}}

\item{...}{additional graphical parameters passed to \code{\link{rect}}
such as \code{border}, \code{density}, \code{lty}, etc.}

\item{reset_par}{logical; if \code{TRUE}, resets \code{\link{par}}
settings to state before function call; setting \code{reset_par = FALSE}
along with the return value are useful for adding to a \code{waffle} plot}

\item{add}{logical; use \code{TRUE} to add to an existing plot; otherwise,
a new frame and window are initialized}
}
\value{
A list of four matrices:

\item{\code{$matrix}}{the input matrix as plotted including inversions}
\item{\code{$origin}}{coordinates of the bottom-left corner for each box}
\item{\code{$centers}}{coordinates for the centers of each box}
\item{\code{$rect}}{coordinates for each corner of each box}
}
\description{
A waffle chart.
}
\examples{
waffle(matrix(1:9, 3))
waffle(matrix(1:9, 3), invert = 'x')
waffle(matrix(1:9, 3), heights = c(.25, .95, .5), border = NA)
waffle(matrix(1:9, 3), xpad = 0, colpad = c(0, .1, 0))


## heatmap
## convert the numeric data to color strings
cols <- c(cor(mtcars))
cols <- tcol(c('blue', 'red')[(cols > 0) + 1L], alpha = c(abs(cols)))

mat <- matrix(cols, 11)
waffle(mat, reset_par = FALSE, invert = 'x')
axis(3, at = 1:11 - .5, labels = names(mtcars), lwd = 0)


## use colpad/rowpad to create sections
w <- waffle(mat, reset_par = FALSE, invert = 'x',
            colpad = rep(c(0, 0.5, 1, 0), c(5, 1, 1, 4)),
            rowpad = rep(c(0, 0.5, 0), c(5, 1, 5)))
axis(3, unique(w$centers[, 'x']), names(mtcars), lwd = 0)


## adding to margins of another plot
set.seed(1)
n  <- 97
ng <- 3
cols <- c('beige', 'dodgerblue2', 'green', 'orange')
x <- sample(cols, n * ng, replace = TRUE, prob = c(.05,.31,.32,.32))
x <- kinda_sort(x, n = 20)

op <- par(fig = c(0, 1, 0.2, 0.9), mar = c(0, 5, 0, 1))
plot(cumsum(rnorm(n)), type = 'l', ann = FALSE, xaxt = 'n')

par(fig = c(0, 1, 0, 0.2), mar = c(1, 5, 0, 1), new = TRUE)
waffle(matrix(x, ng), heights = c(0.95, 0.5, 0.95), border = 'white',
       reset_par = FALSE) -> wf
text(0, rev(unique(wf$centers[, 'y'])), paste('Feature', 1:3),
     xpd = NA, pos = 2L)

par(fig = c(0, 1, 0.9, 1), mar = c(0.5, 5, 0.5, 1), new = TRUE)
waffle(matrix(x, ng)[1L, , drop = FALSE], heights = 0.5,
       border = 'white', reset_par = FALSE)
box()
box('outer')
par(op)


## waffle conveniently returns the centers of the rects
## be sure _not_ to reset pars for proper alignment
w <- waffle(matrix(1:8, 2), reset_par = FALSE, invert = 'xy')
text(w$c[, 'x'], w$c[, 'y'], labels = palette(), col = 'white')


## this is similar to ?rawr::show_colors
col <- colors()[1:25 ^ 2]
w <- waffle(matrix(col, 25), reset_par = FALSE, invert = 'xy', border = NA)
text(w$c[, 'x'], w$c[, 'y'], labels = col, col = 'black', cex = .4)

}
