% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{warnifnot}
\alias{warnifnot}
\title{warnifnot}
\usage{
warnifnot(..., exprs, exprObject, local = TRUE)
}
\arguments{
\item{..., exprs, exprObject, local}{see \code{\link{stopifnot}}}
}
\description{
\code{\link{stopifnot}} but for \code{\link{warnings}}.
}
\examples{
\donttest{
warnifnot(TRUE)
warnifnot(FALSE)
warnifnot(
  'warning1' = FALSE,
  'warning2' = TRUE,
  'warning3' = FALSE,
  FALSE
)
}

}
