% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{waterfall}
\alias{waterfall}
\title{Waterall plot}
\usage{
waterfall(x, type = 1L, col = c("red", "blue"), ..., arrows = TRUE,
  panel.first = NULL, panel.last = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{type}{type of waterfall plot; \code{type = 1} draws sorted bars for
values of \code{x}; \code{type = 2} starts from 0 and draws bars according
to the magnitude and direction of \code{x}}

\item{col}{a vector of colors having 1) the same length as \code{x}; length
2 (for negative and positive values if \code{type = 2}) or more to be used
for color interpolation; see \code{\link{colorRampPalette}}
\code{length(x)}; or 3)}

\item{...}{additional arguments passed to \code{\link{barplot}}, to/from
other methods, or to \code{\link{par}}}

\item{arrows}{logical if \code{TRUE} and \code{type = 2}, arrows are drawn
in the direction of each bar}

\item{panel.first}{an "expression" to be evaluated after the plot axes are
set up but before any plotting takes place; this can be useful for drawing
background grids or scatterplot smooths; note that this works by lazy
evaluation: passing this argument from other plot methods may well not work
since it may be evaluated too early; see also \code{\link{plot.default}}}

\item{panel.last}{an expression to be evaluated after plotting has taken
place but before the axes, title, and box are added; see the comments about
\code{panel.first}}
}
\description{
Draw two types of waterfall plots.
}
\examples{
set.seed(1)
change <- runif(20, -1, 1) * 100
col <- c(PD = 'red', SD = 'blue', CR = 'chartreuse4')

## interpolation
waterfall(change, col = col)

## discrete breaks
waterfall(change, col = as.character(cut(change, c(-Inf, -50, 50, Inf), col)))
legend('top', names(col), fill = col, horiz = TRUE, bty = 'n', border = NA)
title(xlab = 'Patient', ylab = '\% change', main = 'waterfall', line = 1)


## type 2
waterfall(change, col = col, type = 2,
          panel.first = grid(),
          panel.last = {
            axis(2, las = 1)
            title(main = 'waterfall - type 2')
          }
)

}

