% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.min2}
\alias{which.min2}
\alias{which.max2}
\title{\code{which.min} and \code{which.max} for n indices}
\usage{
which.min2(x, n = 1L)

which.max2(x, n = 1L)
}
\arguments{
\item{x}{numeric (logical, integer, or double) vector or an \code{R}
object for which the internal coercion to \code{\link{double}} works
whose \code{\link{min}} or \code{\link{max}} is searched for}

\item{n}{the number of first min or max to return}
}
\description{
Returns the indices of the first \code{n} min and max positions.
}
\examples{
x <- c(3, 1, 2, -2, 5)
stopifnot(identical(which.min(x), which.min2(x)))
stopifnot(identical(which.max(x), which.max2(x)))

which.min2(NA, 3)
which.min2(x, 3)
which.max2(x, 3)

}
\seealso{
\code{\link{which.min}}; \code{\link{which.max}}
}
