% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{winsorize}
\alias{winsorize}
\title{Winsorization}
\usage{
winsorize(x, probs = 1e-03, type = 7L)
}
\arguments{
\item{x}{a numeric vector}

\item{probs}{numeric vector of probability value(s) in \code{[0,1]}}

\item{type}{an integer between 1 and 9 selecting one of the nine quantile
algorithms; see \code{\link{quantile}}}
}
\description{
Winsorize extreme values in a numeric vector.
}
\examples{
x <- c(92, 19, 101, 58, 1053, 91, 26, 78, 10, 13,
       -40, 101, 86, 85, 15, 89, 89, 28, -5, 41)
summary(x)

y <- winsorize(x, 0.1, type = 1L)
summary(y)

}
