% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{write_htmlTable}
\alias{write_htmlTable}
\title{Write an \code{htmlTable} to file}
\usage{
write_htmlTable(x, file = "", attributes = TRUE)
}
\arguments{
\item{x}{a string, e.g., the return of \code{\link[htmlTable]{htmlTable}}}

\item{file}{a character string naming the file to print to; \code{""},
the default, prints to the console}

\item{attributes}{logical; if \code{TRUE}, html attributes are added and
default border color names are replaced with hexadecimal values}
}
\description{
Write an \code{htmlTable} object to a file with optional html attribues.
}
\examples{
\dontrun{
library('htmlTable')
x <- htmlTable(head(cars))

write_htmlTable(x)
write_htmlTable(x, attributes = FALSE)
}

}

