% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{writeftable}
\alias{writeftable}
\title{Write \code{ftable}}
\usage{
writeftable(x, quote = FALSE, digits = getOption("digits"))
}
\arguments{
\item{x}{an object of class \code{ftable}}

\item{quote, digits}{see \code{\link{format.ftable}}}
}
\description{
Write an \code{\link{ftable}} as a matrix.
}
\examples{
## basic usage
writeftable(ftable(1))
writeftable(ftable(mtcars$gear))

## tables write nicer with names
writeftable(ftable(gear = mtcars$gear))
writeftable(ftable(Titanic, row.vars = 1:3))

}
