% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{bincon}
\alias{bincon}
\title{Binomial probability confidence intervals}
\usage{
bincon(r, n, alpha = 0.05, round = NULL, method = c("exact", "wilson",
  "asymptotic", "all"))
}
\arguments{
\item{r}{number of responses (successes)}

\item{n}{number of observations (trials)}

\item{alpha}{type-I error probability}

\item{round}{integer value specifying number of decimal places to round
(default is no rounding); see \code{\link{round}}}

\item{method}{character strings specifying which method to use; see details}
}
\value{
A matrix containing the computed interval(s) and their widths.
}
\description{
Calculates confidence intervals for binomial probabilities for specified
type-I error (\code{alpha}) using exact, Wilson, or asymptotic methods.
   
If \code{method = 'all'}, \code{r} and \code{n} should each be length 1.
The "exact" method uses the \code{\link{df}} distribution to comupte exact
intervals (based on the binomial cdf); the "wilson" interval is score-test-
based; the "asymptotic"  is the asymptotic normal interval.
}
\details{
The "wilson" method has been preferred by Agresti and Coull.
}
\examples{
bincon(0:10, 10)
bincon(5, 10, method = 'all')

}
\author{
Rollin Brant, Frank Harrell, and Brad Biggerstaff; modifications by Robert
Redd
}
\references{
Agresti, A. and B.A. Coull. Approximate is better than "exact"
for interval extimation of binomial proportions. \emph{American
Statistician}. \strong{59}:119-126, 1998.

Brown, L.D., T.T. Cai, and A. Das Gupta. Inverval estimation for
a binomial proportion (with discussion). \emph{Statistical Science}. 
\strong{16}:101-133, 2001.

Newcombe, R.G. Logit confidence intervals and the inverse sinh
transformation, \emph{American Statistician}. \strong{55}:200-202, 2001.
}
\seealso{
\code{\link[Hmisc]{binconf}}; \code{desmon::binci}
}

