% Generated by roxygen2 (4.0.1): do not edit by hand
\name{binconr}
\alias{binconr}
\title{binconr}
\usage{
binconr(r, n, conf = 0.95, digits = 2, est = TRUE, method = 'exact')
}
\arguments{
\item{r}{number of responses (successes)}

\item{n}{number of observations (trials)}

\item{conf}{level of confidence}

\item{digits}{number of digits}

\item{est}{logical; if \code{TRUE}, includes the point estimate}

\item{method}{method to use; see \code{\link{bincon}}}
}
\description{
Binomial confidence interval formatter
}
\examples{
binconr(5, 10, .90, est = FALSE)
binconr(45, 53, .95, digits = 1)
}
\seealso{
\code{\link{bincon}}; \code{\link[Hmisc]{binconf}}
}

