% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{bintest}
\alias{bintest}
\title{Single-stage designs}
\usage{
bintest(p0low, p0high = p0low, p1low, p1high = p1low, n.max,
  r = n.max, alpha = 0.1, beta = 0.1)
}
\arguments{
\item{p0low, p0high}{low and high values for p0}

\item{p1low, p1high}{low and high values for p1}

\item{n.max}{maximum sample size allowed (or feasible)}

\item{r}{cut-off value for responses expected; usually best to leave
\code{r = n.max} (default); useful if the maximum number of responses that
may occur is known}

\item{alpha, beta}{type-I and type-II errors}
}
\value{
A list with the following elements:

\item{\code{$designs}}{a matrix with a row giving a summary of each design
which meets the criteria. The columns are: \code{p0} and \code{p1}, the
null and alternative hypotheses, respectively; \code{n} and \code{r2}, the
overall sample size and minimum number of responses required to reject
\code{p0}; \code{type1} and \code{power}, the alpha and \code{1 - beta} for
the designs; and \code{signal}, the magnitude of difference between
\code{p0} and \code{p1}}
\item{\code{$call}}{the call to \code{bintest}}
\item{\code{$description}}{help text for \code{$designs}}
}
\description{
Function for sample sizes for single-stage designs in early clinical trial
phases (generally phases I and II) based on exact binomial method.

Cycles through possible designs constrained by \code{alpha}, \code{beta},
and \code{n.max} arguments for specified ranges of \code{p0} and \code{p1}.
}
\examples{
bintest(p0low = .2, p1low = .5, n.max = 25)

## example in sas macro
bintest(.1, .15, .2, .2, n.max = 80, alpha = .08, beta = .24)

}
\references{
Khan, Sarker, Hackshaw. Smaller sample sizes for phase II trials
based on exact tests with actual error rates by trading-off their nominal
levels of significance and power. \emph{British J of Cancer} (2012) 107,
1801-9.
}
\seealso{
\code{\link[clinfun]{ph2single}}; \code{desmon::bin1samp}; bintest,
\href{https://github.com/raredd/sascros/blob/master/bintest.sas}{SAS macro}
}
