% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bplot}
\alias{bplot}
\alias{tabler_bincon}
\title{Binomial confidence interval plot}
\usage{
bplot(data, varname, byvar, varname_label = names(varname),
  byvar_label = names(byvar), main = byvar_label, xlab = byvar_label,
  col = c("dodgerblue4", "dodgerblue2"), conf = 0.95, digits = 2L,
  show_overall = TRUE, show_missing = TRUE, alpha_missing = 1)

tabler_bincon(data, varname, byvar, varname_label = names(varname),
  byvar_label = names(byvar), main = byvar_label, xlab = byvar_label,
  col = c("dodgerblue4", "dodgerblue2"), conf = 0.95, digits = 2L,
  show_overall = TRUE, show_missing = TRUE, alpha_missing = 1)
}
\arguments{
\item{data}{a matrix or data frame with variables \code{varname} and
\code{byvar}}

\item{varname, byvar}{one or more variables in \code{data} to calculate
binomial confidence intervals by \code{byvar}; the rows variable(s) of
the table;

note that \code{varname} variables should be factor-like, and \code{byvar}
should be binary}

\item{varname_label, byvar_label}{optional labels for each \code{varname}
and \code{byvar}}

\item{main, xlab}{the x-axis and title labels}

\item{col}{a vector of colors for each \code{varname}; the first will be
used for the overall (if \code{show_overall = TRUE})}

\item{conf}{confidence level; passed to \code{\link{binconr}}}

\item{digits}{the number of places past the decimal to keep}

\item{show_overall, show_missing}{logical; if \code{TRUE}, rows with the
overall and missing value confidence intervals are shown, respectively}

\item{alpha_missing}{if \code{show_missing = TRUE}, the amount of alpha
transparency added to missing value rows; passed to \code{\link{tcol}}}
}
\description{
Wrapper function to create table of binomial confidence intervals and draw
as a forest plot.
}
\examples{
dat <- mtcars[sample(nrow(mtcars), 100L, TRUE), ]
dat[1, 2] <- NA
vv  <- c('cyl', 'vs', 'gear', 'carb')
dat$gear <- factor(dat$gear, 3:6)

bplot(
  dat, vv, 'am',
  col = c(0, seq_along(vv)) + 1L,
  main = 'Manual rate', xlab = 'Proportion of manual'
)

bplot(
  dat, setNames(vv, case(vv)), 'am',
  col = c('red', 'grey50', 'blue', 'grey50', 'blue'),
  conf = 0.9, alpha_missing = 0.4
)

}
