% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{catlist}
\alias{catlist}
\title{Concatenate list for output}
\usage{
catlist(l, sep = " = ", collapse = ", ")
}
\arguments{
\item{l}{a list to concatenate}

\item{sep, collapse}{passed to \code{\link{paste}} controlling the string
to separate name from value and list elements, respectively; if
\code{collapse} is a non character string, the result will be a vector
of strings}
}
\description{
Print a \code{list} (usually named) as a character vector or string.
}
\examples{
l <- list(a = 1:3, b = 2, '4')
catlist(l)
catlist(l, collapse = FALSE)
cat(catlist(par()[1:5], sep = ':\n  ', collapse = '\n'))

}
\seealso{
\code{\link{iprint}}
}
