% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classMethods}
\alias{classMethods}
\title{Show class methods}
\usage{
classMethods(class)
}
\arguments{
\item{class}{an object or classes as a vector of character strings}
}
\description{
Lists available methods for a given class.
}
\examples{
fit <- glm(vs ~ mpg, data = mtcars)
classMethods(fit)
classMethods(c('glm', 'lm'))

}
\references{
\url{https://gist.github.com/MrFlick/55ed854eb935e5c21f71}
}
\seealso{
\code{\link{methods}}, \code{\link{S3Methods}}, \code{\link{class}}
}
