% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classMethods}
\alias{classMethods}
\title{Show or get methods}
\usage{
classMethods(object, generic = NULL)
}
\arguments{
\item{object}{an object or character vector of classes}

\item{generic}{an S3 generic function like \code{plot} or \code{summary}}
}
\description{
List available methods for a given class or identify a specific method
when a \code{generic} is called with \code{object}.
}
\examples{
fit <- glm(vs ~ mpg, data = mtcars)
classMethods(fit)
classMethods(c('glm', 'lm'))

classMethods(1, plot)
classMethods(data.frame(1), plot)
classMethods(density(1:2), plot)

classMethods(1, print)
classMethods(ordered(1), print)
classMethods(mtcars, summary)

}
\references{
\url{https://gist.github.com/MrFlick/55ed854eb935e5c21f71};
\url{https://stackoverflow.com/q/42738851/2994949}
}
\seealso{
\code{\link{methods}}, \code{\link{S3Methods}}, \code{\link{class}}
}
