% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{clc}
\alias{clc}
\title{Clear workspace}
\usage{
clc(all.names = FALSE)
}
\arguments{
\item{all.names}{logical; if \code{TRUE}, also removes hidden (dot) objects}
}
\description{
Clear the workspace by removing all objects in \code{\link{ls}} followed
by \code{\link[=gc]{garbage collection}}.
}
\seealso{
\code{\link{clear}}
}
