% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils2.R
\name{clist}
\alias{clist}
\title{Concatenate a named list for output}
\usage{
clist(l)
}
\arguments{
\item{l}{list to concatenate}
}
\description{
Print a \code{list(a = 1, b = 2)} as \code{(a = 1, b = 2)}
}
\examples{
clist(list(a = 1, b = 2))
clist(par()[1:5])
}

