% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils2.R
\name{clist}
\alias{clist}
\title{Concatenate a named list for output}
\usage{
clist(l)
}
\arguments{
\item{l}{list to concatenate}
}
\description{
Print a \emph{named} \code{list} as a character string with values.
}
\examples{
(l <- list(a = 1, b = 2))

# $a
# [1] 1
#
# $b
# [1] 2

clist(l) # [1] "a = 1, b = 2"
clist(par()[1:5])
}

