% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_levels}
\alias{combine_levels}
\title{Combine factor levels}
\usage{
combine_levels(x, levels, labels)
}
\arguments{
\item{x}{a vector}

\item{levels}{a vector of unique values of \code{x} to combine; to combine
values into multiple groups, use a list}

\item{labels}{a vector of new labels; if \code{levels} is a vector,
\code{labels} should be length 1; if \code{levels} is a list, \code{labels}
(need not be a list but) should have one value for each list element of
\code{levels}}
}
\description{
Convenience function to combine multiple levels of a factor into a new or
existing level(s).
}
\examples{
x <- mtcars$gear
combine_levels(factor(x, ordered = TRUE), 3:4, 6)
combine_levels(x, list(3:4, 5), c(3, 6))

combine_levels(iris$Species, c('setosa', 'virginica'), 'setosa')
combine_levels(iris$Species, list(c('setosa', 'virginica'), 'versicolor'),
               c('Setosa/Virginica','Versicolor'))

}
\seealso{
\code{\link{recoder}}
}

