% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{combine_table}
\alias{combine_table}
\title{Combine html tables}
\usage{
combine_table(l, ...)
}
\arguments{
\item{l}{a list of matrices (preferred) or data frames}

\item{...}{additional arguments passed to \code{\link[htmlTable]{htmlTable}}}
}
\description{
Wrapper to easily combine a list of data frames or matrices into html
tables using the \pkg{htmlTable} package.
}
\examples{
library('htmlTable')
sp <- lapply(split(mtcars, rep(1:3, c(1,11,20))), as.matrix)

## basic table
combine_table(sp)

## adding more options
combine_table(sp, tspanner = c('one','eleven','twenty'),
  css.tspanner = 'text-align: center; color: red; font-style: italic;')

}

