% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils2.R
\name{countr}
\alias{countr}
\title{Count formatter}
\usage{
countr(top, n, lowcase = TRUE)
}
\arguments{
\item{top}{named vector of counts}

\item{n}{total number of observations}

\item{lowcase}{logical; names will be lowercase if \code{TRUE}, upper
case if \code{FALSE}, and unchanged otherwise (do not use \code{\link{NA}})}
}
\description{
Formats and prints a \emph{named} vector of counts with percentages.
}
\examples{
top <- setNames(3:1, c('gold','silver','bronze'))
countr(top, 10)
}

