% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{countr}
\alias{countr}
\title{Count formatter}
\usage{
countr(top, n, lowcase = TRUE, frac = FALSE)
}
\arguments{
\item{top}{named vector of counts (a summary or table) or a long vector of
character strings or factors}

\item{n}{total number of observations; if not given, the length of
\code{top} is used}

\item{lowcase}{logical; names will be lowercase if \code{TRUE}, upper
case if \code{FALSE}, and unchanged for any other value}

\item{frac}{logical; if \code{TRUE}, the numbers are shown as fractions}
}
\description{
Formats and prints a \emph{named} vector of counts with percentages.
}
\examples{
top <- setNames(3:1, c('Gold','Silver','Bronze'))
countr(names(top))
countr(names(top), lowcase = NA)
countr(names(top), frac = TRUE)


countr(top, 10)

}
