% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils2.R
\name{countr}
\alias{countr}
\title{Count formatter}
\usage{
countr(top, n)
}
\arguments{
\item{top}{named vector of counts}

\item{n}{total number of observations}
}
\description{
Formats and prints a \emph{named} vector of counts with percentages.
}
\examples{
top <- setNames(3:1, c('gold','silver','bronze'))
countr(top, 10)
}

