% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ctcae_v3}
\alias{ctcae_v3}
\title{Common Terminology Criteria for Adverse Events, v3}
\format{A data frame with 1059 rows and 4 variables:

\tabular{lll}{
\tab \code{tox_code} \tab alphanumeric toxicity code (\code{AB123}) \cr
\tab \code{tox_desc} \tab description \cr
\tab \code{tox_cat} \tab category \cr
\tab \code{ctc_version} \tab version number \cr
}}
\usage{
data(ctcae_v3)
}
\description{
A dataset containing the CTCAE (version 3) toxicity codes, descriptions, and
categories.
}
\keyword{datasets}

