% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rawr-package.R
\docType{data}
\name{ctcae_v4}
\alias{ctcae_v4}
\title{Common Terminology Criteria for Adverse Events, v4}
\format{A data frame with 790 rows and 5 variables:

\tabular{lll}{
\tab \code{tox_code} \tab alphanumeric toxicity code (\code{AB123}) \cr
\tab \code{tox_desc} \tab description \cr
\tab \code{tox_cat} \tab category \cr
\tab \code{ctc_version} \tab version number \cr
\tab \code{MedDRA_code_v12} \tab Medical Dictionary for Regulatory
Activities codes, v12 \cr
}}
\source{
\href{http://www.dfhcc.harvard.edu/fileadmin/DFHCC_Admin/Clinical_Trials/QACT/Policies_and_Procedures/CTCToxVersion4.pdf}{CTCAE v4}
}
\usage{
ctcae_v4
}
\description{
A dataset containing the CTCAE (version 4) toxicity codes, descriptions, and
categories.
}
\keyword{datasets}

