% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctcae_v4}
\alias{ctcae_v4}
\title{Common Terminology Criteria for Adverse Events, v4}
\format{A data frame with 790 rows and 5 variables:
\tabular{lll}{
\tab \code{tox_code} \tab alphanumeric toxicity code (\code{AB123}) \cr
\tab \code{tox_desc} \tab description \cr
\tab \code{tox_cat} \tab category \cr
\tab \code{ctc_version} \tab version number \cr
\tab \code{MedDRA_code_v12} \tab Medical Dictionary for Regulatory 
Activities codes, v12 \cr
}}
\source{
\url{https://ctep.cancer.gov/protocoldevelopment/electronic_applications/docs/ctcae_4_with_lay_terms.pdf}

\url{https://ctep.cancer.gov/protocoldevelopment/electronic_applications/ctc.htm}
}
\usage{
ctcae_v4
}
\description{
A dataset containing the CTCAE (version 4) toxicity codes, descriptions,
and categories.
}
\seealso{
\code{\link{ctcae_v3}}; \code{\link{match_ctc}}; \code{\link{tox_worst}}
}
\keyword{datasets}
