% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{cuzick.test}
\alias{cuzick.test}
\alias{cuzick.test.default}
\alias{cuzick.test.formula}
\title{Wilcoxon rank sum test for trend of ordered groups}
\usage{
cuzick.test(x, ...)

\method{cuzick.test}{default}(x, g, alternative = c("two.sided", "greater",
  "less"))

\method{cuzick.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a numeric vector of data values or a list of numeric data vectors;
non-numeric elements of a list will be coerced with a warning}

\item{...}{further arguments to be passed to or from methods}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}, ignored with a warning if \code{x} is a list}

\item{alternative}{a character string specifying the alternative hypothesis;
must be one of \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}}

\item{formula}{a formula of the form \code{response ~ group} where
\code{response} gives the data values and \code{group} a vector or factor
of the corresponding groups}

\item{data}{an optional matrix or data frame (or similar: see
\code{\link{model.frame}}) containing the variables in \code{formula}; by
default the variables are taken from \code{environment(formula)}}

\item{subset}{an optional vector specifying a subset of observations to be
used}

\item{na.action}{a function which indicates what should happen when the
data contain \code{NA}s; defaults to \code{getOption("na.action")}}
}
\value{
A list with class "\code{htest}" containing the following components:

\item{statistic}{the value of the test statistic with a name describing it}
\item{parameter}{the parameter(s) for the exact distribution of the test statistic}
\item{p.value}{the p-value for the test}
\item{alternative}{a character string describing the alternative hypothesis}
\item{method}{a character string describing the test used}
\item{data.name}{a character string giving the names of the data}
\item{estimate}{the median by group}
}
\description{
An implementation of Cuzick's extension of the Wilcoxon rank sum test for
trend in cases of three or more \emph{ordered} groups.
}
\examples{
## Altman (1991), 217
## ocular exposure to ultraviolet radiation for 32 pairs of sunglasses
## classified into three groups according to the amount of visible light
## transmitted

x <- list(c(1.4, 1.4, 1.4, 1.6, 2.3, 2.3),
          c(0.9, 1.0, 1.1, 1.1, 1.2, 1.2, 1.5, 1.9, 2.2, 2.6, 2.6, 2.6,
            2.8, 2.8, 3.2, 3.5, 4.3, 5.1),
          c(0.8, 1.7, 1.7, 1.7, 3.4, 7.1, 8.9, 13.5))

## equivalent ways to call cuzick.test
cuzick.test(x)
cuzick.test(unlist(x), rep(seq(x), lengths(x)))
cuzick.test(x ~ g, data.frame(x = unlist(x), g = rep(1:3, lengths(x))))


## Cuzick (1985), 87-90
x <- c(0, 0, 1, 1, 2, 2, 4, 9, 0, 0, 5, 7, 8, 11, 13, 23,
       25, 97, 2, 3, 6, 9, 10, 11, 11, 12, 21, 0, 3, 5, 6,
       10, 19, 56, 100, 132, 2, 4, 6, 6, 6, 7, 18, 39, 60)
g <- rep(1:5, c(8,10,9,9,9))
cuzick.test(x, g)

}
\references{
Altman, D. G. 1991. \emph{Practical Statistics for Medical Research}.
London: Chapman & Hall/CRC.

Cuzick, J. 1985. A Wilcoxon-type test for trend. \emph{Statistics in
Medicine} \strong{4}: 87–90.
}
\seealso{
\url{https://github.com/andrewejaffe/CuzicksTest},
\url{http://r.789695.n4.nabble.com/Cuzick-s-test-for-trend-td807101.html}
}

