% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacamp.R
\name{datacamp}
\alias{datacamp}
\title{Retrieve raw R code from a datacamp tutorial}
\usage{
datacamp(url, method, delimiter)
}
\arguments{
\item{url}{Link to datacamp tutorial}

\item{method}{Not all websites are formatted consistently. To overcome this, try a different
method by setting the method
parameter to integers 2 and greater to try other available methods}

\item{delimiter}{Specify what goes between the last character of one code block and the
first character of the next code block. Default is a two new lines, which appears
visually as one new line between blocks.}
}
\value{
A character vector of length 1 containing the R code from the target url.
}
\description{
Retrieve raw R code from a datacamp tutorial
}
\examples{
\donttest{
library(dplyr)
datacamp("https://www.datacamp.com/community/tutorials/sentiment-analysis-R")
datacamp("https://www.datacamp.com/community/tutorials/R-nlp-machine-learning")

# Same as above but provided to cat for easy viewing
datacamp("https://www.datacamp.com/community/tutorials/sentiment-analysis-R") \%>\%
  cat
datacamp("https://www.datacamp.com/community/tutorials/R-nlp-machine-learning") \%>\%
  cat
}

}
