% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacamp.R
\name{datacamp}
\alias{datacamp}
\title{Retrieve raw R code from a datacamp tutorial}
\usage{
datacamp(url)
}
\arguments{
\item{url}{Link to datacamp tutorial}
}
\description{
Retrieve raw R code from a datacamp tutorial
}
\examples{
library(dplyr)
datacamp("https://www.datacamp.com/community/tutorials/sentiment-analysis-R")
datacamp("https://www.datacamp.com/community/tutorials/R-nlp-machine-learning")

# Same as above but provided to cat for easy viewing
datacamp("https://www.datacamp.com/community/tutorials/sentiment-analysis-R") \%>\%
  cat
datacamp("https://www.datacamp.com/community/tutorials/R-nlp-machine-learning") \%>\%
  cat


}
