% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dlt.table}
\alias{dlt.table}
\title{DLT table}
\usage{
dlt.table(low, high, delta = 10)
}
\arguments{
\item{low}{lowest true DLT rate, percent}

\item{high}{highest true DLT rate, percent}

\item{delta}{interval of DLT rate sequence; default is 10}
}
\description{
Creates a standard dose-limiting toxicity table with probabilities of
dose-escalation.
}
\examples{
dlt.table(10, 50)
dlt.table(10, 50, 5)
}

