% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils2.R
\name{dmy}
\alias{dmy}
\title{Date parse}
\usage{
dmy(d, m, y, origin = 1900)
}
\arguments{
\item{d,m,y}{day, month, year as single integers or vectors}

\item{origin}{the origin for two-digit years; ignored otherwise}
}
\description{
Parses day, month, year columns to the standard date format. Assumes two-
digit years are 1900s dates; converts \code{NA} days, months, and years
to \code{01}, \code{01}, and \code{1900}, respectively.
}
\examples{
dmy(25, 7, 13, origin = 2000)

set.seed(1)
dd <- data.frame(id = 1:10,
                 day = sample(1:31, 10),
                 month = sample(1:12, 10),
                 year = sample(1000:2500, 10))

cbind(dd, dt = with(dd, dmy(day, month, year)))

dd1 <- within(dd, {
  day[sample(1:10, 5)] <- NA
  month[sample(1:10, 2)] <- NA
  year[sample(1:10, 2)] <- NA
})

cbind(dd1, dt = with(dd1, dmy(day, month, year)))
}

