% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{dodge}
\alias{dodge}
\alias{dodge.default}
\alias{dodge.formula}
\title{Point dodge}
\usage{
dodge(x, ...)

\method{dodge}{formula}(formula, data = NULL, ...)

\method{dodge}{default}(x, at, dist, jit, ...)
}
\arguments{
\item{x}{a numeric vector of data}

\item{...}{ignored}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ group}, where
\code{y} is a numeric vector of data values to be split into groups
according to the grouping variable, \code{group}}

\item{data}{optional matrix or data frame containing the variables in
\code{formula}; by default, variables are taken from
\code{environment(formula)}}

\item{at}{grouping variables or, equivalently, positions along x-axis}

\item{jit,dist}{jittering parameters; \code{jit} describes the spread of
close points, and \code{dist} defines a range of data to declare points
"close"}
}
\description{
Dodge and center overlapping points by group. Spreads scattered points
similar to \code{jitter} but symmetrically. Although the default method
can be used, it is recommended to use the formula method for ease of use
and to set useful defaults for \code{jit} and \code{dist}.
}
\examples{
## these are equivalent ways to call dodge:
dodge(mpg ~ gear + vs, mtcars)
with(mtcars, dodge(mpg, list(gear, vs)))
dodge(mtcars$mpg, mtcars[, c('gear', 'vs')])

## compare to overlapping points and jittering
sp <- split(mtcars$mpg, do.call(interaction, mtcars[, c('gear','vs')]))
plot.new()
plot.window(c(.5,6.5),c(10,35))
box()
invisible(sapply(seq_along(sp), function(x)
  points(rep(x, length(sp[[x]])), sp[[x]])))
invisible(sapply(seq_along(sp), function(x)
  points(jitter(rep(x, length(sp[[x]]))), sp[[x]], col = 4, pch = 2)))
points(dodge(mpg ~ gear + vs, data = mtcars), col = 2, pch = 4)
legend('topleft', pch = c(1,1,4), col = c(1,4,2),
       legend = c('overlapping','random jitter','dodging'))

## practical use
boxplot(mpg ~ vs + gear, data = mtcars)
points(dodge(mpg ~ vs + gear, data = mtcars), col = 'red', pch = 19)
}
\seealso{
\code{\link{jitter}}, \code{\link{tplot}}
}

