% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dsplot}
\alias{dsplot}
\alias{dsplot.default}
\alias{dsplot.formula}
\title{Discrete scatter plot}
\usage{
dsplot(x, ...)

\method{dsplot}{default}(
  x,
  y,
  ...,
  bg.col = TRUE,
  col = par("col"),
  xlab = NULL,
  ylab = NULL,
  pch = par("pch"),
  cex = par("cex")
)

\method{dsplot}{formula}(formula, data = NULL, ..., subset, na.action = NULL)
}
\arguments{
\item{x, y}{x- and y-axis variables}

\item{...}{for the \code{formula} method, named arguments to be passed to
the default method

for the default method, unnamed arguments are additional data vectors
(unless x is a list when they are ignored), and named arguments are
arguments and \code{\link{par}}s to be passed to \code{\link{plot}}}

\item{bg.col}{logical or color name to fill boxes with based on density; if
\code{FALSE}, gridlines are drawn to distinguish boxes; if \code{TRUE},
grayscale is used as the default color}

\item{col}{plotting color}

\item{xlab, ylab}{x- and y-axis labels}

\item{pch}{\strong{p}lotting \strong{ch}aracter}

\item{cex}{numerical value giving the amount by which plotting text and
symbols should be magnified relative to the default; this starts as 1
when a device is opened and is reset when the layout is changed, e.g.,
by setting \code{mfrow}}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ group}, where y
is a numeric vector of data values to be split into groups according to the
grouping variable \code{group} (usually a factor)}

\item{data}{a data frame (or list) from which the variables in formula
should be taken}

\item{subset}{an optional vector specifying a subset of observations to be
used for plotting}

\item{na.action}{a function which indicates what should happen when the data
contain \code{\link{NA}}s; the default is to ignore missing values in either
the response or the group}
}
\value{
A table (invisibly) corresponding to the plot cell counts.
}
\description{
This creates a scatter plot (sort of) for discrete, bivariate data; an
alternative to sunflower plots for integer-valued variables.
}
\examples{
set.seed(1)
x <- round(rnorm(400, 100, 4))
y <- round(rnorm(400, 200, 4))
sex <- sample(c('Female', 'Male'), 400, replace = TRUE)

dsplot(x, y, col = (sex \%in\% 'Female') + 1L, bg.col = FALSE)
dsplot(x, y, col = (sex \%in\% 'Female') + 1L, bg.col = 'tomato')

dsplot(y ~ x, pch = 19, col = (sex \%in\% 'Female') + 1L,
       cex = 0.6, bty = 'l', bg.col = 'tomato',
       xlab = 'measurement 1', ylab = 'measurement 2')
legend('bottomright', pch = 19, col = 1:2, bty = 'n',
       legend = c('Male', 'Female'))
       
}
\references{
\href{http://biostat.mc.vanderbilt.edu/wiki/Main/TatsukiRcode}{Tatsuki
\code{dsplot}}
}
