% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fapply}
\alias{fapply}
\title{Apply list of functions over list or vector}
\usage{
fapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link{expression}} object;
other objects (including classed objects) will be coerced by
\code{base::\link{as.list}}.}

\item{FUN}{a list of functions to be applied to each element of \code{X};
see \code{\link{lapply}}:"Details"}

\item{...}{additional arguments passed to \code{FUN}; note that these must
be defined arguments for each function in \code{FUN}; otherwise, you will
need to define a function before passing to \code{FUN}; see examples}
}
\value{
A data frame where \code{nrow} equals the length of \code{X} and
\code{ncol} equals the length of \code{FUN}.
}
\description{
A simple modification to the \code{*apply} functions which allows a list of
functions to be passed simultaneously.
}
\examples{
tmp <- recoder(mtcars, 6, NA)
fapply(tmp, list(mean, median))
fapply(tmp, list(mean, median), na.rm = TRUE)

## define a new function
`95\% CI` <- function(x)
     paste0('(', paste0(quantile(x, c(.025, .975)), collapse = ', '), ')')
fapply(mtcars, list(median, `95\% CI`))

## compare:
t(fapply(mtcars, list(min, mean, max, length)))
summary(mtcars)
}
\seealso{
\code{\link{sapply}}
}

