% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{flatten}
\alias{flatten}
\title{Flatten lists}
\usage{
flatten(x)
}
\arguments{
\item{x}{a list}
}
\description{
Flattens lists and nested lists of vectors, matrices, and/or data frames.
}
\examples{
l <- list(
  matrix(1:3), list(1:3, 'foo'), TRUE, 'hi',
  list(mtcars[1:5, 1:5], list(mtcars[1:5, 1:5]))
)

l
flatten(l)

}
\references{
\url{https://stackoverflow.com/q/8139677/2994949}
}
