% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{get_margs}
\alias{get_margs}
\title{Get arguments from \code{SAS} macros}
\usage{
get_margs(path, name, text)
}
\arguments{
\item{path}{character string of path to \code{.sas} file}

\item{name}{macro name in \code{path} of interest; if missing, returns
all macros found}

\item{text}{(optional) character string(s) of macros given instead of
\code{path}}
}
\value{
A list with macro names in \code{path} and their respective parameters.
}
\description{
Reads a text file (usually \code{.sas}) and extracts \code{SAS} macro names
and parameters with any default values; see tests in examples below.
}
\examples{
get_margs(text = '\%macro macro(a = 1, b = 2); \%mend;')

## $macro
## [1] "a=1, b=2"

path <- system.file('testfiles', package = 'rawr')
get_margs(file.path(path, 'macros.sas'))

## $macro1
## [1] "arg1, arg2"
##
## $macro2
## [1] "arg1=1, arg2=2, arg3=3"
##
## $macro3
## [1] "arg1=, arg2="
##
## $macro4
## [1] "this=, macro=, has=, many=, params=, on=, multiple=, lines="
##
## $macro5
## [1] ""this=, macro=, has=, comments=, between=, each=, parameter="

get_margs(file.path(path, 'macros.sas'), 'macro1')

## $macro1
## [1] "arg1, arg2"

\dontrun{
get_margs(file.path(path, 'macros.sas'), 'no_macro_with_this_name')

Error in get_margs(file.path(path, "macros.sas"), "no_macro_with_this_name") :
 no_macro_with_this_name macro not found in ...

get_margs(file.path(path, 'nomacro.sas'))

## Error in get_margs(file.path(path, 'nomacro.sas')) :
##   no valid macros found in ...

get_margs(file.path(path, 'onemacro.sas'), c('macro2', 'macro5'))

## Error in get_margs(file.path(path, 'onemacro.sas'), c("macro2", "macro5")) :
## macro2, macro5 not found in ...
}

}
\seealso{
\code{\link{sas_path}}, \code{\link{rmacro}}, \code{\link{r2sas}},
\code{\link{sas_mget}}, \code{\link{source_sas}}, \code{\link{parse_formats}}
}

