% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_margs}
\alias{get_margs}
\title{Get arguments from SAS macros}
\usage{
get_margs(mpath, mname)
}
\arguments{
\item{mpath}{character string of path to \code{.sas} file}

\item{mname}{macro name in \code{mpath} of interest; if missing, returns
all macros found}
}
\value{
A list with macro names in \code{mpath} and their respective parameters.
}
\description{
Reads a text file (usually \code{.sas}) and extracts \code{SAS} macro names
and parameters with any default values; see tests in examples below.
}
\examples{
\dontrun{
## tests

get_margs('./tests/testfiles/macros.sas')

## $macro1
## [1] "arg1, arg2"
##
## $macro2
## [1] "arg1=1, arg2=2, arg3=3"
##
## $macro3
## [1] "arg1=, arg2="
##
## $macro4
## [1] "this=, macro=, has=, many=, params=, on=, multiple=, lines="
##
## $macro5
## [1] ""this=, macro=, has=, comments=, between=, each=, parameter="

get_margs('./tests/testfiles/macros.sas', 'macro1')

## $macro1
## [1] "arg1, arg2"

get_margs('./tests/testfiles/macros.sas', 'no_macro_with_this_name')

Error in get_margs("./tests/testfiles/macros.sas", "no_macro_with_this_name") :
 no_macro_with_this_name macro not found in ./tests/testfiles/macros.sas

get_margs('./tests/testfiles/nomacro.sas')

## Error in get_margs("./tests/testfiles/nomacro.sas") :
##   no valid macros found in ./tests/testfiles/nomacro.sas

get_margs('./tests/testfiles/onemacro.sas', c('macro2', 'macro5'))

## Error in get_margs("./tests/testfiles/onemacro.sas", c("macro2", "macro5")) :
## macro2, macro5 not found in ./tests/testfiles/onemacro.sas
}
}
\seealso{
\code{\link{rmacro}}, \code{\link{r2sas}}, \code{\link{sas.mget}},
\code{\link{source_sas}}
}

