% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{ggcols}
\alias{ggcols}
\title{ggplot colors}
\usage{
ggcols(n, l = 65, c = 100)
}
\arguments{
\item{n}{number of colors}

\item{l}{a value in the range \code{[0, 100]} giving the luminance of the 
color; for a given combination of hue and chroma, only a subset of this 
range is possible}

\item{c}{the chroma of the color; the upper bound for chroma depends on hue
and luminance}
}
\description{
A function to replicate default \code{\link[ggplot2]{ggplot}} colors.
}
\examples{
plot(rnorm(1000), col = ggcols(1000), pch = 19)

}
\seealso{
\code{\link{hcl}}
}

