% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github}
\alias{github}
\title{Retrieve raw R code from an .R file hosted on github website}
\usage{
github(url)
}
\arguments{
\item{url}{Link to an R file on github website}
}
\description{
Retrieve raw R code from an .R file hosted on github website
}
\examples{
library(dplyr)
github("https://github.com/hadley/vis-eda/blob/master/travel.R")

# Same as above but provided to cat for easy viewing
github("https://github.com/hadley/vis-eda/blob/master/travel.R") \%>\%
  cat


}
