% Generated by roxygen2 (4.0.2): do not edit by hand
\name{helpExtract}
\alias{helpExtract}
\title{Extract R help files}
\usage{
helpExtract(f, show.sections = FALSE, section = "Usage", type = "m_code",
  ...)
}
\arguments{
\item{f}{a function}

\item{show.sections}{logical; if \code{TRUE}, returns \code{section} options
for \code{f}}

\item{section}{section to extract (default is \code{"Usage"}}

\item{type}{type of character vector you want returned; deefaults is
\code{"m_code"}, see details}

\item{...}{additional arguments passed to \code{utils:::.getHelpFile}}
}
\value{
A character vector to be used in a Sweave or R-markdown document.
}
\description{
Extracts specified portions of R help files (from \emph{loaded} libraries)
for use in Sweave or R-markdown documents.
}
\details{
The \code{type} argument accepts:
\itemize{
\item \code{"m_code"}: Markdown code chunks; for use with markdown documents
when highlighted code is expected.
\item \code{"m_text"}: Markdown plain text; for use with markdown documents
where regular text is expected.
\item \code{"s_code"}: Sweave code chunks; for use with Sweave documents
where highlighted code is expected.
\item \code{"s_text"}: Sweave plain text; for use with Sweave documents
where regular text is expected.
}

To see the results in the console:

\code{cat(helpExtract(print, type = 'm_text'))}

To insert a (highlighted) chunk into a markdown document:

\verb{
```{r, results='asis'}
cat(helpExtract(print), sep ='\n')
```
}

To insert a (highlighted) chunk into a Sweave document:

\verb{
\\Sexpr{knit_child(textConnection(helpExtract(print, type = 's_code')),
     options = list(tidy = FALSE, eval = FALSE))}
}
}
\examples{
cat(helpExtract(print), sep = "\\n")

cat(helpExtract(print, type = 'm_text'))

cat(helpExtract(print, type = 'm_text', section = 'description'))
}
\author{
Ananda Mahto
}

