% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{ht}
\alias{ht}
\title{\code{head}/\code{tail}}
\usage{
ht(x, n = 6L, sep = NULL)
}
\arguments{
\item{x}{an object}

\item{n}{an integer giving the first and last \code{n / 2} elements if
positive or the middle \code{n} elements if negative}

\item{sep}{separator}
}
\description{
\code{\link{rbind}} the \code{\link{head}} and \code{\link{tail}} of an
object.
}
\examples{
ht(letters, 6, '...')
ht(letters, -6)

ht(mtcars)
ht(mtcars, sep = '...')
ht(mtcars, -6, sep = '...')

}

