% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ht}
\alias{ht}
\title{head/tail}
\usage{
ht(x, ..., sep = NULL)
}
\arguments{
\item{x}{an object}

\item{...}{other parameters, such as \code{n}, passed to \code{\link{head}}
or other methods}

\item{sep}{separator}
}
\description{
head/tail
}

