% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{ht}
\alias{ht}
\title{head/tail}
\usage{
ht(x, ..., sep = NULL)
}
\arguments{
\item{x}{an object}

\item{...}{other parameters, such as \code{n}, passed to \code{head},
\code{tail}, or other methods}

\item{sep}{row separator}
}
\description{
\code{\link{rbind}} the \code{\link{head}} and \code{\link{tail}} of an
object.
}
\examples{
ht(mtcars)
ht(mtcars, sep = '...')

}

