% Generated by roxygen2 (4.0.2): do not edit by hand
\name{html.test}
\alias{html.test}
\title{html test}
\usage{
html.test(x)
}
\arguments{
\item{x}{character string of html code}
}
\description{
Render html in rstudio viewer
}
\examples{
\dontrun{
html.test("
<div align = center><h1>A heading<sup>&dagger;</sup><h1></div>
<font size = 1><sup>&dagger;</sup>That was the heading</font>
")

library(Gmisc)
html.test(htmlTable(mtcars, output = FALSE))
}
}

