% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{imgpal}
\alias{imgpal}
\title{Image palettes}
\usage{
imgpal(path, n = 10L, options = "")
}
\arguments{
\item{path}{full file path to image}

\item{n}{maximum number of colors to extract, result will be <= \code{n},
and the calculated number of unique colors will also be provided}

\item{options}{a (optional) character string of additional options passed
to \href{https://www.imagemagick.org/script/command-line-options.php}{\code{magick}}}
}
\value{
A list of class \code{"imgpal"} with the following elements:
\item{filename}{the image file name}
\item{n_unique}{the calculated number of unique colors}
\item{col}{a vector of colors (does not return transparent or white colors)}
\item{counts}{frequency counts for each \code{col}}
\item{call}{the call made to \code{magick}}
\item{magick}{the result of \code{call}}
}
\description{
Extract unique and most commonly-used unique colors from an image file
(requires \href{https://imagemagick.org/index.php}{ImageMagick}).
}
\examples{
go <- 'https://www.google.com/images/branding/googlelogo/2x/googlelogo_color_272x92dp.png'
ip <- imgpal(go)
show_pal(ip, n = 4)

ip <- imgpal(go, options = '-colorize 0,0,50')
show_pal(ip)

\dontrun{
## rawr::rawr_palettes
img <- system.file('fig', package = 'rawr')
img <- list.files(img, full.names = TRUE, pattern = 'g$')
op <- par(mfrow = n2mfrow(length(img)))
sapply(img, function(x) show_pal(imgpal(x), fullrange = TRUE))
par(op)
}

}
\seealso{
\code{\link{show_pal}}; \pkg{\code{magick}} package
}
