% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{inject_div}
\alias{inject_div}
\title{Inject div}
\usage{
inject_div(x, where, style = "")
}
\arguments{
\item{x}{a matrix or data frame}

\item{where}{an \code{nx2} matrix of row and column indices or vector (of
the form c(row, col, row, col, ...)) specifying which cells to select}

\item{style}{vector of character string(s) applied to each cell, recycled
if necessary}
}
\description{
Inject an html division tag with style attribute.
}
\examples{
library('htmlTable')
htmlTable(inject_div(head(cars), c(2,2), style = 'border: dashed 1px;'))

htmlTable(inject_div(head(cars), rbind(c(2,2), c(2,1), c(5,2)),
                     style = 'background-color: yellow;'))

htmlTable(inject_div(head(cars), c(2,2,2,1,5,2),
                     style = c('background-color: red; color: white;',
                               'border: solid 1px;',
                               'font-weight: 900; color: blue;')))

}
\seealso{
\code{\link[htmlTable]{htmlTable}}
}

