% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{inside}
\alias{\%inside\%}
\alias{inside}
\title{Inside}
\usage{
x \%inside\% interval
}
\arguments{
\item{x}{numeric}

\item{interval}{numeric interval}
}
\description{
Return logical vector indicating if \code{x} is inside the \code{interval}
}
\examples{
-5:5 \%inside\% c(0,5)
}

