% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{interleave}
\alias{interleave}
\title{Interleave rows or columns}
\usage{
interleave(..., which)
}
\arguments{
\item{...}{vectors, matrices, or data frames}

\item{which}{joining method to use (\code{'rbind'} or \code{'cbind'}) when
\code{...} are matrices or data frames}
}
\description{
Interleave rows (or columns) of vectors, matrices, or data frames.
}
\examples{
interleave(letters[1:3], LETTERS[3:1], letters[26:24])
interleave(t(matrix(1:9, 3, 3)), t(matrix(letters[1:9], 3, 3)), which = 'rbind')
interleave(matrix(1:9, 3, 3), matrix(letters[1:9], 3, 3), which = 'cbind')
}

