% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{intr}
\alias{intr}
\alias{intr2}
\title{Interval formatter}
\usage{
intr(..., fun = median, conf = NULL, digits = 0L, na.rm = FALSE)

intr2(
  ...,
  fun = median,
  na.rm = FALSE,
  digits = 0L,
  conf.int = 0.95,
  range = TRUE,
  ci = FALSE,
  sd = FALSE,
  mad = FALSE,
  iqr = FALSE
)
}
\arguments{
\item{...}{numeric vector or string of numeric vectors}

\item{fun}{summary stat function, usually \code{\link{mean}} or
\code{\link{median}}}

\item{conf}{width of confidence interval in \code{[0,1]}; if \code{NULL}
(default), returns min and max of \code{...}}

\item{digits}{number of digits (includes trailing 0s)}

\item{na.rm}{logical; if \code{TRUE}, any \code{\link{NA}} and \code{NaN}
are removed from \code{...} before \code{FUN} and \code{\link{quantile}}
are computed}

\item{conf.int}{for \code{intr2}, the confidence level if
\code{conf = TRUE}; default is \code{0.95}}

\item{range, ci, sd, mad, iqr}{logical; for \code{intr2} if \code{TRUE},
the range (default), confidence interval, standard deviation, median
absolute deviation, and/or inner quartile range are shown}
}
\description{
Calculate summary statistic with range or confidence interval.
}
\examples{
x <- 1:10
intr(x)
intr(x, conf = 0.95)

## inner quartile range
cbind(
  'median (IQR)' = lapply(mtcars, intr, conf = 0.5, digits = 2),
  'mean (range)' = lapply(mtcars, intr, fun = mean, digits = 2)
)

## compare to
summary(mtcars)

x <- mtcars$mpg
intr2(x)
intr2(x, fun = mean, sd = TRUE, ci = TRUE, conf.int = 0.9)

## compare
x <- list(mtcars$mpg, mtcars$wt)
intr2(x, range = TRUE, iqr = TRUE, digits = 2)
summary(mtcars[, c('mpg', 'wt')])

}
\seealso{
\code{\link[rawr]{roundr}}
}
