% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{justify}
\alias{justify}
\title{Justify text}
\usage{
justify(
  string,
  width = getOption("width") - 10L,
  fill = c("random", "right", "left")
)
}
\arguments{
\item{string}{a character string}

\item{width}{desired width text in characters given as a positive integer}

\item{fill}{method of adding whitespace, i.e., by starting with the
\code{"right"}- or \code{"left"}-most whitespace or \code{"random"}}
}
\description{
Add whitespace to (monospaced) text for justified or block-style spacing.
}
\examples{
x <- paste(rownames(mtcars), collapse = ' ')
cat(justify(x))

## slight differences in whitespace for fill methods
op <- par(xpd = NA, family = 'mono', cex = 0.8)
plot(0, ann = FALSE, axes = FALSE, type = 'n')
text(1, 1, justify(x, fill = 'random'))
text(1, 0, justify(x, fill = 'right'), col = 2)
text(1, -1, justify(x, fill = 'left'), col = 3)
par(op)

}
\references{
Adapted from \url{https://stackoverflow.com/q/34710597/2994949}
}
\seealso{
\code{\link{strwrap}}
}
