% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle.R
\name{kaggle}
\alias{kaggle}
\title{Retrieve raw R code from a kaggle notebook}
\usage{
kaggle(url)
}
\arguments{
\item{url}{Link to a kaggle notebook}
}
\description{
Retrieve raw R code from a kaggle notebook
}
\examples{
\dontrun{
kaggle("https://www.kaggle.com/vrtjso/mercari-eda-more-info-than-you-can-imagine")
kaggle("https://www.kaggle.com/captcalculator/a-very-extensive-mercari-exploratory-analysis")
kaggle("https://www.kaggle.com/adityaecdrid/mnist-with-keras-for-beginners-99457")

# Same as above but provided to cat for easy viewing
R
kaggle("https://www.kaggle.com/vrtjso/mercari-eda-more-info-than-you-can-imagine")  \%>\%
  cat

rmarkdown
kaggle("https://www.kaggle.com/captcalculator/a-very-extensive-mercari-exploratory-analysis") \%>\%
  cat

python
kaggle("https://www.kaggle.com/adityaecdrid/mnist-with-keras-for-beginners-99457") \%>\%
  cat
}
}
