% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot}
\alias{kmplot}
\title{Survival curves}
\usage{
kmplot(s, lty.surv = 1, lwd.surv = 1, col.surv = col.surv, mark = 3,
  lwd.mark = lwd.surv, lty.ci = 0, lwd.ci = 1, col.ci = col.surv,
  col.band = FALSE, atrisk = TRUE, atrisk.lab = "Number at risk",
  atrisk.lines = TRUE, atrisk.col = !atrisk.lines, strata.lab = NULL,
  strata.expr = NULL, strata.order = seq_along(s$n), extra.margin = 5,
  median = FALSE, xaxs = "s", xlim = c(0, max(s$time)), ylim = c(0, 1),
  xaxis.at = pretty(c(0, s$time)), xaxis.lab = xaxis.at,
  yaxis.at = pretty(ylim), yaxis.lab = yaxis.at, xlab = "Time",
  ylab = "Survival probability", main = "", cex.axis = par("cex.axis"),
  legend = !atrisk && !is.null(s$strata), lr_test = FALSE, mar = NULL,
  add = FALSE, panel.first = NULL, panel.last = NULL, ...)
}
\arguments{
\item{s}{an object of class \code{\link{survfit}} or \code{survfit.cox}}

\item{lty.surv, lwd.surv, col.surv}{line type, width, and color for survival
curve(s); colors may be either numeric, color names as character string(s),
or hexadecimal string(s)}

\item{mark, lwd.mark}{numeric plotting character (\code{\link{pch}}) or
character string, e.g., \code{''}, \code{'|'}; if \code{mark = 'bump'}, a
mark will be drawn only above the curve; \code{lwd.mark} controls the line
width when a \code{pch} or \code{"bump"} is used}

\item{lty.ci, lwd.ci, col.ci}{line type, width, and color for confidence
interval(s); not plotted (i.e., \code{= 0}) by default}

\item{col.band}{color for confidence band(s); either as numeric, color
string(s), or hexadecimal string(s); if \code{TRUE}, \code{col.surv}
values will be used; if \code{FALSE}, \code{NA}, or \code{NULL}, no
bands will be plotted; also note that this is not a true confidence band;
see details}

\item{atrisk}{logical; if \code{TRUE} (default), draws at-risk table}

\item{atrisk.lab}{heading for at-risk table}

\item{atrisk.lines}{logical; draw lines next to strata in at-risk table}

\item{atrisk.col}{logical or a vector with colors for at-risk table text;
if \code{TRUE}, \code{col.surv} will be used}

\item{strata.lab}{labels used in legend and at-risk table for strata; if
\code{NULL} (default), labels created in \code{survfit} are used; if only
one strata is present, "All" is used by default; if \code{FALSE}, labels
are not used; if \code{TRUE}, labels will be stripped of variable names}

\item{strata.expr}{an alternative to \code{strata.lab} which allows for
\code{\link{bquote}} or \code{\link{expression}} to be passed to labels
for at-risk table; note that \code{strata.expr} trumps \code{strata.lab}}

\item{strata.order}{order of strata in legend and at-risk table}

\item{extra.margin}{increase left margin when strata labels in at-risk
table are long}

\item{median}{logical; if \code{TRUE}, median and confidence interval for
each curve is added to at-risk table}

\item{xaxs}{style of axis; see details or \code{\link{par}}}

\item{xlim, ylim}{x- and y-axis limits}

\item{xaxis.at, yaxis.at}{positions for x- and y-axis labels and ticks}

\item{xaxis.lab, yaxis.lab}{x- and y-axis tick labels}

\item{xlab, ylab}{x- and y-axis labels}

\item{main}{title of plot}

\item{cex.axis}{text size for axes labels, legend, at-risk table}

\item{legend}{logical, a vector of x/y coordinates, or a keyword (see
\code{\link{legend}}); if \code{TRUE}, the default position is
\code{"bottomleft"}}

\item{lr_test}{logical or numeric; if \code{TRUE}, a log-rank test will be
performed and the results added to the top-right corner of the plot; if
numeric, the value is passed as \code{rho} controlling the type of test
performed; see \code{\link{survdiff}}}

\item{mar}{margins; see \code{mar} section in \code{\link{par}}}

\item{add}{logical; if \code{TRUE}, \code{par} is not refreshed; allows for
multiple panels, e.g., when using \code{par(mfrow = c(1, 2))}}

\item{panel.first}{an expression to be evaluated after the plot axes are
set up but before any plotting takes place}

\item{panel.last}{an expression to be evaluated after plotting but before
returning from the function}

\item{...}{additional parameters (\code{font}, \code{mfrow}, \code{bty},
\code{tcl}, \code{cex.lab}, \code{xaxs}, etc) passed to \code{par}}
}
\description{
Plot Kaplan-Meier or Cox proportional hazards model with optional at-risk
table.
}
\details{
Line specifications (e.g., \code{lty.surv}, \code{lwd.surv}, etc) will be
recycled as needed.

If \code{col.band} is not \code{NULL}, \code{NA}, or \code{FALSE}, a
confidence band is plotted; however, this is not a confidence band in the
statistical sense, i.e., a xx-percent chance of containing the entire
population of the survival curve which are wider than the point-wise
confidence limits. Rather, it refers to a band of color plotted between
the confidence limits calculated in the survfit object. That is, the xx-\% 
confidence interval (plotted when \code{lty.ci != 0}) and the confidence
bands are identical--just two ways of plotting the same invervals.

\code{xaxs} is the style of the x-axis; see \code{\link{par}}. The default
for \code{kmplot} is \code{"S"} which is equivalent to \code{xaxs = "i"}
but with the maximum \code{xlim} value increased by 4\%. Other styles for
\code{xaxs} currently implemented in \code{R} are \code{"r"} (default for
plotting and the previous value for \code{kmplot}) and \code{"i"} which
will \emph{not} add padding to the ends of the axes.

When saving plots, it is highly recommended to use \code{\link{png}},
\code{\link{svg}}, \code{\link{pdf}}, etc. instead of exporting directly
from the \code{R} graphics device. Doing so may cause the at-risk table or
legend to be mis-aligned.
}
\examples{
library('survival')
km1 <- survfit(Surv(time, status) ~ sex, data = colon)
km2 <- survfit(Surv(time, status) ~ I(rx == "Obs") + adhere, data = colon)

## basic usage
kmplot(km1)
kmplot(km1, atrisk.col = c('grey50','tomato'), lr_test = TRUE)
kmplot(km1, mark = 'bump', atrisk.lines = FALSE, median = TRUE)
kmplot(km2, atrisk = FALSE, lwd.surv = 2, lwd.mark = .5,
       col.surv = 1:4, col.band = c(1,0,0,4))


## expressions in at-risk table (strata.expr takes precedence)
kmplot(km1, strata.lab = c('\\u2640', '\\u2642'))
kmplot(km1, strata.lab = c('\\u2640', '\\u2642'),
               strata.expr = expression(widetilde(ring(Female)),
                                        phantom() >= Male))


## when using mfrow options, use add = TRUE and same mar to align axes
mar <- c(8,6,2,2)
par(mfrow = c(1, 2))
kmplot(km1, add = TRUE, mar = mar)
kmplot(km2, add = TRUE, strata.lab = TRUE, mar = mar)


\dontrun{
pdf(tf <- tempfile(fileext = '.pdf'), height = 8, width = 11,
    pointsize = 12, family = 'serif')
kmplot(survfit(Surv(time, status) ~ rx + adhere, data = colon),
       legend = FALSE,
       panel.first = abline(v = c(0, .5, 1:9) * 365, lty = 3),
       mark = '',                         # no censor mark
       lty.ci = 2, lwd.ci = .3,           # dashed line for CIs
       xaxis.at = c(0, .5, 1:9) * 365,    # change days to years
       xaxis.lab = c(0, .5, 1:9),         # label years
       yaxis.lab = pretty(0:1) * 100,     # change to percent
       xlab = 'Time (years)', ylab = 'Percent survival',
       col.surv = c('blue','red','green','black','purple','orange'),
       col.ci   = c(0,0,0,0,'purple',0),  # CI only for one group
       extra.margin = 6,        # increase margin for long strata labels
       strata.lab = c('Obs','Obs+','Lev','Lev+','Lev5fu','Lev5fu+'),
       strata.order = c(5,6,3,1,4,2),      # order by survival estimates
       font = 2,                # bold table text
       bty = 'l',               # L box type around plot
       tcl = .5)                # change length/direction of ticks
title(main = 'Chemotherapy for stage B/C colon cancer',
      adj = .5, font.main = 1, line = 0.5, cex.main = 1)
dev.off()
file.show(tf)
}

}
\references{
\url{http://biostat.mc.vanderbilt.edu/wiki/Main/TatsukiRcode}
}
\seealso{
\code{\link{kmplot_by}}; \code{survival:::plot.survfit};
\code{\link[plotr]{ggsurv}}
}

