% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot_by}
\alias{kmplot_by}
\title{kmplot_by}
\usage{
kmplot_by(strata, event = "pfs", dat, by, lr_test = TRUE, ..., sub,
  strata_lab, fig_lab)
}
\arguments{
\item{strata}{character string of the strata variable}

\item{event}{character string indicating the event (pfs, os, ttp, etc);
see details}

\item{dat}{data frame to use}

\item{by}{optional character string of stratification variable}

\item{lr_test}{logical; if \code{TRUE}, a log-rank test will be performed
and the results added to the top-right corner of the plot}

\item{...}{additional arguments passed to \code{\link{kmplot}} or
graphical parameters subsequently passed to \code{\link{par}}}

\item{sub}{sub-title displayed in upper left corner; should be a character
vector with length equal to the number of panels (i.e., the number of
levels of \code{by} or length one if \code{by} was not given)}

\item{strata_lab}{at-risk table strata labels; should be a character vector
with length equal to the number of strata; otherwise, the variable labels
will be removed; see examples}

\item{fig_lab}{figure panel labels; should be a character vector with
length equal to the number of panels (i.e., the number of levels of
\code{by} or length one if \code{by} was not given)}
}
\value{
Invisibly returns a list of data frames stratified (optionally) by the
\code{by} variable.
}
\description{
This function helps create stratified \code{\link{kmplot}}s quickly with
panel labels and log-rank tests for the subsets.
}
\details{
The data used should have at least three variables: \code{strata},
\code{*_time}, and \code{*_ind} where \code{*} is \code{event}. For
example, to use progression-free survival, \code{dat} should have columns
\code{"pfs_time"} and \code{"pfs_ind"} and optionally the \code{strata}
column unless \code{strata = "1"}.
}
\examples{
library('survival')
data(colon)
colon <- within(colon[duplicated(colon$id), ], {
  pfs_time <- time
  pfs_ind <- status
  sex <- c('Female','Male')[sex + 1]
})

kmplot_by('1', dat = colon)
kmplot_by('sex', dat = colon, fig_lab = 'B',
  strata_lab = c('F','M'), sub = 'PFS, by sex')
kmplot_by('rx', dat = colon, fig_lab = '', col.surv = 1:3,
  strata_lab = FALSE, col.band = NA)
kmplot_by('rx + sex', dat = colon, strata_lab = '',
  lty.surv = 1:6, col.band = NA)
kmplot_by('rx', dat = colon, by = 'sex', col.surv = 1:3,
  strata_lab = c('Observation','Trt','Trt + 5-FU'))
}
\seealso{
\code{\link{kmplot}}, \code{\link{survdiff}}
}

