% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot_by}
\alias{kmplot_by}
\title{kmplot_by}
\usage{
kmplot_by(strata = "1", event, data, by = NULL, single = TRUE,
  lr_test = TRUE, main = NULL, ylab = NULL, sub = NULL,
  strata_lab = NULL, fig_lab = NULL, col.surv = NULL, map.col = FALSE,
  time = NULL, add = FALSE, plot = TRUE, ...)
}
\arguments{
\item{strata}{character string of the strata variable}

\item{event}{character string indicating the event (pfs, os, ttp, etc);
see details}

\item{data}{data frame to use}

\item{by}{optional character string of stratification variable}

\item{single}{logical; if \code{TRUE}, each level of \code{by} will be
drawn in a separate window}

\item{lr_test}{logical or numeric; if \code{TRUE}, a log-rank test will be
performed and the results added to the top-right corner of the plot; if
numeric, the value is passed as \code{rho} controlling the type of test
performed; see \code{\link{survdiff}}}

\item{main}{title of plot}

\item{ylab}{y-axis label}

\item{sub}{sub-title displayed in upper left corner; should be a character
vector with length equal to the number of panels (i.e., the number of
levels of \code{by} or length one if \code{by} was not given)}

\item{strata_lab}{at-risk table strata labels; should be a character vector
with length equal to the number of strata; \code{TRUE} (or equivalently
missing) is the default, and \code{FALSE} trims the labels; see examples}

\item{fig_lab}{figure panel labels; should be a character vector with
length equal to the number of panels (i.e., the number of levels of
\code{by} or length one if \code{by} was not given)}

\item{col.surv}{color for individual survival curves or for all curves in
a plot if \code{by} is given and \code{map.col = TRUE}; if \code{col.surv}
is a named vector which matches the at risk labels, then colors are mapped
to the corresponding strata; see \code{\link{kmplot}}}

\item{map.col}{logical; if \code{TRUE}, \code{col.surv} will be the color
of all curves in each plot (only used when \code{by} is non-missing)}

\item{time}{character string of the time variable (optional)}

\item{add}{logical; if \code{FALSE} (default), resets graphical parameters
to settings before \code{kmplot_by} was called; set to \code{TRUE} for
adding to existing plots}

\item{plot}{logical; if \code{FALSE}, no plot is created but a list with
\code{survfit}s is returned}

\item{...}{additional arguments passed to \code{\link{kmplot}} or
graphical parameters subsequently passed to \code{\link{par}}}
}
\value{
Invisibly returns a list of \code{\link{survfit}} object(s) used to generate
plot(s). If \code{by} was used, there will be a list element for each unique
value.
}
\description{
This function helps create stratified \code{\link{kmplot}}s quickly with
panel labels and log-rank tests for each plot.

\code{data} should have at least three variables: \code{strata},
\code{*_time}, and \code{*_ind} where \code{*} is \code{event}. For
example, to use progression-free survival, \code{data} should have columns
\code{"pfs_time"} and \code{"pfs_ind"} (in this case the user should use
\code{event = 'pfs'}) and optionally the \code{strata} column unless only
the null model (\code{strata = "1"} is needed (default).

Alternatively, the \code{time} argument may be used instead of following
the above; in this case, \code{time} and \code{event} must be variable
names in \code{data}. However, the method described above is more
efficient and preferred.
}
\examples{
library('survival')
kmplot_by(time = 'time', event = 'status', data = colon)


## create *_ind, *_time variables, see details
colon2 <- within(colon[duplicated(colon$id), ], {
  pfs_time <- time
  pfs_ind  <- status
  sex <- c('Female','Male')[sex + 1L]
  sex1 <- 'Male'
})

## these are equivalent (with minor differences in aesthetics)
kmplot_by(event = 'pfs', data = colon2)
kmplot_by('sex1', 'pfs', colon2)

kmplot_by('rx', 'pfs', data = colon2, col.surv = 1:3,
  strata_lab = FALSE, col.band = NA)


## return value is a list of survfit objects
l <- kmplot_by('sex', 'pfs', colon2, 'rx', plot = FALSE)
str(lapply(l, kmplot))
str(lapply(l, kmplot_by))

## multiple variables can be combined
kmplot_by('rx + sex', 'pfs', colon2, strata_lab = FALSE, lty.surv = 1:6)


## if "by" is given, default is to plot separately
kmplot_by('rx', 'pfs', colon2, by = 'sex', col.surv = 1:3,
  strata_lab = c('Observation','Trt','Trt + 5-FU'))
  
## if "by" is given, use map.col  to map colors to plots
kmplot_by('rx', 'pfs', colon2, by = 'rx', map.col = TRUE, single = FALSE)
kmplot_by('sex', 'pfs', colon2, by = 'rx', map.col = TRUE, single = FALSE)

## to keep colors mapped to the same strata across plots, use named vector
kmplot_by('rx', 'pfs', colon2, by = 'rx', xlim = c(0, 3000),
          col.surv = c(Lev = 'brown', Obs = 'blue', 'Lev+5FU' = 'purple'))

## if single = FALSE, uses n2mfrow function to set par('mfrow')
kmplot_by('rx', 'pfs', colon2, by = 'sex', col.surv = 1:3, single = FALSE,
  strata_lab = c('Observation','Trt','Trt + 5-FU'),
  main = levels(factor(colon2$sex)))
  
## if par('mfrow') is anything other than c(1,1), uses current setting
par(mfrow = c(2,2))
kmplot_by('rx', 'pfs', colon2, by = 'sex', col.surv = 1:3, single = FALSE,
  strata_lab = c('Observation','Trt','Trt + 5-FU'))

## use add = TRUE to add to a figure region without using the by argument
par(mfrow = c(1,2))
mar <- c(8,6,3,2)
kmplot_by( 'rx', 'pfs', colon2, strata_lab = FALSE, add = TRUE, mar = mar)
kmplot_by('sex', 'pfs', colon2, strata_lab = FALSE, add = TRUE, mar = mar)
  
}
\seealso{
\code{\link{kmplot}}, \code{\link{survdiff}}; \code{\link{kmplot_data_}};
\code{\link{lr_text}}; \code{\link{lr_pval}}; \code{\link{points.kmplot}}
}
