% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot_by}
\alias{kmplot_by}
\title{kmplot_by}
\usage{
kmplot_by(strata = "1", event = "pfs", data, by, single = TRUE,
  lr_test = TRUE, ylab, sub, strata_lab, fig_lab, time, add = FALSE, ...)
}
\arguments{
\item{strata}{character string of the strata variable}

\item{event}{character string indicating the event (pfs, os, ttp, etc);
see details}

\item{data}{data frame to use}

\item{by}{optional character string of stratification variable}

\item{single}{logical; if \code{TRUE}, each level of \code{by} will be
drawn in a separate window}

\item{lr_test}{logical; if \code{TRUE}, a log-rank test will be performed
and the results added to the top-right corner of the plot}

\item{ylab}{y-axis label}

\item{sub}{sub-title displayed in upper left corner; should be a character
vector with length equal to the number of panels (i.e., the number of
levels of \code{by} or length one if \code{by} was not given)}

\item{strata_lab}{at-risk table strata labels; should be a character vector
with length equal to the number of strata; otherwise, the variable labels
will be removed; see examples}

\item{fig_lab}{figure panel labels; should be a character vector with
length equal to the number of panels (i.e., the number of levels of
\code{by} or length one if \code{by} was not given)}

\item{time}{character string of the time variable (optional)}

\item{add}{logical; if \code{FALSE} (default), resets graphical parameters
to settings before \code{kmplot_by} was called; set to \code{TRUE} for
adding to existing plots}

\item{...}{additional arguments passed to \code{\link{kmplot}} or
graphical parameters subsequently passed to \code{\link{par}}}
}
\value{
Invisibly returns a list of \code{\link{survfit}} objects for each plot.
}
\description{
This function helps create stratified \code{\link{kmplot}}s quickly with
panel labels and log-rank tests for the subsets.
}
\details{
The data used should have at least three variables: \code{strata},
\code{*_time}, and \code{*_ind} where \code{*} is \code{event}. For
example, to use progression-free survival, \code{dat} should have columns
\code{"pfs_time"} and \code{"pfs_ind"} and optionally the \code{strata}
column unless \code{strata = "1"}.

Alternatively, the \code{time} argument may be used instead of following
the above; in this case, \code{time} and \code{event} must be variable
names in \code{data}.
}
\examples{
library('survival')
data(colon)

kmplot_by(time = 'time', event = 'status', data = colon)

colon <- within(colon[duplicated(colon$id), ], {
  pfs_time <- time
  pfs_ind <- status
  sex <- c('Female','Male')[sex + 1]
})

kmplot_by(data = colon)

## return value is a list of survfit objects
tmp <- kmplot_by('sex', data = colon, fig_lab = 'Figure I',
  strata_lab = c('F','M'), sub = 'PFS, by sex')
kmplot(tmp$sex)
  
kmplot_by('rx', data = colon, col.surv = 1:3,
  strata_lab = FALSE, col.band = NA)

## multiple variables can be combined  
kmplot_by('rx + sex', data = colon, strata_lab = '',
  lty.surv = 1:6, col.band = NA)

## if "by" is given, default is to plot separately
kmplot_by('rx', data = colon, by = 'sex', col.surv = 1:3,
  strata_lab = c('Observation','Trt','Trt + 5-FU'))

## if single = FALSE, uses n2mfrow function to set par('mfrow')
kmplot_by('rx', data = colon, by = 'sex', col.surv = 1:3, single = FALSE,
  strata_lab = c('Observation','Trt','Trt + 5-FU'))
  
## if par('mfrow') is anything other than c(1,1), uses current setting
par(mfrow = c(2,2))
kmplot_by('rx', data = colon, by = 'sex', col.surv = 1:3, single = FALSE,
  strata_lab = c('Observation','Trt','Trt + 5-FU'))

## use add = TRUE to add to a figure region without using the by argument
par(mfrow = c(1,2))
kmplot_by('rx', data = colon, col.surv = 1:3, add = TRUE)
kmplot_by('sex', data = colon, col.surv = 1:3, add = TRUE)
  
}
\seealso{
\code{\link{kmplot}}, \code{\link{survdiff}}
}

