% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{kmplot_data_}
\alias{kmplot_data_}
\title{Create data frame to plot survival data}
\usage{
kmplot_data_(s, strata.lab)
}
\arguments{
\item{s}{a \code{\link{survfit}} object}

\item{strata.lab}{character vector of strata labels; must be same length
as s$n}
}
\description{
Create data frame to plot survival data
}
\seealso{
\code{\link{kmplot}}; \code{\link{kmplot_by}}
}
