% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{labeller}
\alias{labeller}
\title{Add labels to a plot}
\usage{
labeller(
  fig = NULL,
  sub = NULL,
  side = 3L,
  pad = NULL,
  col.text = par("col"),
  cex = par("cex"),
  col.bg = par("bg"),
  col.border = par("col.axis"),
  lty = par("lty"),
  lwd = par("lwd"),
  ...
)
}
\arguments{
\item{fig, sub}{labels for the upper left corner and plot, respectively; note
that \code{side} controls where \code{sub} is placed}

\item{side}{the side of the plot to draw the \code{sub} label}

\item{pad}{(optional) height of the sub label space as a fraction of the
length of the axis perpendicular to \code{side}}

\item{col.text, cex}{the color and character expansion value for text}

\item{col.bg, col.border, lty, lwd}{the background color, border color, border
line type, and border line width of the \code{sub} rectangle}

\item{...}{additional arguments passed to \code{\link{text}} or graphical
parameters passed further to \code{\link{par}}}
}
\description{
Adds a figure or sub label to an existing plot.
}
\examples{
op <- par(mar = c(5, 5, 5, 5))
plot(1)
labeller(fig = 'A', sub = 'plot label')
labeller(sub = 'plot label', side = 4)

par(mfrow = c(2, 3), mar = c(5, 5, 2, 2))
sapply(1:6, function(ii) {
  plot(1)
  fig <- names(plotr::fig(ii, par('mfrow')))
  labeller(LETTERS[ii], sprintf('Panel \%s', ii), side = 3L + grepl('right', fig))
})
par(op)

}
