% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{landmark}
\alias{landmark}
\title{Landmark}
\usage{
landmark(s, times = NULL, lr_test = TRUE, adjust_start = FALSE, ...,
  add = FALSE)
}
\arguments{
\item{s}{a \code{\link[survival]{survfit}} object}

\item{times}{a vector of landmark times}

\item{lr_test}{logical or numeric; if \code{TRUE}, a log-rank test will be
performed and the results added to the top-right corner of the plot; if
numeric, the value is passed as \code{rho} controlling the type of test
performed; see \code{\link{survdiff}}}

\item{adjust_start}{logical; if \code{TRUE}, each landmark plot will begin
at the y-axis}

\item{...}{additional parameters passed to \code{\link{kmplot}}}

\item{add}{logical; if \code{TRUE}, plots are added to current plot}
}
\value{
A data frame with the sample size, chi-square statistic, degrees of
freedom, and p-value for the test (the type of test can be controlled by
using a numeric value for \code{lr_test}, passed as \code{rho} to
\code{\link{survdiff}}).
}
\description{
Fit survival curves for landmark time points.
}
\examples{
library('survival')
s <- survfit(Surv(time, status) ~ node4, colon)
landmark(s, times = c(500, 1000, 2500))

layout(matrix(c(1,1,1,1,2,3), 2))
landmark(s, times = c(500, 2500), adjust_start = TRUE, add = TRUE)

}
