% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{locf}
\alias{locf}
\title{Last observation carried forward}
\usage{
locf(x, fromLast = FALSE)
}
\arguments{
\item{x}{a vector, matrix, or data frame}

\item{fromLast}{logical; if \code{TRUE}, starts from end}
}
\description{
Replaces \code{NA} in vectors, data frames, or matrices with most recent
non-\code{NA} value
}
\examples{
df <- data.frame(V1 = c('Bob', NA, NA, 'Joe', NA, NA),
                 V2 = c(NA, 1, NA, NA, 2, NA))

within(df, {
  V1 <- locf(V1)
  V2 <- locf(V2, fromLast = TRUE)
})

locf(df)
}

