% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locf}
\alias{locf}
\title{Last observation carried forward}
\usage{
locf(x, fromLast = FALSE, na.strings = "")
}
\arguments{
\item{x}{a vector, matrix, or data frame}

\item{fromLast}{logical or vector of logicals; if \code{TRUE}, observations
are carried backward rather than forward; other options such as
\code{c(TRUE, FALSE)} will call \code{locf} first for
\code{fromLast = TRUE} follwed by calling with \code{fromLast = FALSE}
on the result}

\item{na.strings}{a vector of values to be treated as \code{NA}; optionally
a list of single elements can be used for mixed data types; see examples}
}
\description{
Replaces \code{NA} in vectors, data frames, or matrices with most recent
non-\code{NA} value.
}
\examples{
x <- c('', '', 'a', '', 'b', '', '', '', 'c')
locf(x)
locf(x, c(FALSE, TRUE))
locf(x, TRUE)

dd <- data.frame(
  V1 = c('Bob', NA, NA, 'Joe', NA, NA),
  V2 = c(NA, 1, NA, NA, 2, NA),
  stringsAsFactors = FALSE
)

locf(dd)
locf(dd, c(FALSE, TRUE))
locf(dd, na.strings = 2)

## note the differences for numeric and character na.strings
locf(dd, na.strings = c('Joe', 2))
locf(dd, na.strings = list('Joe', 02))
locf(dd, na.strings = list('Joe', '02'))


## with dates
dd$V2 <- as.Date(dd$V2, origin = '2000-01-01')
locf(dd)
locf(dd, TRUE)
locf(dd, c(FALSE, TRUE))

}
