% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{lsf}
\alias{lsf}
\title{List package files}
\usage{
lsf(package, file = "DESCRIPTION")
}
\arguments{
\item{package}{package name, as a \code{\link{name}} or literal character
string}

\item{file}{file to return as a character string; usual options are
\code{'DESCRIPTION'}, \code{'NEWS'}, \code{'INDEX'}, \code{'NAMESPACE'}}
}
\description{
Prints package files (\code{DESCRIPTION}, \code{NEWS}, \code{INDEX},
\code{NAMESPACE}, etc.) to console.
}
\examples{
lsf(rawr, 'DESCRIPTION')
}
\seealso{
\code{\link{lsp}}
}

