% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{lspline}
\alias{lspline}
\title{Linear spline}
\usage{
lspline(x, y, knots = NULL, plot = FALSE, col = NULL)
}
\arguments{
\item{x, y}{the x- and y-axis variables}

\item{knots}{value(s) of \code{x} where knots will be set}

\item{plot}{logical; if \code{TRUE}, fit lines will be plotted}

\item{col}{a vector of colors for each spline}
}
\value{
A list with the following components:

\item{\code{x}}{a data frame used to fit the splines}
\item{\code{y}}{adjusted response values}

Additionally, the linear spline model is returned as an \code{\link{lm}}
object (\code{attr(., "model")}).
}
\description{
Fit a linear spline regression model with pre-specified knots.
}
\examples{
x <- cars$speed
y <- cars$dist

lspline(x, y)
ls <- lspline(x, y, plot = TRUE, knots = c(10, 20))
plot(x, y)
Map('lines', ls$x, ls$y, col = 1:3)


## compare
lspline(x, y, plot = TRUE, knots = 15, col = 2:3)

knot <- 15
fit <- lm(dist ~ speed + I((speed - knot) * (speed >= knot)), cars)
plot(y ~ x)
xx <- seq(min(x), knot, length.out = 1000L)
co <- coef(fit)
lines(xx, co[1] + co[2] * xx, col = 2)
xx <- seq(knot, max(x), length.out = 1000L)
lines(xx, co[1] + co[2] * xx + co[3] * (xx - knot), col = 3)
abline(v = knot)

}
