% Generated by roxygen2 (4.0.1): do not edit by hand
\name{notin}
\alias{\%ni\%}
\alias{notin}
\title{not in}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{vector or NULL; the values to be matched}

\item{y}{vector or NULL; the values to be matched against}
}
\description{
Negation of \code{\link[base]{\%in\%}}
}
\examples{
1:5 \%ni\% 3:5
}

