% Generated by roxygen2 (4.0.0): do not edit by hand
\name{num2char}
\alias{num2char}
\title{Numeric to character string}
\usage{
num2char(num, informal = FALSE, cap = TRUE)
}
\arguments{
\item{num}{number; integer value in \code{(-1e08, 1e08)}}

\item{informal}{logical; if \code{TRUE}, adds "and" before tens or ones}

\item{cap}{logical; if \code{TRUE}, capitalizes the first word}
}
\description{
Convert a number to a character string
}
\details{
Whole numbers twenty-one through ninety-nine are hyphenated when they are
written out whether used alone or as part of a larger number; for example:
twenty-one. Whole numbers are also hyphenated when they are part of larger
numbers that are written out - but not other parts of large numbers; for
example, 5,264 is written "five thousand two hundred sixty-four." The rule
applies only to two-word numbers; for example, 603 is written out, e.g.,
"six hundred three" (formal) or "six hundred and three" (informal). A whole
number followed by hundred, thousand, etc., would be written as, for
example, "one hundred," and not hyphenated. In a phrase like "one hundred
and ten years," no hyphenation should be added.
}
\examples{
num2char(19401, informal = TRUE)

v_num2char <- Vectorize(num2char)
nums <- c(-1000, 100, 10000, 3922, 3012, 201, -152, 1002, 90765432)
v_num2char(nums)
}
\references{
\url{http://dictionary.reference.com/help/faq/language/g80.html}
}

