% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{parse_formats}
\alias{parse_formats}
\title{Get formats from \code{SAS} format files}
\usage{
parse_formats(path)
}
\arguments{
\item{path}{character string of path to \code{.sas} file}
}
\value{
A list with format names in \code{path} and their respective values and
labels as named character vectors.
}
\description{
Reads a text file (usually \code{.sas}) and extracts \code{SAS} format
variable names and format values.
}
\examples{
p <- system.file('testfiles', 'formats.sas', package = 'rawr')
cat(readLines(p), sep = '\\n')
parse_formats(p)

}
\seealso{
\code{\link{sas_path}}, \code{\link{rmacro}}, \code{\link{r2sas}},
\code{\link{get_margs}}, \code{\link{sas_mget}},
\code{\link{source_sas}}, \code{\link{sas_catalog}}
}

