% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{parse_formats}
\alias{parse_formats}
\title{Get formats from \code{SAS} format files}
\usage{
parse_formats(path)
}
\arguments{
\item{path}{character string of path to \code{.sas} file}
}
\value{
A list with format names in \code{path} and their respective values and
labels as named character vectors.
}
\description{
Reads a text file (usually \code{.sas}) and extracts \code{SAS} format
variable names and format values.
}
\examples{
p <- system.file('testfiles', 'formats.sas', package = 'rawr')
cat(readLines(p), sep = '\\n')

#  /*
#    sas format file
#  */
#  
#  proc format;
#    value $jc 'one' = 'management'
#              'two' = 'non-management';
#    value rate
#           -99 = 'not applicable'
#            -1 = 'missing'
#             0 = 'terrible'
#             1 = 'poor'
#             2 = 'fair'
#             3 = 'good'
#             4 = 'excellent';
#    value with_comments
#           -99 = 'not applicable' * comment here;
#            -1 = 'missing'
#             0 = 'terrible' /* comment */
#             1 = 'poor'
#             * comment here;
#             2 = 'fair'
#             3 = 'good'
#             /* 
#               long
#               comment
#             */
#             4 = 'excellent';
#  run;

(pp <- parse_formats(p))

#  $jc
#    'management' 'non-management'; 
#         "'one'"           "'two'"
#  
#  $rate
#  'not applicable'        'missing'       'terrible'           'poor' 
#             "-99"             "-1"              "0"              "1"
#            'fair'           'good'      'excellent' 
#               "2"              "3"              "4"
#  
#  $with_comments
#  'not applicable'        'missing'       'terrible'           'poor' 
#             "-99"             "-1"              "0"              "1"
#            'fair'           'good'      'excellent' 
#               "2"              "3"              "4"

names(pp[[3]][match(c('1','-99','2','0'), pp[[3]])])

}
\seealso{
\code{\link{rmacro}}, \code{\link{r2sas}}, \code{\link{sas_mget}},
\code{\link{source_sas}}
}

