% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{parse_formats}
\alias{apply_formats}
\alias{parse_formats}
\alias{parse_formats2}
\title{Get formats from \code{SAS} format files}
\usage{
parse_formats(path)

parse_formats2(fmt, clean = TRUE)

apply_formats(x, fmt, clean = TRUE, droplevels = FALSE)
}
\arguments{
\item{path}{character string of path to \code{.sas} file}

\item{fmt}{a character string with formats to parse}

\item{clean}{logical; if \code{TRUE} (default), the parsed formats will be
cleaned of quotes and extra whitespace}

\item{x}{a vector of data usually taking a small number of distinct values}

\item{droplevels}{logical; if \code{TRUE}, unused factor levels will be
dropped; default is not to drop unused levels}
}
\value{
A list with format names in \code{path} and their respective values and
labels as named character vectors.
}
\description{
Reads a text file (usually \code{.sas}) and extracts \code{SAS} format
variable names and format values.
}
\examples{
p <- system.file('testfiles', 'formats.sas', package = 'rawr')
cat(readLines(p), sep = '\\n')
parse_formats(p)


## named vector of formats from string
fmt <- '0 = something, 1 = something else; 2 = yada. -9=yadayadayada, .C=blah'
parse_formats2(fmt)

## use to format factor variables
x <- sample(0:2, 10, TRUE)
table(apply_formats(x, fmt), x)
apply_formats(x, fmt, droplevels = TRUE)

## reordered if all levels are numeric and positive
parse_formats2('1=yes, 0=no, 2=maybe')

}
\seealso{
\code{\link{sas_path}}, \code{\link{rmacro}}, \code{\link{r2sas}},
\code{\link{get_margs}}, \code{\link{sas_mget}},
\code{\link{source_sas}}, \code{\link{sas_catalog}}
}

