% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.facet_adjust}
\alias{print.facet_adjust}
\title{facet_adjust print method}
\usage{
\method{print}{facet_adjust}(x, newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{x}{object from \code{\link{facet_adjust}}}

\item{newpage}{draw new (empty) page first; see
\code{\link[ggplot2]{print.ggplot}}}

\item{vp}{viewport to draw plot in; see \code{\link[ggplot2]{print.ggplot}}}
}
\description{
facet_adjust print method
}
\seealso{
\code{\link{facet_adjust}}
}

