% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{progress}
\alias{progress}
\title{Progress function}
\usage{
progress(value, max.value, textbar = FALSE)
}
\arguments{
\item{value}{numeric; i-th iteration or percent completed (values 0-100)}

\item{max.value}{numeric; n-th iteration; if missing, will assume percent
completion is desired}

\item{textbar}{logical; if \code{TRUE}, uses text progress bar which will
span across the console width; see \code{\link{options}}}
}
\description{
Displays the percent (or iterations) completed during some loop.
}
\examples{
\dontrun{
iterations <- 77
## percent completed:
for (ii in 1:iterations) {
   progress(ii / iterations * 100)
   Sys.sleep(.01)
}

## iterations completed
for (ii in 1:iterations) {
   progress(ii, iterations)
   Sys.sleep(.01)
}

## text progress bar
for (ii in 1:iterations) {
   progress(ii, iterations, textbar = TRUE)
   Sys.sleep(.01)
}
}

}
