% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psum}
\alias{psum}
\title{Pairwise sum}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric vectors}

\item{na.rm}{logical; if \code{TRUE}, omits missing values (including
\code{\link{NaN}}) from calculations}
}
\value{
A single vector of element-wise sums.
}
\description{
Compute the pairwise sum of two or more vectors.
}
\details{
Each vector passed in \code{...} must be equal in length. The function
coerces the vectors into a matrix and \code{\link{rowSums}} the rows.
}
\examples{
x <- c(-1, NA, 4, 5)
y <- c(NA, NA, 6, -1)
psum(x, y)
psum(x, y, na.rm = TRUE)

}
\seealso{
\code{\link{pmin}}, \code{\link{pmax}}
}

