% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{pvalr}
\alias{pvalr}
\alias{pvalr2}
\alias{color_pval}
\title{p-value formatter}
\usage{
pvalr(
  pv,
  sig.limit = 1e-03,
  digits = 2L,
  scientific = FALSE,
  html = FALSE,
  show.p = FALSE,
  ...
)

pvalr2(pv, html = FALSE, show.p = FALSE)

color_pval(
  pv,
  breaks = c(0, 0.01, 0.05, 0.1, 0.5, 1),
  cols = colorRampPalette(2:1)(length(breaks)),
  sig.limit = 1e-03,
  digits = 3L,
  show.p = FALSE,
  format_pval = TRUE,
  journal = TRUE,
  ...
)
}
\arguments{
\item{pv}{for \code{pvalr}, a numeric value or vector of p-values; for
\code{pvalr2}, a vector of p-values as character strings}

\item{sig.limit}{lower bound for precision; smaller values will be shown as
\code{< sig.limit}}

\item{digits}{number of significant digits}

\item{scientific}{see \code{\link[base]{format}}}

\item{html}{logical; if \code{TRUE}, uses HTML entities for \code{<}
and \code{>}}

\item{show.p}{logical; if \code{TRUE}, inserts \code{p = }, \code{p < }, or
\code{p > } where appropriate}

\item{...}{ignored}

\item{breaks, cols}{a numeric vector defining breaks in \code{(0,1)} (passed
to \code{\link{findInterval}}) and the corresponding colors}

\item{format_pval}{logical; if \code{TRUE}, p-values will be formatted
using \code{\link{pvalr}}; alternatively, a function may by used which
will be applied to each p-value}
}
\description{
Formats several cases of p-values; see details.

Note that a previous \code{pvalr} function had additional arguments. For
this version, see \code{rawr2::pvalr}.
}
\details{
\code{pvalr} rounds numeric values to a specified number of significant
\code{digits} and add (if necessary) trailing 0s. The result will be a
character string to prevent loss of significant digits.

Note that negative values are treated as small p-values (similar to
\code{format.pval}), but unlike \code{format.pval}, values greater than 1
are formatted as \code{> 1 - 1 / 10 ^ digits}

\code{pvalr2} deals with common cases of character string p-values which are
not ideal (0.000, 1.00, etc.) and will leave others unchanged.
}
\examples{
pv <- c(-1, 0.001, 0.0001, 0.0042, 0.0601, 0.1335, 0.4999, 0.9, 0.997, 1, 2)
format.pval(pv, eps = 0.001)

pvalr(pv)
pvalr(pv, digits = 3)
pvalr(pv, show.p = TRUE, html = TRUE)

\dontrun{
## older version of pvalr in rawr2 package
## note differences in digits used
rawr2::pvalr(pv)
rawr2::pvalr(pv, journal = FALSE)
rawr2::pvalr(pv, limits = 0.1)
}

pv <- c('0.00000', '< 0.001', '0.0', '0.123', '0.6', '1', '1.0', '1.000')
pvalr2(pv)

\dontrun{
pv <- c(0, 0.00001, 0.03, .06, .11, .49, .51, .89, .9, 1)
show_html(color_pval(pv))
show_html(color_pval(pv, format_pval = format.pval))
show_html(iprint(color_pval(pv, show.p = TRUE)))
}

}
\seealso{
\code{\link[rawr]{roundr}}; \code{\link[base]{format.pval}};
\code{rawr:::signif2}
}
