% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{pvalr}
\alias{pvalr}
\alias{pvalr2}
\alias{color_pval}
\title{p-value formatter}
\usage{
pvalr(
  pv,
  sig.limit = 10^-digits,
  digits = 3L,
  html = FALSE,
  show.p = FALSE,
  journal = TRUE,
  breaks = c(0.01, 0.5),
  ...
)

pvalr2(pv, html = FALSE, show.p = FALSE)

color_pval(
  pv,
  breaks = c(0, 0.01, 0.05, 0.1, 0.5, 1),
  cols = colorRampPalette(2:1)(length(breaks)),
  sig.limit = 1e-03,
  digits = 3L,
  show.p = FALSE,
  format_pval = TRUE,
  journal = TRUE,
  ...
)
}
\arguments{
\item{pv}{for \code{pvalr}, a numeric value or vector of p-values; for
\code{pvalr2}, a vector of p-values as character strings}

\item{sig.limit}{lower bound for precision; smaller values will be shown as
\code{< sig.limit}}

\item{digits}{number of digits past the decimal point to keep}

\item{html}{logical; if \code{TRUE}, uses HTML entities for \code{<}
and \code{>}}

\item{show.p}{logical; if \code{TRUE}, inserts \code{p = }, \code{p < }, or
\code{p > } where appropriate}

\item{journal}{logical; if \code{TRUE}, p-values greater than
\code{sig.limit} are rounded to \code{digits}}

\item{breaks}{a vector controlling the number of digits used in p-values

if \code{journal = TRUE}, \code{breaks} should be length one, and p-values
will be rounded to two digits for values greater than \code{breaks} or
\code{digits} for values between \code{sig.limit} and \code{breaks}

if \code{journal = FALSE}, \code{breaks} should be length two, and p-values
will be rounded to one digit for values greater than \code{breaks[2L]},
two digits for values between \code{breaks[1L]} and \code{breaks[2L]},
and \code{digits} for values between \code{sig.limit} and
\code{breaks[1L]}}

\item{...}{additional arguments passed to \code{\link{format.pval}} or
further to \code{\link{format}}}

\item{format_pval}{logical; if \code{TRUE}, p-values will be formatted
using \code{\link{pvalr}}; alternatively, a function may by used which will
be applied to each p-value}

\item{breaks, cols}{a numeric vector defining breaks in \code{(0,1)} (passed
to \code{\link{findInterval}}) and the corresponding colors}
}
\description{
Formats several cases of p-values; see details.
}
\details{
\code{pvalr} will deal with several cases of common p-values: 1) p-values
which are > 0.1 will be rounded to two decimal places (and keep any trailing
0s) since we are not usually interested in precision of insignificant
values; 2) p-values which are less than the specified \code{sig.level} will
be formatted as \code{< 0.001}, for example; 3) p-values that are less than
0.01 but greater than \code{sig.level} will be precise to \code{digits} and
keep any trailing 0s.

\code{pvalr2} deals with common cases of character string p-values which are
not ideal (0.000, 1.00, etc) and will leave others unchanged.
}
\examples{
pv <- c(-1, 0.00001, 0.0042, 0.0601, 0.1335, 0.4999, 0.51, 0.89, 0.9, 1)
format.pval(pv, eps = 0.001)

pvalr(pv)
pvalr(pv, journal = FALSE)
pvalr(pv, breaks = 0.1)
pvalr(pv, show.p = TRUE, html = TRUE)

pv <- c('0.00000', '< 0.001', '0.0', '0.123', '0.6', '1', '1.0', '1.000')
pvalr2(pv)

pv <- c(0, 0.00001, 0.03, .06, .11, .49, .51, .89, .9, 1)

show_html(color_pval(pv))
show_html(color_pval(pv, format_pval = format.pval))
show_html(iprint(color_pval(pv, show.p = TRUE)))

}
\seealso{
\code{\link[rawr]{roundr}}; \code{\link[base]{format.pval}}
}
