% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas.R
\name{r2sas}
\alias{r2sas}
\title{r2sas}
\usage{
r2sas(code, saspath, force = FALSE, out = getwd())
}
\arguments{
\item{code}{character string of valid \code{SAS} code}

\item{saspath}{file path to \code{sas.exe} as character string

Usually \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory

Note that specifying \code{saspath} is not required if v9.2 is installed in
the default directory (i.e., \code{c:/program files/sasfoundation/9.2/sas.exe});

For previous versions or if\code{SAS} is not installed in the default
directory, it will be necessary to supply \code{saspath} to \code{sas.exe}}

\item{force}{logical; by default, user must interactively allow
\code{r2sas} to continue running \code{code}; set to \code{TRUE} to ignore
this or for non-interactive \code{R}}

\item{out}{either \code{FALSE}, directory (as character string) to dump
\code{.sas} program file, \code{.log} file, and any \code{SAS} output, i.e.,
\code{.lst} file(s)}
}
\description{
Write and run \code{SAS} code in \code{R}.

This is meant as a utility function for other \code{SAS} functions in this
package, and users are \emph{not} encouraged to use \code{r2sas} to
interface with \code{SAS} (while possible with this function, see examples).

To source an existing \code{.sas} program, see \code{\link{source_sas}}.

For macros, see \code{\link{get_margs}} and \code{\link{rmacro}}.

To convert \code{.sas7bdat} files, \code{\link{sas.mget}}.
}
\examples{
\dontrun{
code <- "
* this is a sas program file : ;

options nodate nocenter nonumber ;

x 'cd ./newfolder' ;

libname lib './newfolder' ;

data data ;
  set data ;
if x = 1 then delete ;
run ;
"

r2sas(code)

## * this is a sas program file : ;
##
## options nodate nocenter nonumber ;
##
## x 'cd ./newfolder' ;
##
## libname lib './newfolder' ;
##
## data data ;
##   set data ;
## if x = 1 then delete ;
## run ;
##
##
##
## ... will be run. Continue? (y/n):
}
}
\seealso{
\code{\link{rmacro}}, \code{\link{get_margs}}, \code{\link{sas.mget}},
\code{\link{source_sas}}
}

