% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{ransch}
\alias{ransch}
\alias{ranschtbl}
\title{ransch}
\usage{
ransch(n, block, arms, r = 1L, strata = NULL)

ranschtbl(n, block, arms, r = 1L, strata = NULL, write = NULL)
}
\arguments{
\item{n}{sample size of study or each stratum}

\item{block}{block size; note if \code{block} is not a factor of \code{n},
\code{n} will be increased to accommodate a full block}

\item{arms}{names of the treatment arms}

\item{r}{randomization ratio; see examples}

\item{strata}{an optional named list of vectors for each stratum}

\item{write}{a file path to write a directory with csv files for each
randomization table}
}
\description{
Generate block randomization schedule tables.
}
\examples{
## no strata, 2-3 treatments with varying randomization ratios
ransch(24, 4, 1:2) ## 1:1
ransch(24, 6, 1:3) ## 1:1:1
ransch(24, 8, 1:3, c(1, 2, 1)) ## 1:2:1

## one two-level stratum
ransch(24, 4, 1:2, strata = list(Age = c('<65', '>=65')))

## multiple strata
strata <- list(Site = LETTERS[1:3], Age = c('<65', '>=65'))
ransch(24, 4, 1:2, strata = strata)


## tables for printing
ranschtbl(24, 4, c('Pbo', 'Trt'))

ranschtbl(24, 4, c('Pbo', 'Trt'), c(1, 3), strata)

\dontrun{
ranschtbl(24, 4, c('Pbo', 'Trt'), strata = strata, write = '~/desktop')
}

}
