% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawr.R
\docType{package}
\name{rawr-package}
\alias{rawr-package}
\title{Retrieve Raw R Code from Popular Tutorials and Websites}
\description{
Retrieve Raw R Code from Popular Tutorials and Websites
}
\details{
It has the goal of providing a very simple and easy way of quickly retrieving (just the)
    raw R code from popular websites that display R code, including github, kaggle,
    and data camp.
}
\author{
Steve Condylios \email{steve.condylios@gmail.com}
}
