% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawr.R
\name{rawr}
\alias{rawr}
\title{Automatically identify website and retrieve raw R code from it}
\usage{
rawr(url)
}
\arguments{
\item{url}{Link to an R file on supported website (github, kaggle, datacamp, tidytext)}
}
\description{
Automatically identify website and retrieve raw R code from it
}
\examples{
library(dplyr)
rawr("https://github.com/hadley/vis-eda/blob/master/travel.R")

# Same as above but provided to cat for easy viewing
rawr("https://github.com/hadley/vis-eda/blob/master/travel.R") \%>\%
  cat


# Use on multiple urls

domains <- c("https://github.com/hadley/vis-eda/blob/master/travel.R",
"https://www.datacamp.com/community/tutorials/sentiment-analysis-R",
"https://www.tidytextmining.com/sentiment.html",
"https://www.kaggle.com/vrtjso/mercari-eda-more-info-than-you-can-imagine",
"https://www.jtimm.net/2019/04/14/lexical-change-procrustes/")

domains \%>\% sapply(rawr)

}
