% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\docType{data}
\name{rawr_palettes}
\alias{rawr_palettes}
\alias{rawr_pal}
\alias{show_pal}
\title{rawr palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
rawr_palettes

rawr_pal(
  name,
  n = NULL,
  z = n,
  type = c("discrete", "continuous"),
  rev = FALSE
)

show_pal(x, fullrange = FALSE, counts = TRUE)
}
\arguments{
\item{name}{the palette name to be used}

\item{n}{the first \code{n} colors from the palette to use}

\item{z}{for \code{type = 'continuous'}, the number of colors desired}

\item{type}{return a discrete or continuous (gradient) of colors}

\item{rev}{logical; if \code{TRUE}, the palette is reversed}

\item{x}{one of 1) a \code{rawr_palette} name; 2) a vector of two or more
colors; 3) an \code{\link{imgpal}} object}

\item{fullrange}{logical; for \code{\link{imgpal}} objects, if \code{TRUE},
the entire palette is shown; otherwise, only the calculated number of
unique colors}

\item{counts}{logical; for \code{\link{imgpal}} objects, if \code{TRUE},
the color frequencies are shown for each}
}
\description{
rawr palettes
}
\examples{
## some built-in palettes
rawr_palettes

## use or generate new palettes from existing
p <- rawr_pal('dfci')
show_pal(p)
p <- rawr_pal('dfci', 4)
show_pal(p)
p <- rawr_pal('dfci', 4, 100, type = 'continuous')
show_pal(p)

## view palettes from other sources
# show_pal(nord::nord_palettes$afternoon_prarie)
show_pal(rainbow(8))

\dontrun{
filled.contour(volcano, col = rawr_pal('dfci', 4, 21, type = 'c'))
filled.contour(volcano, col = rawr_pal('dfci', z = 21, type = 'c'))
filled.contour(volcano, col = rawr_pal('pokrie', 4, 21, type = 'c'))
}

}
\seealso{
\code{\link{imgpal}}; \code{\link{palette}}; \code{\link{colorRampPalette}};
\code{wesanderson::wes_palettes}; \code{nord::nord_palettes}
}
\keyword{datasets}
