% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rawr_view}
\alias{rawr_view}
\alias{view}
\alias{view2}
\title{View data}
\usage{
view(x, title, ...)

view2(x, use_viewer = FALSE, ...)
}
\arguments{
\item{x}{an \code{R} object which can be coerced to a data frame with
non-zero numbers of rows and columns or an \code{htmlwidget} object}

\item{title}{title for viewer window, defaults to name of \code{x}
prefixed by \code{Data:}}

\item{...}{additional arguments passed to \code{\link[DT]{datatable}}}

\item{use_viewer}{logical; if \code{TRUE}, opens in the
\code{\link[rstudioapi]{viewer}} if available; otherwise, opens in the
default browser}
}
\description{
Convenience functons to use the base \code{R} data viewer (\code{view}
always invokes \code{\link[utils]{View}} instead of the rstudio viewer)
or the default browser (\code{view2} which can open html and widgets in
the browser or to view data frame- or matrix-like objects using
\code{\link[DT]{datatable}}.
}
\examples{
view2(mtcars)
view2(htmlTable::htmlTable(mtcars))

\dontrun{
view2(qtlcharts::iplot(1:5, 1:5))
}

}

