% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{regcaptures}
\alias{regcaptures}
\title{Extract captured substrings}
\usage{
regcaptures(x, m)
}
\arguments{
\item{x}{a character vector}

\item{m}{an object with match data}
}
\value{
A list with a matrix of captures for each string in \code{x}. Note that the
column names of each matrix will be the starting positions of the captures.
}
\description{
Extract the captured substrings from match data obtained by
\code{\link{regexpr}}, \code{\link{gregexpr}}, or \code{\link{regexec}}.
}
\examples{
x <- c('larry:35,M', 'alison:22,F', 'dave:,M', 'lily:55,F', 'no data')
m <- regexpr('(.*):(\\\\d+)?,([MF])?', x, perl = TRUE)
regcaptures(x, m)

x <- 'ACCACCACCAC'
m <- gregexpr('(?=([AC]C))', x, perl = TRUE)
regcaptures(x, m)[[1]]

m <- gregexpr('(?=(CC))', x, perl = TRUE)
regcaptures(x, m)[[1]]

## compare:
mapply(function(xx) substr(x, xx, xx + 1), m[[1]])
}
\references{
\url{https://gist.github.com/MrFlick/10413321}
}
\seealso{
\code{\link{regmatches}}
}

