% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regcaptures}
\alias{regcaptures}
\alias{regcaptures2}
\title{Extract captured substrings}
\usage{
regcaptures(x, m, use.names = TRUE)

regcaptures2(x, pattern, use.names = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{m}{an object with match data}

\item{use.names}{logical; if \code{FALSE}, all names (capture names and
list names) will be stripped; if \code{TRUE} (default) and capture groups
have names, these will be used; otherwise, match start positions will be
used}

\item{pattern}{a character string containing a regular expression}
}
\value{
A list with a matrix of captures for each string in \code{x}. Note that the
column names of each matrix will be the starting positions of the captures.
}
\description{
Extract the captured substrings from match data obtained by
\code{\link{regexpr}}, \code{\link{gregexpr}}, or \code{\link{regexec}}.
}
\examples{
x <- c('larry:35,M', 'alison:22,F', 'dave:,M', 'lily:55,F', 'no data')
p1 <- '(.*):(\\\\d+)?,([MF])?'
p2 <- '(?<name>.*):(?<age>\\\\d+)?,(?<sex>[MF])?'

m <- regexpr(p1, x, perl = TRUE)
regcaptures(x, m)
do.call('rbind.data.frame', regcaptures(x, m, use.names = FALSE))

## regcaptures2 is a convenience function for the two-step
regcaptures2(x, p1)

## both will use named captures (if perl = TRUE)
regcaptures(x, gregexpr(p2, x, perl = TRUE))
do.call('rbind.data.frame', regcaptures2(x, p2))


## capture overlapping matches
x <- 'ACCACCACCCAC'
m <- gregexpr('(?=([AC]C))', x, perl = TRUE)
regcaptures(x, m)[[1]]

m <- gregexpr('(?=(CC))', x, perl = TRUE)
regcaptures(x, m)[[1]]

## compare:
mapply(function(xx) substr(x, xx, xx + 1L), m[[1]])

}
\references{
\url{https://gist.github.com/MrFlick/10413321}
}
\seealso{
\code{\link{regmatches}}; \code{\link{grep}}; \code{\link{regex}}
}

