% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescaler}
\alias{rescaler}
\title{Rescale numeric vector}
\usage{
rescaler(x, to = c(0, 1), from = range(x, na.rm = TRUE))
}
\arguments{
\item{x}{numeric vector of values}

\item{to}{output range (numeric vector of length two)}

\item{from}{input range (numeric vector of length two); if not given,
\code{from} is calculated from the range of \code{x}}
}
\description{
Rescale a numeric vector to have specified maximum and minimum; modified
from the \pkg{scales} package.
}
\examples{
rescaler(1:5)
rescaler(runif(5), c(0.5, 1))
rescaler(1)

}
\seealso{
\code{\link[scales]{rescale}}; \code{\link[scales]{zero_range}}
}
