% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{Restart}
\alias{Restart}
\title{Restart \code{R} session}
\usage{
Restart(afterRestartCommand = "")
}
\arguments{
\item{afterRestartCommand}{character string of command(s) to be
executed after restarting}
}
\description{
Ends current and restarts a clean \code{R} session.
}
\examples{
\dontrun{
Restart("clear(); cat('Here is a clean session just for you')")
}

}
