% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rgene}
\alias{rgene}
\title{Generate random gene names}
\usage{
rgene(n = 1, alpha = LETTERS[1:5], nalpha = 2:5, num = 0:9,
  nnum = 1:5, sep = "-", seed = NULL)
}
\arguments{
\item{n}{number of gene names to return}

\item{alpha}{vector of letters to select from}

\item{nalpha}{range of possible number of \code{alpha} to select}

\item{num}{numerics to select from}

\item{nnum}{range of possible number of \code{num} to select}

\item{sep}{character to separate \code{alpha} and \code{num}}

\item{seed}{seed; integer or \code{NULL}}
}
\description{
Generate random gene names
}
\examples{
rgene()
rgene(5, alpha = 'AB', nalpha = 1, nnum = 5:6)
}

