% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{rgene}
\alias{rgene}
\title{Generate random gene names}
\usage{
rgene(
  n = 1L,
  alpha = LETTERS,
  nalpha = 1:5,
  num = 0:9,
  nnum = nalpha,
  sep = "-"
)
}
\arguments{
\item{n}{number of gene names to return}

\item{alpha, num}{vectors of letters or numerics to select from}

\item{nalpha, nnum}{range of possible number of \code{alpha} or \code{num}
to select}

\item{sep}{character to separate \code{alpha} and \code{num}}
}
\description{
Generate random character strings from pools of letters and digits.
}
\examples{
rgene()

## letters only
rgene(5, alpha = c('A', 'T', 'C', 'G'), nalpha = 10, num = NULL)

## fixed prefix (or suffix)
rgene(5, alpha = 'ABCD', nalpha = 1, nnum = 5)

}
