% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{river}
\alias{check_river_format}
\alias{river}
\alias{river2}
\title{River plots}
\usage{
river(data, id, at = id, legend = "topleft", xlim, ylim)

river2(data, id = 1, tox, legend = "topleft", xlim, ylim)

check_river_format(data, tox)
}
\arguments{
\item{data, tox}{data frames; these should have a specific format, see
details, examples, or run \code{check_river_format()}}

\item{id, at}{optional parameters specifying individuals (rows) from
\code{data} to plot and their positions along the y-axis; if not given,
timelines are plotted sequentially}

\item{legend}{character string giving position (see \code{\link{legend}}
or \code{FALSE}}

\item{xlim, ylim}{x- and y-axis limits}
}
\description{
Summarize a timeline for individuals over the course of a study.
}
\details{
\code{data} and \code{tox} data frames need to have a minimum number of
variables in a specific order, some of which should be \code{\link{Date}}
formats. \code{check_river_format()} without any arguments will give a
summary of the formats.

\code{data} should have at least 12 columns (any additional will be
ignored) in the following order: ID; dates of registration, last treatment
start and end, progression, and response; type of response; dates off
treatment, last contact, off study, and status; and status.

\code{tox} should have at least 5 columns (any additional will be
ignored) in the following order: ID, dates of toxicity start and end,
grade, and description.

Non \code{Date} columns can be character, factor, or numeric. The response
and grade variables should be ordinal and therefore factors with proper
level order since the legend and colors will use this ordering. The
description variable should have relatively short descriptions to avoid
text extending outside the plotting window.
}
\examples{
## generate outcome data for river
dd <- data.frame(id = 1:5, dt_reg = 1:5, dt_txst = 2:6, dt_txend = c(10:13,8),
                 dt_prog = c(11,16,NA,NA,8), dt_resp = c(NA,NA,9,13,NA),
                 resp = factor(c('PD','PD','PR','CR','SD'),
                               levels = c('PD','SD','MR','PR','CR')),
                 dt_offtx = c(12,19,NA,NA,8), dt_last = c(12,19,24,25,NA),
                 dt_off = c(12,19,NA,NA,8),
                 dt_surv = c(12,19,24,25,NA), surv = c(0,1,0,0,NA))

## dates should be date formats
dts <- grep('dt_', names(dd))
dd[, dts] <- lapply(dd[, dts], as.Date.numeric, origin = '1970-01-01')

(river(dd, xlim = c(0,25)))


## generate tox data for river2
tt <- data.frame(id = rep(1:2, times = c(1, 5)), dt_start = c(3,5,5,8,9,11),
                 dt_end = c(NA,5,NA,10,10,NA), grade = c(1,4,3,4,1,2),
                 desc = paste('tox', c(1,1:5)))

## dates should be date formats
dts <- grep('dt_', names(tt))
tt[, dts] <- lapply(tt[, dts], as.Date.numeric, origin = '1970-01-01')

river2(dd, 2, tt)

}

