% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{rm_null}
\alias{rm_null}
\title{Recursive \code{rm} for lists}
\usage{
rm_null(l, rm_list = TRUE)
}
\arguments{
\item{l}{a list}

\item{rm_list}{logical; if \code{FALSE}, lists with only the \code{NULL}
object will not be removed}
}
\description{
Remove \code{NULL} or \code{list(NULL)} objects recursively from a list.
}
\examples{
str(l <- list(list(NULL),list(1),list('a', NULL)))
str(rm_null(l))
str(rm_null(l, FALSE))

}
\references{
\url{http://stackoverflow.com/questions/26539441/r-remove-null-elements-from-list-of-lists}
}

