% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas.R
\name{rmacro}
\alias{rmacro}
\title{Call SAS macros}
\usage{
rmacro(mpath, mname, args, saspath, show.args = FALSE, force = FALSE,
  firstArgs, lastArgs, out = getwd())
}
\arguments{
\item{mpath}{path to macro (\code{.sas}) file as character string}

\item{mname}{macro name; if missing, \code{\link{get_margs}} will search
\code{mpath} for macro names; if missing and \code{get_margs} finds more
than one macro in \code{mpath}, will throw an error}

\item{args}{arguments passed to the macro, separated by commas (do not
include semicolons---\code{args} is passed directly as
\code{\%macro(args);} ); if unsure of the macro parameters, run
\code{rmacro} with \code{show.args = TRUE}}

\item{saspath}{file path to \code{sas.exe} as character string

Usually \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory

Note that specifying \code{saspath} is not required if v9.2 is installed in
the default directory (i.e., \code{c:/program files/sasfoundation/9.2/sas.exe});

For previous versions or if\code{SAS} is not installed in the default
directory, it will be necessary to supply \code{saspath} to \code{sas.exe}
\code{mpath} defaults to the most recent version of \code{SAS}}

\item{show.args}{logical; if \code{TRUE}, \code{rmacro} will only return the
macro names and arguments found in \code{mpath}}

\item{force}{logical; by default, user must interactively allow
\code{r2sas} to continue running \code{code}; set to \code{TRUE} to ignore
this or for non-interactive \code{R}}

\item{firstArgs}{optional character string of (valid) \code{SAS} commands
separated by semicolons to be excuted \emph{before} the macro; for example,
\code{options <OPTIONS>;} or \code{x "cd ..";}}

\item{lastArgs}{optional commands to be executed \emph{after} the macro; see
\code{firstArgs}}

\item{out}{either \code{FALSE}, directory (as character string) to dump
\code{.sas} program file, \code{.log} file, and any \code{SAS} output, i.e.,
\code{.lst} file(s)}
}
\description{
\code{rmacro} runs \code{SAS} macros in \code{.sas} files from \code{R}.
}
\examples{
\dontrun{
rmacro('./tests/testfiles/onemacro.sas',
       args = 'arg1 = 1, arg2 = 2',
       firstArgs = 'options nodate no center; x \\'cd ~/desktop\\';',
       lastArgs = 'endsas;')

## options nodate no center;
## x 'cd ~/desktop';
##
## \%include "./tests/testfiles/onemacro.sas";
## \%macro1(arg1 = 1, arg2 = 2);
## endsas;
##
##
##
##
##
##
## ... will be run. Continue? (y/n):
}
}
\seealso{
\code{\link{get_margs}}, \code{\link{r2sas}}, \code{\link{sas.mget}}
\code{\link{source_sas}}
}

