% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{Round}
\alias{Round}
\title{Round vector to target sum}
\usage{
Round(x, target = NULL)
}
\arguments{
\item{x}{numeric values}

\item{target}{desired sum of \code{x} after rounding}
}
\description{
Rounds a numeric vector constrained to sum to a \code{target} value.
}
\examples{
pcts <- data.frame(
  pct1 = c(33.3, 21.5, 45.51),
  pct2 = c(33.3, 33.3, 33.3)
)

## base round                                     
colSums(mapply(round, pcts))

## round to target
colSums(mapply(Round, pcts, 100))

}
\seealso{
\code{\link{roundr}}; \code{\link{round_to}}
}
