% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{round_to}
\alias{round_to}
\title{Round to}
\usage{
round_to(x, to = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{to}{nearest fraction or integer}
}
\description{
Round numerics to nearest multiple of \code{to}.
}
\examples{
x <- 1:20 / 10
round_to(x, 1)
round_to(x, 0.5)

}
