% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{round_to}
\alias{round_to}
\title{Round to}
\usage{
round_to(x, to = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{to}{nearest fraction or integer}
}
\description{
Round numerics to nearest multiple of \code{to}.
}
\examples{
round_to(mtcars$mpg, 5)
round_to(mtcars$mpg, .5)

}
