% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils2.R
\name{roundr}
\alias{roundr}
\alias{roundr.data.frame}
\alias{roundr.default}
\alias{roundr.matrix}
\title{roundr}
\usage{
roundr(x, digits = 1)

\method{roundr}{default}(x, digits = 1)

\method{roundr}{data.frame}(x, digits = 1)

\method{roundr}{matrix}(x, digits = 1)
}
\arguments{
\item{x}{numeric value, vector, matrix, or data frame}

\item{digits}{number of digits past the decimal point to keep}
}
\value{
An object having the same class as \code{x}.
}
\description{
Improved rounding formatter.
}
\details{
Uses \code{\link[base]{sprintf}} to round numeric value, retaining extra 0s.
}
\examples{
round(0.199, 2)

## compare to
roundr(0.199, 2)

## useful for dropping the negative sign in case 1:
roundr(c(-0.0002, 0.0002, 0.5, -0.5, -0.002), digits = 3)

## for matrices or data frames (including factors and/or characters)
roundr(matrix(1:9, 3))

dd <- within(head(mtcars), {
  mpg <- as.character(mpg)
  cyl <- factor(cyl, labels = LETTERS[1:3])
})

roundr(dd)
}
\seealso{
\code{\link[base]{round}}; \code{\link[base]{sprintf}}
}

