% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rpart_utils}
\alias{parent}
\alias{rpart_utils}
\alias{subset_rpart}
\title{\code{rpart} utilities}
\usage{
parent(node)

subset_rpart(tree, node = 1)
}
\arguments{
\item{node}{an integer representing the node number}

\item{tree}{an object returned from \code{rpart}}
}
\value{
\code{parent} returns a vector representing the path from the root to
\code{node}.

\code{subset_rpart} returns the data frame of observations in \code{node}.
}
\description{
Utilities for the \pkg{\link{rpart}} package:

\code{parent} returns all parent nodes of \code{node}.

\code{subset_rpart} and \code{subset_rpart2} (in examples) return a
subset of the data used in \code{rpart} for any intermediate or terminal
\code{node}.
}
\examples{
\dontrun{
library('rpart')
(fit <- rpart(Kyphosis ~ Age + Number + Start, kyphosis, minsplit = 5))

parent(15)
parent(23)

## children nodes should have identical paths
identical(head(parent(28), -1), head(parent(29), -1))

## terminal nodes should combine to original data
nodes <- as.integer(rownames(fit$frame[fit$frame$var \%in\% '<leaf>', ]))
sum(sapply(nodes, function(x) nrow(subset_rpart(fit, x))))

## all nodes
nodes <- as.integer(rownames(fit$frame))
sapply(nodes, function(x) nrow(subset_rpart(fit, x)))


subset_rpart2 <- function(tree, node = 1) {
  require('partykit')
  ptree <- as.party(tree)
  ptree$data <- model.frame(eval(tree$call$data, parent.frame(1L)))
  data_party(ptree, node)[, seq(ptree$data)]
}

## note differences in nodes labels in party vs rpart
dim(subset_rpart(fit, 4))
dim(subset_rpart2(fit, 3))
}

}
\seealso{
\url{http://stackoverflow.com/questions/36086990/how-to-climb-the-tree-structure-of-rpart-object-using-path-in-order-to-purge-man}

\url{http://stackoverflow.com/questions/36748531/getting-the-observations-in-a-rparts-node-i-e-cart}
}

