% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{sample_each}
\alias{sample_each}
\title{Sample each}
\usage{
sample_each(x, n = 1L)
}
\arguments{
\item{x}{a character, factor, or numeric vector}

\item{n}{number to sample from each unique group in order; if \code{x} is
a factor, \code{n} should correspond to \code{levels(x)}; otherwise,
\code{n} will be matched with the sorted unique groups}
}
\value{
A logical vector the same length as \code{x} identifying selected indices.
}
\description{
Returns a logical vector where \code{n} items are randomly sampled from
each unique value of a vector, \code{x}.
}
\examples{
x <- mtcars$gear

sample_each(x)
mtcars[sample_each(x), ]

## compare numeric vs factor vectors (see description above)
mtcars[sample_each(x, 3:5), ]
X <- factor(x, 5:3)
mtcars[sample_each(X, 3:5), ]

}
\seealso{
\code{\link{sample}}; \code{\link{kinda_sort}}
}
