% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sas.mget}
\alias{sas.mget}
\title{Convert multiple SAS data sets to R data frame}
\usage{
sas.mget(libpath, dsn, saspath, fmtpath, log.file, ..., force = FALSE)
}
\arguments{
\item{libpath}{directory to data set(s) as character string; if missing,
searches the current working directory}

\item{dsn}{data set name(s); either \code{data1} or \code{data1.sas7bdat}
will work; if missing or \code{NULL}, all \code{.sas7bdat} files found in
\code{libpath} will be read}

\item{saspath}{directory to \code{sas.exe} as character string; usually for
windows, the path \code{c:/program files/sashome/sasfoundation/x.x/sas.exe},
where \code{x.x} is the \code{SAS} version number, is the correct directory,
and \code{sas.mget} defaults to this (the most recent version of \code{SAS})}

\item{fmtpath}{path to a format \code{.sas} file as character string;
\code{SAS} throws an error if the host used to make the format catalog was
not on the same platform, so we create a new format catalog either by
running a \code{proc format} macro or by making a copy of the non-native
\code{.sas7bcat} file}

\item{log.file}{name of \code{SAS} log file; default value will create
\code{_temp_.log} in the \code{libpath} directory}

\item{...}{additional arguments passed to \code{\link[Hmisc]{sas.get}}}

\item{force}{logical; by default, user must interactively allow
\code{sas.mget} to continue reading all data sets; set to \code{TRUE} to
ignore this or for non-interactive \code{R}}
}
\value{
A list of data frames resembling the \code{SAS} data sets.
}
\description{
User-friendly wrapper of \code{\link[Hmisc]{sas.get}} to convert one or more
\code{SAS} data sets (\code{.sas7bdat} files) into a list of \code{R} data
frames.
}
\examples{
\dontrun{
data.list <- sas.mget('./tests/testfiles/')

## !!! Two data set(s) will be read !!!
##
##  Size in Kb:
##
##       size (Kb)
## data1         0
## data2         0
##
##
##
##
##
## do you want to continue (y/n)?
}
}
\seealso{
\code{\link{rmacro}}, \code{\link{get_margs}}, \code{\link{r2sas}},
\code{\link{source_sas}}
}

