% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_mget}
\alias{sas_mget}
\title{Convert multiple \code{SAS} data sets to \code{R} data frame}
\usage{
sas_mget(libpath = getwd(), dsn = dsn, saspath = sas_path(),
  fmtpath = NULL, catalog = length(dcf) == 1L, log.file = "_temp_.log",
  ..., force = FALSE, write = FALSE)
}
\arguments{
\item{libpath}{directory to data set(s) as character string; if missing,
searches the current working directory}

\item{dsn}{data set name(s); either \code{data1} or \code{data1.sas7bdat}
will work; if missing or \code{NULL}, all \code{.sas7bdat} files found in
\code{libpath} will be read}

\item{saspath}{file path to sas executable as character string}

\item{fmtpath}{(optional) path to a format \code{.sas} file as character
string; \code{SAS} throws an error if the platform used to make the format
catalog is not the same as the current platform; \code{fmtpath} is passed
to \code{\link{sas_catalog}}}

\item{catalog}{logical; if \code{FALSE} or multiple catalogs are found in
\code{libpath}, catalogs will be ignored; if \code{fmtpath} is used to
create a new catalog, any existing catalogs are moved to a new directory
to avoid overwriting}

\item{log.file}{name of \code{SAS} log file; default value will create
\code{_temp_.log} in the \code{libpath} directory}

\item{...}{additional arguments passed to \code{\link[Hmisc]{sas.get}}}

\item{force}{logical; by default, user must interactively allow
\code{sas_mget} to continue reading all data sets; set to \code{TRUE} to
ignore this or for non-interactive \code{R}}

\item{write}{logical; if \code{TRUE}, each data frame will be written to
a \code{.csv} file in a new sub directory, \code{./_sas_mget_}}
}
\value{
A list of data frames resembling the \code{SAS} data sets.
}
\description{
User-friendly wrapper of \code{\link[Hmisc]{sas.get}} to convert one or
more \code{SAS} data sets (\code{.sas7bdat} files) into a list of \code{R}
data frames.
}
\examples{
\dontrun{
p <- system.file('testfiles', package = 'rawr')
data.list <- sas_mget(p)

## !!! Two data set(s) will be read !!!
## 
##  Size in Kb:
##   
##       size (Kb)
## data1         0
## data2         0
## 
## 
## 
## 
## 
## do you want to continue (y/n)? 
}

}
\seealso{
\code{\link{sas_path}}, \code{\link{rmacro}}, \code{\link{get_margs}},
\code{\link{r2sas}}, \code{\link{source_sas}},
\code{\link{parse_formats}}, \code{\link{sas_catalog}}
}

