% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_path}
\alias{sas_path}
\title{Get SAS path}
\usage{
sas_path(saspath, sashome, version)
}
\arguments{
\item{saspath}{optional full path to \code{sas.exe}; if given, all other
arguments are ignored, and \code{saspath} is only checked for existence}

\item{sashome}{optional full path to directory of version directories; for
\code{>9.2}, the default is \code{c:/program files/sashome/sasfoundation};
for \code{9.2}, the default is \code{c:/program\ files/sas/sasfoundation}}

\item{verison}{optional version to use

\code{\link{r2sas}}, \code{\link{rmacro}}, \code{\link{get_margs}},
\code{\link{sas_mget}}, \code{\link{source_sas}}, \code{\link{parse_formats}}}
}
\description{
Finds versions of sas installed and returns latest version by default.
}
\examples{
\dontrun{
sas_path()
sas_path(version = 9.2)
}

}

