% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search.hist}
\alias{search.hist}
\title{Search history}
\usage{
search.hist(x, ...)
}
\arguments{
\item{x}{numeric or character; if numeric, shows the most recent \code{n}
lines in \code{.Rhistory}; if character, searches for pattern matches}

\item{...}{additional arguments passed to \code{\link{grep}}}
}
\value{
Returns a list of recent commands that match \code{pattern}
}
\description{
Searches Rhistory file for pattern matches
}
\examples{
\dontrun{
search.hist()
search.hist(25)
search.hist('?')
search.hist('?', fixed = TRUE)
search.hist('\\\\?')
}
}

