% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{search_df}
\alias{search_df}
\title{Search function for data frames}
\usage{
search_df(pattern, data, col.name, var = 0, ignore.case = TRUE, ...)
}
\arguments{
\item{pattern}{string to find}

\item{data}{data frame to search}

\item{col.name}{column name in \code{data} to search}

\item{var}{variation; maximum distance allowed for a match; see
\code{\link{agrep}}}

\item{ignore.case}{logical; if \code{FALSE}, the pattern matching is
\emph{case-sensitive}, and if \code{TRUE}, case is ignored during matching}

\item{...}{additional arguments passed to \code{\link{agrep}}}
}
\value{
Subset of the original \code{data} where the \code{pattern} was found in
the specified \code{col.name}.
}
\description{
Searches a data frame column for matches.
}
\examples{
dd <- data.frame(islands = names(islands)[1:32], mtcars)
search_df(New, dd, islands)
search_df(ho, dd, islands, var = 0.2) # too much variation
search_df(ho, dd, islands, var = 0)
search_df('Axel Hieberg', dd, islands) # misspelled, not enough variation
search_df('Axel Hieberg', dd, islands, var = 2)
search_df(19, dd, mpg)

}
