% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_extra.R
\name{show_colors}
\alias{show_colors}
\title{Show colors}
\usage{
show_colors(..., plot = FALSE)
}
\arguments{
\item{...}{integer(s) in \code{1:657} corresponding to the built-in color
name index or color name string(s)}

\item{plot}{logical; if \code{TRUE}, integers or color names in \code{dots}
will be plotted with corresponding number and name}
}
\value{
If \code{...} is missing, a plot will be drawn. If an integer is given, the
color name will be returned; if a color name string is given, the matching
index will be returned. For the latter two options, no plot is drawn.
}
\description{
\code{R} includes 657 named \code{\link{colors}}. This function shows these
and their respective indices. Find a color by number in the plot or by the
name of the color with \code{show_colors(n)} or find the indices with
\code{show_colors('color name')}.
}
\examples{
show_colors()

show_colors(81)
show_colors('darkgreen')

x <- show_colors(sample(657, 10))
stopifnot(identical(x, show_colors(show_colors(x))))

## these plots are identical
show_colors(x, plot = TRUE)
show_colors(show_colors(x), plot = TRUE)

}
\seealso{
\code{\link{show_pch}}, \code{\link{colors}}
}

