% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_extra.R
\name{show_colors}
\alias{show_colors}
\title{Show colors}
\usage{
show_colors()
}
\description{
In \code{R}, there are 657 named colors. This function shows these colors 
and their respective numbers. Find a color by number in the plot or by the
name of the color with \code{colors()[n]}.
}
\examples{
show_colors()
colors()[81]
# [1] "darkgreen"

}
\seealso{
\code{\link{show_pch}}
}

