% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils2.R
\name{show_math}
\alias{show_math}
\title{Show math}
\usage{
show_math(..., css, use_viewer = !is.null(getOption("viewer")))
}
\arguments{
\item{...}{one or more character strings}

\item{css}{optional css formatting}

\item{use_viewer}{logical; if \code{TRUE}, attempts to use
\code{rstudio::viewer} or opens in default browser on error}
}
\description{
Displays math equations in \code{rstudio::viewer} or browser using the
\href{http://www.mathjax.org}{MathJax} javascript engine.
}
\examples{
\dontrun{
form1 <- '$$A=\\\\frac{B}{C}$$'
form2 <- '$$
  \\\\frac{1}{\\\\displaystyle 1+
  \\\\frac{1}{\\\\displaystyle 2+
  \\\\frac{1}{\\\\displaystyle 3+x}}} +
  \\\\frac{1}{1+\\\\frac{1}{2+\\\\frac{1}{3+x}}}
$$'
form3 <- '\\\\frac{d}{dx}\\\\left( \\\\int_{0}^{x} f(u)\\\\,du\\\\right)=f(x)'

show_math(form1)
show_math(form2, use_viewer = FALSE)
show_math(form1, form2, form3, css = 'color: red; font-size: 15px;')
}
}
\seealso{
\code{\link{show_html}},
\href{http://stackoverflow.com/questions/31193843/display-r-formula-elegantly-as-in-latex}{SO question}
}

