% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{simon2}
\alias{simon2}
\title{Simon two-stage designs}
\usage{
simon2(
  p0,
  pa,
  n1max = 0,
  ntmax = 1e+05,
  alpha = 0.1,
  beta = 0.1,
  del = 1,
  minimax = FALSE
)
}
\arguments{
\item{p0, pa}{null and alternative hypothesis response probability}

\item{n1max}{maximum number of subjects entered during the first stage; 
ignored if <= 0}

\item{ntmax}{maximum total number of total subjects}

\item{alpha, beta}{type-I and type-II errors}

\item{del}{searches for designs where the expected number of subjects under
the null with within \code{del} of the minimum possible value}

\item{minimax}{logical; if \code{TRUE}, only searches for designs which will
minimize the maximum sample size}
}
\value{
A list with the following elements:

\item{\code{$designs}}{a matrix with a row giving a summary of each
design which meets the criteria. The columns are: \code{n1}, the number
of subjects entered in the first stage; \code{r1}, the cutoff for stopping
at the first stage (continue if the number of responses is > \code{r1});
\code{n2}, the additional number of subjects enrolled in the second stage;
\code{r2}, the cutoff for inactivity after the second stage (reject the
null if the number of responses is > \code{r2}); \code{Pstop1.H0}, the
probability of stopping after the first stage under H0 (\code{p0});
\code{size}, the actual type-I error; \code{type2}, the actual type-II
error; \code{E.tot.n.H0}, the expected number of subjects under H0;
\code{Pstop1.H1}, the probability of stopping after the first stage under
H1; and \code{E.tot.n.H1}, the expected number of subjects under H1.}
\item{\code{$call}}{the call to \code{simon2}}
\item{\code{$description}}{a text string giving a brief description of
the columns in \code{$designs}}
}
\description{
Function for sample sizes for Simon optimal two-stage, single-arm designs.
}
\details{
For two-stage designs for studies with binary endpoints, searches over
possible two-stage sampling designs to find those that minimize the
expected number of subjects, subject to specified constraints.
}
\examples{
simon2(0.2, c(0.4, 0.5))
simon2(p0 = seq(0.55, 0.6, by = 0.01), pa = 0.75, ntmax = 60)

## compare this function to results from desmon::simon
simon2(0.4, 0.6)
## from desmon package
rawr:::simon(0.4, 0.6)

}
\references{
Simon R (1989). Optimal two-stage designs for phase II clinical trials.
\emph{Controlled Clinical Trials}, 10:1-10.
}
\seealso{
\pkg{desmon}: \code{simon}, \code{twostg}, \code{bin1samp}, \code{pickwin},
\code{rp21}
}
\author{
Robert Gray (\code{desmon::simon}); Robert Redd (\code{simon2})
}
