% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_matrix}
\alias{sort_matrix}
\title{Sort matrix}
\usage{
sort_matrix(m, margin = 1L, order)
}
\arguments{
\item{m}{a matrix}

\item{margin}{margin to sort by; default is to sort based on row values
(\code{margin = 1}) but values by column may be used (\code{margin = 2})}

\item{order}{vector specifying all unique values of \code{m} in the
desired order; if missing, the order will be the sorted unique values
of \code{m}}
}
\description{
Sort a matrix (or an object that can be coerced) by values in rows or
columns.
}
\examples{
set.seed(1)
(m <- matrix(rpois(5 * 10, 1), 5))

## sort columns by decreasing row values
sort_matrix(m)

## sort rows by decreasing column values
sort_matrix(m, 2)

## sort 0s first followed by 3, 1, 2
sort_matrix(m, order = c(0,3,1,2))

}

