% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{source_sas}
\alias{source_sas}
\title{Source \code{SAS} code}
\usage{
source_sas(path, ...)
}
\arguments{
\item{path}{path to \code{.sas} file as character string}

\item{...}{additional parameters passed to \code{\link{r2sas}}}
}
\description{
\code{\link{source}}-esque function for \code{.sas} files.
}
\examples{
\dontrun{
source_sas(system.file('testfiles', 'onemacro.sas', package = 'rawr'))

## \%macro macro1(arg1, arg2);
## data data;
##   set data;
## run;
## \%mend macro1;
##
##
##
## ... will be run. Continue? (y/n): 
}

}
\seealso{
\code{\link{sas_path}}, \code{\link{r2sas}}, \code{\link{rmacro}},
\code{\link{get_margs}}, \code{\link{sas_mget}},
\code{\link{parse_formats}}, \code{\link{sas_catalog}}
}

