% Generated by roxygen2 (4.0.1): do not edit by hand
\name{surv_cp}
\alias{surv_cp}
\title{Create counting process data}
\usage{
surv_cp(dat, time.var, status.var,
        covars = setdiff(names(dat), c(time.var, status.var)))
}
\arguments{
\item{dat}{data frame with survival time, survival status, and other
covariates}

\item{time.var}{\code{dat} variable name representing survival time}

\item{status.var}{\code{dat} variable name representing status}

\item{covars}{other covariates to retain}
}
\value{
A data frame with events in counting process notation.
}
\description{
Converts a data frame to counting process notation and allows for time-
dependent variables to be introduced.
}
\examples{
library(survival)

cp <- surv_cp(aml, 'time', 'status')
coxph(Surv(start, stop, status) ~ x, data = cp)

## compare to
coxph(Surv(time, status) ~ x, data = aml)
}
\author{
Beau Benjamin Bruce <bbbruce@emory.edu>
}
\references{
\url{http://www.ddiez.com/teac/surv/}
}

