% Generated by roxygen2 (4.0.2): do not edit by hand
\name{surv_summary}
\alias{surv_summary}
\title{Summary of a survival curve}
\usage{
surv_summary(s, digits = max(options()$digits - 4, 3), ...)
}
\arguments{
\item{s}{\code{\link[survival]{survfit}} object}

\item{digits}{number of digits to use in printing numbers}

\item{...}{additional arguments passed to
\code{\link[survival]{summary.survfit}}}
}
\value{
A list with summaries for each strata; see
\code{\link[survival]{summary.survfit}}
}
\description{
Prints and returns a list containing the survival curve, confidence limits
for the curve, and other information.
}
\examples{
library(survival)
data(cancer)
fit1 <- survfit(coxph(Surv(time, status) ~ strata(I(age > 60)),
                      data = cancer),
                conf.type = 'log-log')
surv_summary(fit1, times = c(0, 100, 200))
}
\seealso{
\code{\link[survival]{survfit}},
\code{\link[survival]{print.summary.survfit}}
}

