% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_table}
\alias{surv_table}
\title{Summary table}
\usage{
surv_table(s, digits = 3, times = pretty(range(s$time)), ...)
}
\arguments{
\item{s}{\code{\link[survival]{survfit}} object}

\item{digits}{number of digits to use in printing numbers}

\item{times}{vector of times}

\item{...}{additional arguments passed to
\code{\link[survival]{summary.survfit}}}
}
\value{
A matrix (or list of matrices) with formatted summaries for each strata; see
\code{\link[survival]{summary.survfit}}
}
\description{
Prints a formatted summary table for \code{\link[survival]{survfit}} objects
}
\examples{
library(survival)
data(cancer)

fit0 <- survfit(coxph(Surv(time, status) ~ 1,
                      data = cancer),
                conf.type = 'log-log')
surv_table(fit0, times = c(0, 100, 200))

## also works for list of tables
fit1 <- survfit(coxph(Surv(time, status) ~ strata(I(age > 60)),
                      data = cancer),
                conf.type = 'log-log', conf.int = 0.9)
surv_table(fit1)

\dontrun{
library(Gmisc)
s <- `colnames<-`(surv_table(fit0, times = c(0, 200, 400, 600, 800),
                             digits = 2)[ , -4],
                  c('Time','No. at risk','No. of events','OS (95\% CI)'))
htmlTable(s)
}
}
\seealso{
\code{\link[survival]{survfit}},
\code{\link[survival]{print.summary.survfit}}
}

