% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_table}
\alias{surv_table}
\title{Summary table}
\usage{
surv_table(
  object,
  digits = ifelse(percent, 0L, 3L),
  times = pretty(object$time),
  maxtime = FALSE,
  percent = FALSE,
  locf = FALSE,
  median = FALSE,
  ci.median = TRUE,
  digits.median = 0L,
  na.median = "NR",
  ...
)
}
\arguments{
\item{object}{an object of class \code{\link[survival]{survfit}}}

\item{digits}{number of digits to use in printing numbers}

\item{times}{vector of times}

\item{maxtime}{logical; if \code{TRUE}, adds the maximum time for which an
even occurs; if \code{FALSE}, number of events may not sum to total}

\item{percent}{logical; if \code{TRUE}, percentages are shown instead of
probabilities}

\item{locf}{logical; if \code{TRUE}, any \code{NA} probabilities will be
carried forward, e.g., if no events occurred during two consecutive time
intervals}

\item{median, ci.median}{logical; if \code{TRUE}, the median (and confidence
interval) will be added as a separate column}

\item{digits.median}{for the median, the number of digits past the decimal
point to keep}

\item{na.median}{a string to use in place of "NA" for the median text;
default is "NR" for "not reached"}

\item{...}{additional arguments passed to \code{\link{summary.survfit}}}
}
\value{
A matrix (or list of matrices) with formatted summaries for each strata; see
\code{\link{summary.survfit}}
}
\description{
Prints a formatted summary table for \code{\link{survfit}} objects.
}
\examples{
library('survival')
fit0 <- survfit(Surv(time, status == 2) ~ 1, data = cancer)
surv_table(fit0, times = 0:2 * 100)
surv_table(fit0, times = 0:2 * 100, median = TRUE)

## also works for list of tables
fit1 <- survfit(Surv(time, status == 2) ~ sex, data = cancer)
surv_table(fit1, median = TRUE, digits.median = 2, ci.median = FALSE)
rawr::combine_table(surv_table(fit1, median = TRUE))

}
\seealso{
\code{\link{survfit}}; \code{\link{print.summary.survfit}}
}
