% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{surv_test}
\alias{surv_test}
\alias{lr_pval}
\alias{lr_text}
\alias{tt_pval}
\alias{tt_text}
\alias{hr_pval}
\alias{hr_text}
\alias{pw_pval}
\alias{pw_text}
\title{Survival curve tests}
\usage{
lr_pval(object, details = FALSE, data = NULL, ...)

lr_text(formula, data, rho = 0, ..., details = TRUE, pFUN = NULL)

tt_pval(object, details = FALSE, data = NULL, ...)

tt_text(formula, data, ..., details = TRUE, pFUN = NULL)

hr_pval(object, details = FALSE, data = NULL, ...)

hr_text(formula, data, ..., details = TRUE, pFUN = NULL)

pw_pval(object, details = FALSE, data = NULL, ..., method = "none")

pw_text(formula, data, ..., details = TRUE, pFUN = NULL, method = "none")
}
\arguments{
\item{object}{a \code{\link{survfit}}, \code{\link{survdiff}}, or
\code{\link{coxph}} object; alternatively a \code{\link[=Surv]{survival
formula}} (\code{data} must be given)}

\item{details}{logical; \code{TRUE} returns statistic, degrees of freedom,
and p-value where \code{FALSE} returns only a pvalue}

\item{formula, data, rho, ...}{passed to \code{\link{survdiff}} or
\code{\link{coxph}}}

\item{pFUN}{logical; if \code{TRUE}, p-values are formatted with
\code{\link{pvalr}}; if \code{FALSE}, no formatting is performed;
alternatively, a function can be passed which should take a numeric value
and return a character string (or a value to be coerced) for printing}

\item{method}{for \code{pw_*}, the method used to adjust p-values for
multiple comparisons (default is \code{"none"}); see
\code{\link{p.adjust.methods}}}
}
\description{
Internal functions for \code{\link{survdiff}} and \code{\link{survfit}}
objects.

\code{*_pval} functions take (\code{survfit} or \code{survdiff}) objects
or formulas and compute test statistics, p-values, etc. and return a
numeric vector or list.

\code{*_text} functions format the test results for plotting and return
an expression or vector of character strings.

Note that \code{pw_pval} and \code{pw_text} do not support formulas
with more than one predictor, e.g., \code{y ~ a + b}. An equivalent
formula is acceptable, e.g., \code{y ~ x} where \code{x} is the
\code{interaction(a, b)} or similar. See \code{\link{survdiff_pairs}}
for more details and examples.
}
\examples{
\dontrun{
library('survival')
data('larynx', package = 'KMsurv')
larynx$stage <- factor(larynx$stage)

form <- Surv(time, delta) ~ stage
sf <- survfit(form, larynx)
sd <- survdiff(form, larynx)

kmplot(sf, lr_test = TRUE)


## log-rank
rawr:::lr_pval(sf)
rawr:::lr_pval(sd, TRUE)
rawr:::lr_text(Surv(time, delta) ~ stage, larynx)


## pairwise log-rank
sf$call$formula <- form
rawr:::pw_pval(sf)
rawr:::pw_text(sf)


## tarone trend
rawr:::tt_pval(sf)
rawr:::tt_pval(sd, TRUE)
rawr:::tt_text(Surv(time, delta) ~ stage, larynx)

## compare
chi <- coxph(Surv(time, delta) ~ stage, larynx)$score
list(chi = chi, p.value = pchisq(chi, 1, lower.tail = FALSE))


## hazard ratio/wald p-values
rawr:::hr_pval(sf)
rawr:::hr_pval(sd, TRUE)
rawr:::hr_text(Surv(time, delta) ~ stage, larynx)
}

}
\references{
Tarone, Robert E. Tests for Trend in Life Table Analysis. \emph{Biometrika}
\strong{62} vol. 62 (Dec 1975), 679-82.

\url{https://stat.ethz.ch/pipermail/r-help/2008-April/160209.html}
}
\seealso{
\code{\link{survdiff_pairs}}
}
