% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{survdiff_pairs}
\alias{survdiff_pairs}
\title{Pairwise survival group comparisons}
\usage{
survdiff_pairs(s, ..., method = "bonferroni", digits = getOption("digits"))
}
\arguments{
\item{s}{a \code{\link[survival]{survdiff}} object}

\item{...}{additional arguments passed to \code{\link{survdiff}} such as
\code{na.action} to filter missing data or \code{rho} to control the type
of test}

\item{method}{p-value correction method; see \code{\link{p.adjust}}}

\item{digits}{integer indicating the number of decimal places to be used}
}
\description{
Evaluate differences in survival curves for all pairs of a grouping
variable. This function currently works for one \code{factor}-like
variable, and all unqiue levels are treated as a sub-group.
}
\examples{
library('survival')
fit1 <- survdiff(Surv(time, status) ~ sex, data = lung)
survdiff_pairs(fit1)

## note that despite a numeric group variable, contrasts
## are calculated for each unique level combination
dd <- lung[lung$ph.ecog \%in\% 0:2, ]
fit2 <- survdiff(Surv(time, status) ~ ph.ecog, data = dd)
survdiff_pairs(fit2)

## compare
survdiff(Surv(time, status) ~ ph.ecog, data = dd[dd$ph.ecog \%in\% 0:1, ])

}
\seealso{
\code{\link[rawr]{pvalr}}; \code{\link{survdiff}}; \code{\link{p.adjust}};
\code{\link[rms]{contrast}}
}

