% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{sym_sort}
\alias{sym_sort}
\title{Symmetrical sort}
\usage{
sym_sort(x, rev = FALSE)
}
\arguments{
\item{x}{a numeric, complex, character, or logical vector}

\item{rev}{logical; if \code{TRUE}, vectors are sorted in reverse}
}
\description{
Sort a vector symmetrically, i.e., the two most extreme values are put
at opposite ends, followed by the next two, etc.
}
\examples{
sym_sort(letters)
sym_sort(letters, rev = TRUE)

plot(sym_sort(runif(50)))
plot(sym_sort(rnorm(99), rev = TRUE))

}
\seealso{
\code{\link{kinda_sort}}
}
