% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabler}
\alias{tabler}
\title{Tabler}
\usage{
tabler(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{lm}}, \code{\link{glm}},
\code{\link{survfit}}}

\item{...}{additional parameters passed to other methods}
}
\description{
Extracts coefficients, standard errors, odds ratios, confidence intervals,
p-values, etc. from model fits.
}
\examples{
lmfit <- lm(mpg ~ hp + disp + wt, data = mtcars)
tabler(lmfit)

glmfit <- glm(vs ~ drat + factor(gear), data = mtcars, family = 'binomial')
tabler(glmfit, type = 'or')

\donttest{
library(survival)
sfit <- survfit(Surv(time, status) ~ 1, data = cancer, conf.int = 0.9)
tabler(sfit)
}
}

