% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler}
\alias{tabler}
\alias{tabler.default}
\alias{tabler.lm}
\alias{tabler.glm}
\alias{tabler.survfit}
\alias{tabler.coxph}
\title{Tabler}
\usage{
tabler(x, ...)

\method{tabler}{default}(x, ...)

\method{tabler}{lm}(x, digits = 3L, add_reference, ...)

\method{tabler}{glm}(x, digits = 3L, level = 0.95, exp = TRUE, add_reference = FALSE, ...)

\method{tabler}{survfit}(x, add_reference, ...)

\method{tabler}{coxph}(x, digits = 3L, level = 0.95, exp = TRUE, add_reference = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{\link{lm}}, \code{\link{glm}},
\code{\link[survival]{survfit}}, \code{\link[survival]{coxph}}}

\item{...}{additional arguments passed to or from other methods}

\item{digits}{number of digits printed}

\item{add_reference}{logical; if \code{TRUE}, adds row for each reference
group if applicable}

\item{level}{confidence level; default is \code{0.95}}

\item{exp}{logical; if \code{TRUE}, estimates and confidence intervals are
exponentiated (for \code{glm} or \code{coxph} methods only)}
}
\description{
Extracts coefficients, standard errors, odds ratios, confidence intervals,
p-values, etc. from model fits.
}
\examples{
lmfit <- lm(mpg ~ hp + disp + wt, mtcars)
tabler(lmfit)

glmfit <- glm(vs ~ drat + factor(gear), mtcars, family = 'binomial')
tabler(glmfit)
tabler(glmfit, add_reference = TRUE)
tabler(glmfit, exp = FALSE)

library('survival')
sfit <- survfit(Surv(time, status) ~ 1, cancer, conf.int = 0.9)
tabler(sfit)

cphfit <- coxph(Surv(time, status) ~ factor(sex) + age, cancer)
tabler(cphfit)
tabler(cphfit, add_reference = TRUE)

}
\seealso{
\code{\link{surv_table}}

Other tabler: 
\code{\link{tabler_by}()},
\code{\link{tabler_resp}()},
\code{\link{tabler_stat2}()},
\code{\link{tabler_stat}()}
}
\concept{tabler}
