% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{tabler}
\alias{tabler}
\alias{tabler.default}
\alias{tabler.glm}
\alias{tabler.lm}
\alias{tabler.survfit}
\title{Tabler}
\usage{
tabler(x, ...)

\method{tabler}{default}(x, ...)

\method{tabler}{lm}(x, digits = 3, ...)

\method{tabler}{glm}(x, digits = 3, level = 0.95, type = "", ...)

\method{tabler}{survfit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{lm}}, \code{\link{glm}},
\code{\link{survfit}}}

\item{...}{additional parameters passed to other methods}

\item{digits}{number of digits printed}

\item{level}{confidence level; default is \code{0.95}}

\item{type}{use \code{"or"} for odds ratios; others may be added later}
}
\description{
Extracts coefficients, standard errors, odds ratios, confidence intervals,
p-values, etc. from model fits.
}
\examples{
lmfit <- lm(mpg ~ hp + disp + wt, data = mtcars)
tabler(lmfit)

glmfit <- glm(vs ~ drat + factor(gear), data = mtcars, family = 'binomial')
tabler(glmfit, type = 'or')

\donttest{
library(survival)
sfit <- survfit(Surv(time, status) ~ 1, data = cancer, conf.int = 0.9)
tabler(sfit)
}
}

