% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_resp}
\alias{tabler_resp}
\title{Response table}
\usage{
tabler_resp(
  x,
  r_or_better = levels(x)[3:2],
  conf = 0.95,
  digits = 0L,
  frac = TRUE,
  show_conf = TRUE,
  pct.sign = TRUE,
  total = FALSE,
  two_stage = FALSE
)
}
\arguments{
\item{x}{a factor variable of responses; responses should be ordered as
CR, PR, SD, PD, NE or similar (i.e., best to worst)}

\item{r_or_better}{if integer(s), indicates the levels of \code{x} that
are to be combined with better responses; for example, if
\code{r_or_better = 3} (default), then any occurrence of level 1, 2, or 3
of \code{x} is treated as a response, and the proportion and confidence
interval are calculated for the aggregate

if \code{FALSE}; levels of \code{x} are estimated independently}

\item{conf, frac, show_conf, pct.sign}{additional arguments passed to
\code{\link{binconr}}}

\item{digits}{number of digits past the decimal point to keep}

\item{total}{logical or numeric; if \code{TRUE}, a column with the total,
i.e., \code{length(x)} is added; if numeric, \code{length(x)} and,
optionally, fraction and percent out of \code{total} is added}

\item{two_stage}{\code{FALSE} (default, assumes exact binomial CIs are
desired) or a vector of length 3 with the 1) maximum number responses in
the first stage that can be observed \emph{without} continuing; 2) the
number entered in the first stage; and 3) the additional number entered
in the second stage

if more than three integers are given, the remaining should indicate the
columns which should be calculated as two-stage CIs; usually this is only
one value but multiple are accepted}
}
\description{
Convenience function to calculate proportions and confidence invervals and
format for easy display.
}
\examples{
set.seed(1)
r <- c('CR', 'PR', 'SD', 'PD', 'NE')
x <- factor(sample(r, 30, replace = TRUE), r)

tabler_resp(x)
tabler_resp(x, 3)
tabler_resp(x, 'PR')
tabler_resp(x, 'PR', total = 50)

## note NAs are removed
y <- replace(x, 1:10, value = NA)
tabler_resp(x, FALSE)
tabler_resp(y, FALSE)


## two-stage designs
## use two-stage CI in "PR" column
tabler_resp(x)

two_idx <- 1:2
two_stage <- c(r1 = 2, n1 = 10, n2 = 20)
tabler_resp(x, two_stage = c(two_stage, two_idx))

## compare
bincon(c(2, 4),  c(10, 20), method = 'two-stage') ## CRs
bincon(c(2, 11), c(10, 20), method = 'two-stage') ## PRs
## one-stage methods should not be used
bincon(c(4, 11), 30, method = 'exact')


## typical usage
ht <- htmlTable::htmlTable(
  rbind(
    tabler_resp(x),
    tabler_resp(x, conf = 0.9),
    tabler_resp(x, frac = FALSE, pct.sign = FALSE,
                show_conf = FALSE, digits = 1),
    tabler_resp(x, two_stage = c(2, 10, 20, 1))
  ),
  caption = 'Table of responses with exact binomial and
    two-stage<sup>&dagger;</sup>confidence intervals.',
  css.cell = 'padding: 0 10 0px; white-space: nowrap;',
  cgroup = c('Evaluation', 'Outcome (95\% CI)'),
  n.cgroup = c(nlevels(x), 3L)
)
structure(ht, class = 'htmlTable')

}
\seealso{
\code{\link{bincon}}; \code{\link{binconr}}

Other tabler: 
\code{\link{tabler_by}()},
\code{\link{tabler_stat2}()},
\code{\link{tabler_stat}()},
\code{\link{tabler}()}
}
\concept{tabler}
