% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_resp}
\alias{tabler_resp}
\title{Response table}
\usage{
tabler_resp(x, r_or_better = 3L, conf = 0.95, frac = TRUE,
  show_conf = TRUE)
}
\arguments{
\item{x}{a factor variable of responses; responses should be ordered as
CR, PR, SD, PD, NE or similar}

\item{r_or_better}{if an integer, the first \code{r_or_better} levels of
\code{x} will be combined and proportions and confidence intervals will be
calculated for the aggregate; if \code{FALSE}; this is not included}

\item{conf, frac, show_conf}{arguments passed to \code{\link{binconr}}}
}
\description{
Convenience function to calculate proportions and confidence invervals and
format for easy display.
}
\examples{
set.seed(1)
r <- c('CR','PR','SD','PD','NE')
x <- factor(sample(r, 30, replace = TRUE), r)
tabler_resp(x)

y <- `[<-`(x, 1:10, value = NA)
tabler_resp(y, FALSE)

library('htmlTable')
htmlTable(t(as.matrix(tabler_resp(x, show_conf = FALSE))),
          caption = 'Table of responses with 95\% confidence intervals.',
          css.cell = 'padding: .1em .5em .1em;',
          cgroup = c('Evaluation', 'Outcome (95\% CI)'),
          n.cgroup = c(nlevels(x), 3))

}
\seealso{
\code{\link{bincon}}; \code{\link{binconr}}
}

