% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_resp}
\alias{tabler_resp}
\title{Response table}
\usage{
tabler_resp(x, r_or_better = 3L, conf = 0.95, digits = 0L, frac = TRUE,
  show_conf = TRUE, pct.sign = TRUE, total = FALSE)
}
\arguments{
\item{x}{a factor variable of responses; responses should be ordered as
CR, PR, SD, PD, NE or similar}

\item{r_or_better}{if an integer, the first \code{r_or_better} levels of
\code{x} will be combined and proportions and confidence intervals will be
calculated for the aggregate; if \code{FALSE}; this is not included}

\item{conf, frac, show_conf, pct.sign}{additional arguments passed to
\code{\link{binconr}}}

\item{digits}{number of digits past the decimal point to keep}

\item{total}{logical or numeric; if \code{TRUE}, a column with the total,
i.e., \code{length(x)} is added; if numeric, \code{length(x)} and,
optionally, fracton and percent out of \code{total} is added}
}
\description{
Convenience function to calculate proportions and confidence invervals and
format for easy display.
}
\examples{
set.seed(1)
r <- c('CR','PR','SD','PD','NE')
x <- factor(sample(r, 30, replace = TRUE), r)
tabler_resp(x, 3)
tabler_resp(x, 'PR')
tabler_resp(x, 'PR', total = 50)

## NAs are removed
y <- `[<-`(x, 1:10, value = NA)
tabler_resp(x, FALSE)
tabler_resp(y, FALSE)

library('htmlTable')
htmlTable(
  rbind(
    tabler_resp(x),
    tabler_resp(x, conf = 0.9),
    tabler_resp(x, frac = FALSE, pct.sign = FALSE,
                show_conf = FALSE, digits = 1)
  ),
  caption = 'Table of responses with confidence intervals.',
  css.cell = 'padding: 0 10 0px; white-space: nowrap;',
  cgroup = c('Evaluation', 'Outcome (95\% CI)'),
  n.cgroup = c(nlevels(x), 3L)
)

}
\seealso{
\code{\link{bincon}}; \code{\link{binconr}}
}

