% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_stat}
\alias{tabler_stat}
\title{Description statistics tabler}
\usage{
tabler_stat(data, varname, byvar, digits = 0L, FUN = NULL,
  color_pval = TRUE, color_missing = TRUE, dagger = TRUE, ...)
}
\arguments{
\item{data}{a matrix or data frame with variables \code{varname} and
\code{byvar}}

\item{varname, byvar}{the row and column variable, respectively}

\item{digits}{number of digits past the decimal point to keep}

\item{FUN}{a function performing the test of association between
\code{varname} and \code{byvar}; \code{FALSE} will suppress the test but
keep a column for p-values; \code{NA} will suppress the test and drop the
column for p-values; see details}

\item{color_pval}{logical; if \code{TRUE}, p-values will be colored
by significance; see \code{\link{color_pval}}}

\item{color_missing}{logical; if \code{TRUE}, rows summarizing missing
values will be shown in light grey; a color string can be used for a
custom color}

\item{dagger}{logical or a character string giving the character to
associate with \code{FUN}; if \code{FALSE}, none are used; if \code{TRUE},
the defaults are used (\code{"*"} is used if \code{FUN} is given)}

\item{...}{additional arguments passed to
\code{\link[Gmisc]{getDescriptionStatsBy}}}
}
\value{
A matrix with additional attributes:

\item{\code{attr(,"FUN")}}{the test passed to \code{FUN} or the test
selected based on \code{varname} and \code{byvar} if \code{FUN = NULL}}
\item{\code{attr(,"p.value")}}{the numeric p-value returned by \code{FUN}}
\item{\code{attr(,"fnames")}}{a vector of the default \code{FUN} options
with names to match the appropriate \code{dagger} character; see examples;
if \code{FUN} is given, the function name will be added with a new dagger
symbol (\code{"*"} by default or \code{dagger} if given)}
}
\description{
Wrapper function to create table of description statistics with tests
of association.
}
\details{
If \code{FUN} is \code{FALSE}, no test will be performed. If \code{FUN}
is a function, it must take two vector arguments: the row variable vector,
\code{data$varname}, and the column variable vector, \code{data$byvar},
and return a numeric p-value. Otherwise, \code{FUN} will be selected
based on the input vectors:

For a continuous \code{varname}, \code{\link{wilcox.test}} or
\code{\link{kruskal.test}} if \code{byvar} has two or more than two levels,
respectively. If \code{varname} is not numeric (or has few unique values),
\code{fisher.test} is used.
}
\examples{
tabler_stat(mtcars, 'mpg', 'cyl')
tabler_stat(mtcars, 'mpg', 'cyl', FUN = NA)
tabler_stat(mtcars, 'mpg', 'cyl', FUN = FALSE)

tabler_stat(mtcars, 'mpg', 'cyl',
  FUN = function(x, y)
    cuzick.test(x ~ y, data.frame(x, y))$p.value)


mt <- within(mtcars, {
  mpg[1:5] <- carb[1:5] <- drat[1:20] <- NA
  carb <- factor(carb)
  cyl  <- factor(cyl)
})

tbl <- lapply(names(mt)[-10L], function(x)
  tabler_stat(mt, x, 'gear', percentage_sign = FALSE))

htmlTable::htmlTable(
  do.call('rbind', tbl),
  cgroup = c('', 'Gear', ''), n.cgroup = c(1, 3, 1),
  rgroup = names(mt)[-10L], n.rgroup = sapply(tbl, nrow),
  tfoot = toString(sprintf('<sup>\%s</sup>\%s',
    names(attr(tbl[[1L]], 'fnames')), attr(tbl[[1L]], 'fnames'))[-1L])
)

}
\seealso{
\code{\link{tabler}}, \code{\link{tabler_by}}
}
