% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_stat}
\alias{tabler_stat}
\title{Description statistics \code{tabler}}
\usage{
tabler_stat(data, varname, byvar, digits = 0L, FUN = NULL,
  color_pval = TRUE, color_missing = TRUE, dagger = TRUE, ...)
}
\arguments{
\item{data}{a matrix or data frame with variables \code{varname} and
\code{byvar}}

\item{varname, byvar}{the row and column variable, respectively}

\item{digits}{number of digits past the decimal point to keep}

\item{FUN}{a function performing the test of association between
\code{varname} and \code{byvar}; \code{FALSE} will suppress the test but
keep a column for p-values; \code{NA} will suppress the test and drop the
column for p-values; or a character string; see details}

\item{color_pval}{logical; if \code{TRUE}, p-values will be colored
by significance; see \code{\link{color_pval}}}

\item{color_missing}{logical; if \code{TRUE}, rows summarizing missing
values will be shown in light grey; a color string can be used for a
custom color}

\item{dagger}{logical or a character string giving the character to
associate with \code{FUN}; if \code{FALSE}, none are used; if \code{TRUE},
the defaults are used (\code{"*"} is used if \code{FUN} is given)}

\item{...}{additional arguments passed to
\code{\link[Gmisc]{getDescriptionStatsBy}}}
}
\value{
A matrix with additional attributes:

\item{\code{attr(,"FUN")}}{the test passed to \code{FUN} or the test
selected based on \code{varname} and \code{byvar} if \code{FUN = NULL}}
\item{\code{attr(,"p.value")}}{the numeric p-value returned by \code{FUN}}
\item{\code{attr(,"fnames")}}{a vector of the default \code{FUN} options
with names to match the appropriate \code{dagger} character; see examples;
if \code{FUN} is given, the function name will be added with a new dagger
symbol (\code{"*"} by default or \code{dagger} if given)}
\item{\code{attr(,"tfoot")}}{a footnote for the table using each dagger
and corresponding test name}
}
\description{
Wrapper function to create table of description statistics with optional
tests of association.
}
\details{
If \code{FUN} is \code{FALSE}, no test will be performed but a column for
p-values will still be added; \code{FUN = NA} will prevent both the test
from being performed and the column from being added to the output.

For \code{FUN = NULL} (default), the correct test will be guessed based on
the row and column data. The current options are \code{\link{fisher.test}},
\code{\link{wilcox.test}}, and \code{\link{kruskal.test}}. If the row data
is continuous, one of the latter two tests will be used based on the number
of unique values in the column data.

For special cases, the function is not always guessed correctly (e.g., if
the row data contains few unique values, a Fisher test may be used where
not appropriate). One of the default tests can be given explicitly with a
character string, one of \code{"fisher"}, \code{"wilcoxon"},
\code{"kruskal"}, \code{"chisq"}, or \code{"anova"} (can be abbreviated).

If \code{FUN} is a function, it must take two vector arguments: the row
variable vector, \code{data$varname}, and the column variable vector,
\code{data$byvar}, in this order and return a numeric p-value only.
}
\examples{
tabler_stat(mtcars, 'mpg', 'cyl') ## picks kruskal-wallis
tabler_stat(mtcars, 'mpg', 'cyl', FUN = NA)
tabler_stat(mtcars, 'mpg', 'cyl', FUN = FALSE)
tabler_stat(mtcars, 'mpg', 'cyl', FUN = 'fisher')
tabler_stat(mtcars, 'mpg', 'cyl', FUN = 'anova')


## use of a custom function - see ?rawr::cuzick.test
tabler_stat(mtcars, 'mpg', 'cyl',
  FUN = function(x, y)
    cuzick.test(x ~ y, data.frame(x, y))$p.value)


## typical usage
mt <- within(mtcars, {
  mpg[1:5] <- carb[1:5] <- drat[1:20] <- NA
  carb <- factor(carb)
  cyl  <- factor(cyl)
})

tbl <- lapply(names(mt)[-10L], function(x)
  tabler_stat(mt, x, 'gear', percentage_sign = FALSE))

htmlTable::htmlTable(
  do.call('rbind', tbl),
  cgroup = c('', 'Gear', ''), n.cgroup = c(1, 3, 1),
  rgroup = names(mt)[-10L], n.rgroup = sapply(tbl, nrow),
  tfoot = attr(tbl[[1L]], 'tfoot')
)

}
\seealso{
\code{\link{tabler}}; \code{\link{tabler_by}}
}
