% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tabler_stat2}
\alias{tabler_stat2}
\title{\code{tabler_stat} wrappers}
\usage{
tabler_stat2(data, varname, byvar, varname_label = varname,
  byvar_label = byvar, digits = NULL, FUN = NULL, color_pval = TRUE,
  color_missing = TRUE, dagger = TRUE, statArgs = NULL, align = NULL,
  rgroup = NULL, cgroup = NULL, tfoot = NULL, htmlArgs = NULL)
}
\arguments{
\item{data}{a matrix or data frame with variables \code{varname} and
\code{byvar}}

\item{varname}{one or more variables in \code{data} to calculate
statistics by \code{byvar}}

\item{byvar}{the column or stratification variable}

\item{varname_label, byvar_label}{optional labels for \code{varname} and
\code{byvar}}

\item{digits}{\code{NULL} or a vector of digits past the decimal point to
keep for each \code{varname}; if \code{NULL}, these will be guessed using
\code{rawr:::guess_digits}}

\item{FUN}{\code{NULL} or a list of functions performing the test of
association between each \code{varname} and \code{byvar}; see
\code{\link{tabler_stat}}}

\item{color_pval, color_missing, dagger}{\code{NULL} or vectors, recycled as
needed for each \code{varname}; see \code{\link{tabler_stat}}}

\item{statArgs}{a named list of additional arguments passed to
\code{\link[Gmisc]{getDescriptionStatsBy}}}

\item{align, rgroup, cgroup, tfoot}{optional arguments passed to
\code{\link[htmlTable]{htmlTable}}}

\item{htmlArgs}{a named list of additional arguments passed to
\code{\link[htmlTable]{htmlTable}}}
}
\description{
Helper functions for using \code{tabler_stat}.
}
\examples{
sapply(mtcars, rawr:::guess_digits)

rawr:::get_tabler_stat_n(mtcars$gear)

## typical usage
mt <- within(mtcars, {
  cyl  <- factor(cyl)
  gear <- factor(gear)
  vs2  <- factor(vs, 1:0)
})

tabler_stat2(mt, c('mpg', 'cyl', 'wt'), 'vs')

tabler_stat2(
 mt, c('mpg', 'cyl', 'wt'), 'vs',
 cgroup = c('', 'V/S engine', ''),
 rgroup = c('Miles/gallon', 'No. cylinders', 'Weight (1000 lbs)')
)

tabler_stat2(
  mt, c('mpg', 'cyl', 'wt'), 'vs',
  byvar_label = 'V/S engine',
  varname_label = c('MPG', 'Cylinders', 'Weight')
)

}
\seealso{
\code{\link{tabler_stat}}
}
