% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{tcol}
\alias{tcol}
\title{Transparent colors}
\usage{
tcol(color, trans = 255)
}
\arguments{
\item{color}{vector of color names (or hexadecimal)}

\item{trans}{transparency defined as an integer in the range
\code{[0, 255]} where \code{0} is fully transparent and \code{255} is fully
visible; see details}
}
\description{
Add transparency to colors.
}
\details{
This is a vectorized function to add transparency to colors.
\code{color} and \code{trans} must either be the same length or one of the
two must have length one.

The function adds integers (in hex) between 0 (fully transparent) and 255
(fully visible) to the color(s) given. \code{color} values are converted to
RGB with transparency.
}
\examples{
cols <- c('red','green','blue')

# a normal plot
plot(rnorm(100), col = tcol(cols), pch = 16, cex = 4)

# more transparent
plot(rnorm(100), col = tcol(cols, 100), pch = 16, cex = 4)

# hexadecimal colors also work
cols <- c('#FF0000','#00FF00','#0000FF')
plot(rnorm(100), col = tcol(cols, c(50, 100, 255)), pch= 16, cex = 4)
}
\seealso{
\code{\link{as.hexmode}}, \code{\link{col2rgb}},
\code{\link{adjustcolor}}
}

