% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_extra.R
\name{tcol}
\alias{tcol}
\title{Transparent colors}
\usage{
tcol(color, trans = 255, alpha)
}
\arguments{
\item{color}{vector of color names (or hexadecimal) as character strings
or integers corresponding to colors in the current \code{\link{palette}}}

\item{trans}{alpha transparency defined as an integer in the range
\code{[0, 255]} where \code{0} is fully transparent and \code{255} is fully
visible; see details}

\item{alpha}{the alpha transparency in \code{[0,1]}; \code{trans} is
ignored if \code{alpha} is given}
}
\description{
Add alpha transparency to colors.
}
\details{
This is a vectorized function to add transparency to colors. \code{color}
and \code{trans} (or \code{alpha}) must either be the same length or one
of the two must have length one.

The function adds integers (in hexadecimal) between 0 (fully transparent)
and 255 (fully visible) to the color(s) given. \code{color} values are
converted to RGB with transparency.
}
\examples{
cols <- c('red','green','blue')

## a normal plot
plot(rnorm(100), col = tcol(cols), pch = 16, cex = 4)

## more transparent
plot(x <- rnorm(100), col = tcol(cols, 100), pch = 16, cex = 4)
## or equivalently using alpha
plot(x, col = tcol(cols, alpha = .4), pch = 16, cex = 4)

## hexadecimal colors also work
cols <- c('#FF0000','#00FF00','#0000FF')
plot(rnorm(100), col = tcol(cols, c(50, 100, 255)), pch= 16, cex = 4)

}
\seealso{
\code{\link{as.hexmode}}, \code{\link{col2rgb}}, \code{\link{adjustcolor}}
}

