% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytext.R
\name{tidytext}
\alias{tidytext}
\alias{tidytextmining}
\title{Retrieve raw R code tidytext tutorial}
\usage{
tidytext(url, method, padding)

tidytextmining(url, method, padding = "\\n\\n")
}
\arguments{
\item{url}{Link to tidytext tutorial}

\item{method}{Not all websites are formatted consistently. To overcome this, try a different
method by setting the method
parameter to integers 2 and greater to try other available methods}

\item{padding}{Specify what goes between the last character of one code block and the
first character of the next code block. Default is a two new lines, which appears
visually as one new line between codeblocks.}
}
\value{
A character vector of length 1 containing the R code from the target url.
}
\description{
Retrieve raw R code tidytext tutorial
}
\examples{

\dontrun{
library(dplyr)

constructed_url <- paste0("https://cran.r-project.org/web/packages/",
 "tidytext/vignettes/tidytext.html")
tidytext(constructed_url) \%>\%
  cat
}

}
