% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{tox_worst}
\alias{tox_worst}
\title{Find highest grade toxicities}
\usage{
tox_worst(data, id = "id", desc = "desc", grade = "grade", code, version = 4L)
}
\arguments{
\item{data}{toxicity data frame}

\item{id}{column name with identifier}

\item{desc}{column name with toxicity descriptions (or codes)}

\item{grade}{column name with toxiticity grades; should be a factor
with the desired order to be properly sorted, i.e., least to most severe}

\item{code, version}{if \code{code} is given, \code{\link{match_ctc}} will
match this column from \code{data} with the CTCAE \code{version} given and
return toxicity descriptions rather than codes; if showing toxicity codes
is desired, use \code{desc} instead

note that \code{version} should be either \code{3} or \code{4} (default)
corresponding to \code{\link{ctcae_v3}} or \code{\link{ctcae_v4}},
respectively}
}
\value{
A filtered data frame with attributes:

\item{\code{attr(., "data")}}{the input data frame sorted by \code{id},
\code{desc}, and \code{grade}}
\item{\code{attr(., "duplicates"}}{the indices of rows removed from
\code{attr(., "data")} which correspond to duplicate \code{desc} per
\code{id} with equal or lesser \code{grade}s}
}
\description{
Returns a subset of input data with worst toxicity grade per toxicity
code per patient, ie, sorts and removes duplicates.
}
\examples{
set.seed(1)
f <- function(x, ...) sample(x, 100, replace = TRUE, ...)

tox <- data.frame(id = rep(1:10, 10), phase = 1:2,
                  code = f(rawr::ctcae_v4$tox_code[1:10]),
                  grade = factor(f(1:3, prob = c(.3, .4, .3))))

## get worst tox by CTCAE code
## this will convert the code to description strings
tox1 <- tox_worst(tox, id = 'id', grade = 'grade', code = 'code')

## or by formatted descriptions and grade
tox$desc <- factor(match_ctc(tox$code)$tox_desc)
tox2 <- tox_worst(tox, id = 'id', grade = 'grade', desc = 'desc')

## both methods are equivalent
stopifnot(identical(tox1, tox2))


## these rows have been removed from attr(tox1, 'data')
attr(tox1, 'duplicates')

stopifnot(
  all.equal(tox1, attr(tox1, 'data')[-attr(tox1, 'duplicates'), ],
            check.attributes = FALSE)
)


## use tabler_by/tabler_by2 to summarize
tabler_by(tox1, 'desc', 'grade', n = 10, pct.sign = FALSE)
tabler_by2(tox1, 'desc', 'grade', stratvar = 'phase')

}
\seealso{
\code{\link{match_ctc}}, \code{\link{tabler_by}}; \code{\link{tabler_by2}}
}
