% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tplot}
\alias{tplot}
\alias{tplot.default}
\alias{tplot.formula}
\title{tplot (boxplot)}
\usage{
tplot(x, ...)

\method{tplot}{default}(x, ..., type = c("d", "db", "bd", "b"), jit = 0.1,
  dist, main, sub, xlab, ylab, names, xlim, ylim, col, group.col = TRUE,
  boxcol, boxborder, pch = par("pch"), group.pch = TRUE,
  median.line = FALSE, mean.line = FALSE, median.pars = list(col =
  par("col")), mean.pars = median.pars, boxplot.pars, show.n = TRUE,
  show.na = TRUE, cex.n, my.gray = gray(0.75), ann = par("ann"),
  axes = TRUE, frame.plot = axes, add = FALSE, at, horizontal = FALSE,
  panel.first = NULL, panel.last = NULL)

\method{tplot}{formula}(formula, data = NULL, ..., subset, na.action = NULL,
  panel.first = NULL, panel.last = NULL)
}
\arguments{
\item{x}{for specifying data from which the boxplots are to be produced; 
either a numeric vector or a single list containing such vectors; additional
unnamed arguments specify further data as separate vectors (each 
corresponding to a component boxplot); \code{NA}s are allowed in the data}

\item{...}{for the \code{formula} method, named arguments to be passed to 
the default method

for the default method, unnamed arguments are additional data vectors
(unless x is a list when they are ignored), and named arguments are 
arguments and \code{\link{par}}s to be passed to \code{\link{bxp}}}

\item{type}{type of plot (dot, dot-box, box-dot, box)}

\item{jit}{numeric; x-axis jitter parameter for overlapping pts (\code{0} for 
overlap and higher values for more distance between points)}

\item{dist}{numeric; additional jitter parameter}

\item{main}{overall title for the plot}

\item{sub}{sub-title for the plot (below x-axis)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{names}{group labels}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{col}{plotting color}

\item{group.col}{logical; if \code{TRUE}, color by group; o/w by order}

\item{boxcol}{fill color}

\item{boxborder}{border color}

\item{pch}{plotting character}

\item{group.pch}{logical; if \code{TRUE}, \code{pch} by group; o/w, by order}

\item{median.line}{logical; draw median line}

\item{mean.line}{logical; draw mean line}

\item{median.pars}{list of graphical parameters for median line}

\item{mean.pars}{see above}

\item{boxplot.pars}{additional list of graphical parameters for box plots}

\item{show.n}{show number in each group}

\item{show.na}{show number missing in each group}

\item{cex.n}{text size for \code{show.n} and \code{show.na}}

\item{my.gray}{default border color}

\item{ann}{logical; annotate plot}

\item{axes}{logical; draw axes}

\item{frame.plot}{logical; draw box around \code{x-y} plot}

\item{add}{logical; add to an existing plot}

\item{at}{group positions on axis}

\item{horizontal}{logical; flip axes}

\item{panel.first}{an "expression" to be evaluated after the plot axes are 
set up but before any plotting takes place; this can be useful for drawing 
background grids or scatterplot smooths; note that this works by lazy 
evaluation: passing this argument from other plot methods may well not work 
since it may be evaluated too early; see also \code{\link{plot.default}}}

\item{panel.last}{an expression to be evaluated after plotting has taken 
place but before the axes, title, and box are added; see the comments about 
\code{panel.first}}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ grp}, where y is 
a numeric vector of data values to be split into groups according to the 
grouping variable \code{grp} (usually a factor)}

\item{data}{a data frame (or list) from which the variables in formula 
should be taken}

\item{subset}{an optional vector specifying a subset of observations to be 
used for plotting}

\item{na.action}{a function which indicates what should happen when the data
contain \code{\link{NA}}s; the default is to ignore missing values in either
the response or the group}
}
\description{
An alternative to \code{\link{boxplot}}. The individual data can be shown 
(either in the foreground or background) with jittering if necessary.
}
\examples{
## equivalent ways to call tplot
## the formula method is a convenience function for the first case
tplot(split(mtcars$mpg, mtcars$gear))
tplot(mpg ~ gear, mtcars)

## use panel.first/panel.last like in `plot` (unavailable in `boxplot`)
tplot(mpg ~ gear, data = mtcars, col = 1:3, type = 'd',
      panel.last = legend('topleft', legend = 3:5, col = 1:3, pch = 1),
      panel.first = {
        abline(h = mean(mtcars$mpg))
        abline(h = 1:6 * 5 + 5, lty = 'dotted', col = 'grey70')
      })

set.seed(1)
dat <- data.frame(age = rnorm(80, rep(c(26, 36), c(70, 10)), 4),
                  sex = sample(c('Female', 'Male'), 80, replace = TRUE),
                  group = paste0('Group ', 
                                 sample(1:4, 40, prob = c(2, 5, 4, 1), 
                                        replace = TRUE)))
dat[1:5, 'age'] <- NA

tplot(age ~ group, data = dat, las = 1, cex.n = .8, cex.axis = 1, bty = 'L',
      type = c('db', 'db', 'db', 'd'), names = LETTERS[1:4],
      group.pch = TRUE, pch = c(15, 17, 19, 8),
      group.col = FALSE, col = c('darkred', 'darkblue')[c(sex)],
      boxcol = c('lightsteelblue1', 'lightyellow1', grey(.9)),
      boxplot.pars = list(notch = TRUE, boxwex = .5))

}
\seealso{
\href{http://biostat.mc.vanderbilt.edu/wiki/Main/TatsukiRcode}{Tatsuki
\code{tplot}}; \href{http://data.vanderbilt.edu/~graywh/dotplot/}{web app
for Tatsuki \code{tplot}}
}

