% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tplot}
\alias{tplot}
\alias{tplot.default}
\alias{tplot.formula}
\title{tplot}
\usage{
tplot(x, ...)

\method{tplot}{default}(x, g, ..., type = "db", jit = NULL, dist = NULL,
  main = NULL, sub = NULL, xlab = NULL, ylab = NULL, xlim = NULL,
  ylim = NULL, names, col, group.col = TRUE, boxcol = "grey90",
  bordercol = par("fg"), pch = par("pch"), group.pch = TRUE,
  cex = par("cex"), group.cex = FALSE, median.line = FALSE,
  mean.line = FALSE, median.pars = list(), mean.pars = list(),
  boxplot.pars = list(), show.n = TRUE, show.na = show.n, cex.n = cex,
  text.na = "missing", test = FALSE, args.test = list(),
  format_pval = TRUE, ann = par("ann"), axes = TRUE, frame.plot = axes,
  add = FALSE, at, horizontal = FALSE, panel.first = NULL,
  panel.last = NULL)

\method{tplot}{formula}(formula, data = NULL, ..., subset, na.action = NULL,
  panel.first = NULL, panel.last = NULL)
}
\arguments{
\item{x}{a numeric vector or a single list containing such vectors}

\item{...}{additional graphical parameters passed to \code{\link{par}}}

\item{g}{a vector or factor object giving the group for the corresponding
elements of \code{x}, ignored with a warning if \code{x} is a list}

\item{type}{type of plot (\code{"d"} for dot, \code{"db"} for dot-box,
\code{"bd"} for box-dot, or \code{"b"} box; \code{"v"} may be used instead
of \code{"b"} for a violin plot)}

\item{jit, dist}{jitter parameters for overlapping points (use \code{0} for
no jitter (i.e., points may overlap) and values > 0 for more distance
between points); both can be length 1 (recycled as needed for groups) or
equal to the number of groups (useful if one group has more points and
needs more jittering than other groups)

\code{jit} controls the amount of spread in a group of neighboring points,
and \code{dist} controls the size of the interval to group neighboring
points, i.e., a group of sequential points that are no more than
\code{dist} apart are considered neighbors and will be jittered}

\item{main, sub}{overall title and sub-title for the plot (below x-axis)}

\item{xlab, ylab}{x- and y-axis labels}

\item{xlim, ylim}{x- and y-axis limits}

\item{names}{group labels}

\item{col}{plotting color}

\item{group.col}{logical; if \code{TRUE}, color by group; otherwise by order}

\item{boxcol, bordercol}{box fill and border colors}

\item{pch}{plotting character}

\item{group.pch}{logical; if \code{TRUE}, \code{pch} by group; otherwise
by order}

\item{cex}{\strong{c}haracter \strong{ex}pansion value}

\item{group.cex}{logical; if \code{TRUE}, groups use the same \code{cex}
value; otherwise, points have individual values, recycled if necessary}

\item{median.line, mean.line}{logical; draw median, mean lines}

\item{median.pars, mean.pars}{lists of graphical parameters for median and
mean lines}

\item{boxplot.pars}{additional list of graphical parameters for box plots}

\item{show.n, show.na}{logical; show total and missing in each group}

\item{cex.n}{character expansion for \code{show.n} and \code{show.na}}

\item{text.na}{label for missing values (default is "missing")}

\item{test}{logical or function; if \code{TRUE}, a rank-sum p-value is
added to the plot (\code{\link{wilcox.test}} or \code{\link{kruskal.test}}
based on the number of groups)

Alternatively, a function (or function name as a character string) can be
used, e.g., \code{test = cuzick.test} or \code{function(x, g)
cuzick.test(x, g)}; note that if \code{test} is a function, it must have
at least two arguments with the numeric data values and group}

\item{args.test}{an optional \emph{named} list of \code{\link{mtext}}
arguments controlling the \code{test} text}

\item{format_pval}{logical; if \code{TRUE}, p-values are formatted with
\code{\link{pvalr}}; if \code{FALSE}, no formatting is performed;
alternatively, a function can be passed which should take a numeric value
and return a character string (or a value to be coerced) for printing}

\item{ann}{logical; annotate plot}

\item{axes}{logical; draw axes}

\item{frame.plot}{logical; draw box around \code{x-y} plot}

\item{add}{logical; add to an existing plot}

\item{at}{group positions on axis}

\item{horizontal}{logical; flip axes}

\item{panel.first}{an "expression" to be evaluated after the plot axes are
set up but before any plotting takes place; this can be useful for drawing
background grids or scatterplot smooths; note that this works by lazy
evaluation: passing this argument from other plot methods may well not work
since it may be evaluated too early; see also \code{\link{plot.default}}}

\item{panel.last}{an expression to be evaluated after plotting has taken
place but before the axes, title, and box are added; see the comments about
\code{panel.first}}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ grp}, where y is
a numeric vector of data values to be split into groups according to the
grouping variable \code{grp} (usually a factor)}

\item{data}{a data frame (or list) from which the variables in formula
should be taken}

\item{subset}{an optional vector specifying a subset of observations to be
used for plotting}

\item{na.action}{a function which indicates what should happen when the data
contain \code{\link{NA}}s; the default is to ignore missing values in either
the response or the group}
}
\value{
A list with the following components (see \code{\link{boxplot}}:
\item{\code{$stats}}{a matrix, each column contains the extreme of the lower
whisker, the lower hinge, the median, the upper hinge and the extreme of the
upper whisker for one group/plot. If all the inputs have the same class
attribute, so will this component.}
\item{\code{$n}}{a vector with the number of observations in each group.}
\item{\code{$conf}}{a matrix where each column contains the lower and upper
extremes of the notch.}
\item{\code{$out}}{the values of any data points which lie beyond the
extremes of the whiskers.}
\item{\code{$group}}{a vector of the same length as \code{out} whose
elements indicate to which group the outlier belongs.}
\item{\code{$names}}{a vector of names for the groups.}
}
\description{
An alternative to \code{\link{boxplot}}. The individual data can be shown
(either in the foreground or background) with jittering if necessary.
}
\examples{
## these are equivalent ways to call tplot
x <- mtcars$mpg
g <- interaction(mtcars$gear, mtcars$vs)

tplot(x, g)
tplot(split(x, g))
tplot(x ~ g)
tplot(mpg ~ gear + vs, mtcars)


## options for box, violin, dots
types <- c('d', 'db', 'bd', 'b', 'v', 'vd', 'dv', 'n')
l <- lapply(types, function(...) mtcars$mpg)
tplot(l, type = types, names = types, xlab = 'type = \\'x\\'')


## horizontal plots may cut off show.n/show.na text
tplot(x, g, horizontal = TRUE)

op <- par(mar = par('mar') + c(0, 0, 0, 2))
tplot(x, g, horizontal = TRUE)

## and/or rotate labels
tplot(x, g, horizontal = TRUE, srt = 45)
par(op)


## add rank-sum or custom test to plot
tplot(mpg ~ vs, mtcars, test = TRUE)   ## two groups - wilcox.test
tplot(mpg ~ gear, mtcars, test = TRUE) ## >=2 groups - kruskal.test
tplot(mpg ~ gear, mtcars, test = rawr::cuzick.test) ## trend test

## custom test/text formatting
tplot(mtcars$mpg, 1:2, test = function(x, g)
  wilcox.test(x ~ g, data.frame(x, g), exact = FALSE, paired = TRUE),
  args.test = list(col = 2, at = 1.5, adj = 0.5, line = -3, cex = 2))


## tplot has the same return value as boxplot
identical(tplot(mtcars$mpg), boxplot(mtcars$mpg))


## use panel.first/panel.last like in `plot` (unavailable in `boxplot`)
tplot(mpg ~ gear, data = mtcars, col = 1:3, type = 'd', show.na = FALSE,
      cex = c(1,5)[(mtcars$mpg > 30) + 1L],
      panel.last = legend('topleft', legend = 3:5, col = 1:3, pch = 1),
      panel.first = {
        abline(h = mean(mtcars$mpg))
        abline(h = 1:6 * 5 + 5, lty = 'dotted', col = 'grey70')
      })


## example with missing data
set.seed(1)
dat <- data.frame(
  age   = rnorm(80, rep(c(26, 36), c(70, 10)), 4),
  sex   = factor(sample(c('Female', 'Male'), 80, replace = TRUE)),
  group = paste0('Group ', sample(1:4, 80, prob = c(2, 5, 4, 1),
                                  replace = TRUE))
)
dat[1:5, 'age'] <- NA

tplot(age ~ group, data = dat, las = 1, cex.axis = 1, bty = 'L',
      type = c('db', 'db', 'db', 'd'), names = LETTERS[1:4],
      text.na = 'n/a', ## default is 'missing'
      group.pch = TRUE, pch = c(15, 17, 19, 8),
      group.col = FALSE, col = c('darkred', 'darkblue')[sex],
      boxcol = c('lightsteelblue1', 'lightyellow1', grey(.9)),
      boxplot.pars = list(notch = TRUE, boxwex = .5))
legend(par('usr')[1], par('usr')[3], xpd = NA, bty = 'n',
       legend = levels(dat$sex), col = c('darkred', 'darkblue'), pch = 19)

}
\seealso{
\href{http://biostat.mc.vanderbilt.edu/wiki/Main/TatsukiRcode}{Tatsuki
\code{tplot}}; \href{http://data.vanderbilt.edu/~graywh/dotplot/}{web app
for Tatsuki \code{tplot}}; \code{\link{boxplot}}; \code{\link{jmplot}}
}
