% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zxx.R
\name{try_require}
\alias{try_require}
\title{Quietly try to require a package}
\usage{
try_require(package)
}
\arguments{
\item{package}{name of package as name or character string}
}
\description{
Quietly require a package, returning an error message if not installed.
}
