% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryCatch2}
\alias{tryCatch2}
\title{tryCatch2}
\usage{
tryCatch2(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
A list of at least \code{$value}. If \code{expr} is evaluated with no error,
warning, messages, or interrupts, then a list with a single element,
\code{$value}, is returned with the value of \code{expr}. If any other
condition is triggered, a corresponding element will be appended to the
list with \code{$value}. For errors, \code{$value} will be \code{NULL}.
}
\description{
Modification of \code{\link{tryCatch}} which allows recovery from errors,
warnings, messages, and interrupts while returning the value from
\code{expr}.
}
\examples{
tryCatch2(1)
tryCatch2(stop('halt at once!'))

l <- alist(1, simpleError('oops'), stop('halt!'),
           warning('hmm'), message('hey'))
lapply(l, function(x) tryCatch2(eval(x)))

l <- lapply(list(1,-1,'a'), function(x) tryCatch2(log(x)))
rm_null(l)

}
\seealso{
\href{http://stackoverflow.com/questions/4948361/how-do-i-save-warnings-and-errors-as-output-from-a-function}{SO question};
\code{\link[logR]{tryCatch2}}; \href{https://github.com/jangorecki/logR}{
\pkg{logR} github repo}
}
\author{
Martin Morgan, Jan Gorecki, Robert Redd (modifications)
}
