% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{twostg.test}
\alias{twostg.test}
\title{Two-stage trials}
\usage{
twostg.test(r, r1, n1, n2, p0 = NULL, conf = 0.95, dp = 1)
}
\arguments{
\item{r, r1}{total observed responses and maximum number of responses in
first stage to declare treatment inactive}

\item{n1, n2}{sample size of the first and second stage}

\item{p0}{(optional) null hypothesis}

\item{conf}{confidence level}

\item{dp}{affects the ordering of outcomes within the sample space, see
\code{\link[desmon]{twocon}}}
}
\description{
Calculate confidence intervals, estimators, and p-value from a two-stage
trial.
}
\examples{
simon2(0.1, 0.3)

# $designs$`p0 = 0.1, pa = 0.3`
#      n1 r1 n2 r2 Pstop1.H0       size      type2 E.tot.n.H0  Pstop1.H1 E.tot.n.H1
# [1,] 12  1 23  5 0.6590023 0.09771828 0.09855051   19.84295 0.08502505   33.04442
# 
# ...
# 
# $description
# [1] "n1, n2 = cases 1st stage and additional # in 2nd"                    
# [2] "r1, r2 = max # responses 1st stage and total to declare trt inactive"

twostg.test(5, 1, 12, 23)
twostg.test(5, 1, 12, 23, 0.1)
twostg.test(8, 1, 12, 23, 0.1)$p.value

}
\seealso{
\code{\link[desmon]{twocon}}; \code{\link[clinfun]{twostage.inference}}
}
