% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{waffle}
\alias{waffle}
\title{waffle chart}
\usage{
waffle(mat, xpad = 0, ypad = 0.05, ..., reset_par = TRUE)
}
\arguments{
\item{mat}{a matrix of integers or character strings of color names; if
\code{mat} is a matrix of integers, the colors used will correspond to
the current \code{\link{palette}}}

\item{xpad, ypad}{amount of padding between \code{rect}s along axes}

\item{...}{additional graphical parameters passed to \code{\link{par}}}

\item{reset_par}{logical; if \code{TRUE}, resets \code{\link{par}} settings
to state before function call}
}
\value{
A list of three matrices: \code{matrix}, the input matrix; \code{origin},
coordinates of the bottom-left corner for each box; and \code{centers},
the coordinates for the centers of each box adjusted for \code{xpad} and
\code{ypad}.
}
\description{
A waffle chart.
}
\examples{
waffle(matrix(1:8, 2))

## heatmap
cols <- c(cor(mtcars))
cols <- tcol(c('blue','red')[(cols > 0) + 1L], alpha = c(abs(cols)))
waffle(matrix(cols, 11)[11:1, ], reset_par = FALSE, ypad = 0)
axis(3, at = 1:11 - .5, labels = names(mtcars), cex.axis = .8, lwd = 0)

## adding to margins of another plot
set.seed(1)
n <- 100
ng <- 3
cols <- c('beige','dodgerblue2','green','orange')
x <- sample(cols, n * ng, replace = TRUE, prob = c(.05,.31,.32,.32))
x <- rawr::kinda_sort(x, n = 20)

par(fig = c(0,1,.2,.9), mar = c(0,5,0,1))
plot(cumsum(rnorm(n)), type = 'l', ann = FALSE, xaxt = 'n')
par(fig = c(0,1,0,.2), mar = c(1,5,0,1), new = TRUE)
waffle(matrix(x, ng))
par(fig = c(0,1,.9,1), mar = c(.5,5,.5,1), new = TRUE)
waffle(matrix(x, ng)[1, , drop = FALSE], ypad = 0, reset_par = FALSE)
box()

## waffle conveniently returns the centers of the rects
## be sure /not/ to reset pars on exit for proper alignment
(w <- waffle(matrix(1:8, 2), reset_par = FALSE))
text(w$c[, 'x'], w$c[, 'y'], labels = palette(), col = 'white')

}

