% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{waterfall}
\alias{waterfall}
\title{Waterall plot}
\usage{
waterfall(x, type = 1L, col = c("red", "blue"), ..., arrows = type == 2L,
  rev = FALSE, plot = TRUE, panel.first = NULL, panel.last = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{type}{type of waterfall plot; \code{type = 1} draws sorted bars for
values of \code{x}; \code{type = 2} starts from 0 and draws bars according
to the magnitude and direction of \code{x}}

\item{col}{a vector of colors having 1) the same length as \code{x}; length
2 (for negative and positive values if \code{type = 2}) or otherwise to be
used for color interpolation (\code{\link{colorRampPalette}})}

\item{...}{additional arguments passed to \code{\link{barplot}}, to/from
other methods, or to \code{\link{par}}}

\item{arrows}{logical; if \code{TRUE}, arrows are drawn in the direction
of each bar}

\item{rev}{logical; if \code{TRUE}, the order along the x-axis is reversed}

\item{plot}{logical; if \code{FALSE}, nothing is plotted}

\item{panel.first}{an "expression" to be evaluated after the plot axes are
set up but before any plotting takes place; this can be useful for drawing
background grids or scatterplot smooths; note that this works by lazy
evaluation: passing this argument from other plot methods may well not work
since it may be evaluated too early; see also \code{\link{plot.default}}}

\item{panel.last}{an expression to be evaluated after plotting has taken
place but before the axes, title, and box are added; see the comments about
\code{panel.first}}
}
\value{
A matrix giving the coordinates of \emph{all} the bar midpoints drawn (see
\code{\link{barplot}}) and the \code{\link{order}} of \code{x}.
}
\description{
Draw two types of waterfall plots.
}
\examples{
set.seed(1)
change <- runif(20, -1, 1) * 100
col <- c(PD = 'red', SD = 'blue', CR = 'chartreuse4')

## interpolation
waterfall(change, col = col)

## discrete breaks
waterfall(change, col = as.character(cut(change, c(-Inf, -50, 50, Inf), col)))
legend('top', names(col), fill = col, horiz = TRUE, bty = 'n', border = NA)
title(xlab = 'Patient', ylab = '\% change', main = 'waterfall', line = 2)

## use return value to add an existing plot
wf <- waterfall(change, plot = FALSE)
text(wf[, 1L], 0, labels = sprintf('\%0.1f', change[wf[, 2L]]),
     pos = ifelse(change[wf[, 2L]] > 0, 1, 3), cex = 0.5)


## type 2
waterfall(change, type = 2, col = col, panel.first = grid())
title(main = 'waterfall - type 2')
axis(2, las = 1)

}
