% Generated by roxygen2 (4.0.0): do not edit by hand
\name{witchcraft}
\alias{witchcraft}
\title{Simple memory game}
\usage{
witchcraft(level = 1, seed = NULL)
}
\arguments{
\item{level}{difficulty; choose levels 0 (easy) to 3 (difficult)}

\item{seed}{seed; see \code{\link{set.seed}}}
}
\value{
Interactive game

A prize.
}
\description{
Pick boxes in the correct numerical order to win a prize.
}
\details{
Set the seed to play the same game.
}
\examples{
\dontrun{
witchcraft()
witchcraft(level = 3)
}
}

