% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeftable}
\alias{writeftable}
\title{Write ftable}
\usage{
writeftable(x, quote = FALSE, digits = getOption('digits'), ...)
}
\arguments{
\item{x}{an object of class \code{ftable}}

\item{quote}{logical; if \code{TRUE}, strings will be surrounded by double
quotes}

\item{digits}{integer giving the number of significant digits to use for
the cell entries of \code{x}}

\item{...}{additional parameters passed to \code{\link{format.ftable}}}
}
\value{
A matrix formatted as \code{ftable} would print.
}
\description{
\code{\link{ftable}}s can look nice, but are only \code{\link{cat}}'d to
the console and, thus, not easily used (or manipulated).
\code{\link{write.ftable}} does not write \code{ftable}s as they print,
so here we are.
}
\examples{
x <- ftable(Titanic, row.vars = 1:3)
writeftable(x)
}

